The Causal Rule Ensemble

Performs the Causal Rule Ensemble on a data set with a response variable, a treatment variable, and various features

Usage

cre(y, z, X, method_params, hyper_params)

Arguments

y

The observed response vector.

z

The treatment vector.

X

The covariate matrix.

method_params

Parameters for individual treatment effect, includig:

Parameters for Discovery

ratio_dis: The ratio of data delegated to the discovery sub-sample.

ite_method_dis: The method to estimate the discovery sample ITE.

include_ps_dis: Whether or not to include propensity score estimate as a covariate in discovery ITE estimation, considered only for BART, or CF.

ps_method_dis: The estimation model for the propensity score on the discovery subsample.

or_method_dis: The estimation model for the outcome regressions estimate_ite_aipw on the discovery subsample.

Parameters for Inference

ite_method_inf: The method to estimate the inference sample ITE.

include_ps_inf: Whether or not to include propensity score estimate as a covariate in inference ITE estimation, considered only for BART, or CF.

include_ps_inf: Whether or not to include propensity score estimate as a covariate in inference ITE estimation, considered only for BART, or CF.

ps_method_inf: The estimation model for the propensity score on the inference subsample.

or_method_inf: The estimation model for the outcome regressions in estimate_ite_aipw on the inference subsample.

Other Parameters:

include_offset: Whether or not to include an offset when estimating the ITE, for Poisson only.

offset_name: The name of the offset, if it is to be included.

cate_method: The method to estimate the CATE values.

cate_SL_library: The library used if cate_method is set to DRLearner.

filter_cate: Whether or not to filter rules with p-value <= 0.05.

hyper_params

The list of parameters required to tune the functions, including:

effect_modifiers: Effect Modifiers for Rules Generation.

ntrees_rf: The number of decision trees for randomForest.

ntrees_gbm: The number of decision trees for gradient boosting.

node_size: The minimum size of the trees' terminal nodes.

max_nodes: The maximum number of terminal nodes trees in the forest can have.

max_depth: The number of top levels from each tree considered to extract conditions.

max_decay: Decay Threshold for pruning the rules.

type_decay: Decay Type for pruning the rules (1: relative error; 2: error).

t: The common support used in generating the causal rules matrix.

q: The selection threshold used in selecting the causal rules.

stability_selection: Whether or not using stability selection for selecting the causal rules.

pfer_val: The Per-Family Error Rate, the expected number of false discoveries.

Value

An S3 object containing the matrix of Conditional Average Treatment Effect estimates