Estimate the Individual Treatment Effect (ITE)

Estimates the Individual Treatment Effect given a response vector, a treatment vector, a covariate matrix, and a desired algorithm.

Usage

estimate_ite(y, z, X, ite_method, ...)

Arguments

y

An observed response vector.

z

An observed treatment vector.

X

A covariate matrix.

ite_method

A method for estimating the Individual Treatment Effect. Some methods requires additional parameters. These parameters are mentioned in the indented blocks for each method and their definitions are provided at the end of this parameters list.

"slearner": S-Learner. - learner_y

"tlearner": T-Learner.

learner_y

"tpoisson": T-Poisson.

offset

"xlearner": X-Learner.

learner_y

"aipw": Augmented Inverse Probability Weighting.

learner_ps and learner_y

"bart": Bayesian Additive Regression Trees.

learner_ps

"cf": Causal Forest.

learner_ps

...

Additional parameters passed to different models.

Value

A list of ITE estimates.

Details

Additional parameters

learner_ps: An estimation method for the propensity score. This includes libraries for the SuperLearner package.

learner_y: An estimation model for the outcome. This includes libraries for the SuperLearner package.

offset: Name of the covariate to use as offset (i.e. "x1") for Poisson ITE Estimation. NULL if offset is not used.