Estimate the Individual Treatment Effect using Outcome Regression

Method for estimating the Individual Treatment Effect using Outcome Regression given a response vector, a treatment vector, and a features matrix

estimate_ite_or(y, z, X)

Arguments

y
the observed response vector

z
the treatment vector

X
the features matrix

Value

a vector of ITE estimates

Examples

dataset <- generate_cre_dataset(n = 1000, rho = 0, n_rules = 2, p = 10,
                                effect_size = 2, binary = FALSE)

# Initialize parameters
y <- dataset[["y"]]
z <- dataset[["z"]]
X <- as.data.frame(dataset[["X"]])

ite_list <- estimate_ite_or(y, z, X)

#> *****Into main of wbart
#> *****Data:
#> data:n,p,np: 387, 10, 613
#> y1,yn: 0.533299, 2.248639
#> x1,x[n*p]: 0.000000, 1.000000
#> xp1,xp[np*p]: 0.000000, 1.000000
#> *****Number of Trees: 200
#> *****Number of Cut Points: 1 ... 1
#> *****burn and ndpost: 100, 1000
#> *****Prior:beta,alpha,tau,nu,lambda: 2.000000,0.950000,0.161492,3.000000,0.298833
#> *****sigma: 1.238594
#> *****w (weights): 1.000000 ... 1.000000
#> *****Dirichlet:sparse,theta,omega,a,b,rho,augment: 0,0,1,0.5,1,10,0
#> *****nkeeptrain,nkeeptest,nkeeptestme,nkeeptreedraws: 1000,1000,1000,1000
#> *****printevery: 100
#> *****skiptr,skipte,skipteme,skiptreedraws: 1,1,1,1
#> 
#> MCMC
#> done 0 (out of 1100)
#> done 100 (out of 1100)
#> done 200 (out of 1100)
#> done 300 (out of 1100)
#> done 400 (out of 1100)
#> done 500 (out of 1100)
#> done 600 (out of 1100)
#> done 700 (out of 1100)
#> done 800 (out of 1100)
#> done 900 (out of 1100)
#> done 1000 (out of 1100)
#> time: 13s
#> check counts
#> trcnt,tecnt,temecnt,treedrawscnt: 1000,1000,1000,1000
#> *****Into main of wbart
#> *****Data:
#> data:n,p,np: 613, 10, 387
#> y1,yn: -0.574499, -1.331884
#> x1,x[n*p]: 0.000000, 1.000000
#> xp1,xp[np*p]: 0.000000, 1.000000
#> *****Number of Trees: 200
#> *****Number of Cut Points: 1 ... 1
#> *****burn and ndpost: 100, 1000
#> *****Prior:beta,alpha,tau,nu,lambda: 2.000000,0.950000,0.141730,3.000000,0.200642
#> *****sigma: 1.014906
#> *****w (weights): 1.000000 ... 1.000000
#> *****Dirichlet:sparse,theta,omega,a,b,rho,augment: 0,0,1,0.5,1,10,0
#> *****nkeeptrain,nkeeptest,nkeeptestme,nkeeptreedraws: 1000,1000,1000,1000
#> *****printevery: 100
#> *****skiptr,skipte,skipteme,skiptreedraws: 1,1,1,1
#> 
#> MCMC
#> done 0 (out of 1100)
#> done 100 (out of 1100)
#> done 200 (out of 1100)
#> done 300 (out of 1100)
#> done 400 (out of 1100)
#> done 500 (out of 1100)
#> done 600 (out of 1100)
#> done 700 (out of 1100)
#> done 800 (out of 1100)
#> done 900 (out of 1100)
#> done 1000 (out of 1100)
#> time: 20s
#> check counts
#> trcnt,tecnt,temecnt,treedrawscnt: 1000,1000,1000,1000