Estimate the Individual Treatment Effect using Poisson Regression

Estimates the Individual Treatment Effect using Poisson Regression given a response vector, a treatment vector, and a features matrix.

Usage

estimate_ite_poisson(y, z, X, X_names, include_offset, offset_name)

Arguments

y

the observed response vector

z

the treatment vector

X

the features matrix

X_names

the names of the covariates

include_offset

whether or not to include an offset when estimating the ITE, for poisson only

offset_name

the name of the offset, if it is to be included

Value

a vector of ITE estimates

Examples

dataset <- generate_cre_dataset(n = 1000, rho = 0, n_rules = 2, p = 10,
                                effect_size = 2, binary = FALSE)

# Initialize parameters
y <- abs(dataset[["y"]])
z <- dataset[["z"]]
X <- as.data.frame(dataset[["X"]])
X_names <- names(as.data.frame(X))
include_offset <- FALSE
offset_name <- NA

ite_list <- estimate_ite_poisson(y, z, X, X_names, include_offset, offset_name)
#> Warning: non-integer x = 1.811321
#> Warning: non-integer x = 0.402978
#> Warning: non-integer x = 2.876139
#> Warning: non-integer x = 2.408807
#> Warning: non-integer x = 0.357838
#> Warning: non-integer x = 1.336929
#> Warning: non-integer x = 3.945956
#> Warning: non-integer x = 2.684246
#> Warning: non-integer x = 2.243681
#> Warning: non-integer x = 0.170480
#> Warning: non-integer x = 1.344869
#> Warning: non-integer x = 0.662835
#> Warning: non-integer x = 3.528969
#> Warning: non-integer x = 0.964597
#> Warning: non-integer x = 1.941161
#> Warning: non-integer x = 2.408888
#> Warning: non-integer x = 0.212344
#> Warning: non-integer x = 0.148301
#> Warning: non-integer x = 2.013081
#> Warning: non-integer x = 0.435544
#> Warning: non-integer x = 2.287071
#> Warning: non-integer x = 0.370544
#> Warning: non-integer x = 2.038821
#> Warning: non-integer x = 0.515737
#> Warning: non-integer x = 0.308658
#> Warning: non-integer x = 0.062227
#> Warning: non-integer x = 2.491228
#> Warning: non-integer x = 0.414125
#> Warning: non-integer x = 2.442464
#> Warning: non-integer x = 1.035821
#> Warning: non-integer x = 0.585792
#> Warning: non-integer x = 3.193124
#> Warning: non-integer x = 2.837734
#> Warning: non-integer x = 1.845290
#> Warning: non-integer x = 1.674083
#> Warning: non-integer x = 1.221545
#> Warning: non-integer x = 0.339619
#> Warning: non-integer x = 1.868222
#> Warning: non-integer x = 2.241728
#> Warning: non-integer x = 2.437055
#> Warning: non-integer x = 1.872839
#> Warning: non-integer x = 0.532658
#> Warning: non-integer x = 1.368085
#> Warning: non-integer x = 1.545631
#> Warning: non-integer x = 0.354074
#> Warning: non-integer x = 1.276886
#> Warning: non-integer x = 2.679913
#> Warning: non-integer x = 0.253203
#> Warning: non-integer x = 3.082656
#> Warning: non-integer x = 1.583860
#> Warning: non-integer x = 0.066011
#> Warning: non-integer x = 2.706818
#> Warning: non-integer x = 2.188576
#> Warning: non-integer x = 2.980204
#> Warning: non-integer x = 2.335458
#> Warning: non-integer x = 2.471019
#> Warning: non-integer x = 0.449494
#> Warning: non-integer x = 0.907493
#> Warning: non-integer x = 0.338388
#> Warning: non-integer x = 2.437266
#> Warning: non-integer x = 0.768220
#> Warning: non-integer x = 1.374907
#> Warning: non-integer x = 2.119770
#> Warning: non-integer x = 1.549855
#> Warning: non-integer x = 2.487495
#> Warning: non-integer x = 0.506368
#> Warning: non-integer x = 0.644261
#> Warning: non-integer x = 2.345061
#> Warning: non-integer x = 2.355267
#> Warning: non-integer x = 0.625035
#> Warning: non-integer x = 0.052257
#> Warning: non-integer x = 0.436357
#> Warning: non-integer x = 2.798103
#> Warning: non-integer x = 3.131918
#> Warning: non-integer x = 3.396179
#> Warning: non-integer x = 0.351526
#> Warning: non-integer x = 1.996589
#> Warning: non-integer x = 2.963073
#> Warning: non-integer x = 0.072248
#> Warning: non-integer x = 0.021404
#> Warning: non-integer x = 1.221154
#> Warning: non-integer x = 1.388892
#> Warning: non-integer x = 0.972414
#> Warning: non-integer x = 0.864488
#> Warning: non-integer x = 1.304118
#> Warning: non-integer x = 1.967445
#> Warning: non-integer x = 1.147591
#> Warning: non-integer x = 2.155882
#> Warning: non-integer x = 0.759964
#> Warning: non-integer x = 2.192800
#> Warning: non-integer x = 5.227975
#> Warning: non-integer x = 1.176923
#> Warning: non-integer x = 0.774662
#> Warning: non-integer x = 2.584772
#> Warning: non-integer x = 0.760929
#> Warning: non-integer x = 2.315145
#> Warning: non-integer x = 1.151050
#> Warning: non-integer x = 1.470236
#> Warning: non-integer x = 0.516981
#> Warning: non-integer x = 2.440481
#> Warning: non-integer x = 0.685891
#> Warning: non-integer x = 1.977194
#> Warning: non-integer x = 0.390499
#> Warning: non-integer x = 0.681978
#> Warning: non-integer x = 2.466543
#> Warning: non-integer x = 1.111120
#> Warning: non-integer x = 0.116521
#> Warning: non-integer x = 2.337713
#> Warning: non-integer x = 0.858393
#> Warning: non-integer x = 2.209866
#> Warning: non-integer x = 1.292269
#> Warning: non-integer x = 0.890498
#> Warning: non-integer x = 1.854488
#> Warning: non-integer x = 1.031281
#> Warning: non-integer x = 1.017427
#> Warning: non-integer x = 2.532175
#> Warning: non-integer x = 1.942313
#> Warning: non-integer x = 1.781780
#> Warning: non-integer x = 0.673546
#> Warning: non-integer x = 1.898985
#> Warning: non-integer x = 1.561444
#> Warning: non-integer x = 1.171635
#> Warning: non-integer x = 2.685081
#> Warning: non-integer x = 3.559755
#> Warning: non-integer x = 2.068064
#> Warning: non-integer x = 2.408011
#> Warning: non-integer x = 1.175937
#> Warning: non-integer x = 2.219180
#> Warning: non-integer x = 0.191457
#> Warning: non-integer x = 3.136382
#> Warning: non-integer x = 2.887508
#> Warning: non-integer x = 2.424338
#> Warning: non-integer x = 1.391766
#> Warning: non-integer x = 0.755931
#> Warning: non-integer x = 1.125265
#> Warning: non-integer x = 2.983222
#> Warning: non-integer x = 1.847202
#> Warning: non-integer x = 1.624173
#> Warning: non-integer x = 3.756333
#> Warning: non-integer x = 3.534151
#> Warning: non-integer x = 1.847347
#> Warning: non-integer x = 2.478934
#> Warning: non-integer x = 2.514235
#> Warning: non-integer x = 1.479387
#> Warning: non-integer x = 2.926424
#> Warning: non-integer x = 0.866346
#> Warning: non-integer x = 0.748412
#> Warning: non-integer x = 3.650418
#> Warning: non-integer x = 0.663321
#> Warning: non-integer x = 4.068631
#> Warning: non-integer x = 3.388173
#> Warning: non-integer x = 2.285441
#> Warning: non-integer x = 1.067351
#> Warning: non-integer x = 2.287179
#> Warning: non-integer x = 1.508514
#> Warning: non-integer x = 1.269710
#> Warning: non-integer x = 2.182401
#> Warning: non-integer x = 2.146538
#> Warning: non-integer x = 1.846581
#> Warning: non-integer x = 0.293662
#> Warning: non-integer x = 1.150022
#> Warning: non-integer x = 1.560245
#> Warning: non-integer x = 1.755726
#> Warning: non-integer x = 3.271130
#> Warning: non-integer x = 0.237816
#> Warning: non-integer x = 0.730243
#> Warning: non-integer x = 2.167781
#> Warning: non-integer x = 0.196512
#> Warning: non-integer x = 2.106356
#> Warning: non-integer x = 0.774102
#> Warning: non-integer x = 2.445719
#> Warning: non-integer x = 1.778391
#> Warning: non-integer x = 0.687449
#> Warning: non-integer x = 3.615182
#> Warning: non-integer x = 3.565063
#> Warning: non-integer x = 1.521265
#> Warning: non-integer x = 0.428525
#> Warning: non-integer x = 1.684572
#> Warning: non-integer x = 4.442046
#> Warning: non-integer x = 0.356518
#> Warning: non-integer x = 2.747963
#> Warning: non-integer x = 2.052432
#> Warning: non-integer x = 2.137385
#> Warning: non-integer x = 0.445494
#> Warning: non-integer x = 0.410527
#> Warning: non-integer x = 1.626585
#> Warning: non-integer x = 0.531767
#> Warning: non-integer x = 2.554216
#> Warning: non-integer x = 2.684457
#> Warning: non-integer x = 4.445293
#> Warning: non-integer x = 1.399976
#> Warning: non-integer x = 1.743784
#> Warning: non-integer x = 2.213675
#> Warning: non-integer x = 2.654350
#> Warning: non-integer x = 2.151711
#> Warning: non-integer x = 1.001007
#> Warning: non-integer x = 2.131367
#> Warning: non-integer x = 1.258023
#> Warning: non-integer x = 0.386452
#> Warning: non-integer x = 0.890881
#> Warning: non-integer x = 1.043185
#> Warning: non-integer x = 1.883935
#> Warning: non-integer x = 0.904574
#> Warning: non-integer x = 2.546531
#> Warning: non-integer x = 1.427413
#> Warning: non-integer x = 0.698458
#> Warning: non-integer x = 3.215466
#> Warning: non-integer x = 0.705474
#> Warning: non-integer x = 1.838338
#> Warning: non-integer x = 0.256245
#> Warning: non-integer x = 1.184976
#> Warning: non-integer x = 1.333282
#> Warning: non-integer x = 0.067899
#> Warning: non-integer x = 1.245266
#> Warning: non-integer x = 2.060634
#> Warning: non-integer x = 2.572179
#> Warning: non-integer x = 0.478816
#> Warning: non-integer x = 1.761145
#> Warning: non-integer x = 1.199226
#> Warning: non-integer x = 2.430488
#> Warning: non-integer x = 0.013079
#> Warning: non-integer x = 1.465290
#> Warning: non-integer x = 0.594974
#> Warning: non-integer x = 4.133461
#> Warning: non-integer x = 0.599107
#> Warning: non-integer x = 1.706068
#> Warning: non-integer x = 2.128078
#> Warning: non-integer x = 0.152785
#> Warning: non-integer x = 4.310444
#> Warning: non-integer x = 3.068875
#> Warning: non-integer x = 0.512134
#> Warning: non-integer x = 0.691414
#> Warning: non-integer x = 1.266906
#> Warning: non-integer x = 2.763631
#> Warning: non-integer x = 0.740044
#> Warning: non-integer x = 1.499575
#> Warning: non-integer x = 0.287604
#> Warning: non-integer x = 0.636437
#> Warning: non-integer x = 2.927018
#> Warning: non-integer x = 2.255280
#> Warning: non-integer x = 2.415432
#> Warning: non-integer x = 1.844213
#> Warning: non-integer x = 2.094884
#> Warning: non-integer x = 1.041040
#> Warning: non-integer x = 0.865697
#> Warning: non-integer x = 1.330157
#> Warning: non-integer x = 1.683726
#> Warning: non-integer x = 2.282117
#> Warning: non-integer x = 3.003350
#> Warning: non-integer x = 4.740002
#> Warning: non-integer x = 3.436858
#> Warning: non-integer x = 0.159710
#> Warning: non-integer x = 0.792204
#> Warning: non-integer x = 4.531186
#> Warning: non-integer x = 0.469389
#> Warning: non-integer x = 0.789828
#> Warning: non-integer x = 3.817006
#> Warning: non-integer x = 2.785523
#> Warning: non-integer x = 2.326080
#> Warning: non-integer x = 1.003240
#> Warning: non-integer x = 3.198166
#> Warning: non-integer x = 1.092057
#> Warning: non-integer x = 2.072204
#> Warning: non-integer x = 1.419909
#> Warning: non-integer x = 2.984094
#> Warning: non-integer x = 0.271057
#> Warning: non-integer x = 1.577730
#> Warning: non-integer x = 1.855936
#> Warning: non-integer x = 4.399383
#> Warning: non-integer x = 3.960950
#> Warning: non-integer x = 0.725617
#> Warning: non-integer x = 1.658524
#> Warning: non-integer x = 4.239269
#> Warning: non-integer x = 5.192036
#> Warning: non-integer x = 3.303426
#> Warning: non-integer x = 2.455569
#> Warning: non-integer x = 3.107453
#> Warning: non-integer x = 2.543214
#> Warning: non-integer x = 1.498907
#> Warning: non-integer x = 0.993210
#> Warning: non-integer x = 2.503563
#> Warning: non-integer x = 0.514358
#> Warning: non-integer x = 0.211079
#> Warning: non-integer x = 2.665219
#> Warning: non-integer x = 2.309704
#> Warning: non-integer x = 2.847989
#> Warning: non-integer x = 0.753943
#> Warning: non-integer x = 4.946808
#> Warning: non-integer x = 0.215894
#> Warning: non-integer x = 3.764788
#> Warning: non-integer x = 0.390025
#> Warning: non-integer x = 2.235716
#> Warning: non-integer x = 2.998953
#> Warning: non-integer x = 3.283973
#> Warning: non-integer x = 2.490380
#> Warning: non-integer x = 1.214851
#> Warning: non-integer x = 1.028036
#> Warning: non-integer x = 1.046206
#> Warning: non-integer x = 1.735289
#> Warning: non-integer x = 4.533920
#> Warning: non-integer x = 0.847782
#> Warning: non-integer x = 3.244135
#> Warning: non-integer x = 1.310847
#> Warning: non-integer x = 2.744297
#> Warning: non-integer x = 2.889311
#> Warning: non-integer x = 0.022800
#> Warning: non-integer x = 3.968284
#> Warning: non-integer x = 0.138680
#> Warning: non-integer x = 1.437920
#> Warning: non-integer x = 0.919927
#> Warning: non-integer x = 1.162750
#> Warning: non-integer x = 3.029174
#> Warning: non-integer x = 0.070587
#> Warning: non-integer x = 0.659886
#> Warning: non-integer x = 0.261616
#> Warning: non-integer x = 1.196862
#> Warning: non-integer x = 1.384004
#> Warning: non-integer x = 0.679283
#> Warning: non-integer x = 1.273673
#> Warning: non-integer x = 3.712009
#> Warning: non-integer x = 1.542673
#> Warning: non-integer x = 2.069361
#> Warning: non-integer x = 1.501044
#> Warning: non-integer x = 2.252588
#> Warning: non-integer x = 1.632254
#> Warning: non-integer x = 3.685548
#> Warning: non-integer x = 1.394071
#> Warning: non-integer x = 1.480494
#> Warning: non-integer x = 0.309203
#> Warning: non-integer x = 0.015933
#> Warning: non-integer x = 0.406096
#> Warning: non-integer x = 2.473021
#> Warning: non-integer x = 1.916792
#> Warning: non-integer x = 4.886948
#> Warning: non-integer x = 1.559623
#> Warning: non-integer x = 1.988430
#> Warning: non-integer x = 0.434419
#> Warning: non-integer x = 2.591340
#> Warning: non-integer x = 2.162249
#> Warning: non-integer x = 2.529812
#> Warning: non-integer x = 2.282178
#> Warning: non-integer x = 2.910876
#> Warning: non-integer x = 0.337505
#> Warning: non-integer x = 3.067463
#> Warning: non-integer x = 0.606175
#> Warning: non-integer x = 2.919348
#> Warning: non-integer x = 1.356596
#> Warning: non-integer x = 2.650600
#> Warning: non-integer x = 1.006820
#> Warning: non-integer x = 1.544564
#> Warning: non-integer x = 0.942775
#> Warning: non-integer x = 1.135867
#> Warning: non-integer x = 0.050734
#> Warning: non-integer x = 0.652343
#> Warning: non-integer x = 1.859880
#> Warning: non-integer x = 0.159837
#> Warning: non-integer x = 1.779218
#> Warning: non-integer x = 0.929937
#> Warning: non-integer x = 1.566552
#> Warning: non-integer x = 2.269824
#> Warning: non-integer x = 2.884645
#> Warning: non-integer x = 1.879433
#> Warning: non-integer x = 2.108757
#> Warning: non-integer x = 0.737308
#> Warning: non-integer x = 1.153935
#> Warning: non-integer x = 1.850725
#> Warning: non-integer x = 1.780725
#> Warning: non-integer x = 0.817213
#> Warning: non-integer x = 2.540120
#> Warning: non-integer x = 2.349655
#> Warning: non-integer x = 1.678941
#> Warning: non-integer x = 1.446670
#> Warning: non-integer x = 0.635193
#> Warning: non-integer x = 4.022439
#> Warning: non-integer x = 2.991824
#> Warning: non-integer x = 3.923875
#> Warning: non-integer x = 2.437185
#> Warning: non-integer x = 1.170869
#> Warning: non-integer x = 2.672040
#> Warning: non-integer x = 2.787092
#> Warning: non-integer x = 0.927012
#> Warning: non-integer x = 1.863308
#> Warning: non-integer x = 1.181084
#> Warning: non-integer x = 3.942717
#> Warning: non-integer x = 4.197875
#> Warning: non-integer x = 2.208039
#> Warning: non-integer x = 0.970345
#> Warning: non-integer x = 1.884407
#> Warning: non-integer x = 3.139019
#> Warning: non-integer x = 2.574226
#> Warning: non-integer x = 0.686628
#> Warning: non-integer x = 2.269507
#> Warning: non-integer x = 1.206327
#> Warning: non-integer x = 3.642441
#> Warning: non-integer x = 3.700953
#> Warning: non-integer x = 0.406080
#> Warning: non-integer x = 2.902100
#> Warning: non-integer x = 0.033302
#> Warning: non-integer x = 1.227464
#> Warning: non-integer x = 2.515663
#> Warning: non-integer x = 1.331277
#> Warning: non-integer x = 1.048197
#> Warning: non-integer x = 1.005127
#> Warning: non-integer x = 0.511778
#> Warning: non-integer x = 0.557616
#> Warning: non-integer x = 0.085993
#> Warning: non-integer x = 2.631829
#> Warning: non-integer x = 0.351812
#> Warning: non-integer x = 2.431039
#> Warning: non-integer x = 2.352838
#> Warning: non-integer x = 3.077986
#> Warning: non-integer x = 2.209853
#> Warning: non-integer x = 0.676625
#> Warning: non-integer x = 0.914336
#> Warning: non-integer x = 2.078656
#> Warning: non-integer x = 0.297591
#> Warning: non-integer x = 0.986750
#> Warning: non-integer x = 2.413845
#> Warning: non-integer x = 0.041317
#> Warning: non-integer x = 1.752123
#> Warning: non-integer x = 3.476740
#> Warning: non-integer x = 0.791227
#> Warning: non-integer x = 0.444165
#> Warning: non-integer x = 1.219084
#> Warning: non-integer x = 0.466300
#> Warning: non-integer x = 0.987727
#> Warning: non-integer x = 2.347621
#> Warning: non-integer x = 0.713311
#> Warning: non-integer x = 1.179696
#> Warning: non-integer x = 1.897540
#> Warning: non-integer x = 0.257800
#> Warning: non-integer x = 1.703760
#> Warning: non-integer x = 1.891310
#> Warning: non-integer x = 0.426073
#> Warning: non-integer x = 1.970909
#> Warning: non-integer x = 0.791241
#> Warning: non-integer x = 0.049308
#> Warning: non-integer x = 3.164076
#> Warning: non-integer x = 1.446738
#> Warning: non-integer x = 1.384661
#> Warning: non-integer x = 0.714478
#> Warning: non-integer x = 1.442256
#> Warning: non-integer x = 0.367181
#> Warning: non-integer x = 1.683106
#> Warning: non-integer x = 3.685367
#> Warning: non-integer x = 0.721637
#> Warning: non-integer x = 1.549349
#> Warning: non-integer x = 2.303494
#> Warning: non-integer x = 1.842247
#> Warning: non-integer x = 1.839033
#> Warning: non-integer x = 0.602547
#> Warning: non-integer x = 1.772852
#> Warning: non-integer x = 1.796549
#> Warning: non-integer x = 1.871688
#> Warning: non-integer x = 1.705009
#> Warning: non-integer x = 3.265737
#> Warning: non-integer x = 0.716093
#> Warning: non-integer x = 1.015557
#> Warning: non-integer x = 1.607450
#> Warning: non-integer x = 2.180438
#> Warning: non-integer x = 1.219131
#> Warning: non-integer x = 3.255608
#> Warning: non-integer x = 0.266078
#> Warning: non-integer x = 2.726361
#> Warning: non-integer x = 2.060233
#> Warning: non-integer x = 0.541573
#> Warning: non-integer x = 0.576189
#> Warning: non-integer x = 2.275459
#> Warning: non-integer x = 0.454250
#> Warning: non-integer x = 0.000939
#> Warning: non-integer x = 1.271669
#> Warning: non-integer x = 0.142531
#> Warning: non-integer x = 1.398623
#> Warning: non-integer x = 1.389279
#> Warning: non-integer x = 0.020984
#> Warning: non-integer x = 3.310727
#> Warning: non-integer x = 2.694556
#> Warning: non-integer x = 0.292179
#> Warning: non-integer x = 0.106045
#> Warning: non-integer x = 0.902488
#> Warning: non-integer x = 1.860137
#> Warning: non-integer x = 1.513131
#> Warning: non-integer x = 0.031068
#> Warning: non-integer x = 1.021233
#> Warning: non-integer x = 1.518979
#> Warning: non-integer x = 1.790902
#> Warning: non-integer x = 0.061636
#> Warning: non-integer x = 0.497601
#> Warning: non-integer x = 0.253649
#> Warning: non-integer x = 2.024047
#> Warning: non-integer x = 0.729017
#> Warning: non-integer x = 2.514535
#> Warning: non-integer x = 0.621949
#> Warning: non-integer x = 1.296580
#> Warning: non-integer x = 2.376486
#> Warning: non-integer x = 1.096194
#> Warning: non-integer x = 1.610441
#> Warning: non-integer x = 0.511896
#> Warning: non-integer x = 1.831171
#> Warning: non-integer x = 0.186867
#> Warning: non-integer x = 0.850806
#> Warning: non-integer x = 3.148455
#> Warning: non-integer x = 1.188028
#> Warning: non-integer x = 2.449674
#> Warning: non-integer x = 1.086529
#> Warning: non-integer x = 0.439225
#> Warning: non-integer x = 2.280495
#> Warning: non-integer x = 0.971463
#> Warning: non-integer x = 2.143243
#> Warning: non-integer x = 0.340437
#> Warning: non-integer x = 0.079155
#> Warning: non-integer x = 0.425964
#> Warning: non-integer x = 0.990164
#> Warning: non-integer x = 0.441359
#> Warning: non-integer x = 0.565445
#> Warning: non-integer x = 1.461147
#> Warning: non-integer x = 0.521359
#> Warning: non-integer x = 4.164797
#> Warning: non-integer x = 0.620732
#> Warning: non-integer x = 0.628574
#> Warning: non-integer x = 0.812847
#> Warning: non-integer x = 0.345913
#> Warning: non-integer x = 1.511421
#> Warning: non-integer x = 1.650038
#> Warning: non-integer x = 0.115486
#> Warning: non-integer x = 1.397526
#> Warning: non-integer x = 2.390938
#> Warning: non-integer x = 0.344453
#> Warning: non-integer x = 2.396301
#> Warning: non-integer x = 1.213021
#> Warning: non-integer x = 2.111374
#> Warning: non-integer x = 1.743519
#> Warning: non-integer x = 0.022096
#> Warning: non-integer x = 0.335515
#> Warning: non-integer x = 0.598244
#> Warning: non-integer x = 0.563408
#> Warning: non-integer x = 3.955268
#> Warning: non-integer x = 1.464649
#> Warning: non-integer x = 1.962592
#> Warning: non-integer x = 2.335008
#> Warning: non-integer x = 3.526250
#> Warning: non-integer x = 2.481792
#> Warning: non-integer x = 0.198204
#> Warning: non-integer x = 0.249789
#> Warning: non-integer x = 1.041020
#> Warning: non-integer x = 0.256853
#> Warning: non-integer x = 0.345000
#> Warning: non-integer x = 1.753234
#> Warning: non-integer x = 1.007537
#> Warning: non-integer x = 2.046938
#> Warning: non-integer x = 1.177740
#> Warning: non-integer x = 1.893066
#> Warning: non-integer x = 2.051258
#> Warning: non-integer x = 2.824226
#> Warning: non-integer x = 1.362686
#> Warning: non-integer x = 1.779151
#> Warning: non-integer x = 1.633506
#> Warning: non-integer x = 2.168596
#> Warning: non-integer x = 0.729835
#> Warning: non-integer x = 0.933531
#> Warning: non-integer x = 2.352099
#> Warning: non-integer x = 0.133924
#> Warning: non-integer x = 0.448395
#> Warning: non-integer x = 1.416079
#> Warning: non-integer x = 1.818055
#> Warning: non-integer x = 3.025708
#> Warning: non-integer x = 0.999708
#> Warning: non-integer x = 0.146390
#> Warning: non-integer x = 2.631906
#> Warning: non-integer x = 0.311277
#> Warning: non-integer x = 1.639706
#> Warning: non-integer x = 0.436385
#> Warning: non-integer x = 0.482999
#> Warning: non-integer x = 1.502462
#> Warning: non-integer x = 0.647491
#> Warning: non-integer x = 0.373012
#> Warning: non-integer x = 2.695598
#> Warning: non-integer x = 2.266243
#> Warning: non-integer x = 0.108003
#> Warning: non-integer x = 2.570106
#> Warning: non-integer x = 0.572936
#> Warning: non-integer x = 0.920188
#> Warning: non-integer x = 0.652441
#> Warning: non-integer x = 0.125570
#> Warning: non-integer x = 1.634292
#> Warning: non-integer x = 2.223252
#> Warning: non-integer x = 2.070712
#> Warning: non-integer x = 1.072993
#> Warning: non-integer x = 0.110738
#> Warning: non-integer x = 2.361786
#> Warning: non-integer x = 1.106977
#> Warning: non-integer x = 0.351592
#> Warning: non-integer x = 1.657330
#> Warning: non-integer x = 1.149711
#> Warning: non-integer x = 2.196100
#> Warning: non-integer x = 1.524231
#> Warning: non-integer x = 1.213767
#> Warning: non-integer x = 1.629793
#> Warning: non-integer x = 1.050726
#> Warning: non-integer x = 1.136924
#> Warning: non-integer x = 2.656435
#> Warning: non-integer x = 0.982761
#> Warning: non-integer x = 1.201459
#> Warning: non-integer x = 1.777437
#> Warning: non-integer x = 2.250934
#> Warning: non-integer x = 1.840399
#> Warning: non-integer x = 0.949736
#> Warning: non-integer x = 1.506164
#> Warning: non-integer x = 1.962270
#> Warning: non-integer x = 2.158635
#> Warning: non-integer x = 0.351279
#> Warning: non-integer x = 1.205593
#> Warning: non-integer x = 1.392497
#> Warning: non-integer x = 2.775814
#> Warning: non-integer x = 0.591809
#> Warning: non-integer x = 2.331062
#> Warning: non-integer x = 0.155472
#> Warning: non-integer x = 0.610799
#> Warning: non-integer x = 0.934344
#> Warning: non-integer x = 2.598756
#> Warning: non-integer x = 1.547697
#> Warning: non-integer x = 1.133724
#> Warning: non-integer x = 0.610598
#> Warning: non-integer x = 2.379808
#> Warning: non-integer x = 1.479523
#> Warning: non-integer x = 0.165710
#> Warning: non-integer x = 1.127334
#> Warning: non-integer x = 0.772779
#> Warning: non-integer x = 0.934682
#> Warning: non-integer x = 1.406653
#> Warning: non-integer x = 1.713273
#> Warning: non-integer x = 0.471408
#> Warning: non-integer x = 1.768477
#> Warning: non-integer x = 0.920648
#> Warning: non-integer x = 1.384440
#> Warning: non-integer x = 0.578176
#> Warning: non-integer x = 0.272812
#> Warning: non-integer x = 2.712894
#> Warning: non-integer x = 0.530438
#> Warning: non-integer x = 2.640830
#> Warning: non-integer x = 0.040933
#> Warning: non-integer x = 1.206204
#> Warning: non-integer x = 0.646406
#> Warning: non-integer x = 1.420669
#> Warning: non-integer x = 0.318936
#> Warning: non-integer x = 2.069708
#> Warning: non-integer x = 3.007374
#> Warning: non-integer x = 1.739233
#> Warning: non-integer x = 1.763749
#> Warning: non-integer x = 1.557491
#> Warning: non-integer x = 3.359080
#> Warning: non-integer x = 0.797194
#> Warning: non-integer x = 2.168540
#> Warning: non-integer x = 0.485111
#> Warning: non-integer x = 4.400428
#> Warning: non-integer x = 2.818312
#> Warning: non-integer x = 1.521102
#> Warning: non-integer x = 2.141516
#> Warning: non-integer x = 3.317008
#> Warning: non-integer x = 3.430950
#> Warning: non-integer x = 1.161540
#> Warning: non-integer x = 1.064932
#> Warning: non-integer x = 0.784124
#> Warning: non-integer x = 1.899658
#> Warning: non-integer x = 2.786715
#> Warning: non-integer x = 0.127222
#> Warning: non-integer x = 0.364693
#> Warning: non-integer x = 2.288273
#> Warning: non-integer x = 3.004282
#> Warning: non-integer x = 0.572063
#> Warning: non-integer x = 1.478702
#> Warning: non-integer x = 1.763893
#> Warning: non-integer x = 1.473539
#> Warning: non-integer x = 2.310031
#> Warning: non-integer x = 0.354857
#> Warning: non-integer x = 0.925184
#> Warning: non-integer x = 0.943008
#> Warning: non-integer x = 1.540975
#> Warning: non-integer x = 0.130588
#> Warning: non-integer x = 0.542053
#> Warning: non-integer x = 1.372097
#> Warning: non-integer x = 0.293183
#> Warning: non-integer x = 0.954370
#> Warning: non-integer x = 1.408175
#> Warning: non-integer x = 1.409517
#> Warning: non-integer x = 1.952018
#> Warning: non-integer x = 0.524326
#> Warning: non-integer x = 0.886976
#> Warning: non-integer x = 1.653801
#> Warning: non-integer x = 1.126097
#> Warning: non-integer x = 1.055408
#> Warning: non-integer x = 2.654028
#> Warning: non-integer x = 1.637222
#> Warning: non-integer x = 2.497854
#> Warning: non-integer x = 2.926418
#> Warning: non-integer x = 1.571851
#> Warning: non-integer x = 0.730362
#> Warning: non-integer x = 1.411166
#> Warning: non-integer x = 1.748589
#> Warning: non-integer x = 0.408430
#> Warning: non-integer x = 1.745905
#> Warning: non-integer x = 1.226640
#> Warning: non-integer x = 0.766251
#> Warning: non-integer x = 1.286902
#> Warning: non-integer x = 2.674897
#> Warning: non-integer x = 0.523407
#> Warning: non-integer x = 2.807961
#> Warning: non-integer x = 0.160094
#> Warning: non-integer x = 1.345653
#> Warning: non-integer x = 0.240899
#> Warning: non-integer x = 2.022164
#> Warning: non-integer x = 1.185270
#> Warning: non-integer x = 2.102250
#> Warning: non-integer x = 1.376266
#> Warning: non-integer x = 0.665022
#> Warning: non-integer x = 0.368905
#> Warning: non-integer x = 1.212801
#> Warning: non-integer x = 1.060837
#> Warning: non-integer x = 2.631218
#> Warning: non-integer x = 2.047334
#> Warning: non-integer x = 4.227462
#> Warning: non-integer x = 0.849294
#> Warning: non-integer x = 0.537380
#> Warning: non-integer x = 3.533471
#> Warning: non-integer x = 2.901415
#> Warning: non-integer x = 2.810559
#> Warning: non-integer x = 1.489831
#> Warning: non-integer x = 0.119330
#> Warning: non-integer x = 0.364186
#> Warning: non-integer x = 3.041136
#> Warning: non-integer x = 1.096049
#> Warning: non-integer x = 4.557915
#> Warning: non-integer x = 1.219065
#> Warning: non-integer x = 2.212033
#> Warning: non-integer x = 0.546436
#> Warning: non-integer x = 0.982338
#> Warning: non-integer x = 1.765667
#> Warning: non-integer x = 0.335833
#> Warning: non-integer x = 0.035220
#> Warning: non-integer x = 0.129806
#> Warning: non-integer x = 1.780831
#> Warning: non-integer x = 0.620160
#> Warning: non-integer x = 1.121712
#> Warning: non-integer x = 0.994837
#> Warning: non-integer x = 0.722416
#> Warning: non-integer x = 0.871324
#> Warning: non-integer x = 2.042384
#> Warning: non-integer x = 1.628334
#> Warning: non-integer x = 3.234717
#> Warning: non-integer x = 1.105599
#> Warning: non-integer x = 0.545209
#> Warning: non-integer x = 1.806404
#> Warning: non-integer x = 1.005476
#> Warning: non-integer x = 2.541061
#> Warning: non-integer x = 0.242856
#> Warning: non-integer x = 1.881070
#> Warning: non-integer x = 2.297701
#> Warning: non-integer x = 0.022009
#> Warning: non-integer x = 0.371664
#> Warning: non-integer x = 2.247009
#> Warning: non-integer x = 1.695631
#> Warning: non-integer x = 0.833746
#> Warning: non-integer x = 0.942880
#> Warning: non-integer x = 2.009552
#> Warning: non-integer x = 1.692380
#> Warning: non-integer x = 0.187420
#> Warning: non-integer x = 1.187244
#> Warning: non-integer x = 1.158997
#> Warning: non-integer x = 0.106272
#> Warning: non-integer x = 1.000775
#> Warning: non-integer x = 0.040605
#> Warning: non-integer x = 0.873697
#> Warning: non-integer x = 3.566889
#> Warning: non-integer x = 1.245664
#> Warning: non-integer x = 0.022107
#> Warning: non-integer x = 0.714279
#> Warning: non-integer x = 0.924813
#> Warning: non-integer x = 0.355359
#> Warning: non-integer x = 0.327769
#> Warning: non-integer x = 2.068934
#> Warning: non-integer x = 2.152933
#> Warning: non-integer x = 2.359257
#> Warning: non-integer x = 0.687897
#> Warning: non-integer x = 0.737922
#> Warning: non-integer x = 0.641984
#> Warning: non-integer x = 2.513746
#> Warning: non-integer x = 0.194606
#> Warning: non-integer x = 0.735758
#> Warning: non-integer x = 1.068786
#> Warning: non-integer x = 2.478733
#> Warning: non-integer x = 2.779832
#> Warning: non-integer x = 2.025496
#> Warning: non-integer x = 1.596470
#> Warning: non-integer x = 1.593473
#> Warning: non-integer x = 1.400946
#> Warning: non-integer x = 2.763246
#> Warning: non-integer x = 1.104044
#> Warning: non-integer x = 0.531352
#> Warning: non-integer x = 1.083590
#> Warning: non-integer x = 2.097538
#> Warning: non-integer x = 2.550511
#> Warning: non-integer x = 2.064918
#> Warning: non-integer x = 2.350270
#> Warning: non-integer x = 1.209870
#> Warning: non-integer x = 1.561182
#> Warning: non-integer x = 0.224339
#> Warning: non-integer x = 0.621237
#> Warning: non-integer x = 0.041844
#> Warning: non-integer x = 2.057544
#> Warning: non-integer x = 2.105946
#> Warning: non-integer x = 5.419621
#> Warning: non-integer x = 0.613647
#> Warning: non-integer x = 0.982654
#> Warning: non-integer x = 0.132952
#> Warning: non-integer x = 3.390031
#> Warning: non-integer x = 2.249082
#> Warning: non-integer x = 0.121236
#> Warning: non-integer x = 0.768459
#> Warning: non-integer x = 0.552996
#> Warning: non-integer x = 2.197158
#> Warning: non-integer x = 0.389910
#> Warning: non-integer x = 1.782909
#> Warning: non-integer x = 1.130850
#> Warning: non-integer x = 0.208696
#> Warning: non-integer x = 1.879849
#> Warning: non-integer x = 1.061829
#> Warning: non-integer x = 2.981935
#> Warning: non-integer x = 0.112895
#> Warning: non-integer x = 0.276054
#> Warning: non-integer x = 0.461625
#> Warning: non-integer x = 1.418066
#> Warning: non-integer x = 2.335000
#> Warning: non-integer x = 0.636566
#> Warning: non-integer x = 0.775204
#> Warning: non-integer x = 1.595814
#> Warning: non-integer x = 3.787323
#> Warning: non-integer x = 1.604918
#> Warning: non-integer x = 2.502658
#> Warning: non-integer x = 0.411077
#> Warning: non-integer x = 1.529769
#> Warning: non-integer x = 2.760107
#> Warning: non-integer x = 1.648151
#> Warning: non-integer x = 1.072123
#> Warning: non-integer x = 0.936606
#> Warning: non-integer x = 1.709868
#> Warning: non-integer x = 3.221467
#> Warning: non-integer x = 0.628952
#> Warning: non-integer x = 0.540224
#> Warning: non-integer x = 1.104237
#> Warning: non-integer x = 1.237571
#> Warning: non-integer x = 3.379860
#> Warning: non-integer x = 2.783149
#> Warning: non-integer x = 0.922066
#> Warning: non-integer x = 1.254711
#> Warning: non-integer x = 2.236330
#> Warning: non-integer x = 2.302737
#> Warning: non-integer x = 0.611972
#> Warning: non-integer x = 1.545367
#> Warning: non-integer x = 3.013720
#> Warning: non-integer x = 3.340279
#> Warning: non-integer x = 0.771766
#> Warning: non-integer x = 0.300477
#> Warning: non-integer x = 0.028984
#> Warning: non-integer x = 3.257702
#> Warning: non-integer x = 0.139954
#> Warning: non-integer x = 0.998798
#> Warning: non-integer x = 1.759124
#> Warning: non-integer x = 1.311243
#> Warning: non-integer x = 0.847060
#> Warning: non-integer x = 2.856804
#> Warning: non-integer x = 1.778279
#> Warning: non-integer x = 0.843557
#> Warning: non-integer x = 2.204666
#> Warning: non-integer x = 1.542416
#> Warning: non-integer x = 3.735130
#> Warning: non-integer x = 2.390096
#> Warning: non-integer x = 0.210086
#> Warning: non-integer x = 2.343145
#> Warning: non-integer x = 3.316654
#> Warning: non-integer x = 1.078759
#> Warning: non-integer x = 1.258058
#> Warning: non-integer x = 0.348955
#> Warning: non-integer x = 1.735436
#> Warning: non-integer x = 1.044042
#> Warning: non-integer x = 1.145560
#> Warning: non-integer x = 2.700642
#> Warning: non-integer x = 0.493360
#> Warning: non-integer x = 0.068797
#> Warning: non-integer x = 0.980919
#> Warning: non-integer x = 0.354744
#> Warning: non-integer x = 2.537138
#> Warning: non-integer x = 0.682589
#> Warning: non-integer x = 0.023230
#> Warning: non-integer x = 1.868301
#> Warning: non-integer x = 2.570437
#> Warning: non-integer x = 0.753302
#> Warning: non-integer x = 1.199218
#> Warning: non-integer x = 2.393968
#> Warning: non-integer x = 2.177657
#> Warning: non-integer x = 2.019127
#> Warning: non-integer x = 1.150734
#> Warning: non-integer x = 0.077946
#> Warning: non-integer x = 0.648289
#> Warning: non-integer x = 0.669635
#> Warning: non-integer x = 0.895714
#> Warning: non-integer x = 2.404010
#> Warning: non-integer x = 0.886056
#> Warning: non-integer x = 0.907458
#> Warning: non-integer x = 0.436971
#> Warning: non-integer x = 0.121128
#> Warning: non-integer x = 0.870904
#> Warning: non-integer x = 3.082177
#> Warning: non-integer x = 0.632006
#> Warning: non-integer x = 0.515760
#> Warning: non-integer x = 0.762327
#> Warning: non-integer x = 0.124290
#> Warning: non-integer x = 1.056961
#> Warning: non-integer x = 0.200669
#> Warning: non-integer x = 2.426039
#> Warning: non-integer x = 2.338450
#> Warning: non-integer x = 0.349574
#> Warning: non-integer x = 1.654968
#> Warning: non-integer x = 0.695555
#> Warning: non-integer x = 0.345348
#> Warning: non-integer x = 1.094006
#> Warning: non-integer x = 1.845598
#> Warning: non-integer x = 2.952405
#> Warning: non-integer x = 1.309676
#> Warning: non-integer x = 0.299359
#> Warning: non-integer x = 1.905475
#> Warning: non-integer x = 0.259562
#> Warning: non-integer x = 1.757606
#> Warning: non-integer x = 2.096863
#> Warning: non-integer x = 0.092091
#> Warning: non-integer x = 0.868524
#> Warning: non-integer x = 0.688834
#> Warning: non-integer x = 0.509115
#> Warning: non-integer x = 1.145511
#> Warning: non-integer x = 0.641868
#> Warning: non-integer x = 0.588055
#> Warning: non-integer x = 2.310425
#> Warning: non-integer x = 1.496815
#> Warning: non-integer x = 1.647379
#> Warning: non-integer x = 0.455689
#> Warning: non-integer x = 3.258638
#> Warning: non-integer x = 1.015282
#> Warning: non-integer x = 0.897098
#> Warning: non-integer x = 1.726756
#> Warning: non-integer x = 0.327566
#> Warning: non-integer x = 1.044553
#> Warning: non-integer x = 2.309919
#> Warning: non-integer x = 0.092297
#> Warning: non-integer x = 2.291330
#> Warning: non-integer x = 1.525829
#> Warning: non-integer x = 4.014748
#> Warning: non-integer x = 1.253764
#> Warning: non-integer x = 0.871730
#> Warning: non-integer x = 2.012943
#> Warning: non-integer x = 0.641051
#> Warning: non-integer x = 1.930948
#> Warning: non-integer x = 0.039382
#> Warning: non-integer x = 1.593252
#> Warning: non-integer x = 0.911335
#> Warning: non-integer x = 0.490971
#> Warning: non-integer x = 0.632203
#> Warning: non-integer x = 3.105344
#> Warning: non-integer x = 2.584626
#> Warning: non-integer x = 1.265790
#> Warning: non-integer x = 0.029568
#> Warning: non-integer x = 0.209247
#> Warning: non-integer x = 2.181798
#> Warning: non-integer x = 0.967470
#> Warning: non-integer x = 0.070559
#> Warning: non-integer x = 3.459624
#> Warning: non-integer x = 0.971967
#> Warning: non-integer x = 0.053702
#> Warning: non-integer x = 2.860963
#> Warning: non-integer x = 0.567361
#> Warning: non-integer x = 0.090476
#> Warning: non-integer x = 0.915675
#> Warning: non-integer x = 1.280881
#> Warning: non-integer x = 0.822601
#> Warning: non-integer x = 3.208953
#> Warning: non-integer x = 2.075421
#> Warning: non-integer x = 0.280640
#> Warning: non-integer x = 1.341315
#> Warning: non-integer x = 2.700691
#> Warning: non-integer x = 0.005814
#> Warning: non-integer x = 0.296816
#> Warning: non-integer x = 0.144097
#> Warning: non-integer x = 0.835942
#> Warning: non-integer x = 1.223834
#> Warning: non-integer x = 0.087645
#> Warning: non-integer x = 0.485701
#> Warning: non-integer x = 0.556330
#> Warning: non-integer x = 1.057356
#> Warning: non-integer x = 1.321310
#> Warning: non-integer x = 1.044956
#> Warning: non-integer x = 2.261254
#> Warning: non-integer x = 3.073290