Extract (causal) decision rules

Extracts causal rules from the random forest or the gradient boosting algorithms.

Usage

extract_rules(treelist, X, max_depth, digits = 2)

Arguments

treelist

A list of decision trees.

X

Features matrix.

max_depth

A number of top levels from each tree considered.

digits

A Number of digits for rounding decision rules to extract conditions.

Value

A vector of (causal) decision rules.