The Causal Rule Ensemble

Performs the Causal Rule Ensemble on a dataset with a response variable, a treatment variable, and various features

Usage

cre(
  y,
  z,
  X,
  ratio_dis,
  ite_method_dis,
  include_ps_dis = NA,
  ps_method_dis = "SL.xgboost",
  or_method_dis = NA,
  ite_method_inf,
  include_ps_inf = NA,
  ps_method_inf = "SL.xgboost",
  or_method_inf = NA,
  ntrees_rf,
  ntrees_gbm,
  min_nodes,
  max_nodes,
  t,
  q,
  rules_method = NA,
  include_offset = FALSE,
  offset_name = NA,
  cate_method = "DRLearner",
  cate_SL_library = "SL.xgboost",
  filter_cate = FALSE
)

Arguments

y

the observed response vector

z

the treatment vector

X

the covariate matrix

ratio_dis

the ratio of data delegated to the discovery sub-sample

ite_method_dis

the method to estimate the discovery sample ITE

include_ps_dis

whether or not to include propensity score estimate as a covariate in discovery ITE estimation, considered only for BART, XBART, or CF

ps_method_dis

the estimation model for the propensity score on the discovery subsample

or_method_dis

the estimation model for the outcome regressions in estimate_ite_aipw on the discovery subsample

ite_method_inf

the method to estimate the inference sample ITE

include_ps_inf

whether or not to include propensity score estimate as a covariate in inference ITE estimation, considered only for BART, XBART, or CF

ps_method_inf

the estimation model for the propensity score on the inference subsample

or_method_inf

the estimation model for the outcome regressions in estimate_ite_aipw on the inference subsample

ntrees_rf

the number of decision trees for randomForest

ntrees_gbm

the number of decision trees for gradient boosting

min_nodes

the minimum size of the trees' terminal nodes

max_nodes

the maximum size of the trees' terminal nodes

t

the common support used in generating the causal rules matrix

q

the selection threshold used in selecting the causal rules

rules_method

the method for selecting causal rules with binary outcomes

include_offset

whether or not to include an offset when estimating the ITE, for poisson only

offset_name

the name of the offset, if it is to be included

cate_method

the method to estimate the CATE values

cate_SL_library

the library used if cate_method is set to DRLearner

filter_cate

whether or not to filter rules with p-value <= 0.05

Value

an S3 object containing the matrix of Conditional Average Treatment Effect estimates

Examples

dataset <- generate_cre_dataset(n = 1000, rho = 0, n_rules = 2, p = 10,
                                effect_size = 2, binary = FALSE)

cre_results <- cre(y = dataset[["y"]], z = dataset[["z"]],
                   X = as.data.frame(dataset[["X"]]), ratio_dis = 0.25,
                   ite_method_dis = "bart", include_ps_dis = TRUE,
                   ite_method_inf = "bart", include_ps_inf = TRUE,
                   ntrees_rf = 100, ntrees_gbm = 50, min_nodes = 20,
                   max_nodes = 5, t = 0.025, q = 0.8)
#> Loading required package: nnls
#> Loading required namespace: xgboost
#> fitting treatment model via method 'bart'
#> fitting response model via method 'bart'
#> Registered S3 method overwritten by 'RRF':
#>   method      from        
#>   plot.margin randomForest
#> 417 rules (length<=15) were extracted from the first 100 trees.
#> 100 rules (length<=15) were extracted from the first 50 trees.
#> Conducting Inference Subsample Analysis
#> fitting treatment model via method 'bart'
#> fitting response model via method 'bart'