Estimate the Individual Treatment Effect (ITE)

Estimates the Individual Treatment Effect given a response vector, a treatment vector, a covariate matrix, and a desired algorithm.

Usage

estimate_ite(y, z, X, ite_method, is_y_binary, ...)

Arguments

y

The observed response vector.

z

The observed treatment vector.

X

The covariate matrix.

ite_method

the method for estimating the Individual Treatment Effect. Some methods requires additional parameters. These parameters are mentioned in the indented blocks for each method and their definitions are provided at the end of this parameters list.

ipw: Inverse Propensity Weighting

ps_method

sipw: Stabilized Inverse Propensity Weighting

ps_method

aipw: Augmented Inverse Propensity Weighting

ps_method and oreg_method

oreg: Outcome Regression

bart: Bayesian Additive Regression Trees

include_ps and ps_method

xbart: Accelerated Bayesian Additive Regression Trees

include_ps and ps_method

bcf: Bayesian Causal Forest

ps_method

xbcf: Accelerated Bayesian Causal Forest

ps_method

cf: Causal Forest

include_ps and ps_method

poisson: Poisson Estimation

X_names, include_offset, offset_name

is_y_binary

whether or not the outcome is binary

...

A dditional parameters passed to different models.

Value

A list that includes:

raw ITE estimates

standardized ITE estimates, and

standard deviations for the ITE estimates.

Details

Additional parameters

include_ps: Whether or not to include propensity score estimate as a covariate in ITE estimation.

ps_method: Estimation method for the propensity score. This include libraries for the SuperLearner package.

oreg_method: The estimation model for the outcome regressions. This include libraries for the SuperLearner package.

X_names: The names of the covariates. (TODO: Remove from input params.)

include_offset: Whether or not to include an offset when estimating the ITE.

offset_name: The name of the offset.