Estimate the Individual Treatment Effect

Estimates the Individual Treatment Effect given a response vector, a treatment vector, a features matrix, and a desired algorithm.

Usage

estimate_ite(
  y,
  z,
  X,
  ite_method,
  include_ps,
  ps_method,
  or_method,
  binary,
  X_names,
  include_offset,
  offset_name
)

Arguments

y

the observed response vector

z

the treatment vector

X

the features matrix

ite_method

the method for estimating the Individual Treatment Effect:

ipw: Inverse Propensity Weighting

sipw: Stabilized Inverse Propensity Weighting

aipw: Augmented Inverse Propensity Weighting

or: Outcome Regression, TODO: change this into a non reserved term.

bart: BART

xbart: Accelerated BART

bcf: Bayesian Causal Forest

xbcf: Accelerated Bayesian Causal Forest

cf: Causal Forest

poisson: Poisson Estimation

include_ps

whether or not to include propensity score estimate as a covariate in ITE estimation

ps_method

estimation method for the propensity score

or_method

the estimation model for the outcome regressions in estimate_ite_aipw

binary

whether or not the outcome is binary

X_names

the names of the covariates

include_offset

whether or not to include an offset when estimating the ITE, for poisson only

offset_name

the name of the offset, if it is to be included

Value

a list that includes:

raw ITE estimates

standardized ITE estimates, and

standard deviations for the ITE estimates.

Examples

dataset <- generate_cre_dataset(n = 1000, rho = 0, n_rules = 2, p = 10,
                                effect_size = 2, binary = FALSE)

# Initialize parameters
y <- dataset[["y"]]
z <- dataset[["z"]]
X <- as.data.frame(dataset[["X"]])
ite_method <- "bart"
include_ps <- TRUE
ps_method <- "SL.xgboost"
or_method <- NA
binary <- FALSE
X_names <- names(as.data.frame(X))
include_offset <- FALSE
offset_name <- NA

ite_list <- estimate_ite(y, z, X, ite_method, include_ps, ps_method, or_method,
                         binary, X_names, include_offset, offset_name)
#> fitting treatment model via method 'bart'
#> fitting response model via method 'bart'