Estimate the Individual Treatment Effect using Poisson Regression

Estimates the Individual Treatment Effect using Poisson Regression given a response vector, a treatment vector, and a features matrix.

estimate_ite_poisson(y, z, X, X_names, include_offset, offset_name)

Arguments

y
the observed response vector

z
the treatment vector

X
the features matrix

X_names
the names of the covariates

include_offset
whether or not to include an offset when estimating the ITE, for poisson only

offset_name
the name of the offset, if it is to be included

Value

a vector of ITE estimates

Examples

dataset <- generate_cre_dataset(n = 1000, rho = 0, n_rules = 2, p = 10,
                                effect_size = 2, binary = FALSE)

# Initialize parameters
y <- abs(dataset[["y"]])
z <- dataset[["z"]]
X <- as.data.frame(dataset[["X"]])
X_names <- names(as.data.frame(X))
include_offset <- FALSE
offset_name <- NA

ite_list <- estimate_ite_poisson(y, z, X, X_names, include_offset, offset_name)

#> Warning: non-integer x = 0.561275

#> Warning: non-integer x = 0.648203

#> Warning: non-integer x = 3.919623

#> Warning: non-integer x = 2.657466

#> Warning: non-integer x = 1.644176

#> Warning: non-integer x = 3.317281

#> Warning: non-integer x = 2.143061

#> Warning: non-integer x = 3.437782

#> Warning: non-integer x = 1.241689

#> Warning: non-integer x = 3.757908

#> Warning: non-integer x = 0.209552

#> Warning: non-integer x = 2.152663

#> Warning: non-integer x = 2.371491

#> Warning: non-integer x = 2.698321

#> Warning: non-integer x = 1.378461

#> Warning: non-integer x = 1.596039

#> Warning: non-integer x = 0.946965

#> Warning: non-integer x = 1.356498

#> Warning: non-integer x = 1.236284

#> Warning: non-integer x = 1.759058

#> Warning: non-integer x = 2.567343

#> Warning: non-integer x = 1.097644

#> Warning: non-integer x = 1.440120

#> Warning: non-integer x = 0.208551

#> Warning: non-integer x = 2.855150

#> Warning: non-integer x = 2.020701

#> Warning: non-integer x = 0.156327

#> Warning: non-integer x = 2.827326

#> Warning: non-integer x = 0.574688

#> Warning: non-integer x = 1.539955

#> Warning: non-integer x = 1.051813

#> Warning: non-integer x = 1.040030

#> Warning: non-integer x = 1.268571

#> Warning: non-integer x = 0.869475

#> Warning: non-integer x = 1.983412

#> Warning: non-integer x = 4.478592

#> Warning: non-integer x = 1.916429

#> Warning: non-integer x = 2.812762

#> Warning: non-integer x = 2.133844

#> Warning: non-integer x = 1.358142

#> Warning: non-integer x = 0.402085

#> Warning: non-integer x = 3.311727

#> Warning: non-integer x = 0.526314

#> Warning: non-integer x = 2.167979

#> Warning: non-integer x = 2.228035

#> Warning: non-integer x = 0.973401

#> Warning: non-integer x = 0.678107

#> Warning: non-integer x = 3.112933

#> Warning: non-integer x = 1.478960

#> Warning: non-integer x = 1.824967

#> Warning: non-integer x = 0.562808

#> Warning: non-integer x = 1.575223

#> Warning: non-integer x = 0.651571

#> Warning: non-integer x = 3.849445

#> Warning: non-integer x = 2.977266

#> Warning: non-integer x = 0.960863

#> Warning: non-integer x = 4.260476

#> Warning: non-integer x = 2.576696

#> Warning: non-integer x = 1.137835

#> Warning: non-integer x = 1.475499

#> Warning: non-integer x = 3.714272

#> Warning: non-integer x = 3.766654

#> Warning: non-integer x = 2.031047

#> Warning: non-integer x = 1.439177

#> Warning: non-integer x = 1.862285

#> Warning: non-integer x = 1.499477

#> Warning: non-integer x = 0.258627

#> Warning: non-integer x = 0.698785

#> Warning: non-integer x = 0.015993

#> Warning: non-integer x = 1.006601

#> Warning: non-integer x = 2.618881

#> Warning: non-integer x = 2.645140

#> Warning: non-integer x = 3.149717

#> Warning: non-integer x = 2.039342

#> Warning: non-integer x = 1.175812

#> Warning: non-integer x = 2.129836

#> Warning: non-integer x = 2.549794

#> Warning: non-integer x = 0.767117

#> Warning: non-integer x = 0.031623

#> Warning: non-integer x = 2.746417

#> Warning: non-integer x = 1.938351

#> Warning: non-integer x = 2.331134

#> Warning: non-integer x = 0.010759

#> Warning: non-integer x = 2.088310

#> Warning: non-integer x = 3.355772

#> Warning: non-integer x = 1.824964

#> Warning: non-integer x = 2.829548

#> Warning: non-integer x = 1.410531

#> Warning: non-integer x = 0.393027

#> Warning: non-integer x = 2.517333

#> Warning: non-integer x = 0.908666

#> Warning: non-integer x = 2.143447

#> Warning: non-integer x = 3.414454

#> Warning: non-integer x = 2.426473

#> Warning: non-integer x = 2.060160

#> Warning: non-integer x = 1.055323

#> Warning: non-integer x = 1.370427

#> Warning: non-integer x = 0.751894

#> Warning: non-integer x = 2.081515

#> Warning: non-integer x = 0.920682

#> Warning: non-integer x = 1.396991

#> Warning: non-integer x = 0.914326

#> Warning: non-integer x = 1.735770

#> Warning: non-integer x = 3.484263

#> Warning: non-integer x = 1.155243

#> Warning: non-integer x = 1.782704

#> Warning: non-integer x = 0.024143

#> Warning: non-integer x = 2.559072

#> Warning: non-integer x = 0.323339

#> Warning: non-integer x = 0.815259

#> Warning: non-integer x = 1.725269

#> Warning: non-integer x = 0.667095

#> Warning: non-integer x = 1.444870

#> Warning: non-integer x = 1.311868

#> Warning: non-integer x = 0.988491

#> Warning: non-integer x = 2.296640

#> Warning: non-integer x = 2.863578

#> Warning: non-integer x = 3.469705

#> Warning: non-integer x = 3.711573

#> Warning: non-integer x = 2.867492

#> Warning: non-integer x = 2.478267

#> Warning: non-integer x = 2.688372

#> Warning: non-integer x = 1.823532

#> Warning: non-integer x = 0.922272

#> Warning: non-integer x = 0.469490

#> Warning: non-integer x = 4.219692

#> Warning: non-integer x = 0.796669

#> Warning: non-integer x = 3.153070

#> Warning: non-integer x = 1.867147

#> Warning: non-integer x = 3.224987

#> Warning: non-integer x = 0.915748

#> Warning: non-integer x = 0.419327

#> Warning: non-integer x = 1.064850

#> Warning: non-integer x = 0.875286

#> Warning: non-integer x = 0.376500

#> Warning: non-integer x = 1.741592

#> Warning: non-integer x = 1.406075

#> Warning: non-integer x = 2.798934

#> Warning: non-integer x = 1.190486

#> Warning: non-integer x = 1.188992

#> Warning: non-integer x = 0.230422

#> Warning: non-integer x = 2.183167

#> Warning: non-integer x = 3.540589

#> Warning: non-integer x = 0.433063

#> Warning: non-integer x = 0.402775

#> Warning: non-integer x = 1.299260

#> Warning: non-integer x = 0.134568

#> Warning: non-integer x = 0.854101

#> Warning: non-integer x = 0.634740

#> Warning: non-integer x = 2.604213

#> Warning: non-integer x = 2.036294

#> Warning: non-integer x = 0.235343

#> Warning: non-integer x = 1.004692

#> Warning: non-integer x = 1.993279

#> Warning: non-integer x = 0.858994

#> Warning: non-integer x = 0.864824

#> Warning: non-integer x = 1.096887

#> Warning: non-integer x = 0.277522

#> Warning: non-integer x = 0.488706

#> Warning: non-integer x = 0.417216

#> Warning: non-integer x = 0.814416

#> Warning: non-integer x = 4.548176

#> Warning: non-integer x = 1.461451

#> Warning: non-integer x = 1.425597

#> Warning: non-integer x = 1.576737

#> Warning: non-integer x = 3.118761

#> Warning: non-integer x = 0.786696

#> Warning: non-integer x = 2.506151

#> Warning: non-integer x = 1.606704

#> Warning: non-integer x = 0.947415

#> Warning: non-integer x = 1.601665

#> Warning: non-integer x = 3.017111

#> Warning: non-integer x = 0.853756

#> Warning: non-integer x = 3.459178

#> Warning: non-integer x = 1.278983

#> Warning: non-integer x = 3.046017

#> Warning: non-integer x = 0.196567

#> Warning: non-integer x = 3.502443

#> Warning: non-integer x = 3.641926

#> Warning: non-integer x = 2.656236

#> Warning: non-integer x = 1.058761

#> Warning: non-integer x = 3.208190

#> Warning: non-integer x = 0.071950

#> Warning: non-integer x = 2.451656

#> Warning: non-integer x = 3.376296

#> Warning: non-integer x = 0.717065

#> Warning: non-integer x = 2.263023

#> Warning: non-integer x = 0.949068

#> Warning: non-integer x = 0.360216

#> Warning: non-integer x = 0.649213

#> Warning: non-integer x = 4.088461

#> Warning: non-integer x = 1.095757

#> Warning: non-integer x = 1.950895

#> Warning: non-integer x = 2.711752

#> Warning: non-integer x = 1.853521

#> Warning: non-integer x = 0.573233

#> Warning: non-integer x = 0.543856

#> Warning: non-integer x = 0.307939

#> Warning: non-integer x = 3.864515

#> Warning: non-integer x = 4.536054

#> Warning: non-integer x = 0.573625

#> Warning: non-integer x = 1.964911

#> Warning: non-integer x = 2.773251

#> Warning: non-integer x = 3.082338

#> Warning: non-integer x = 2.017575

#> Warning: non-integer x = 1.076067

#> Warning: non-integer x = 1.211578

#> Warning: non-integer x = 0.885016

#> Warning: non-integer x = 2.655167

#> Warning: non-integer x = 2.252974

#> Warning: non-integer x = 2.597602

#> Warning: non-integer x = 1.394409

#> Warning: non-integer x = 0.558752

#> Warning: non-integer x = 4.061368

#> Warning: non-integer x = 2.816387

#> Warning: non-integer x = 1.287018

#> Warning: non-integer x = 0.573912

#> Warning: non-integer x = 0.995611

#> Warning: non-integer x = 1.041603

#> Warning: non-integer x = 2.072771

#> Warning: non-integer x = 0.367764

#> Warning: non-integer x = 0.026254

#> Warning: non-integer x = 2.416716

#> Warning: non-integer x = 2.648032

#> Warning: non-integer x = 3.338643

#> Warning: non-integer x = 0.631858

#> Warning: non-integer x = 0.364245

#> Warning: non-integer x = 3.204283

#> Warning: non-integer x = 1.357256

#> Warning: non-integer x = 0.438706

#> Warning: non-integer x = 1.553354

#> Warning: non-integer x = 1.905059

#> Warning: non-integer x = 2.737527

#> Warning: non-integer x = 1.054186

#> Warning: non-integer x = 0.989710

#> Warning: non-integer x = 3.422654

#> Warning: non-integer x = 1.119097

#> Warning: non-integer x = 0.667936

#> Warning: non-integer x = 1.002793

#> Warning: non-integer x = 0.481181

#> Warning: non-integer x = 1.502825

#> Warning: non-integer x = 1.506203

#> Warning: non-integer x = 0.440435

#> Warning: non-integer x = 2.361688

#> Warning: non-integer x = 0.070704

#> Warning: non-integer x = 1.454799

#> Warning: non-integer x = 2.250302

#> Warning: non-integer x = 3.650746

#> Warning: non-integer x = 2.493065

#> Warning: non-integer x = 0.296348

#> Warning: non-integer x = 0.278062

#> Warning: non-integer x = 3.039420

#> Warning: non-integer x = 1.121392

#> Warning: non-integer x = 2.933193

#> Warning: non-integer x = 2.883757

#> Warning: non-integer x = 1.529175

#> Warning: non-integer x = 0.077913

#> Warning: non-integer x = 0.389776

#> Warning: non-integer x = 1.781993

#> Warning: non-integer x = 2.451721

#> Warning: non-integer x = 3.964669

#> Warning: non-integer x = 3.697422

#> Warning: non-integer x = 2.397255

#> Warning: non-integer x = 1.864896

#> Warning: non-integer x = 3.760344

#> Warning: non-integer x = 3.533307

#> Warning: non-integer x = 0.171729

#> Warning: non-integer x = 2.198218

#> Warning: non-integer x = 1.042268

#> Warning: non-integer x = 0.159717

#> Warning: non-integer x = 0.605871

#> Warning: non-integer x = 1.221511

#> Warning: non-integer x = 3.423210

#> Warning: non-integer x = 0.081013

#> Warning: non-integer x = 1.995103

#> Warning: non-integer x = 1.285678

#> Warning: non-integer x = 0.726278

#> Warning: non-integer x = 2.015768

#> Warning: non-integer x = 1.299780

#> Warning: non-integer x = 0.681435

#> Warning: non-integer x = 1.201008

#> Warning: non-integer x = 0.510039

#> Warning: non-integer x = 2.644764

#> Warning: non-integer x = 1.726107

#> Warning: non-integer x = 0.763052

#> Warning: non-integer x = 1.373281

#> Warning: non-integer x = 0.243965

#> Warning: non-integer x = 2.306897

#> Warning: non-integer x = 1.054610

#> Warning: non-integer x = 0.010584

#> Warning: non-integer x = 0.001366

#> Warning: non-integer x = 3.359558

#> Warning: non-integer x = 0.739824

#> Warning: non-integer x = 1.765881

#> Warning: non-integer x = 0.821337

#> Warning: non-integer x = 0.406412

#> Warning: non-integer x = 1.798472

#> Warning: non-integer x = 2.177113

#> Warning: non-integer x = 3.183758

#> Warning: non-integer x = 3.224639

#> Warning: non-integer x = 0.091018

#> Warning: non-integer x = 2.753354

#> Warning: non-integer x = 1.860367

#> Warning: non-integer x = 1.206660

#> Warning: non-integer x = 0.457996

#> Warning: non-integer x = 0.142643

#> Warning: non-integer x = 0.047671

#> Warning: non-integer x = 1.066248

#> Warning: non-integer x = 3.115742

#> Warning: non-integer x = 0.533851

#> Warning: non-integer x = 1.101264

#> Warning: non-integer x = 2.138435

#> Warning: non-integer x = 0.394769

#> Warning: non-integer x = 0.631309

#> Warning: non-integer x = 2.670547

#> Warning: non-integer x = 1.252463

#> Warning: non-integer x = 1.638265

#> Warning: non-integer x = 0.522995

#> Warning: non-integer x = 3.128659

#> Warning: non-integer x = 2.998500

#> Warning: non-integer x = 2.922597

#> Warning: non-integer x = 2.639912

#> Warning: non-integer x = 1.366159

#> Warning: non-integer x = 0.956259

#> Warning: non-integer x = 0.585657

#> Warning: non-integer x = 1.836694

#> Warning: non-integer x = 3.171444

#> Warning: non-integer x = 0.744341

#> Warning: non-integer x = 2.492741

#> Warning: non-integer x = 0.308974

#> Warning: non-integer x = 2.894241

#> Warning: non-integer x = 1.909796

#> Warning: non-integer x = 2.454694

#> Warning: non-integer x = 1.390263

#> Warning: non-integer x = 2.322672

#> Warning: non-integer x = 0.339824

#> Warning: non-integer x = 0.556165

#> Warning: non-integer x = 0.915597

#> Warning: non-integer x = 1.768398

#> Warning: non-integer x = 1.329308

#> Warning: non-integer x = 1.017010

#> Warning: non-integer x = 0.930497

#> Warning: non-integer x = 2.679409

#> Warning: non-integer x = 0.385744

#> Warning: non-integer x = 0.608932

#> Warning: non-integer x = 4.355693

#> Warning: non-integer x = 0.836605

#> Warning: non-integer x = 2.118055

#> Warning: non-integer x = 0.318716

#> Warning: non-integer x = 2.005969

#> Warning: non-integer x = 2.596457

#> Warning: non-integer x = 1.877758

#> Warning: non-integer x = 1.429710

#> Warning: non-integer x = 2.331957

#> Warning: non-integer x = 1.337181

#> Warning: non-integer x = 0.427220

#> Warning: non-integer x = 2.134714

#> Warning: non-integer x = 1.159701

#> Warning: non-integer x = 3.258810

#> Warning: non-integer x = 1.712425

#> Warning: non-integer x = 2.108391

#> Warning: non-integer x = 0.168044

#> Warning: non-integer x = 1.832262

#> Warning: non-integer x = 1.966553

#> Warning: non-integer x = 2.767968

#> Warning: non-integer x = 2.249412

#> Warning: non-integer x = 1.333172

#> Warning: non-integer x = 1.076620

#> Warning: non-integer x = 3.417189

#> Warning: non-integer x = 2.387812

#> Warning: non-integer x = 6.402624

#> Warning: non-integer x = 1.972176

#> Warning: non-integer x = 0.480543

#> Warning: non-integer x = 2.701043

#> Warning: non-integer x = 2.529010

#> Warning: non-integer x = 1.401318

#> Warning: non-integer x = 1.032199

#> Warning: non-integer x = 0.434504

#> Warning: non-integer x = 1.615053

#> Warning: non-integer x = 0.585753

#> Warning: non-integer x = 0.411261

#> Warning: non-integer x = 0.438237

#> Warning: non-integer x = 0.025492

#> Warning: non-integer x = 1.880133

#> Warning: non-integer x = 1.984105

#> Warning: non-integer x = 0.271907

#> Warning: non-integer x = 1.973085

#> Warning: non-integer x = 4.026740

#> Warning: non-integer x = 2.049262

#> Warning: non-integer x = 2.790617

#> Warning: non-integer x = 1.151451

#> Warning: non-integer x = 2.074055

#> Warning: non-integer x = 0.137724

#> Warning: non-integer x = 1.617200

#> Warning: non-integer x = 2.995280

#> Warning: non-integer x = 1.086417

#> Warning: non-integer x = 1.509486

#> Warning: non-integer x = 0.206514

#> Warning: non-integer x = 0.560215

#> Warning: non-integer x = 0.420912

#> Warning: non-integer x = 2.079624

#> Warning: non-integer x = 1.005399

#> Warning: non-integer x = 0.701766

#> Warning: non-integer x = 1.357641

#> Warning: non-integer x = 0.419849

#> Warning: non-integer x = 0.346029

#> Warning: non-integer x = 2.044869

#> Warning: non-integer x = 0.512253

#> Warning: non-integer x = 2.893960

#> Warning: non-integer x = 2.416155

#> Warning: non-integer x = 0.529860

#> Warning: non-integer x = 1.028049

#> Warning: non-integer x = 0.397016

#> Warning: non-integer x = 2.391038

#> Warning: non-integer x = 2.978235

#> Warning: non-integer x = 2.406964

#> Warning: non-integer x = 0.714409

#> Warning: non-integer x = 2.150702

#> Warning: non-integer x = 1.419674

#> Warning: non-integer x = 1.899990

#> Warning: non-integer x = 1.725677

#> Warning: non-integer x = 1.383888

#> Warning: non-integer x = 1.442961

#> Warning: non-integer x = 1.012655

#> Warning: non-integer x = 0.268726

#> Warning: non-integer x = 0.818513

#> Warning: non-integer x = 2.015010

#> Warning: non-integer x = 2.080204

#> Warning: non-integer x = 3.323597

#> Warning: non-integer x = 2.078948

#> Warning: non-integer x = 0.048203

#> Warning: non-integer x = 1.043572

#> Warning: non-integer x = 2.270624

#> Warning: non-integer x = 0.925994

#> Warning: non-integer x = 0.565932

#> Warning: non-integer x = 0.808024

#> Warning: non-integer x = 2.015787

#> Warning: non-integer x = 0.178305

#> Warning: non-integer x = 1.389613

#> Warning: non-integer x = 1.407830

#> Warning: non-integer x = 1.845111

#> Warning: non-integer x = 2.635657

#> Warning: non-integer x = 1.915537

#> Warning: non-integer x = 1.442094

#> Warning: non-integer x = 0.727311

#> Warning: non-integer x = 0.462535

#> Warning: non-integer x = 0.582087

#> Warning: non-integer x = 1.748971

#> Warning: non-integer x = 2.845638

#> Warning: non-integer x = 2.355894

#> Warning: non-integer x = 0.524354

#> Warning: non-integer x = 1.083035

#> Warning: non-integer x = 0.534086

#> Warning: non-integer x = 0.573029

#> Warning: non-integer x = 0.813183

#> Warning: non-integer x = 1.795335

#> Warning: non-integer x = 4.043724

#> Warning: non-integer x = 1.467377

#> Warning: non-integer x = 0.239162

#> Warning: non-integer x = 0.350398

#> Warning: non-integer x = 2.837647

#> Warning: non-integer x = 0.978767

#> Warning: non-integer x = 1.439926

#> Warning: non-integer x = 1.019367

#> Warning: non-integer x = 0.179580

#> Warning: non-integer x = 1.719098

#> Warning: non-integer x = 0.313407

#> Warning: non-integer x = 0.206228

#> Warning: non-integer x = 0.302462

#> Warning: non-integer x = 0.732433

#> Warning: non-integer x = 0.526161

#> Warning: non-integer x = 1.788174

#> Warning: non-integer x = 0.265164

#> Warning: non-integer x = 1.416717

#> Warning: non-integer x = 1.804409

#> Warning: non-integer x = 1.139371

#> Warning: non-integer x = 0.949356

#> Warning: non-integer x = 0.716454

#> Warning: non-integer x = 3.085556

#> Warning: non-integer x = 0.862935

#> Warning: non-integer x = 0.308127

#> Warning: non-integer x = 3.556577

#> Warning: non-integer x = 1.058434

#> Warning: non-integer x = 2.365110

#> Warning: non-integer x = 0.735202

#> Warning: non-integer x = 1.182655

#> Warning: non-integer x = 1.573537

#> Warning: non-integer x = 1.193837

#> Warning: non-integer x = 2.395599

#> Warning: non-integer x = 2.406869

#> Warning: non-integer x = 0.154285

#> Warning: non-integer x = 0.062661

#> Warning: non-integer x = 3.324814

#> Warning: non-integer x = 1.970802

#> Warning: non-integer x = 2.270548

#> Warning: non-integer x = 1.399041

#> Warning: non-integer x = 1.370060

#> Warning: non-integer x = 0.733201

#> Warning: non-integer x = 1.091762

#> Warning: non-integer x = 0.296077

#> Warning: non-integer x = 1.240090

#> Warning: non-integer x = 0.957003

#> Warning: non-integer x = 0.487121

#> Warning: non-integer x = 1.758579

#> Warning: non-integer x = 0.972771

#> Warning: non-integer x = 1.081262

#> Warning: non-integer x = 2.366925

#> Warning: non-integer x = 0.190761

#> Warning: non-integer x = 3.455169

#> Warning: non-integer x = 1.003795

#> Warning: non-integer x = 0.017111

#> Warning: non-integer x = 0.202555

#> Warning: non-integer x = 0.275759

#> Warning: non-integer x = 1.788149

#> Warning: non-integer x = 1.123812

#> Warning: non-integer x = 1.503862

#> Warning: non-integer x = 0.062893

#> Warning: non-integer x = 1.989920

#> Warning: non-integer x = 0.661500

#> Warning: non-integer x = 0.994982

#> Warning: non-integer x = 0.965442

#> Warning: non-integer x = 1.679900

#> Warning: non-integer x = 0.795131

#> Warning: non-integer x = 0.933394

#> Warning: non-integer x = 0.003111

#> Warning: non-integer x = 2.383446

#> Warning: non-integer x = 4.969472

#> Warning: non-integer x = 1.853547

#> Warning: non-integer x = 0.221632

#> Warning: non-integer x = 1.740241

#> Warning: non-integer x = 0.886171

#> Warning: non-integer x = 1.162764

#> Warning: non-integer x = 1.083105

#> Warning: non-integer x = 3.703167

#> Warning: non-integer x = 0.602410

#> Warning: non-integer x = 0.182199

#> Warning: non-integer x = 2.537718

#> Warning: non-integer x = 3.225499

#> Warning: non-integer x = 2.738140

#> Warning: non-integer x = 0.194857

#> Warning: non-integer x = 2.301274

#> Warning: non-integer x = 1.477650

#> Warning: non-integer x = 1.391471

#> Warning: non-integer x = 2.258710

#> Warning: non-integer x = 0.740356

#> Warning: non-integer x = 0.324923

#> Warning: non-integer x = 2.012341

#> Warning: non-integer x = 0.495526

#> Warning: non-integer x = 2.011970

#> Warning: non-integer x = 0.974678

#> Warning: non-integer x = 0.503162

#> Warning: non-integer x = 0.215773

#> Warning: non-integer x = 0.706025

#> Warning: non-integer x = 0.245284

#> Warning: non-integer x = 2.408367

#> Warning: non-integer x = 2.485184

#> Warning: non-integer x = 1.614253

#> Warning: non-integer x = 0.240348

#> Warning: non-integer x = 1.224021

#> Warning: non-integer x = 1.638117

#> Warning: non-integer x = 0.833458

#> Warning: non-integer x = 2.025408

#> Warning: non-integer x = 0.831238

#> Warning: non-integer x = 1.141662

#> Warning: non-integer x = 0.285627

#> Warning: non-integer x = 0.152443

#> Warning: non-integer x = 1.757957

#> Warning: non-integer x = 0.196758

#> Warning: non-integer x = 1.782209

#> Warning: non-integer x = 0.613597

#> Warning: non-integer x = 1.050665

#> Warning: non-integer x = 1.873756

#> Warning: non-integer x = 0.080972

#> Warning: non-integer x = 1.088141

#> Warning: non-integer x = 1.656653

#> Warning: non-integer x = 1.985469

#> Warning: non-integer x = 1.071798

#> Warning: non-integer x = 1.514095

#> Warning: non-integer x = 0.221625

#> Warning: non-integer x = 1.810763

#> Warning: non-integer x = 1.453060

#> Warning: non-integer x = 2.080906

#> Warning: non-integer x = 1.443401

#> Warning: non-integer x = 0.959385

#> Warning: non-integer x = 0.811163

#> Warning: non-integer x = 0.702506

#> Warning: non-integer x = 4.409851

#> Warning: non-integer x = 0.101447

#> Warning: non-integer x = 2.677502

#> Warning: non-integer x = 1.347748

#> Warning: non-integer x = 0.778868

#> Warning: non-integer x = 2.379174

#> Warning: non-integer x = 0.602709

#> Warning: non-integer x = 0.156130

#> Warning: non-integer x = 1.415724

#> Warning: non-integer x = 0.392297

#> Warning: non-integer x = 0.781541

#> Warning: non-integer x = 2.594697

#> Warning: non-integer x = 0.327573

#> Warning: non-integer x = 0.561694

#> Warning: non-integer x = 0.234533

#> Warning: non-integer x = 0.764599

#> Warning: non-integer x = 1.770984

#> Warning: non-integer x = 1.597642

#> Warning: non-integer x = 1.224971

#> Warning: non-integer x = 1.892355

#> Warning: non-integer x = 2.196432

#> Warning: non-integer x = 1.633378

#> Warning: non-integer x = 1.490926

#> Warning: non-integer x = 0.106213

#> Warning: non-integer x = 1.698273

#> Warning: non-integer x = 1.480762

#> Warning: non-integer x = 2.437305

#> Warning: non-integer x = 2.186821

#> Warning: non-integer x = 1.982947

#> Warning: non-integer x = 0.658019

#> Warning: non-integer x = 3.784500

#> Warning: non-integer x = 0.537030

#> Warning: non-integer x = 0.224424

#> Warning: non-integer x = 1.223102

#> Warning: non-integer x = 0.596544

#> Warning: non-integer x = 0.997720

#> Warning: non-integer x = 3.587546

#> Warning: non-integer x = 2.285230

#> Warning: non-integer x = 1.695676

#> Warning: non-integer x = 0.999106

#> Warning: non-integer x = 1.435912

#> Warning: non-integer x = 1.110552

#> Warning: non-integer x = 4.130309

#> Warning: non-integer x = 2.371888

#> Warning: non-integer x = 1.650927

#> Warning: non-integer x = 0.249972

#> Warning: non-integer x = 2.900909

#> Warning: non-integer x = 0.363867

#> Warning: non-integer x = 1.228149

#> Warning: non-integer x = 0.315205

#> Warning: non-integer x = 0.851799

#> Warning: non-integer x = 2.507286

#> Warning: non-integer x = 0.305590

#> Warning: non-integer x = 1.605569

#> Warning: non-integer x = 3.230436

#> Warning: non-integer x = 0.294035

#> Warning: non-integer x = 1.344837

#> Warning: non-integer x = 0.980088

#> Warning: non-integer x = 2.742282

#> Warning: non-integer x = 2.189645

#> Warning: non-integer x = 0.791832

#> Warning: non-integer x = 0.358068

#> Warning: non-integer x = 1.240844

#> Warning: non-integer x = 0.931036

#> Warning: non-integer x = 1.270891

#> Warning: non-integer x = 2.014536

#> Warning: non-integer x = 0.219323

#> Warning: non-integer x = 3.468618

#> Warning: non-integer x = 2.520014

#> Warning: non-integer x = 2.489984

#> Warning: non-integer x = 1.153649

#> Warning: non-integer x = 0.023818

#> Warning: non-integer x = 0.158727

#> Warning: non-integer x = 1.198333

#> Warning: non-integer x = 1.682740

#> Warning: non-integer x = 0.321944

#> Warning: non-integer x = 1.039386

#> Warning: non-integer x = 1.929611

#> Warning: non-integer x = 3.254488

#> Warning: non-integer x = 3.995900

#> Warning: non-integer x = 2.014101

#> Warning: non-integer x = 0.781540

#> Warning: non-integer x = 1.672439

#> Warning: non-integer x = 1.003934

#> Warning: non-integer x = 0.566552

#> Warning: non-integer x = 0.858286

#> Warning: non-integer x = 1.066234

#> Warning: non-integer x = 0.088543

#> Warning: non-integer x = 2.119596

#> Warning: non-integer x = 1.250340

#> Warning: non-integer x = 1.028532

#> Warning: non-integer x = 2.683470

#> Warning: non-integer x = 0.362005

#> Warning: non-integer x = 1.632195

#> Warning: non-integer x = 1.028016

#> Warning: non-integer x = 3.700344

#> Warning: non-integer x = 1.158768

#> Warning: non-integer x = 1.654461

#> Warning: non-integer x = 1.146283

#> Warning: non-integer x = 1.121929

#> Warning: non-integer x = 0.854290

#> Warning: non-integer x = 0.785346

#> Warning: non-integer x = 0.014775

#> Warning: non-integer x = 2.531130

#> Warning: non-integer x = 0.426088

#> Warning: non-integer x = 1.838845

#> Warning: non-integer x = 0.567641

#> Warning: non-integer x = 0.605491

#> Warning: non-integer x = 1.250539

#> Warning: non-integer x = 1.621749

#> Warning: non-integer x = 1.405260

#> Warning: non-integer x = 1.008870

#> Warning: non-integer x = 0.717028

#> Warning: non-integer x = 0.254702

#> Warning: non-integer x = 0.041641

#> Warning: non-integer x = 0.568269

#> Warning: non-integer x = 2.384006

#> Warning: non-integer x = 0.325058

#> Warning: non-integer x = 3.686110

#> Warning: non-integer x = 1.952153

#> Warning: non-integer x = 2.250390

#> Warning: non-integer x = 1.913224

#> Warning: non-integer x = 0.961030

#> Warning: non-integer x = 2.557309

#> Warning: non-integer x = 0.580333

#> Warning: non-integer x = 2.121277

#> Warning: non-integer x = 0.539043

#> Warning: non-integer x = 1.547484

#> Warning: non-integer x = 0.042680

#> Warning: non-integer x = 1.071391

#> Warning: non-integer x = 0.646901

#> Warning: non-integer x = 2.532075

#> Warning: non-integer x = 3.190464

#> Warning: non-integer x = 2.873519

#> Warning: non-integer x = 1.895732

#> Warning: non-integer x = 0.596674

#> Warning: non-integer x = 0.016044

#> Warning: non-integer x = 1.043490

#> Warning: non-integer x = 0.007305

#> Warning: non-integer x = 0.780799

#> Warning: non-integer x = 0.113082

#> Warning: non-integer x = 2.329890

#> Warning: non-integer x = 0.076449

#> Warning: non-integer x = 2.394202

#> Warning: non-integer x = 0.935515

#> Warning: non-integer x = 0.852043

#> Warning: non-integer x = 1.707374

#> Warning: non-integer x = 1.229389

#> Warning: non-integer x = 0.753765

#> Warning: non-integer x = 1.414979

#> Warning: non-integer x = 0.594140

#> Warning: non-integer x = 3.465153

#> Warning: non-integer x = 1.298714

#> Warning: non-integer x = 1.176342

#> Warning: non-integer x = 1.363999

#> Warning: non-integer x = 2.083245

#> Warning: non-integer x = 1.104046

#> Warning: non-integer x = 0.284069

#> Warning: non-integer x = 1.209928

#> Warning: non-integer x = 2.690316

#> Warning: non-integer x = 1.370323

#> Warning: non-integer x = 0.794983

#> Warning: non-integer x = 0.487244

#> Warning: non-integer x = 0.238356

#> Warning: non-integer x = 1.099733

#> Warning: non-integer x = 1.194490

#> Warning: non-integer x = 0.522042

#> Warning: non-integer x = 0.354482

#> Warning: non-integer x = 1.818426

#> Warning: non-integer x = 3.272200

#> Warning: non-integer x = 3.633312

#> Warning: non-integer x = 1.016527

#> Warning: non-integer x = 0.998017

#> Warning: non-integer x = 1.552041

#> Warning: non-integer x = 3.366086

#> Warning: non-integer x = 2.195876

#> Warning: non-integer x = 0.245704

#> Warning: non-integer x = 0.902733

#> Warning: non-integer x = 0.611279

#> Warning: non-integer x = 1.073064

#> Warning: non-integer x = 3.122896

#> Warning: non-integer x = 3.289383

#> Warning: non-integer x = 1.362005

#> Warning: non-integer x = 1.226182

#> Warning: non-integer x = 1.673576

#> Warning: non-integer x = 2.319741

#> Warning: non-integer x = 3.318039

#> Warning: non-integer x = 1.171827

#> Warning: non-integer x = 2.402918

#> Warning: non-integer x = 2.328164

#> Warning: non-integer x = 2.085228

#> Warning: non-integer x = 3.148231

#> Warning: non-integer x = 1.397113

#> Warning: non-integer x = 1.607181

#> Warning: non-integer x = 0.547809

#> Warning: non-integer x = 1.477165

#> Warning: non-integer x = 3.503793

#> Warning: non-integer x = 2.747797

#> Warning: non-integer x = 0.354085

#> Warning: non-integer x = 2.610756

#> Warning: non-integer x = 0.501000

#> Warning: non-integer x = 1.590403

#> Warning: non-integer x = 2.153105

#> Warning: non-integer x = 1.650260

#> Warning: non-integer x = 2.160881

#> Warning: non-integer x = 0.356180

#> Warning: non-integer x = 2.649653

#> Warning: non-integer x = 0.369972

#> Warning: non-integer x = 2.187013

#> Warning: non-integer x = 2.033029

#> Warning: non-integer x = 1.869599

#> Warning: non-integer x = 0.062174

#> Warning: non-integer x = 0.415602

#> Warning: non-integer x = 0.178065

#> Warning: non-integer x = 4.154189

#> Warning: non-integer x = 0.985877

#> Warning: non-integer x = 0.568747

#> Warning: non-integer x = 0.652447

#> Warning: non-integer x = 0.108887

#> Warning: non-integer x = 0.255063

#> Warning: non-integer x = 0.090761

#> Warning: non-integer x = 1.487675

#> Warning: non-integer x = 0.050814

#> Warning: non-integer x = 0.334425

#> Warning: non-integer x = 1.632813

#> Warning: non-integer x = 1.754326

#> Warning: non-integer x = 0.599684

#> Warning: non-integer x = 0.670335

#> Warning: non-integer x = 0.172537

#> Warning: non-integer x = 1.056144

#> Warning: non-integer x = 1.063414

#> Warning: non-integer x = 1.897777

#> Warning: non-integer x = 3.371699

#> Warning: non-integer x = 2.750189

#> Warning: non-integer x = 0.880520

#> Warning: non-integer x = 0.993584

#> Warning: non-integer x = 0.729309

#> Warning: non-integer x = 0.962651

#> Warning: non-integer x = 1.593494

#> Warning: non-integer x = 1.434694

#> Warning: non-integer x = 0.547087

#> Warning: non-integer x = 2.285665

#> Warning: non-integer x = 1.340578

#> Warning: non-integer x = 0.766256

#> Warning: non-integer x = 1.032638

#> Warning: non-integer x = 2.291937

#> Warning: non-integer x = 0.458387

#> Warning: non-integer x = 1.542869

#> Warning: non-integer x = 2.761907

#> Warning: non-integer x = 2.124701

#> Warning: non-integer x = 3.175064

#> Warning: non-integer x = 1.461625

#> Warning: non-integer x = 2.421410

#> Warning: non-integer x = 1.388629

#> Warning: non-integer x = 0.588777

#> Warning: non-integer x = 0.772240

#> Warning: non-integer x = 1.897591

#> Warning: non-integer x = 1.437158

#> Warning: non-integer x = 0.535339

#> Warning: non-integer x = 0.414032

#> Warning: non-integer x = 3.231389

#> Warning: non-integer x = 0.643210

#> Warning: non-integer x = 2.453036

#> Warning: non-integer x = 0.816121

#> Warning: non-integer x = 1.109269

#> Warning: non-integer x = 0.461943

#> Warning: non-integer x = 2.396525

#> Warning: non-integer x = 0.725003

#> Warning: non-integer x = 1.269596

#> Warning: non-integer x = 0.881810

#> Warning: non-integer x = 0.298783

#> Warning: non-integer x = 2.537973

#> Warning: non-integer x = 1.993396

#> Warning: non-integer x = 1.278817

#> Warning: non-integer x = 1.436672

#> Warning: non-integer x = 1.784021

#> Warning: non-integer x = 1.835774

#> Warning: non-integer x = 0.370742

#> Warning: non-integer x = 0.434732

#> Warning: non-integer x = 0.822324

#> Warning: non-integer x = 1.683792

#> Warning: non-integer x = 1.512521

#> Warning: non-integer x = 0.294523

#> Warning: non-integer x = 1.461478

#> Warning: non-integer x = 0.034762

#> Warning: non-integer x = 0.136582

#> Warning: non-integer x = 2.809777

#> Warning: non-integer x = 0.656851

#> Warning: non-integer x = 1.879677

#> Warning: non-integer x = 1.727809

#> Warning: non-integer x = 2.192786

#> Warning: non-integer x = 2.051932

#> Warning: non-integer x = 2.263427

#> Warning: non-integer x = 0.313490

#> Warning: non-integer x = 1.991597

#> Warning: non-integer x = 1.800275

#> Warning: non-integer x = 0.560866

#> Warning: non-integer x = 1.745964

#> Warning: non-integer x = 1.854971

#> Warning: non-integer x = 2.730219

#> Warning: non-integer x = 1.190449

#> Warning: non-integer x = 1.085223

#> Warning: non-integer x = 0.126387

#> Warning: non-integer x = 0.034626

#> Warning: non-integer x = 2.352709

#> Warning: non-integer x = 1.652628

#> Warning: non-integer x = 0.486552

#> Warning: non-integer x = 0.474583

#> Warning: non-integer x = 1.649086

#> Warning: non-integer x = 0.038674

#> Warning: non-integer x = 0.564086

#> Warning: non-integer x = 0.956350

#> Warning: non-integer x = 0.498292

#> Warning: non-integer x = 3.444968

#> Warning: non-integer x = 0.345586

#> Warning: non-integer x = 3.806156

#> Warning: non-integer x = 1.096546

#> Warning: non-integer x = 0.209458

#> Warning: non-integer x = 1.985461

#> Warning: non-integer x = 2.566107

#> Warning: non-integer x = 0.366918

#> Warning: non-integer x = 2.312529

#> Warning: non-integer x = 0.230595

#> Warning: non-integer x = 0.203112

#> Warning: non-integer x = 0.033968

#> Warning: non-integer x = 3.807750

#> Warning: non-integer x = 0.946132

#> Warning: non-integer x = 0.169267

#> Warning: non-integer x = 2.186852

#> Warning: non-integer x = 2.761185

#> Warning: non-integer x = 3.820612

#> Warning: non-integer x = 1.116838

#> Warning: non-integer x = 2.271299

#> Warning: non-integer x = 1.667447

#> Warning: non-integer x = 2.289011

#> Warning: non-integer x = 0.642307

#> Warning: non-integer x = 0.055070

#> Warning: non-integer x = 0.334731

#> Warning: non-integer x = 3.679664

#> Warning: non-integer x = 0.962674

#> Warning: non-integer x = 0.576832

#> Warning: non-integer x = 1.895526

#> Warning: non-integer x = 0.661811

#> Warning: non-integer x = 0.715476

#> Warning: non-integer x = 2.482528

#> Warning: non-integer x = 0.660958

#> Warning: non-integer x = 0.498407

#> Warning: non-integer x = 0.695529

#> Warning: non-integer x = 0.543350

#> Warning: non-integer x = 2.388218

#> Warning: non-integer x = 0.620985

#> Warning: non-integer x = 1.398758

#> Warning: non-integer x = 0.995798

#> Warning: non-integer x = 0.818043

#> Warning: non-integer x = 1.410983

#> Warning: non-integer x = 0.470503

#> Warning: non-integer x = 1.789305

#> Warning: non-integer x = 2.211965

#> Warning: non-integer x = 0.619030

#> Warning: non-integer x = 0.069857

#> Warning: non-integer x = 1.821570

#> Warning: non-integer x = 0.439926

#> Warning: non-integer x = 1.389763

#> Warning: non-integer x = 0.352174

#> Warning: non-integer x = 3.608563

#> Warning: non-integer x = 1.073748

#> Warning: non-integer x = 1.516657

#> Warning: non-integer x = 0.344700

#> Warning: non-integer x = 2.122679

#> Warning: non-integer x = 0.463898

#> Warning: non-integer x = 0.180553

#> Warning: non-integer x = 1.980955

#> Warning: non-integer x = 0.817329

#> Warning: non-integer x = 0.170171

#> Warning: non-integer x = 2.368046

#> Warning: non-integer x = 0.683003

#> Warning: non-integer x = 3.715714

#> Warning: non-integer x = 0.032344

#> Warning: non-integer x = 1.714037

#> Warning: non-integer x = 1.640732

#> Warning: non-integer x = 1.016832

#> Warning: non-integer x = 1.288359

#> Warning: non-integer x = 0.754651

#> Warning: non-integer x = 1.027619

#> Warning: non-integer x = 0.013008

#> Warning: non-integer x = 0.004020

#> Warning: non-integer x = 0.680096

#> Warning: non-integer x = 0.293095

#> Warning: non-integer x = 0.297801

#> Warning: non-integer x = 2.408213

#> Warning: non-integer x = 1.022336

#> Warning: non-integer x = 3.682257

#> Warning: non-integer x = 2.026801

#> Warning: non-integer x = 0.963139

#> Warning: non-integer x = 3.174991

#> Warning: non-integer x = 2.605916

#> Warning: non-integer x = 0.257111

#> Warning: non-integer x = 2.330878

#> Warning: non-integer x = 2.841589

#> Warning: non-integer x = 2.354104

#> Warning: non-integer x = 2.588199

#> Warning: non-integer x = 0.239448

#> Warning: non-integer x = 1.697559

#> Warning: non-integer x = 0.283268

#> Warning: non-integer x = 1.564758

#> Warning: non-integer x = 1.709531

#> Warning: non-integer x = 3.370166

#> Warning: non-integer x = 1.262489

#> Warning: non-integer x = 3.040476

#> Warning: non-integer x = 0.494064

#> Warning: non-integer x = 3.178133

#> Warning: non-integer x = 1.654066

#> Warning: non-integer x = 2.703021

#> Warning: non-integer x = 2.468365