Estimate the Individual Treatment Effect using Poisson Regression

Estimates the Individual Treatment Effect using Poisson Regression given a response vector, a treatment vector, and a features matrix.

estimate_ite_poisson(y, z, X, X_names, include_offset, offset_name)

Arguments

y
the observed response vector

z
the treatment vector

X
the features matrix

X_names
the names of the covariates

include_offset
whether or not to include an offset when estimating the ITE, for poisson only

offset_name
the name of the offset, if it is to be included

Value

a vector of ITE estimates

Examples

dataset <- generate_cre_dataset(n = 1000, rho = 0, n_rules = 2, p = 10,
                                effect_size = 2, binary = FALSE)

# Initialize parameters
y <- abs(dataset[["y"]])
z <- dataset[["z"]]
X <- as.data.frame(dataset[["X"]])
X_names <- names(as.data.frame(X))
include_offset <- FALSE
offset_name <- NA

ite_list <- estimate_ite_poisson(y, z, X, X_names, include_offset, offset_name)

#> Warning: non-integer x = 2.821240

#> Warning: non-integer x = 1.871303

#> Warning: non-integer x = 3.184374

#> Warning: non-integer x = 0.473848

#> Warning: non-integer x = 3.006992

#> Warning: non-integer x = 1.530212

#> Warning: non-integer x = 1.793594

#> Warning: non-integer x = 3.970080

#> Warning: non-integer x = 2.052994

#> Warning: non-integer x = 1.378239

#> Warning: non-integer x = 1.375607

#> Warning: non-integer x = 1.456961

#> Warning: non-integer x = 3.170255

#> Warning: non-integer x = 0.074252

#> Warning: non-integer x = 0.748259

#> Warning: non-integer x = 0.080865

#> Warning: non-integer x = 1.114466

#> Warning: non-integer x = 1.605019

#> Warning: non-integer x = 0.521452

#> Warning: non-integer x = 0.463659

#> Warning: non-integer x = 0.057071

#> Warning: non-integer x = 0.838180

#> Warning: non-integer x = 0.886262

#> Warning: non-integer x = 2.950515

#> Warning: non-integer x = 0.686190

#> Warning: non-integer x = 1.846463

#> Warning: non-integer x = 3.485210

#> Warning: non-integer x = 2.461706

#> Warning: non-integer x = 0.913516

#> Warning: non-integer x = 2.422835

#> Warning: non-integer x = 2.983278

#> Warning: non-integer x = 0.692317

#> Warning: non-integer x = 1.667309

#> Warning: non-integer x = 1.806191

#> Warning: non-integer x = 1.004392

#> Warning: non-integer x = 0.356220

#> Warning: non-integer x = 1.463589

#> Warning: non-integer x = 1.533261

#> Warning: non-integer x = 4.405756

#> Warning: non-integer x = 1.117455

#> Warning: non-integer x = 0.197018

#> Warning: non-integer x = 3.440676

#> Warning: non-integer x = 2.267769

#> Warning: non-integer x = 0.013696

#> Warning: non-integer x = 1.244637

#> Warning: non-integer x = 2.280290

#> Warning: non-integer x = 1.333677

#> Warning: non-integer x = 1.510757

#> Warning: non-integer x = 1.173431

#> Warning: non-integer x = 4.046538

#> Warning: non-integer x = 2.165331

#> Warning: non-integer x = 2.258528

#> Warning: non-integer x = 1.113233

#> Warning: non-integer x = 0.256166

#> Warning: non-integer x = 2.147550

#> Warning: non-integer x = 2.118960

#> Warning: non-integer x = 0.368129

#> Warning: non-integer x = 1.139190

#> Warning: non-integer x = 0.723559

#> Warning: non-integer x = 2.230221

#> Warning: non-integer x = 2.140524

#> Warning: non-integer x = 1.276897

#> Warning: non-integer x = 0.675420

#> Warning: non-integer x = 2.173552

#> Warning: non-integer x = 3.335058

#> Warning: non-integer x = 2.450581

#> Warning: non-integer x = 2.723006

#> Warning: non-integer x = 0.985171

#> Warning: non-integer x = 2.744872

#> Warning: non-integer x = 1.196658

#> Warning: non-integer x = 2.928118

#> Warning: non-integer x = 0.847691

#> Warning: non-integer x = 1.824637

#> Warning: non-integer x = 1.457522

#> Warning: non-integer x = 0.276685

#> Warning: non-integer x = 0.338425

#> Warning: non-integer x = 1.321730

#> Warning: non-integer x = 2.182585

#> Warning: non-integer x = 2.030642

#> Warning: non-integer x = 2.704143

#> Warning: non-integer x = 1.909168

#> Warning: non-integer x = 1.017741

#> Warning: non-integer x = 2.686367

#> Warning: non-integer x = 0.864794

#> Warning: non-integer x = 1.454177

#> Warning: non-integer x = 1.096647

#> Warning: non-integer x = 1.234520

#> Warning: non-integer x = 0.515779

#> Warning: non-integer x = 1.563777

#> Warning: non-integer x = 1.687494

#> Warning: non-integer x = 1.326493

#> Warning: non-integer x = 2.646712

#> Warning: non-integer x = 1.700763

#> Warning: non-integer x = 1.670873

#> Warning: non-integer x = 0.205994

#> Warning: non-integer x = 0.789056

#> Warning: non-integer x = 0.319380

#> Warning: non-integer x = 1.438456

#> Warning: non-integer x = 0.799538

#> Warning: non-integer x = 1.711313

#> Warning: non-integer x = 2.373721

#> Warning: non-integer x = 0.370020

#> Warning: non-integer x = 1.502907

#> Warning: non-integer x = 2.307693

#> Warning: non-integer x = 2.533543

#> Warning: non-integer x = 0.775556

#> Warning: non-integer x = 0.665348

#> Warning: non-integer x = 0.814628

#> Warning: non-integer x = 3.169916

#> Warning: non-integer x = 0.095997

#> Warning: non-integer x = 1.463771

#> Warning: non-integer x = 2.183749

#> Warning: non-integer x = 1.888531

#> Warning: non-integer x = 4.402693

#> Warning: non-integer x = 1.266543

#> Warning: non-integer x = 1.089881

#> Warning: non-integer x = 1.041424

#> Warning: non-integer x = 0.562069

#> Warning: non-integer x = 0.172169

#> Warning: non-integer x = 1.954794

#> Warning: non-integer x = 0.294418

#> Warning: non-integer x = 1.828571

#> Warning: non-integer x = 4.466330

#> Warning: non-integer x = 2.746779

#> Warning: non-integer x = 0.434328

#> Warning: non-integer x = 2.687160

#> Warning: non-integer x = 1.680366

#> Warning: non-integer x = 2.077631

#> Warning: non-integer x = 1.377442

#> Warning: non-integer x = 0.794760

#> Warning: non-integer x = 3.962162

#> Warning: non-integer x = 0.644992

#> Warning: non-integer x = 1.999248

#> Warning: non-integer x = 0.409524

#> Warning: non-integer x = 2.357616

#> Warning: non-integer x = 2.753630

#> Warning: non-integer x = 1.816632

#> Warning: non-integer x = 0.922271

#> Warning: non-integer x = 3.256690

#> Warning: non-integer x = 2.313854

#> Warning: non-integer x = 2.882227

#> Warning: non-integer x = 0.347592

#> Warning: non-integer x = 1.285646

#> Warning: non-integer x = 1.119128

#> Warning: non-integer x = 2.065025

#> Warning: non-integer x = 1.833449

#> Warning: non-integer x = 0.900549

#> Warning: non-integer x = 1.451767

#> Warning: non-integer x = 1.708001

#> Warning: non-integer x = 0.966711

#> Warning: non-integer x = 0.720772

#> Warning: non-integer x = 4.008636

#> Warning: non-integer x = 4.369970

#> Warning: non-integer x = 0.266407

#> Warning: non-integer x = 0.669097

#> Warning: non-integer x = 0.889942

#> Warning: non-integer x = 0.530214

#> Warning: non-integer x = 1.247237

#> Warning: non-integer x = 0.246045

#> Warning: non-integer x = 3.014504

#> Warning: non-integer x = 1.893551

#> Warning: non-integer x = 0.816621

#> Warning: non-integer x = 1.651740

#> Warning: non-integer x = 3.292410

#> Warning: non-integer x = 2.063026

#> Warning: non-integer x = 0.726618

#> Warning: non-integer x = 0.666182

#> Warning: non-integer x = 1.926062

#> Warning: non-integer x = 0.579346

#> Warning: non-integer x = 0.483306

#> Warning: non-integer x = 3.065686

#> Warning: non-integer x = 2.430039

#> Warning: non-integer x = 2.268221

#> Warning: non-integer x = 1.192281

#> Warning: non-integer x = 1.761568

#> Warning: non-integer x = 2.604838

#> Warning: non-integer x = 1.956961

#> Warning: non-integer x = 2.793194

#> Warning: non-integer x = 0.849732

#> Warning: non-integer x = 4.614708

#> Warning: non-integer x = 0.559342

#> Warning: non-integer x = 0.229379

#> Warning: non-integer x = 2.140047

#> Warning: non-integer x = 1.594034

#> Warning: non-integer x = 2.848907

#> Warning: non-integer x = 2.597335

#> Warning: non-integer x = 1.740352

#> Warning: non-integer x = 0.655371

#> Warning: non-integer x = 1.028525

#> Warning: non-integer x = 0.396142

#> Warning: non-integer x = 2.187956

#> Warning: non-integer x = 0.926266

#> Warning: non-integer x = 1.826496

#> Warning: non-integer x = 2.792366

#> Warning: non-integer x = 1.107947

#> Warning: non-integer x = 1.923699

#> Warning: non-integer x = 0.239378

#> Warning: non-integer x = 0.049576

#> Warning: non-integer x = 3.744822

#> Warning: non-integer x = 0.250671

#> Warning: non-integer x = 0.878495

#> Warning: non-integer x = 1.124239

#> Warning: non-integer x = 2.763751

#> Warning: non-integer x = 4.728096

#> Warning: non-integer x = 0.299353

#> Warning: non-integer x = 1.567619

#> Warning: non-integer x = 4.237075

#> Warning: non-integer x = 0.548155

#> Warning: non-integer x = 4.099125

#> Warning: non-integer x = 1.063078

#> Warning: non-integer x = 0.109915

#> Warning: non-integer x = 1.786826

#> Warning: non-integer x = 1.453813

#> Warning: non-integer x = 2.044390

#> Warning: non-integer x = 0.282087

#> Warning: non-integer x = 2.493440

#> Warning: non-integer x = 0.681776

#> Warning: non-integer x = 1.086071

#> Warning: non-integer x = 1.528183

#> Warning: non-integer x = 0.688518

#> Warning: non-integer x = 0.594432

#> Warning: non-integer x = 1.776459

#> Warning: non-integer x = 1.666207

#> Warning: non-integer x = 1.143660

#> Warning: non-integer x = 2.583836

#> Warning: non-integer x = 0.582422

#> Warning: non-integer x = 0.756451

#> Warning: non-integer x = 3.336784

#> Warning: non-integer x = 2.332686

#> Warning: non-integer x = 2.981670

#> Warning: non-integer x = 0.348265

#> Warning: non-integer x = 1.113551

#> Warning: non-integer x = 0.689458

#> Warning: non-integer x = 2.550907

#> Warning: non-integer x = 2.386449

#> Warning: non-integer x = 0.480989

#> Warning: non-integer x = 0.900736

#> Warning: non-integer x = 3.646812

#> Warning: non-integer x = 0.449925

#> Warning: non-integer x = 2.905561

#> Warning: non-integer x = 1.382410

#> Warning: non-integer x = 1.542043

#> Warning: non-integer x = 1.379430

#> Warning: non-integer x = 0.543885

#> Warning: non-integer x = 1.096024

#> Warning: non-integer x = 2.762151

#> Warning: non-integer x = 0.285428

#> Warning: non-integer x = 1.970801

#> Warning: non-integer x = 0.765844

#> Warning: non-integer x = 0.235602

#> Warning: non-integer x = 2.200564

#> Warning: non-integer x = 1.679118

#> Warning: non-integer x = 2.673066

#> Warning: non-integer x = 1.036057

#> Warning: non-integer x = 2.873234

#> Warning: non-integer x = 0.798695

#> Warning: non-integer x = 1.422857

#> Warning: non-integer x = 3.805520

#> Warning: non-integer x = 1.312795

#> Warning: non-integer x = 0.852667

#> Warning: non-integer x = 0.270364

#> Warning: non-integer x = 0.928123

#> Warning: non-integer x = 0.400859

#> Warning: non-integer x = 2.129069

#> Warning: non-integer x = 1.531622

#> Warning: non-integer x = 0.477100

#> Warning: non-integer x = 0.305622

#> Warning: non-integer x = 1.210127

#> Warning: non-integer x = 2.389164

#> Warning: non-integer x = 2.134019

#> Warning: non-integer x = 1.233266

#> Warning: non-integer x = 2.124995

#> Warning: non-integer x = 2.818531

#> Warning: non-integer x = 1.527485

#> Warning: non-integer x = 1.900107

#> Warning: non-integer x = 2.896905

#> Warning: non-integer x = 0.140638

#> Warning: non-integer x = 0.633084

#> Warning: non-integer x = 1.974449

#> Warning: non-integer x = 3.140207

#> Warning: non-integer x = 0.746453

#> Warning: non-integer x = 1.751670

#> Warning: non-integer x = 3.659662

#> Warning: non-integer x = 1.444508

#> Warning: non-integer x = 2.280786

#> Warning: non-integer x = 1.380459

#> Warning: non-integer x = 2.515922

#> Warning: non-integer x = 1.221692

#> Warning: non-integer x = 1.197841

#> Warning: non-integer x = 1.970569

#> Warning: non-integer x = 3.752796

#> Warning: non-integer x = 2.259405

#> Warning: non-integer x = 1.669592

#> Warning: non-integer x = 4.166514

#> Warning: non-integer x = 0.045215

#> Warning: non-integer x = 2.620239

#> Warning: non-integer x = 1.981251

#> Warning: non-integer x = 2.812900

#> Warning: non-integer x = 1.536778

#> Warning: non-integer x = 2.958547

#> Warning: non-integer x = 2.619667

#> Warning: non-integer x = 1.565209

#> Warning: non-integer x = 1.232460

#> Warning: non-integer x = 0.757032

#> Warning: non-integer x = 1.301728

#> Warning: non-integer x = 2.169092

#> Warning: non-integer x = 0.575196

#> Warning: non-integer x = 0.872869

#> Warning: non-integer x = 2.399398

#> Warning: non-integer x = 0.496109

#> Warning: non-integer x = 0.067492

#> Warning: non-integer x = 3.068367

#> Warning: non-integer x = 0.784201

#> Warning: non-integer x = 0.206503

#> Warning: non-integer x = 1.142608

#> Warning: non-integer x = 3.063056

#> Warning: non-integer x = 0.866581

#> Warning: non-integer x = 0.745939

#> Warning: non-integer x = 0.026894

#> Warning: non-integer x = 3.270425

#> Warning: non-integer x = 2.894403

#> Warning: non-integer x = 3.352579

#> Warning: non-integer x = 3.608044

#> Warning: non-integer x = 3.176444

#> Warning: non-integer x = 2.007949

#> Warning: non-integer x = 2.055686

#> Warning: non-integer x = 1.923755

#> Warning: non-integer x = 1.195024

#> Warning: non-integer x = 2.527413

#> Warning: non-integer x = 2.277970

#> Warning: non-integer x = 0.263288

#> Warning: non-integer x = 1.515048

#> Warning: non-integer x = 1.007403

#> Warning: non-integer x = 2.050987

#> Warning: non-integer x = 3.544725

#> Warning: non-integer x = 1.432824

#> Warning: non-integer x = 0.148888

#> Warning: non-integer x = 0.179578

#> Warning: non-integer x = 1.841631

#> Warning: non-integer x = 0.056197

#> Warning: non-integer x = 0.245178

#> Warning: non-integer x = 3.199764

#> Warning: non-integer x = 0.719069

#> Warning: non-integer x = 1.224794

#> Warning: non-integer x = 0.262161

#> Warning: non-integer x = 1.201808

#> Warning: non-integer x = 3.153965

#> Warning: non-integer x = 2.982670

#> Warning: non-integer x = 1.398824

#> Warning: non-integer x = 3.209126

#> Warning: non-integer x = 0.753205

#> Warning: non-integer x = 0.586020

#> Warning: non-integer x = 0.706543

#> Warning: non-integer x = 2.338601

#> Warning: non-integer x = 1.653786

#> Warning: non-integer x = 0.452273

#> Warning: non-integer x = 4.381501

#> Warning: non-integer x = 2.450526

#> Warning: non-integer x = 1.234159

#> Warning: non-integer x = 1.359084

#> Warning: non-integer x = 2.649693

#> Warning: non-integer x = 0.965345

#> Warning: non-integer x = 1.378141

#> Warning: non-integer x = 0.861343

#> Warning: non-integer x = 2.044118

#> Warning: non-integer x = 1.189257

#> Warning: non-integer x = 1.484190

#> Warning: non-integer x = 1.879165

#> Warning: non-integer x = 1.874385

#> Warning: non-integer x = 2.038485

#> Warning: non-integer x = 0.470061

#> Warning: non-integer x = 0.040949

#> Warning: non-integer x = 2.254795

#> Warning: non-integer x = 1.378036

#> Warning: non-integer x = 1.602441

#> Warning: non-integer x = 2.286895

#> Warning: non-integer x = 1.381048

#> Warning: non-integer x = 2.261538

#> Warning: non-integer x = 0.810985

#> Warning: non-integer x = 2.590255

#> Warning: non-integer x = 0.457222

#> Warning: non-integer x = 1.041443

#> Warning: non-integer x = 0.081160

#> Warning: non-integer x = 1.745241

#> Warning: non-integer x = 2.829400

#> Warning: non-integer x = 0.315944

#> Warning: non-integer x = 1.292629

#> Warning: non-integer x = 0.399428

#> Warning: non-integer x = 0.977719

#> Warning: non-integer x = 1.144453

#> Warning: non-integer x = 0.602010

#> Warning: non-integer x = 2.396437

#> Warning: non-integer x = 2.823982

#> Warning: non-integer x = 2.259414

#> Warning: non-integer x = 2.450181

#> Warning: non-integer x = 0.576435

#> Warning: non-integer x = 1.807273

#> Warning: non-integer x = 0.632190

#> Warning: non-integer x = 0.879270

#> Warning: non-integer x = 1.624481

#> Warning: non-integer x = 2.105027

#> Warning: non-integer x = 1.627419

#> Warning: non-integer x = 0.895909

#> Warning: non-integer x = 2.181009

#> Warning: non-integer x = 2.034330

#> Warning: non-integer x = 0.371792

#> Warning: non-integer x = 0.330919

#> Warning: non-integer x = 0.200740

#> Warning: non-integer x = 3.384355

#> Warning: non-integer x = 0.582994

#> Warning: non-integer x = 0.867523

#> Warning: non-integer x = 2.835560

#> Warning: non-integer x = 1.793714

#> Warning: non-integer x = 0.246174

#> Warning: non-integer x = 1.372442

#> Warning: non-integer x = 1.116379

#> Warning: non-integer x = 0.028232

#> Warning: non-integer x = 0.986100

#> Warning: non-integer x = 2.131737

#> Warning: non-integer x = 2.102982

#> Warning: non-integer x = 2.246521

#> Warning: non-integer x = 2.077202

#> Warning: non-integer x = 1.663117

#> Warning: non-integer x = 0.697889

#> Warning: non-integer x = 0.107579

#> Warning: non-integer x = 1.499027

#> Warning: non-integer x = 0.496078

#> Warning: non-integer x = 3.989741

#> Warning: non-integer x = 0.853683

#> Warning: non-integer x = 0.044147

#> Warning: non-integer x = 3.079336

#> Warning: non-integer x = 1.381499

#> Warning: non-integer x = 1.351017

#> Warning: non-integer x = 0.494436

#> Warning: non-integer x = 3.484486

#> Warning: non-integer x = 0.857702

#> Warning: non-integer x = 1.754628

#> Warning: non-integer x = 4.061817

#> Warning: non-integer x = 0.606390

#> Warning: non-integer x = 1.167116

#> Warning: non-integer x = 0.659374

#> Warning: non-integer x = 1.295283

#> Warning: non-integer x = 0.509252

#> Warning: non-integer x = 1.889992

#> Warning: non-integer x = 0.982366

#> Warning: non-integer x = 0.935921

#> Warning: non-integer x = 1.366890

#> Warning: non-integer x = 2.119836

#> Warning: non-integer x = 1.400995

#> Warning: non-integer x = 2.116614

#> Warning: non-integer x = 0.491162

#> Warning: non-integer x = 1.419852

#> Warning: non-integer x = 2.366675

#> Warning: non-integer x = 0.254447

#> Warning: non-integer x = 0.866857

#> Warning: non-integer x = 0.378627

#> Warning: non-integer x = 2.615807

#> Warning: non-integer x = 1.210403

#> Warning: non-integer x = 0.766307

#> Warning: non-integer x = 1.285588

#> Warning: non-integer x = 0.618310

#> Warning: non-integer x = 1.143065

#> Warning: non-integer x = 0.726441

#> Warning: non-integer x = 1.044387

#> Warning: non-integer x = 0.872296

#> Warning: non-integer x = 0.846827

#> Warning: non-integer x = 2.226753

#> Warning: non-integer x = 1.316470

#> Warning: non-integer x = 2.884608

#> Warning: non-integer x = 2.389568

#> Warning: non-integer x = 1.680915

#> Warning: non-integer x = 1.449747

#> Warning: non-integer x = 0.291505

#> Warning: non-integer x = 1.308059

#> Warning: non-integer x = 3.332529

#> Warning: non-integer x = 2.064713

#> Warning: non-integer x = 0.196534

#> Warning: non-integer x = 2.958555

#> Warning: non-integer x = 2.042333

#> Warning: non-integer x = 0.917149

#> Warning: non-integer x = 2.234042

#> Warning: non-integer x = 0.981606

#> Warning: non-integer x = 0.971940

#> Warning: non-integer x = 1.796618

#> Warning: non-integer x = 0.105724

#> Warning: non-integer x = 0.090315

#> Warning: non-integer x = 0.056894

#> Warning: non-integer x = 0.877379

#> Warning: non-integer x = 0.615992

#> Warning: non-integer x = 3.082717

#> Warning: non-integer x = 2.234895

#> Warning: non-integer x = 2.294941

#> Warning: non-integer x = 1.848858

#> Warning: non-integer x = 0.346259

#> Warning: non-integer x = 1.617344

#> Warning: non-integer x = 0.675078

#> Warning: non-integer x = 2.362806

#> Warning: non-integer x = 0.248195

#> Warning: non-integer x = 2.531280

#> Warning: non-integer x = 0.737194

#> Warning: non-integer x = 0.390946

#> Warning: non-integer x = 2.314524

#> Warning: non-integer x = 0.052639

#> Warning: non-integer x = 0.500654

#> Warning: non-integer x = 0.186076

#> Warning: non-integer x = 1.530216

#> Warning: non-integer x = 1.397110

#> Warning: non-integer x = 1.354720

#> Warning: non-integer x = 0.849997

#> Warning: non-integer x = 3.428810

#> Warning: non-integer x = 3.128183

#> Warning: non-integer x = 0.422460

#> Warning: non-integer x = 1.190714

#> Warning: non-integer x = 0.855767

#> Warning: non-integer x = 3.738169

#> Warning: non-integer x = 0.364286

#> Warning: non-integer x = 3.221500

#> Warning: non-integer x = 2.105453

#> Warning: non-integer x = 0.646795

#> Warning: non-integer x = 1.368636

#> Warning: non-integer x = 2.197104

#> Warning: non-integer x = 1.396437

#> Warning: non-integer x = 0.221544

#> Warning: non-integer x = 4.544378

#> Warning: non-integer x = 1.563935

#> Warning: non-integer x = 0.530627

#> Warning: non-integer x = 1.683839

#> Warning: non-integer x = 1.049242

#> Warning: non-integer x = 0.536363

#> Warning: non-integer x = 0.793782

#> Warning: non-integer x = 1.723484

#> Warning: non-integer x = 0.820824

#> Warning: non-integer x = 0.188218

#> Warning: non-integer x = 0.555019

#> Warning: non-integer x = 0.816305

#> Warning: non-integer x = 0.266256

#> Warning: non-integer x = 1.628790

#> Warning: non-integer x = 0.402230

#> Warning: non-integer x = 0.195390

#> Warning: non-integer x = 0.430887

#> Warning: non-integer x = 0.666261

#> Warning: non-integer x = 0.257086

#> Warning: non-integer x = 0.358362

#> Warning: non-integer x = 2.688084

#> Warning: non-integer x = 5.104047

#> Warning: non-integer x = 0.419566

#> Warning: non-integer x = 2.638260

#> Warning: non-integer x = 0.048389

#> Warning: non-integer x = 1.541057

#> Warning: non-integer x = 0.088157

#> Warning: non-integer x = 2.385553

#> Warning: non-integer x = 0.753832

#> Warning: non-integer x = 0.615377

#> Warning: non-integer x = 0.911862

#> Warning: non-integer x = 2.009813

#> Warning: non-integer x = 3.221909

#> Warning: non-integer x = 0.189698

#> Warning: non-integer x = 1.672611

#> Warning: non-integer x = 1.659763

#> Warning: non-integer x = 0.204117

#> Warning: non-integer x = 0.023246

#> Warning: non-integer x = 1.213692

#> Warning: non-integer x = 1.935575

#> Warning: non-integer x = 0.370774

#> Warning: non-integer x = 1.941258

#> Warning: non-integer x = 1.806611

#> Warning: non-integer x = 1.422672

#> Warning: non-integer x = 2.476337

#> Warning: non-integer x = 0.060460

#> Warning: non-integer x = 1.907843

#> Warning: non-integer x = 1.119708

#> Warning: non-integer x = 2.180062

#> Warning: non-integer x = 1.283875

#> Warning: non-integer x = 1.519126

#> Warning: non-integer x = 0.577437

#> Warning: non-integer x = 0.949831

#> Warning: non-integer x = 0.882733

#> Warning: non-integer x = 0.765698

#> Warning: non-integer x = 0.019769

#> Warning: non-integer x = 1.264672

#> Warning: non-integer x = 1.081652

#> Warning: non-integer x = 3.146639

#> Warning: non-integer x = 0.606420

#> Warning: non-integer x = 0.867417

#> Warning: non-integer x = 0.895279

#> Warning: non-integer x = 0.549944

#> Warning: non-integer x = 1.878410

#> Warning: non-integer x = 0.572456

#> Warning: non-integer x = 1.226961

#> Warning: non-integer x = 4.804715

#> Warning: non-integer x = 3.031405

#> Warning: non-integer x = 0.125153

#> Warning: non-integer x = 1.493466

#> Warning: non-integer x = 3.433982

#> Warning: non-integer x = 0.299194

#> Warning: non-integer x = 1.223499

#> Warning: non-integer x = 1.037598

#> Warning: non-integer x = 0.723458

#> Warning: non-integer x = 0.510981

#> Warning: non-integer x = 0.779857

#> Warning: non-integer x = 2.079247

#> Warning: non-integer x = 0.298196

#> Warning: non-integer x = 1.770402

#> Warning: non-integer x = 2.100515

#> Warning: non-integer x = 0.788432

#> Warning: non-integer x = 1.067597

#> Warning: non-integer x = 2.280560

#> Warning: non-integer x = 0.650468

#> Warning: non-integer x = 0.137436

#> Warning: non-integer x = 1.788617

#> Warning: non-integer x = 2.327950

#> Warning: non-integer x = 2.798806

#> Warning: non-integer x = 0.739369

#> Warning: non-integer x = 2.129540

#> Warning: non-integer x = 0.031336

#> Warning: non-integer x = 1.400744

#> Warning: non-integer x = 1.480134

#> Warning: non-integer x = 0.391965

#> Warning: non-integer x = 1.314782

#> Warning: non-integer x = 0.555357

#> Warning: non-integer x = 2.450957

#> Warning: non-integer x = 2.511721

#> Warning: non-integer x = 1.152063

#> Warning: non-integer x = 2.377895

#> Warning: non-integer x = 0.559029

#> Warning: non-integer x = 0.330475

#> Warning: non-integer x = 0.730257

#> Warning: non-integer x = 0.806630

#> Warning: non-integer x = 0.042926

#> Warning: non-integer x = 1.890460

#> Warning: non-integer x = 1.718006

#> Warning: non-integer x = 1.824343

#> Warning: non-integer x = 0.461340

#> Warning: non-integer x = 0.821806

#> Warning: non-integer x = 2.637744

#> Warning: non-integer x = 0.631506

#> Warning: non-integer x = 0.621134

#> Warning: non-integer x = 2.514317

#> Warning: non-integer x = 2.014923

#> Warning: non-integer x = 0.886230

#> Warning: non-integer x = 2.109528

#> Warning: non-integer x = 1.553941

#> Warning: non-integer x = 0.191572

#> Warning: non-integer x = 3.089501

#> Warning: non-integer x = 1.943704

#> Warning: non-integer x = 2.326277

#> Warning: non-integer x = 1.321283

#> Warning: non-integer x = 2.296705

#> Warning: non-integer x = 3.179443

#> Warning: non-integer x = 0.958842

#> Warning: non-integer x = 0.649562

#> Warning: non-integer x = 0.952752

#> Warning: non-integer x = 0.263645

#> Warning: non-integer x = 0.751708

#> Warning: non-integer x = 2.473375

#> Warning: non-integer x = 0.087952

#> Warning: non-integer x = 0.048995

#> Warning: non-integer x = 0.696182

#> Warning: non-integer x = 1.440396

#> Warning: non-integer x = 1.775096

#> Warning: non-integer x = 0.147807

#> Warning: non-integer x = 0.024404

#> Warning: non-integer x = 3.205359

#> Warning: non-integer x = 2.446985

#> Warning: non-integer x = 0.779037

#> Warning: non-integer x = 0.303738

#> Warning: non-integer x = 1.651685

#> Warning: non-integer x = 1.182267

#> Warning: non-integer x = 1.739975

#> Warning: non-integer x = 2.518590

#> Warning: non-integer x = 2.977612

#> Warning: non-integer x = 0.421048

#> Warning: non-integer x = 0.148595

#> Warning: non-integer x = 1.167033

#> Warning: non-integer x = 1.241697

#> Warning: non-integer x = 2.173566

#> Warning: non-integer x = 1.901948

#> Warning: non-integer x = 0.866666

#> Warning: non-integer x = 1.506437

#> Warning: non-integer x = 0.947213

#> Warning: non-integer x = 1.111729

#> Warning: non-integer x = 1.397284

#> Warning: non-integer x = 0.169590

#> Warning: non-integer x = 1.246248

#> Warning: non-integer x = 0.001718

#> Warning: non-integer x = 0.956335

#> Warning: non-integer x = 0.679882

#> Warning: non-integer x = 2.034472

#> Warning: non-integer x = 0.172327

#> Warning: non-integer x = 0.957445

#> Warning: non-integer x = 2.822933

#> Warning: non-integer x = 2.286198

#> Warning: non-integer x = 1.217884

#> Warning: non-integer x = 1.795442

#> Warning: non-integer x = 1.698592

#> Warning: non-integer x = 2.641608

#> Warning: non-integer x = 3.857759

#> Warning: non-integer x = 1.282324

#> Warning: non-integer x = 2.123700

#> Warning: non-integer x = 1.284288

#> Warning: non-integer x = 0.538010

#> Warning: non-integer x = 2.073051

#> Warning: non-integer x = 0.239400

#> Warning: non-integer x = 2.720404

#> Warning: non-integer x = 1.741134

#> Warning: non-integer x = 0.421617

#> Warning: non-integer x = 0.287403

#> Warning: non-integer x = 1.242056

#> Warning: non-integer x = 1.547841

#> Warning: non-integer x = 0.681047

#> Warning: non-integer x = 1.686157

#> Warning: non-integer x = 1.233068

#> Warning: non-integer x = 0.181950

#> Warning: non-integer x = 1.537473

#> Warning: non-integer x = 0.150423

#> Warning: non-integer x = 0.593252

#> Warning: non-integer x = 2.234081

#> Warning: non-integer x = 0.787435

#> Warning: non-integer x = 1.440296

#> Warning: non-integer x = 1.549382

#> Warning: non-integer x = 1.134087

#> Warning: non-integer x = 0.094761

#> Warning: non-integer x = 0.941190

#> Warning: non-integer x = 2.100934

#> Warning: non-integer x = 1.859120

#> Warning: non-integer x = 1.157608

#> Warning: non-integer x = 1.323636

#> Warning: non-integer x = 0.271853

#> Warning: non-integer x = 1.614866

#> Warning: non-integer x = 1.744023

#> Warning: non-integer x = 0.625880

#> Warning: non-integer x = 1.185063

#> Warning: non-integer x = 2.822321

#> Warning: non-integer x = 0.779632

#> Warning: non-integer x = 1.117508

#> Warning: non-integer x = 0.960430

#> Warning: non-integer x = 1.843258

#> Warning: non-integer x = 0.479439

#> Warning: non-integer x = 1.718777

#> Warning: non-integer x = 0.833932

#> Warning: non-integer x = 0.436677

#> Warning: non-integer x = 4.578682

#> Warning: non-integer x = 0.225630

#> Warning: non-integer x = 1.198798

#> Warning: non-integer x = 0.534517

#> Warning: non-integer x = 0.441780

#> Warning: non-integer x = 0.690734

#> Warning: non-integer x = 1.200570

#> Warning: non-integer x = 0.081384

#> Warning: non-integer x = 2.924589

#> Warning: non-integer x = 2.532380

#> Warning: non-integer x = 0.714781

#> Warning: non-integer x = 0.356381

#> Warning: non-integer x = 0.579491

#> Warning: non-integer x = 3.274690

#> Warning: non-integer x = 1.406847

#> Warning: non-integer x = 0.247337

#> Warning: non-integer x = 1.108102

#> Warning: non-integer x = 2.558911

#> Warning: non-integer x = 1.382794

#> Warning: non-integer x = 3.585926

#> Warning: non-integer x = 0.741759

#> Warning: non-integer x = 2.167183

#> Warning: non-integer x = 0.165456

#> Warning: non-integer x = 1.515894

#> Warning: non-integer x = 1.842133

#> Warning: non-integer x = 1.362668

#> Warning: non-integer x = 1.155581

#> Warning: non-integer x = 1.303115

#> Warning: non-integer x = 0.011678

#> Warning: non-integer x = 0.206601

#> Warning: non-integer x = 1.692131

#> Warning: non-integer x = 1.749794

#> Warning: non-integer x = 1.443542

#> Warning: non-integer x = 1.446366

#> Warning: non-integer x = 1.767716

#> Warning: non-integer x = 0.170033

#> Warning: non-integer x = 0.212553

#> Warning: non-integer x = 0.996197

#> Warning: non-integer x = 1.296718

#> Warning: non-integer x = 4.099017

#> Warning: non-integer x = 1.845568

#> Warning: non-integer x = 1.318010

#> Warning: non-integer x = 0.044532

#> Warning: non-integer x = 1.154815

#> Warning: non-integer x = 1.328871

#> Warning: non-integer x = 1.537857

#> Warning: non-integer x = 1.503979

#> Warning: non-integer x = 0.505006

#> Warning: non-integer x = 0.180829

#> Warning: non-integer x = 1.925449

#> Warning: non-integer x = 1.291883

#> Warning: non-integer x = 0.315456

#> Warning: non-integer x = 2.508162

#> Warning: non-integer x = 1.673738

#> Warning: non-integer x = 0.941612

#> Warning: non-integer x = 0.640275

#> Warning: non-integer x = 1.918573

#> Warning: non-integer x = 1.788664

#> Warning: non-integer x = 3.216706

#> Warning: non-integer x = 1.470175

#> Warning: non-integer x = 1.276465

#> Warning: non-integer x = 0.881197

#> Warning: non-integer x = 2.315243

#> Warning: non-integer x = 1.208224

#> Warning: non-integer x = 1.222929

#> Warning: non-integer x = 2.045819

#> Warning: non-integer x = 0.224180

#> Warning: non-integer x = 1.555959

#> Warning: non-integer x = 0.343269

#> Warning: non-integer x = 3.444866

#> Warning: non-integer x = 1.320110

#> Warning: non-integer x = 1.445516

#> Warning: non-integer x = 0.405187

#> Warning: non-integer x = 2.423733

#> Warning: non-integer x = 0.968552

#> Warning: non-integer x = 1.400214

#> Warning: non-integer x = 0.060320

#> Warning: non-integer x = 0.527369

#> Warning: non-integer x = 2.447325

#> Warning: non-integer x = 1.525436

#> Warning: non-integer x = 2.784213

#> Warning: non-integer x = 1.326148

#> Warning: non-integer x = 2.150157

#> Warning: non-integer x = 1.184313

#> Warning: non-integer x = 0.446875

#> Warning: non-integer x = 3.538198

#> Warning: non-integer x = 2.253286

#> Warning: non-integer x = 0.287689

#> Warning: non-integer x = 1.320391

#> Warning: non-integer x = 1.712309

#> Warning: non-integer x = 1.541336

#> Warning: non-integer x = 2.654415

#> Warning: non-integer x = 0.580130

#> Warning: non-integer x = 0.565259

#> Warning: non-integer x = 0.505895

#> Warning: non-integer x = 1.049195

#> Warning: non-integer x = 1.430011

#> Warning: non-integer x = 0.042383

#> Warning: non-integer x = 0.886917

#> Warning: non-integer x = 0.749413

#> Warning: non-integer x = 1.409527

#> Warning: non-integer x = 1.565757

#> Warning: non-integer x = 1.837868

#> Warning: non-integer x = 2.924488

#> Warning: non-integer x = 0.875960

#> Warning: non-integer x = 3.266555

#> Warning: non-integer x = 1.713856

#> Warning: non-integer x = 1.677043

#> Warning: non-integer x = 0.990932

#> Warning: non-integer x = 3.118034

#> Warning: non-integer x = 2.541240

#> Warning: non-integer x = 1.459932

#> Warning: non-integer x = 0.487788

#> Warning: non-integer x = 1.357400

#> Warning: non-integer x = 3.502660

#> Warning: non-integer x = 4.007909

#> Warning: non-integer x = 1.736820

#> Warning: non-integer x = 2.147781

#> Warning: non-integer x = 0.696726

#> Warning: non-integer x = 4.119644

#> Warning: non-integer x = 1.160620

#> Warning: non-integer x = 0.987903

#> Warning: non-integer x = 3.354153

#> Warning: non-integer x = 0.335099

#> Warning: non-integer x = 1.046642

#> Warning: non-integer x = 2.286205

#> Warning: non-integer x = 0.495154

#> Warning: non-integer x = 0.588323

#> Warning: non-integer x = 0.844998

#> Warning: non-integer x = 0.916545

#> Warning: non-integer x = 0.613232

#> Warning: non-integer x = 1.971939

#> Warning: non-integer x = 2.001538

#> Warning: non-integer x = 0.396205

#> Warning: non-integer x = 2.418077

#> Warning: non-integer x = 0.658572

#> Warning: non-integer x = 2.240157

#> Warning: non-integer x = 1.596304

#> Warning: non-integer x = 0.099798

#> Warning: non-integer x = 0.452997

#> Warning: non-integer x = 0.368594

#> Warning: non-integer x = 1.516164

#> Warning: non-integer x = 0.924206

#> Warning: non-integer x = 1.487889

#> Warning: non-integer x = 2.270865

#> Warning: non-integer x = 0.917971

#> Warning: non-integer x = 1.481014

#> Warning: non-integer x = 0.108268

#> Warning: non-integer x = 2.341050

#> Warning: non-integer x = 1.359348

#> Warning: non-integer x = 0.229524

#> Warning: non-integer x = 0.649961

#> Warning: non-integer x = 1.444271

#> Warning: non-integer x = 1.673006

#> Warning: non-integer x = 0.053329

#> Warning: non-integer x = 1.377877

#> Warning: non-integer x = 0.055418

#> Warning: non-integer x = 0.279269

#> Warning: non-integer x = 2.694755

#> Warning: non-integer x = 2.451023

#> Warning: non-integer x = 0.413568

#> Warning: non-integer x = 1.312358

#> Warning: non-integer x = 2.720383

#> Warning: non-integer x = 0.330115

#> Warning: non-integer x = 1.331469

#> Warning: non-integer x = 2.898801

#> Warning: non-integer x = 2.605300

#> Warning: non-integer x = 2.966324

#> Warning: non-integer x = 1.227319

#> Warning: non-integer x = 1.347755

#> Warning: non-integer x = 2.363302

#> Warning: non-integer x = 2.480050

#> Warning: non-integer x = 0.649734

#> Warning: non-integer x = 0.303888

#> Warning: non-integer x = 0.158660

#> Warning: non-integer x = 0.755480

#> Warning: non-integer x = 1.346927

#> Warning: non-integer x = 3.126898

#> Warning: non-integer x = 0.088848

#> Warning: non-integer x = 1.795156

#> Warning: non-integer x = 0.593982

#> Warning: non-integer x = 2.735089

#> Warning: non-integer x = 0.199367

#> Warning: non-integer x = 0.698136

#> Warning: non-integer x = 0.080052

#> Warning: non-integer x = 0.511096

#> Warning: non-integer x = 0.017359

#> Warning: non-integer x = 0.816844

#> Warning: non-integer x = 1.999494

#> Warning: non-integer x = 0.402487

#> Warning: non-integer x = 1.956858

#> Warning: non-integer x = 1.203716

#> Warning: non-integer x = 4.627803

#> Warning: non-integer x = 3.133854

#> Warning: non-integer x = 1.690041

#> Warning: non-integer x = 1.204287

#> Warning: non-integer x = 1.593982

#> Warning: non-integer x = 4.019270

#> Warning: non-integer x = 1.442435

#> Warning: non-integer x = 2.737290

#> Warning: non-integer x = 1.334106

#> Warning: non-integer x = 0.166764

#> Warning: non-integer x = 2.495023

#> Warning: non-integer x = 1.140248

#> Warning: non-integer x = 0.524937

#> Warning: non-integer x = 1.588634

#> Warning: non-integer x = 0.432746

#> Warning: non-integer x = 1.147922

#> Warning: non-integer x = 0.971342

#> Warning: non-integer x = 2.717373

#> Warning: non-integer x = 0.960976

#> Warning: non-integer x = 1.016417

#> Warning: non-integer x = 1.634454

#> Warning: non-integer x = 0.453101

#> Warning: non-integer x = 1.882491

#> Warning: non-integer x = 1.271709

#> Warning: non-integer x = 2.182601

#> Warning: non-integer x = 0.355554

#> Warning: non-integer x = 1.103075

#> Warning: non-integer x = 0.161321

#> Warning: non-integer x = 0.179790

#> Warning: non-integer x = 0.812729

#> Warning: non-integer x = 2.208072

#> Warning: non-integer x = 1.741000

#> Warning: non-integer x = 0.475284

#> Warning: non-integer x = 0.946706

#> Warning: non-integer x = 0.128211

#> Warning: non-integer x = 1.924827

#> Warning: non-integer x = 1.570672

#> Warning: non-integer x = 2.563128

#> Warning: non-integer x = 1.842296

#> Warning: non-integer x = 0.074228

#> Warning: non-integer x = 0.520915

#> Warning: non-integer x = 0.363553

#> Warning: non-integer x = 1.666457

#> Warning: non-integer x = 0.702063

#> Warning: non-integer x = 1.776431

#> Warning: non-integer x = 1.689492

#> Warning: non-integer x = 0.473826

#> Warning: non-integer x = 0.463824

#> Warning: non-integer x = 1.605645

#> Warning: non-integer x = 3.584949

#> Warning: non-integer x = 1.253056

#> Warning: non-integer x = 0.422828

#> Warning: non-integer x = 0.528553

#> Warning: non-integer x = 2.840799

#> Warning: non-integer x = 1.706947

#> Warning: non-integer x = 0.149630

#> Warning: non-integer x = 0.202213

#> Warning: non-integer x = 0.436208

#> Warning: non-integer x = 2.152498

#> Warning: non-integer x = 1.579600

#> Warning: non-integer x = 0.204896

#> Warning: non-integer x = 2.344654

#> Warning: non-integer x = 0.967629

#> Warning: non-integer x = 1.241201

#> Warning: non-integer x = 1.685621

#> Warning: non-integer x = 1.206556

#> Warning: non-integer x = 2.160344