Estimate the Individual Treatment Effect using Poisson Regression

Estimates the Individual Treatment Effect using Poisson Regression given a response vector, a treatment vector, and a features matrix.

Usage

estimate_ite_poisson(y, z, X, X_names, include_offset, offset_name)

Arguments

y

the observed response vector

z

the treatment vector

X

the features matrix

X_names

the names of the covariates

include_offset

whether or not to include an offset when estimating the ITE, for poisson only

offset_name

the name of the offset, if it is to be included

Value

a vector of ITE estimates

Examples

dataset <- generate_cre_dataset(n = 1000, rho = 0, n_rules = 2, p = 10,
                                effect_size = 2, binary = FALSE)

# Initialize parameters
y <- abs(dataset[["y"]])
z <- dataset[["z"]]
X <- as.data.frame(dataset[["X"]])
X_names <- names(as.data.frame(X))
include_offset <- FALSE
offset_name <- NA

ite_list <- estimate_ite_poisson(y, z, X, X_names, include_offset, offset_name)
#> Warning: non-integer x = 1.329571
#> Warning: non-integer x = 2.189896
#> Warning: non-integer x = 1.287482
#> Warning: non-integer x = 0.735469
#> Warning: non-integer x = 2.488721
#> Warning: non-integer x = 0.880358
#> Warning: non-integer x = 0.519948
#> Warning: non-integer x = 1.828772
#> Warning: non-integer x = 1.284664
#> Warning: non-integer x = 0.306996
#> Warning: non-integer x = 0.467918
#> Warning: non-integer x = 0.659164
#> Warning: non-integer x = 3.324860
#> Warning: non-integer x = 0.796441
#> Warning: non-integer x = 1.181647
#> Warning: non-integer x = 1.304421
#> Warning: non-integer x = 2.903059
#> Warning: non-integer x = 3.453037
#> Warning: non-integer x = 3.458396
#> Warning: non-integer x = 1.700984
#> Warning: non-integer x = 0.758284
#> Warning: non-integer x = 4.044729
#> Warning: non-integer x = 1.326068
#> Warning: non-integer x = 1.453030
#> Warning: non-integer x = 2.834663
#> Warning: non-integer x = 1.294680
#> Warning: non-integer x = 0.549818
#> Warning: non-integer x = 2.246125
#> Warning: non-integer x = 1.985035
#> Warning: non-integer x = 0.110684
#> Warning: non-integer x = 4.247741
#> Warning: non-integer x = 2.379606
#> Warning: non-integer x = 1.836316
#> Warning: non-integer x = 2.609234
#> Warning: non-integer x = 3.689985
#> Warning: non-integer x = 1.341409
#> Warning: non-integer x = 1.740274
#> Warning: non-integer x = 1.035601
#> Warning: non-integer x = 2.922537
#> Warning: non-integer x = 2.637817
#> Warning: non-integer x = 1.535721
#> Warning: non-integer x = 0.570078
#> Warning: non-integer x = 0.364946
#> Warning: non-integer x = 3.305451
#> Warning: non-integer x = 2.056802
#> Warning: non-integer x = 3.960289
#> Warning: non-integer x = 1.051362
#> Warning: non-integer x = 2.309853
#> Warning: non-integer x = 4.443743
#> Warning: non-integer x = 2.142585
#> Warning: non-integer x = 2.945074
#> Warning: non-integer x = 2.968780
#> Warning: non-integer x = 1.073678
#> Warning: non-integer x = 0.620527
#> Warning: non-integer x = 2.573914
#> Warning: non-integer x = 1.278210
#> Warning: non-integer x = 1.261083
#> Warning: non-integer x = 3.281321
#> Warning: non-integer x = 1.331748
#> Warning: non-integer x = 2.004762
#> Warning: non-integer x = 3.895473
#> Warning: non-integer x = 0.037035
#> Warning: non-integer x = 1.395328
#> Warning: non-integer x = 2.053835
#> Warning: non-integer x = 1.717341
#> Warning: non-integer x = 2.659821
#> Warning: non-integer x = 1.473770
#> Warning: non-integer x = 1.944458
#> Warning: non-integer x = 1.972292
#> Warning: non-integer x = 2.106960
#> Warning: non-integer x = 0.438694
#> Warning: non-integer x = 2.003201
#> Warning: non-integer x = 1.753143
#> Warning: non-integer x = 0.655473
#> Warning: non-integer x = 3.676312
#> Warning: non-integer x = 1.216046
#> Warning: non-integer x = 1.350381
#> Warning: non-integer x = 1.531590
#> Warning: non-integer x = 1.540915
#> Warning: non-integer x = 1.160809
#> Warning: non-integer x = 2.941365
#> Warning: non-integer x = 4.081367
#> Warning: non-integer x = 3.475381
#> Warning: non-integer x = 2.745020
#> Warning: non-integer x = 1.167866
#> Warning: non-integer x = 0.906803
#> Warning: non-integer x = 3.485918
#> Warning: non-integer x = 2.066702
#> Warning: non-integer x = 2.234537
#> Warning: non-integer x = 2.636447
#> Warning: non-integer x = 3.136702
#> Warning: non-integer x = 0.559504
#> Warning: non-integer x = 1.126949
#> Warning: non-integer x = 2.619008
#> Warning: non-integer x = 0.839602
#> Warning: non-integer x = 1.154089
#> Warning: non-integer x = 0.458613
#> Warning: non-integer x = 0.373811
#> Warning: non-integer x = 1.113616
#> Warning: non-integer x = 3.947293
#> Warning: non-integer x = 0.146936
#> Warning: non-integer x = 2.192763
#> Warning: non-integer x = 3.150438
#> Warning: non-integer x = 1.627988
#> Warning: non-integer x = 0.789033
#> Warning: non-integer x = 2.202775
#> Warning: non-integer x = 2.034492
#> Warning: non-integer x = 1.056333
#> Warning: non-integer x = 3.214146
#> Warning: non-integer x = 0.714821
#> Warning: non-integer x = 2.267365
#> Warning: non-integer x = 0.670210
#> Warning: non-integer x = 2.404886
#> Warning: non-integer x = 2.077945
#> Warning: non-integer x = 0.160204
#> Warning: non-integer x = 4.084334
#> Warning: non-integer x = 1.208865
#> Warning: non-integer x = 0.028907
#> Warning: non-integer x = 0.035857
#> Warning: non-integer x = 2.185836
#> Warning: non-integer x = 3.269074
#> Warning: non-integer x = 1.896031
#> Warning: non-integer x = 3.620886
#> Warning: non-integer x = 1.147475
#> Warning: non-integer x = 0.457849
#> Warning: non-integer x = 3.013442
#> Warning: non-integer x = 1.779966
#> Warning: non-integer x = 3.317831
#> Warning: non-integer x = 1.882980
#> Warning: non-integer x = 1.678939
#> Warning: non-integer x = 3.234481
#> Warning: non-integer x = 0.073992
#> Warning: non-integer x = 1.038379
#> Warning: non-integer x = 2.166452
#> Warning: non-integer x = 0.851961
#> Warning: non-integer x = 2.114656
#> Warning: non-integer x = 3.702354
#> Warning: non-integer x = 1.772296
#> Warning: non-integer x = 5.314622
#> Warning: non-integer x = 1.400958
#> Warning: non-integer x = 3.813560
#> Warning: non-integer x = 2.203894
#> Warning: non-integer x = 1.629901
#> Warning: non-integer x = 0.289500
#> Warning: non-integer x = 2.407846
#> Warning: non-integer x = 1.583661
#> Warning: non-integer x = 1.136858
#> Warning: non-integer x = 0.698629
#> Warning: non-integer x = 1.086932
#> Warning: non-integer x = 0.438418
#> Warning: non-integer x = 0.477113
#> Warning: non-integer x = 1.823942
#> Warning: non-integer x = 1.427729
#> Warning: non-integer x = 1.815190
#> Warning: non-integer x = 3.223051
#> Warning: non-integer x = 1.224414
#> Warning: non-integer x = 0.626682
#> Warning: non-integer x = 1.100952
#> Warning: non-integer x = 0.012043
#> Warning: non-integer x = 3.530782
#> Warning: non-integer x = 0.318798
#> Warning: non-integer x = 0.453158
#> Warning: non-integer x = 0.514988
#> Warning: non-integer x = 1.245641
#> Warning: non-integer x = 2.338443
#> Warning: non-integer x = 2.752109
#> Warning: non-integer x = 0.366639
#> Warning: non-integer x = 2.079527
#> Warning: non-integer x = 4.757859
#> Warning: non-integer x = 1.680902
#> Warning: non-integer x = 1.709257
#> Warning: non-integer x = 2.111334
#> Warning: non-integer x = 2.825051
#> Warning: non-integer x = 0.899906
#> Warning: non-integer x = 0.226665
#> Warning: non-integer x = 0.064263
#> Warning: non-integer x = 0.758886
#> Warning: non-integer x = 1.667866
#> Warning: non-integer x = 0.155145
#> Warning: non-integer x = 0.356762
#> Warning: non-integer x = 0.590464
#> Warning: non-integer x = 0.186678
#> Warning: non-integer x = 2.626737
#> Warning: non-integer x = 0.729664
#> Warning: non-integer x = 2.528784
#> Warning: non-integer x = 2.086795
#> Warning: non-integer x = 4.843675
#> Warning: non-integer x = 1.373479
#> Warning: non-integer x = 3.057146
#> Warning: non-integer x = 0.618311
#> Warning: non-integer x = 3.200146
#> Warning: non-integer x = 2.376844
#> Warning: non-integer x = 1.194534
#> Warning: non-integer x = 0.998220
#> Warning: non-integer x = 2.255313
#> Warning: non-integer x = 0.120223
#> Warning: non-integer x = 0.664395
#> Warning: non-integer x = 1.870127
#> Warning: non-integer x = 2.327493
#> Warning: non-integer x = 2.475789
#> Warning: non-integer x = 3.228630
#> Warning: non-integer x = 3.428875
#> Warning: non-integer x = 2.647217
#> Warning: non-integer x = 1.471603
#> Warning: non-integer x = 2.133074
#> Warning: non-integer x = 2.274002
#> Warning: non-integer x = 2.757754
#> Warning: non-integer x = 0.552779
#> Warning: non-integer x = 0.200422
#> Warning: non-integer x = 3.156177
#> Warning: non-integer x = 1.906934
#> Warning: non-integer x = 2.726832
#> Warning: non-integer x = 3.244819
#> Warning: non-integer x = 0.710727
#> Warning: non-integer x = 2.077611
#> Warning: non-integer x = 1.956459
#> Warning: non-integer x = 1.187087
#> Warning: non-integer x = 0.350815
#> Warning: non-integer x = 3.858940
#> Warning: non-integer x = 0.885126
#> Warning: non-integer x = 1.848904
#> Warning: non-integer x = 0.899218
#> Warning: non-integer x = 2.618327
#> Warning: non-integer x = 1.649866
#> Warning: non-integer x = 1.800849
#> Warning: non-integer x = 1.551172
#> Warning: non-integer x = 2.651637
#> Warning: non-integer x = 1.170895
#> Warning: non-integer x = 0.657469
#> Warning: non-integer x = 1.546834
#> Warning: non-integer x = 2.262746
#> Warning: non-integer x = 1.562570
#> Warning: non-integer x = 1.215104
#> Warning: non-integer x = 1.501948
#> Warning: non-integer x = 0.685832
#> Warning: non-integer x = 3.777817
#> Warning: non-integer x = 2.441608
#> Warning: non-integer x = 0.307254
#> Warning: non-integer x = 2.033328
#> Warning: non-integer x = 1.882738
#> Warning: non-integer x = 1.693565
#> Warning: non-integer x = 1.877838
#> Warning: non-integer x = 0.462917
#> Warning: non-integer x = 1.312570
#> Warning: non-integer x = 1.346346
#> Warning: non-integer x = 1.002497
#> Warning: non-integer x = 3.294643
#> Warning: non-integer x = 4.191884
#> Warning: non-integer x = 2.189188
#> Warning: non-integer x = 2.318160
#> Warning: non-integer x = 2.725388
#> Warning: non-integer x = 0.173341
#> Warning: non-integer x = 0.666871
#> Warning: non-integer x = 3.346673
#> Warning: non-integer x = 1.640996
#> Warning: non-integer x = 3.211367
#> Warning: non-integer x = 1.150017
#> Warning: non-integer x = 0.996334
#> Warning: non-integer x = 1.068718
#> Warning: non-integer x = 3.229792
#> Warning: non-integer x = 3.017414
#> Warning: non-integer x = 1.366899
#> Warning: non-integer x = 3.049920
#> Warning: non-integer x = 2.205284
#> Warning: non-integer x = 1.919425
#> Warning: non-integer x = 0.310802
#> Warning: non-integer x = 0.137888
#> Warning: non-integer x = 0.083380
#> Warning: non-integer x = 0.227755
#> Warning: non-integer x = 0.584425
#> Warning: non-integer x = 2.144172
#> Warning: non-integer x = 0.254569
#> Warning: non-integer x = 3.002859
#> Warning: non-integer x = 0.994020
#> Warning: non-integer x = 1.219983
#> Warning: non-integer x = 1.570373
#> Warning: non-integer x = 2.240283
#> Warning: non-integer x = 2.219969
#> Warning: non-integer x = 4.431013
#> Warning: non-integer x = 1.827039
#> Warning: non-integer x = 0.520558
#> Warning: non-integer x = 2.231749
#> Warning: non-integer x = 0.681340
#> Warning: non-integer x = 1.471941
#> Warning: non-integer x = 0.091797
#> Warning: non-integer x = 0.723308
#> Warning: non-integer x = 0.759667
#> Warning: non-integer x = 3.786566
#> Warning: non-integer x = 0.587318
#> Warning: non-integer x = 1.416693
#> Warning: non-integer x = 0.034415
#> Warning: non-integer x = 0.945919
#> Warning: non-integer x = 2.317807
#> Warning: non-integer x = 0.549304
#> Warning: non-integer x = 0.665374
#> Warning: non-integer x = 0.168938
#> Warning: non-integer x = 0.156257
#> Warning: non-integer x = 1.029625
#> Warning: non-integer x = 2.713802
#> Warning: non-integer x = 0.334573
#> Warning: non-integer x = 0.673675
#> Warning: non-integer x = 0.944028
#> Warning: non-integer x = 1.714400
#> Warning: non-integer x = 3.151964
#> Warning: non-integer x = 3.051687
#> Warning: non-integer x = 5.229452
#> Warning: non-integer x = 0.398482
#> Warning: non-integer x = 2.884310
#> Warning: non-integer x = 2.521273
#> Warning: non-integer x = 1.721287
#> Warning: non-integer x = 0.487984
#> Warning: non-integer x = 0.330848
#> Warning: non-integer x = 3.184230
#> Warning: non-integer x = 0.421406
#> Warning: non-integer x = 0.735826
#> Warning: non-integer x = 1.461207
#> Warning: non-integer x = 4.244981
#> Warning: non-integer x = 1.678278
#> Warning: non-integer x = 1.214535
#> Warning: non-integer x = 0.170869
#> Warning: non-integer x = 0.053716
#> Warning: non-integer x = 2.898706
#> Warning: non-integer x = 0.824945
#> Warning: non-integer x = 0.392674
#> Warning: non-integer x = 0.214908
#> Warning: non-integer x = 0.421323
#> Warning: non-integer x = 0.665862
#> Warning: non-integer x = 0.020666
#> Warning: non-integer x = 1.714416
#> Warning: non-integer x = 2.779896
#> Warning: non-integer x = 1.648572
#> Warning: non-integer x = 0.485786
#> Warning: non-integer x = 1.552321
#> Warning: non-integer x = 3.007464
#> Warning: non-integer x = 0.720714
#> Warning: non-integer x = 0.392189
#> Warning: non-integer x = 1.132481
#> Warning: non-integer x = 0.086866
#> Warning: non-integer x = 0.713999
#> Warning: non-integer x = 0.380821
#> Warning: non-integer x = 0.441633
#> Warning: non-integer x = 1.122680
#> Warning: non-integer x = 1.236014
#> Warning: non-integer x = 0.616119
#> Warning: non-integer x = 2.027739
#> Warning: non-integer x = 1.533247
#> Warning: non-integer x = 0.128062
#> Warning: non-integer x = 0.556850
#> Warning: non-integer x = 1.644630
#> Warning: non-integer x = 0.193830
#> Warning: non-integer x = 0.546893
#> Warning: non-integer x = 3.139376
#> Warning: non-integer x = 1.724523
#> Warning: non-integer x = 0.322784
#> Warning: non-integer x = 2.222362
#> Warning: non-integer x = 1.569536
#> Warning: non-integer x = 2.749234
#> Warning: non-integer x = 1.567895
#> Warning: non-integer x = 2.168407
#> Warning: non-integer x = 0.636803
#> Warning: non-integer x = 1.462914
#> Warning: non-integer x = 1.733413
#> Warning: non-integer x = 1.724415
#> Warning: non-integer x = 1.976634
#> Warning: non-integer x = 3.610316
#> Warning: non-integer x = 0.682310
#> Warning: non-integer x = 1.321001
#> Warning: non-integer x = 2.802347
#> Warning: non-integer x = 3.074841
#> Warning: non-integer x = 1.038285
#> Warning: non-integer x = 3.635853
#> Warning: non-integer x = 1.230419
#> Warning: non-integer x = 3.200732
#> Warning: non-integer x = 4.631856
#> Warning: non-integer x = 1.037452
#> Warning: non-integer x = 1.025884
#> Warning: non-integer x = 3.179240
#> Warning: non-integer x = 3.520128
#> Warning: non-integer x = 1.375080
#> Warning: non-integer x = 2.020608
#> Warning: non-integer x = 1.692921
#> Warning: non-integer x = 2.152017
#> Warning: non-integer x = 3.276854
#> Warning: non-integer x = 1.160086
#> Warning: non-integer x = 4.167709
#> Warning: non-integer x = 0.234885
#> Warning: non-integer x = 5.552545
#> Warning: non-integer x = 2.355944
#> Warning: non-integer x = 0.162108
#> Warning: non-integer x = 3.294212
#> Warning: non-integer x = 2.575123
#> Warning: non-integer x = 1.923723
#> Warning: non-integer x = 4.909810
#> Warning: non-integer x = 1.212109
#> Warning: non-integer x = 1.448722
#> Warning: non-integer x = 1.512221
#> Warning: non-integer x = 1.237451
#> Warning: non-integer x = 2.478864
#> Warning: non-integer x = 3.226985
#> Warning: non-integer x = 1.046891
#> Warning: non-integer x = 1.708950
#> Warning: non-integer x = 0.488246
#> Warning: non-integer x = 0.162010
#> Warning: non-integer x = 0.159803
#> Warning: non-integer x = 1.402852
#> Warning: non-integer x = 1.130816
#> Warning: non-integer x = 1.812764
#> Warning: non-integer x = 2.561296
#> Warning: non-integer x = 1.390579
#> Warning: non-integer x = 0.792036
#> Warning: non-integer x = 0.271825
#> Warning: non-integer x = 2.813388
#> Warning: non-integer x = 0.817812
#> Warning: non-integer x = 0.810648
#> Warning: non-integer x = 2.818694
#> Warning: non-integer x = 0.264354
#> Warning: non-integer x = 1.322655
#> Warning: non-integer x = 0.741461
#> Warning: non-integer x = 1.368218
#> Warning: non-integer x = 1.072807
#> Warning: non-integer x = 1.254035
#> Warning: non-integer x = 0.619939
#> Warning: non-integer x = 1.818116
#> Warning: non-integer x = 2.055195
#> Warning: non-integer x = 3.362195
#> Warning: non-integer x = 1.168124
#> Warning: non-integer x = 1.996263
#> Warning: non-integer x = 2.353299
#> Warning: non-integer x = 0.394179
#> Warning: non-integer x = 2.309642
#> Warning: non-integer x = 2.270143
#> Warning: non-integer x = 0.593311
#> Warning: non-integer x = 1.607826
#> Warning: non-integer x = 0.201186
#> Warning: non-integer x = 0.094587
#> Warning: non-integer x = 2.160983
#> Warning: non-integer x = 0.798766
#> Warning: non-integer x = 1.556299
#> Warning: non-integer x = 2.174091
#> Warning: non-integer x = 2.885618
#> Warning: non-integer x = 3.107788
#> Warning: non-integer x = 1.829297
#> Warning: non-integer x = 3.020078
#> Warning: non-integer x = 0.638336
#> Warning: non-integer x = 0.827952
#> Warning: non-integer x = 0.278304
#> Warning: non-integer x = 0.710136
#> Warning: non-integer x = 1.900141
#> Warning: non-integer x = 2.037689
#> Warning: non-integer x = 3.191655
#> Warning: non-integer x = 0.964017
#> Warning: non-integer x = 1.733823
#> Warning: non-integer x = 1.857941
#> Warning: non-integer x = 0.557960
#> Warning: non-integer x = 1.143669
#> Warning: non-integer x = 1.948901
#> Warning: non-integer x = 1.818547
#> Warning: non-integer x = 1.808165
#> Warning: non-integer x = 0.073860
#> Warning: non-integer x = 0.824902
#> Warning: non-integer x = 2.307884
#> Warning: non-integer x = 0.525466
#> Warning: non-integer x = 0.813766
#> Warning: non-integer x = 4.598934
#> Warning: non-integer x = 0.214996
#> Warning: non-integer x = 1.341302
#> Warning: non-integer x = 0.001911
#> Warning: non-integer x = 3.435400
#> Warning: non-integer x = 3.321828
#> Warning: non-integer x = 0.240147
#> Warning: non-integer x = 2.099812
#> Warning: non-integer x = 1.014693
#> Warning: non-integer x = 0.367795
#> Warning: non-integer x = 2.189881
#> Warning: non-integer x = 0.722903
#> Warning: non-integer x = 0.510296
#> Warning: non-integer x = 0.515780
#> Warning: non-integer x = 0.067732
#> Warning: non-integer x = 1.102481
#> Warning: non-integer x = 0.937289
#> Warning: non-integer x = 0.805774
#> Warning: non-integer x = 0.017423
#> Warning: non-integer x = 1.028829
#> Warning: non-integer x = 0.321000
#> Warning: non-integer x = 0.502617
#> Warning: non-integer x = 2.771242
#> Warning: non-integer x = 1.731909
#> Warning: non-integer x = 2.156022
#> Warning: non-integer x = 0.249141
#> Warning: non-integer x = 0.157857
#> Warning: non-integer x = 0.602808
#> Warning: non-integer x = 0.000760
#> Warning: non-integer x = 0.235168
#> Warning: non-integer x = 0.368571
#> Warning: non-integer x = 0.829675
#> Warning: non-integer x = 2.718089
#> Warning: non-integer x = 0.615118
#> Warning: non-integer x = 0.365444
#> Warning: non-integer x = 2.145224
#> Warning: non-integer x = 1.162593
#> Warning: non-integer x = 0.901296
#> Warning: non-integer x = 0.093317
#> Warning: non-integer x = 1.394206
#> Warning: non-integer x = 0.710294
#> Warning: non-integer x = 3.351070
#> Warning: non-integer x = 0.146058
#> Warning: non-integer x = 1.684819
#> Warning: non-integer x = 1.092994
#> Warning: non-integer x = 1.955170
#> Warning: non-integer x = 2.487545
#> Warning: non-integer x = 0.343713
#> Warning: non-integer x = 3.078645
#> Warning: non-integer x = 1.026158
#> Warning: non-integer x = 2.194328
#> Warning: non-integer x = 0.504527
#> Warning: non-integer x = 2.112691
#> Warning: non-integer x = 1.060668
#> Warning: non-integer x = 0.527781
#> Warning: non-integer x = 0.235660
#> Warning: non-integer x = 0.740483
#> Warning: non-integer x = 3.738769
#> Warning: non-integer x = 1.286623
#> Warning: non-integer x = 0.597422
#> Warning: non-integer x = 0.878121
#> Warning: non-integer x = 2.021323
#> Warning: non-integer x = 1.317948
#> Warning: non-integer x = 1.201859
#> Warning: non-integer x = 0.762728
#> Warning: non-integer x = 0.545340
#> Warning: non-integer x = 1.059633
#> Warning: non-integer x = 4.405935
#> Warning: non-integer x = 1.920892
#> Warning: non-integer x = 0.909815
#> Warning: non-integer x = 0.349017
#> Warning: non-integer x = 4.185620
#> Warning: non-integer x = 0.564956
#> Warning: non-integer x = 1.946870
#> Warning: non-integer x = 1.763208
#> Warning: non-integer x = 0.588598
#> Warning: non-integer x = 0.705893
#> Warning: non-integer x = 1.163487
#> Warning: non-integer x = 2.524362
#> Warning: non-integer x = 0.484212
#> Warning: non-integer x = 2.586228
#> Warning: non-integer x = 1.523825
#> Warning: non-integer x = 1.409023
#> Warning: non-integer x = 1.424484
#> Warning: non-integer x = 1.580643
#> Warning: non-integer x = 1.612213
#> Warning: non-integer x = 2.060044
#> Warning: non-integer x = 3.902895
#> Warning: non-integer x = 0.573934
#> Warning: non-integer x = 1.499951
#> Warning: non-integer x = 1.032124
#> Warning: non-integer x = 0.222054
#> Warning: non-integer x = 0.285793
#> Warning: non-integer x = 0.136120
#> Warning: non-integer x = 3.451553
#> Warning: non-integer x = 0.716869
#> Warning: non-integer x = 0.176208
#> Warning: non-integer x = 3.212441
#> Warning: non-integer x = 0.552134
#> Warning: non-integer x = 0.345488
#> Warning: non-integer x = 1.084298
#> Warning: non-integer x = 0.925055
#> Warning: non-integer x = 1.381896
#> Warning: non-integer x = 1.581426
#> Warning: non-integer x = 1.816198
#> Warning: non-integer x = 0.802127
#> Warning: non-integer x = 2.290932
#> Warning: non-integer x = 2.474868
#> Warning: non-integer x = 1.606652
#> Warning: non-integer x = 1.431761
#> Warning: non-integer x = 2.133935
#> Warning: non-integer x = 1.179918
#> Warning: non-integer x = 1.841247
#> Warning: non-integer x = 2.236606
#> Warning: non-integer x = 2.122096
#> Warning: non-integer x = 0.582886
#> Warning: non-integer x = 1.447313
#> Warning: non-integer x = 0.170132
#> Warning: non-integer x = 2.800608
#> Warning: non-integer x = 0.513621
#> Warning: non-integer x = 1.143932
#> Warning: non-integer x = 1.276814
#> Warning: non-integer x = 0.065751
#> Warning: non-integer x = 0.463142
#> Warning: non-integer x = 0.559481
#> Warning: non-integer x = 3.249188
#> Warning: non-integer x = 2.591897
#> Warning: non-integer x = 0.261284
#> Warning: non-integer x = 1.678569
#> Warning: non-integer x = 0.994199
#> Warning: non-integer x = 0.335383
#> Warning: non-integer x = 1.673583
#> Warning: non-integer x = 1.792937
#> Warning: non-integer x = 0.038116
#> Warning: non-integer x = 0.809958
#> Warning: non-integer x = 0.295913
#> Warning: non-integer x = 2.344434
#> Warning: non-integer x = 0.688101
#> Warning: non-integer x = 1.222716
#> Warning: non-integer x = 1.515733
#> Warning: non-integer x = 1.887436
#> Warning: non-integer x = 0.506291
#> Warning: non-integer x = 2.172225
#> Warning: non-integer x = 0.226810
#> Warning: non-integer x = 0.037731
#> Warning: non-integer x = 0.318881
#> Warning: non-integer x = 3.161386
#> Warning: non-integer x = 0.522305
#> Warning: non-integer x = 1.437263
#> Warning: non-integer x = 1.275640
#> Warning: non-integer x = 1.746967
#> Warning: non-integer x = 1.174976
#> Warning: non-integer x = 1.064491
#> Warning: non-integer x = 0.725949
#> Warning: non-integer x = 0.274344
#> Warning: non-integer x = 4.852752
#> Warning: non-integer x = 1.280314
#> Warning: non-integer x = 1.233029
#> Warning: non-integer x = 1.412020
#> Warning: non-integer x = 1.730538
#> Warning: non-integer x = 1.162671
#> Warning: non-integer x = 1.420716
#> Warning: non-integer x = 0.940845
#> Warning: non-integer x = 2.986006
#> Warning: non-integer x = 1.996705
#> Warning: non-integer x = 0.418428
#> Warning: non-integer x = 0.687063
#> Warning: non-integer x = 2.434311
#> Warning: non-integer x = 2.383170
#> Warning: non-integer x = 0.218393
#> Warning: non-integer x = 1.824025
#> Warning: non-integer x = 0.422035
#> Warning: non-integer x = 0.904037
#> Warning: non-integer x = 0.745648
#> Warning: non-integer x = 2.563060
#> Warning: non-integer x = 1.389049
#> Warning: non-integer x = 0.572684
#> Warning: non-integer x = 0.433092
#> Warning: non-integer x = 0.924505
#> Warning: non-integer x = 1.814068
#> Warning: non-integer x = 0.292342
#> Warning: non-integer x = 2.102607
#> Warning: non-integer x = 1.188533
#> Warning: non-integer x = 1.816206
#> Warning: non-integer x = 0.249429
#> Warning: non-integer x = 0.962807
#> Warning: non-integer x = 0.750108
#> Warning: non-integer x = 2.230288
#> Warning: non-integer x = 1.786108
#> Warning: non-integer x = 0.119713
#> Warning: non-integer x = 1.719878
#> Warning: non-integer x = 2.799691
#> Warning: non-integer x = 0.341616
#> Warning: non-integer x = 0.236638
#> Warning: non-integer x = 1.399837
#> Warning: non-integer x = 1.789743
#> Warning: non-integer x = 2.071843
#> Warning: non-integer x = 1.531466
#> Warning: non-integer x = 0.956859
#> Warning: non-integer x = 1.901960
#> Warning: non-integer x = 1.105011
#> Warning: non-integer x = 2.788297
#> Warning: non-integer x = 3.157137
#> Warning: non-integer x = 0.100474
#> Warning: non-integer x = 2.298146
#> Warning: non-integer x = 1.261649
#> Warning: non-integer x = 1.374975
#> Warning: non-integer x = 0.484534
#> Warning: non-integer x = 1.208577
#> Warning: non-integer x = 3.351584
#> Warning: non-integer x = 1.118379
#> Warning: non-integer x = 0.554395
#> Warning: non-integer x = 0.893971
#> Warning: non-integer x = 3.170424
#> Warning: non-integer x = 2.679245
#> Warning: non-integer x = 5.052856
#> Warning: non-integer x = 1.246792
#> Warning: non-integer x = 0.982492
#> Warning: non-integer x = 2.390092
#> Warning: non-integer x = 0.156827
#> Warning: non-integer x = 2.819136
#> Warning: non-integer x = 0.107942
#> Warning: non-integer x = 0.084462
#> Warning: non-integer x = 0.768311
#> Warning: non-integer x = 1.999355
#> Warning: non-integer x = 1.339688
#> Warning: non-integer x = 2.902143
#> Warning: non-integer x = 0.499101
#> Warning: non-integer x = 2.280589
#> Warning: non-integer x = 1.329264
#> Warning: non-integer x = 1.180541
#> Warning: non-integer x = 3.241522
#> Warning: non-integer x = 0.860787
#> Warning: non-integer x = 2.813117
#> Warning: non-integer x = 0.604113
#> Warning: non-integer x = 0.167992
#> Warning: non-integer x = 3.118191
#> Warning: non-integer x = 0.433551
#> Warning: non-integer x = 0.483398
#> Warning: non-integer x = 1.245271
#> Warning: non-integer x = 0.112844
#> Warning: non-integer x = 0.906722
#> Warning: non-integer x = 2.086175
#> Warning: non-integer x = 2.698640
#> Warning: non-integer x = 0.740164
#> Warning: non-integer x = 0.009737
#> Warning: non-integer x = 1.633594
#> Warning: non-integer x = 0.722542
#> Warning: non-integer x = 1.935087
#> Warning: non-integer x = 2.566064
#> Warning: non-integer x = 0.766833
#> Warning: non-integer x = 2.926219
#> Warning: non-integer x = 1.494171
#> Warning: non-integer x = 0.434807
#> Warning: non-integer x = 2.514482
#> Warning: non-integer x = 0.455157
#> Warning: non-integer x = 1.623872
#> Warning: non-integer x = 2.385460
#> Warning: non-integer x = 2.857592
#> Warning: non-integer x = 2.599914
#> Warning: non-integer x = 0.933575
#> Warning: non-integer x = 1.783123
#> Warning: non-integer x = 0.425374
#> Warning: non-integer x = 1.694395
#> Warning: non-integer x = 2.296386
#> Warning: non-integer x = 1.669315
#> Warning: non-integer x = 0.732893
#> Warning: non-integer x = 2.302233
#> Warning: non-integer x = 1.599598
#> Warning: non-integer x = 0.691962
#> Warning: non-integer x = 1.593602
#> Warning: non-integer x = 0.206204
#> Warning: non-integer x = 1.382440
#> Warning: non-integer x = 3.725809
#> Warning: non-integer x = 2.531508
#> Warning: non-integer x = 2.542395
#> Warning: non-integer x = 2.282299
#> Warning: non-integer x = 0.881408
#> Warning: non-integer x = 0.761000
#> Warning: non-integer x = 0.511661
#> Warning: non-integer x = 1.048245
#> Warning: non-integer x = 0.431301
#> Warning: non-integer x = 0.067271
#> Warning: non-integer x = 0.387583
#> Warning: non-integer x = 0.661303
#> Warning: non-integer x = 1.547032
#> Warning: non-integer x = 1.733702
#> Warning: non-integer x = 2.275651
#> Warning: non-integer x = 0.046162
#> Warning: non-integer x = 1.613047
#> Warning: non-integer x = 1.667834
#> Warning: non-integer x = 2.508734
#> Warning: non-integer x = 0.828235
#> Warning: non-integer x = 0.665935
#> Warning: non-integer x = 2.311628
#> Warning: non-integer x = 0.163262
#> Warning: non-integer x = 2.357081
#> Warning: non-integer x = 3.164769
#> Warning: non-integer x = 0.279349
#> Warning: non-integer x = 0.096730
#> Warning: non-integer x = 1.738815
#> Warning: non-integer x = 0.333873
#> Warning: non-integer x = 2.147894
#> Warning: non-integer x = 0.683551
#> Warning: non-integer x = 1.069778
#> Warning: non-integer x = 0.942528
#> Warning: non-integer x = 2.348088
#> Warning: non-integer x = 0.132644
#> Warning: non-integer x = 0.321495
#> Warning: non-integer x = 1.839039
#> Warning: non-integer x = 2.624133
#> Warning: non-integer x = 1.290232
#> Warning: non-integer x = 1.457443
#> Warning: non-integer x = 1.878697
#> Warning: non-integer x = 0.909673
#> Warning: non-integer x = 1.916071
#> Warning: non-integer x = 1.749127
#> Warning: non-integer x = 0.002462
#> Warning: non-integer x = 0.011991
#> Warning: non-integer x = 1.678201
#> Warning: non-integer x = 0.416279
#> Warning: non-integer x = 1.955562
#> Warning: non-integer x = 0.215679
#> Warning: non-integer x = 0.325276
#> Warning: non-integer x = 1.986417
#> Warning: non-integer x = 0.016506
#> Warning: non-integer x = 0.769151
#> Warning: non-integer x = 0.880962
#> Warning: non-integer x = 0.161565
#> Warning: non-integer x = 2.940597
#> Warning: non-integer x = 0.695814
#> Warning: non-integer x = 0.861236
#> Warning: non-integer x = 1.048896
#> Warning: non-integer x = 0.488319
#> Warning: non-integer x = 1.297642
#> Warning: non-integer x = 2.211211
#> Warning: non-integer x = 2.190148
#> Warning: non-integer x = 1.348658
#> Warning: non-integer x = 0.534714
#> Warning: non-integer x = 0.197772
#> Warning: non-integer x = 2.078391
#> Warning: non-integer x = 1.677447
#> Warning: non-integer x = 1.057935
#> Warning: non-integer x = 0.980078
#> Warning: non-integer x = 1.328674
#> Warning: non-integer x = 1.352447
#> Warning: non-integer x = 1.039486
#> Warning: non-integer x = 1.802786
#> Warning: non-integer x = 0.128861
#> Warning: non-integer x = 2.780723
#> Warning: non-integer x = 1.512874
#> Warning: non-integer x = 0.767836
#> Warning: non-integer x = 1.010714
#> Warning: non-integer x = 1.389610
#> Warning: non-integer x = 1.470807
#> Warning: non-integer x = 4.135946
#> Warning: non-integer x = 0.498677
#> Warning: non-integer x = 2.142441
#> Warning: non-integer x = 0.588185
#> Warning: non-integer x = 2.208697
#> Warning: non-integer x = 0.900989
#> Warning: non-integer x = 0.500300
#> Warning: non-integer x = 1.398623
#> Warning: non-integer x = 0.929733
#> Warning: non-integer x = 2.142239
#> Warning: non-integer x = 0.330088
#> Warning: non-integer x = 2.562044
#> Warning: non-integer x = 2.224235
#> Warning: non-integer x = 0.625417
#> Warning: non-integer x = 1.224930
#> Warning: non-integer x = 0.899040
#> Warning: non-integer x = 1.803463
#> Warning: non-integer x = 1.833755
#> Warning: non-integer x = 0.136810
#> Warning: non-integer x = 2.713252
#> Warning: non-integer x = 1.111207
#> Warning: non-integer x = 1.137737
#> Warning: non-integer x = 0.543472
#> Warning: non-integer x = 2.484199
#> Warning: non-integer x = 0.501798
#> Warning: non-integer x = 1.721950
#> Warning: non-integer x = 2.614900
#> Warning: non-integer x = 0.148443
#> Warning: non-integer x = 0.164479
#> Warning: non-integer x = 0.086149
#> Warning: non-integer x = 1.442559
#> Warning: non-integer x = 2.406178
#> Warning: non-integer x = 2.305585
#> Warning: non-integer x = 1.557366
#> Warning: non-integer x = 0.113194
#> Warning: non-integer x = 0.407487
#> Warning: non-integer x = 0.432342
#> Warning: non-integer x = 2.507135
#> Warning: non-integer x = 0.877554
#> Warning: non-integer x = 0.674376
#> Warning: non-integer x = 1.668023
#> Warning: non-integer x = 1.896652
#> Warning: non-integer x = 2.663225
#> Warning: non-integer x = 0.982606
#> Warning: non-integer x = 0.154592
#> Warning: non-integer x = 2.445623
#> Warning: non-integer x = 0.931500
#> Warning: non-integer x = 1.619289
#> Warning: non-integer x = 2.569988
#> Warning: non-integer x = 2.674770
#> Warning: non-integer x = 1.134307
#> Warning: non-integer x = 0.986554
#> Warning: non-integer x = 0.457461
#> Warning: non-integer x = 0.792374
#> Warning: non-integer x = 0.210747
#> Warning: non-integer x = 0.085532
#> Warning: non-integer x = 1.990365
#> Warning: non-integer x = 3.033013
#> Warning: non-integer x = 1.374513
#> Warning: non-integer x = 2.514397
#> Warning: non-integer x = 0.091459
#> Warning: non-integer x = 2.314268
#> Warning: non-integer x = 0.198675
#> Warning: non-integer x = 1.578032
#> Warning: non-integer x = 0.230152
#> Warning: non-integer x = 1.272149
#> Warning: non-integer x = 2.033988
#> Warning: non-integer x = 0.935346
#> Warning: non-integer x = 0.829831
#> Warning: non-integer x = 3.965484
#> Warning: non-integer x = 1.106812
#> Warning: non-integer x = 1.011340
#> Warning: non-integer x = 0.900805
#> Warning: non-integer x = 0.786001
#> Warning: non-integer x = 2.925151
#> Warning: non-integer x = 2.669469
#> Warning: non-integer x = 1.696853
#> Warning: non-integer x = 0.283370
#> Warning: non-integer x = 2.685794
#> Warning: non-integer x = 0.676350
#> Warning: non-integer x = 0.779267
#> Warning: non-integer x = 1.256227
#> Warning: non-integer x = 0.319383
#> Warning: non-integer x = 0.219189
#> Warning: non-integer x = 0.306087
#> Warning: non-integer x = 1.933455
#> Warning: non-integer x = 2.059780
#> Warning: non-integer x = 2.668588
#> Warning: non-integer x = 1.339339
#> Warning: non-integer x = 1.920804
#> Warning: non-integer x = 2.159238
#> Warning: non-integer x = 1.821151
#> Warning: non-integer x = 0.587294
#> Warning: non-integer x = 1.568148
#> Warning: non-integer x = 0.279041
#> Warning: non-integer x = 2.412869
#> Warning: non-integer x = 4.295267
#> Warning: non-integer x = 0.592557
#> Warning: non-integer x = 2.541790
#> Warning: non-integer x = 0.959111
#> Warning: non-integer x = 1.756377
#> Warning: non-integer x = 0.827387
#> Warning: non-integer x = 0.345098
#> Warning: non-integer x = 1.388741
#> Warning: non-integer x = 3.532591
#> Warning: non-integer x = 1.584149
#> Warning: non-integer x = 2.489761
#> Warning: non-integer x = 0.204349
#> Warning: non-integer x = 1.011077
#> Warning: non-integer x = 2.141552
#> Warning: non-integer x = 1.600280
#> Warning: non-integer x = 0.317164
#> Warning: non-integer x = 0.000543
#> Warning: non-integer x = 0.305344
#> Warning: non-integer x = 0.695445
#> Warning: non-integer x = 1.564343
#> Warning: non-integer x = 2.205547
#> Warning: non-integer x = 1.760839
#> Warning: non-integer x = 0.735551
#> Warning: non-integer x = 0.454861
#> Warning: non-integer x = 0.289815
#> Warning: non-integer x = 2.031473
#> Warning: non-integer x = 0.193821
#> Warning: non-integer x = 0.921161
#> Warning: non-integer x = 1.594805
#> Warning: non-integer x = 0.634290
#> Warning: non-integer x = 1.561961
#> Warning: non-integer x = 0.212439
#> Warning: non-integer x = 0.878972
#> Warning: non-integer x = 1.738874
#> Warning: non-integer x = 0.811879
#> Warning: non-integer x = 1.566012
#> Warning: non-integer x = 0.234847
#> Warning: non-integer x = 3.108150
#> Warning: non-integer x = 0.071681
#> Warning: non-integer x = 0.316996
#> Warning: non-integer x = 1.424033
#> Warning: non-integer x = 0.306778
#> Warning: non-integer x = 1.497658
#> Warning: non-integer x = 2.430384
#> Warning: non-integer x = 0.374603
#> Warning: non-integer x = 3.108138
#> Warning: non-integer x = 1.360578
#> Warning: non-integer x = 0.316087
#> Warning: non-integer x = 0.792448
#> Warning: non-integer x = 0.040997
#> Warning: non-integer x = 1.588111
#> Warning: non-integer x = 3.072961
#> Warning: non-integer x = 2.072158
#> Warning: non-integer x = 2.864207
#> Warning: non-integer x = 1.106493
#> Warning: non-integer x = 1.076578
#> Warning: non-integer x = 1.147324
#> Warning: non-integer x = 0.684265
#> Warning: non-integer x = 0.900117
#> Warning: non-integer x = 3.249697
#> Warning: non-integer x = 0.723661
#> Warning: non-integer x = 2.430967
#> Warning: non-integer x = 0.521749
#> Warning: non-integer x = 0.745270
#> Warning: non-integer x = 0.110260
#> Warning: non-integer x = 0.796252
#> Warning: non-integer x = 0.538714
#> Warning: non-integer x = 1.887448
#> Warning: non-integer x = 1.138972
#> Warning: non-integer x = 1.012998
#> Warning: non-integer x = 1.013776
#> Warning: non-integer x = 1.021292
#> Warning: non-integer x = 0.726669
#> Warning: non-integer x = 2.102350
#> Warning: non-integer x = 1.983957
#> Warning: non-integer x = 0.958118
#> Warning: non-integer x = 2.872688
#> Warning: non-integer x = 1.774191
#> Warning: non-integer x = 1.672768
#> Warning: non-integer x = 0.013686
#> Warning: non-integer x = 1.143923
#> Warning: non-integer x = 1.970211
#> Warning: non-integer x = 0.490305
#> Warning: non-integer x = 0.165751
#> Warning: non-integer x = 1.174306
#> Warning: non-integer x = 0.048667