Estimate the Individual Treatment Effect using Poisson Regression

Estimates the Individual Treatment Effect using Poisson Regression given a response vector, a treatment vector, and a features matrix.

Usage

estimate_ite_poisson(y, z, X, X_names, include_offset, offset_name)

Arguments

y

the observed response vector

z

the treatment vector

X

the features matrix

X_names

the names of the covariates

include_offset

whether or not to include an offset when estimating the ITE, for poisson only

offset_name

the name of the offset, if it is to be included

Value

a vector of ITE estimates

Examples

dataset <- generate_cre_dataset(n = 1000, rho = 0, n_rules = 2, p = 10,
                                effect_size = 2, binary = FALSE)

# Initialize parameters
y <- abs(dataset[["y"]])
z <- dataset[["z"]]
X <- as.data.frame(dataset[["X"]])
X_names <- names(as.data.frame(X))
include_offset <- FALSE
offset_name <- NA

ite_list <- estimate_ite_poisson(y, z, X, X_names, include_offset, offset_name)
#> Warning: non-integer x = 0.531051
#> Warning: non-integer x = 0.137855
#> Warning: non-integer x = 1.666939
#> Warning: non-integer x = 1.154956
#> Warning: non-integer x = 0.774683
#> Warning: non-integer x = 1.278338
#> Warning: non-integer x = 1.338639
#> Warning: non-integer x = 0.336560
#> Warning: non-integer x = 3.209410
#> Warning: non-integer x = 0.387837
#> Warning: non-integer x = 3.274734
#> Warning: non-integer x = 2.172643
#> Warning: non-integer x = 2.280190
#> Warning: non-integer x = 0.804160
#> Warning: non-integer x = 1.280454
#> Warning: non-integer x = 3.931554
#> Warning: non-integer x = 2.147411
#> Warning: non-integer x = 0.661066
#> Warning: non-integer x = 0.968824
#> Warning: non-integer x = 2.390063
#> Warning: non-integer x = 2.630768
#> Warning: non-integer x = 1.040762
#> Warning: non-integer x = 2.162851
#> Warning: non-integer x = 3.464486
#> Warning: non-integer x = 2.113291
#> Warning: non-integer x = 1.097068
#> Warning: non-integer x = 1.424617
#> Warning: non-integer x = 2.433979
#> Warning: non-integer x = 1.095910
#> Warning: non-integer x = 0.474419
#> Warning: non-integer x = 2.678390
#> Warning: non-integer x = 2.085590
#> Warning: non-integer x = 1.224533
#> Warning: non-integer x = 1.706777
#> Warning: non-integer x = 1.016028
#> Warning: non-integer x = 1.302974
#> Warning: non-integer x = 2.211423
#> Warning: non-integer x = 1.390800
#> Warning: non-integer x = 1.971374
#> Warning: non-integer x = 0.190329
#> Warning: non-integer x = 2.152201
#> Warning: non-integer x = 0.145882
#> Warning: non-integer x = 0.273705
#> Warning: non-integer x = 1.924918
#> Warning: non-integer x = 0.121305
#> Warning: non-integer x = 2.635421
#> Warning: non-integer x = 1.262622
#> Warning: non-integer x = 1.247847
#> Warning: non-integer x = 2.084197
#> Warning: non-integer x = 1.345789
#> Warning: non-integer x = 2.859653
#> Warning: non-integer x = 1.862679
#> Warning: non-integer x = 0.826953
#> Warning: non-integer x = 2.762272
#> Warning: non-integer x = 3.567017
#> Warning: non-integer x = 0.457197
#> Warning: non-integer x = 0.046938
#> Warning: non-integer x = 2.327082
#> Warning: non-integer x = 3.471134
#> Warning: non-integer x = 2.215619
#> Warning: non-integer x = 5.140459
#> Warning: non-integer x = 2.735973
#> Warning: non-integer x = 2.448612
#> Warning: non-integer x = 3.392929
#> Warning: non-integer x = 2.371478
#> Warning: non-integer x = 4.023620
#> Warning: non-integer x = 1.569757
#> Warning: non-integer x = 1.555053
#> Warning: non-integer x = 2.068417
#> Warning: non-integer x = 2.817476
#> Warning: non-integer x = 1.305412
#> Warning: non-integer x = 0.943105
#> Warning: non-integer x = 1.192176
#> Warning: non-integer x = 1.110854
#> Warning: non-integer x = 0.387163
#> Warning: non-integer x = 2.963179
#> Warning: non-integer x = 1.285740
#> Warning: non-integer x = 3.286838
#> Warning: non-integer x = 1.924012
#> Warning: non-integer x = 1.732936
#> Warning: non-integer x = 0.713829
#> Warning: non-integer x = 2.043968
#> Warning: non-integer x = 2.358049
#> Warning: non-integer x = 1.867296
#> Warning: non-integer x = 1.139718
#> Warning: non-integer x = 0.911935
#> Warning: non-integer x = 1.907172
#> Warning: non-integer x = 0.581254
#> Warning: non-integer x = 0.331094
#> Warning: non-integer x = 0.776051
#> Warning: non-integer x = 1.670838
#> Warning: non-integer x = 0.824352
#> Warning: non-integer x = 2.630450
#> Warning: non-integer x = 0.608392
#> Warning: non-integer x = 1.919431
#> Warning: non-integer x = 2.540546
#> Warning: non-integer x = 0.183089
#> Warning: non-integer x = 0.913834
#> Warning: non-integer x = 2.820290
#> Warning: non-integer x = 3.534297
#> Warning: non-integer x = 0.554842
#> Warning: non-integer x = 0.674563
#> Warning: non-integer x = 2.652939
#> Warning: non-integer x = 2.336628
#> Warning: non-integer x = 3.076617
#> Warning: non-integer x = 0.638637
#> Warning: non-integer x = 0.843201
#> Warning: non-integer x = 1.278915
#> Warning: non-integer x = 3.927151
#> Warning: non-integer x = 3.835507
#> Warning: non-integer x = 1.983245
#> Warning: non-integer x = 3.555659
#> Warning: non-integer x = 0.777034
#> Warning: non-integer x = 0.316611
#> Warning: non-integer x = 1.973456
#> Warning: non-integer x = 1.670443
#> Warning: non-integer x = 2.541588
#> Warning: non-integer x = 0.745027
#> Warning: non-integer x = 0.034385
#> Warning: non-integer x = 1.560076
#> Warning: non-integer x = 0.542473
#> Warning: non-integer x = 0.335017
#> Warning: non-integer x = 2.361190
#> Warning: non-integer x = 0.474786
#> Warning: non-integer x = 0.149474
#> Warning: non-integer x = 1.902958
#> Warning: non-integer x = 1.687462
#> Warning: non-integer x = 1.360814
#> Warning: non-integer x = 3.085990
#> Warning: non-integer x = 2.207734
#> Warning: non-integer x = 0.520627
#> Warning: non-integer x = 1.318946
#> Warning: non-integer x = 1.823914
#> Warning: non-integer x = 2.729625
#> Warning: non-integer x = 1.340309
#> Warning: non-integer x = 1.126298
#> Warning: non-integer x = 2.266419
#> Warning: non-integer x = 0.193238
#> Warning: non-integer x = 1.597545
#> Warning: non-integer x = 3.062701
#> Warning: non-integer x = 2.594543
#> Warning: non-integer x = 0.218106
#> Warning: non-integer x = 3.733062
#> Warning: non-integer x = 3.352289
#> Warning: non-integer x = 1.417046
#> Warning: non-integer x = 2.024484
#> Warning: non-integer x = 0.150710
#> Warning: non-integer x = 0.559834
#> Warning: non-integer x = 0.765940
#> Warning: non-integer x = 2.659229
#> Warning: non-integer x = 0.174358
#> Warning: non-integer x = 2.052876
#> Warning: non-integer x = 2.020731
#> Warning: non-integer x = 3.883953
#> Warning: non-integer x = 1.213211
#> Warning: non-integer x = 0.322274
#> Warning: non-integer x = 4.018216
#> Warning: non-integer x = 1.786435
#> Warning: non-integer x = 0.699775
#> Warning: non-integer x = 1.648058
#> Warning: non-integer x = 1.511190
#> Warning: non-integer x = 1.847233
#> Warning: non-integer x = 1.808680
#> Warning: non-integer x = 0.562790
#> Warning: non-integer x = 2.908143
#> Warning: non-integer x = 2.105060
#> Warning: non-integer x = 3.430335
#> Warning: non-integer x = 0.352059
#> Warning: non-integer x = 2.361010
#> Warning: non-integer x = 1.460500
#> Warning: non-integer x = 3.089511
#> Warning: non-integer x = 1.829122
#> Warning: non-integer x = 1.862490
#> Warning: non-integer x = 1.648370
#> Warning: non-integer x = 1.091796
#> Warning: non-integer x = 2.342979
#> Warning: non-integer x = 2.371741
#> Warning: non-integer x = 1.465225
#> Warning: non-integer x = 1.192605
#> Warning: non-integer x = 1.082156
#> Warning: non-integer x = 3.128071
#> Warning: non-integer x = 0.013548
#> Warning: non-integer x = 1.700527
#> Warning: non-integer x = 0.335070
#> Warning: non-integer x = 3.293491
#> Warning: non-integer x = 3.184536
#> Warning: non-integer x = 2.458887
#> Warning: non-integer x = 1.371116
#> Warning: non-integer x = 0.322702
#> Warning: non-integer x = 0.185020
#> Warning: non-integer x = 1.193954
#> Warning: non-integer x = 3.257502
#> Warning: non-integer x = 1.084550
#> Warning: non-integer x = 1.025529
#> Warning: non-integer x = 1.302506
#> Warning: non-integer x = 0.816067
#> Warning: non-integer x = 4.859060
#> Warning: non-integer x = 2.079181
#> Warning: non-integer x = 2.706149
#> Warning: non-integer x = 4.435999
#> Warning: non-integer x = 1.436084
#> Warning: non-integer x = 2.070155
#> Warning: non-integer x = 1.121332
#> Warning: non-integer x = 2.316864
#> Warning: non-integer x = 3.224818
#> Warning: non-integer x = 1.828132
#> Warning: non-integer x = 0.457708
#> Warning: non-integer x = 2.283273
#> Warning: non-integer x = 1.067907
#> Warning: non-integer x = 1.374317
#> Warning: non-integer x = 0.954771
#> Warning: non-integer x = 0.615520
#> Warning: non-integer x = 3.025006
#> Warning: non-integer x = 0.016269
#> Warning: non-integer x = 2.751511
#> Warning: non-integer x = 0.145061
#> Warning: non-integer x = 3.429307
#> Warning: non-integer x = 0.660833
#> Warning: non-integer x = 1.031497
#> Warning: non-integer x = 0.196866
#> Warning: non-integer x = 0.749406
#> Warning: non-integer x = 1.745168
#> Warning: non-integer x = 2.088754
#> Warning: non-integer x = 1.722572
#> Warning: non-integer x = 2.016559
#> Warning: non-integer x = 2.054746
#> Warning: non-integer x = 1.715591
#> Warning: non-integer x = 2.607147
#> Warning: non-integer x = 0.389083
#> Warning: non-integer x = 1.284691
#> Warning: non-integer x = 0.379871
#> Warning: non-integer x = 1.176023
#> Warning: non-integer x = 0.809431
#> Warning: non-integer x = 1.092861
#> Warning: non-integer x = 2.013352
#> Warning: non-integer x = 1.324931
#> Warning: non-integer x = 0.213574
#> Warning: non-integer x = 1.381900
#> Warning: non-integer x = 2.502924
#> Warning: non-integer x = 1.474801
#> Warning: non-integer x = 3.168020
#> Warning: non-integer x = 1.421413
#> Warning: non-integer x = 0.190859
#> Warning: non-integer x = 1.219245
#> Warning: non-integer x = 2.332766
#> Warning: non-integer x = 0.687017
#> Warning: non-integer x = 0.538733
#> Warning: non-integer x = 0.543798
#> Warning: non-integer x = 0.402882
#> Warning: non-integer x = 1.555018
#> Warning: non-integer x = 0.614344
#> Warning: non-integer x = 0.764379
#> Warning: non-integer x = 3.858211
#> Warning: non-integer x = 0.137957
#> Warning: non-integer x = 2.972301
#> Warning: non-integer x = 4.935363
#> Warning: non-integer x = 4.453008
#> Warning: non-integer x = 0.461420
#> Warning: non-integer x = 1.807266
#> Warning: non-integer x = 0.315498
#> Warning: non-integer x = 1.259955
#> Warning: non-integer x = 1.592242
#> Warning: non-integer x = 0.493670
#> Warning: non-integer x = 2.700188
#> Warning: non-integer x = 0.719372
#> Warning: non-integer x = 1.358004
#> Warning: non-integer x = 2.597198
#> Warning: non-integer x = 1.279211
#> Warning: non-integer x = 2.088745
#> Warning: non-integer x = 1.593095
#> Warning: non-integer x = 1.130645
#> Warning: non-integer x = 0.757529
#> Warning: non-integer x = 3.677390
#> Warning: non-integer x = 0.530438
#> Warning: non-integer x = 2.110319
#> Warning: non-integer x = 2.083862
#> Warning: non-integer x = 1.128803
#> Warning: non-integer x = 4.672121
#> Warning: non-integer x = 0.513766
#> Warning: non-integer x = 3.339449
#> Warning: non-integer x = 0.951182
#> Warning: non-integer x = 2.951415
#> Warning: non-integer x = 0.025557
#> Warning: non-integer x = 1.930739
#> Warning: non-integer x = 1.668343
#> Warning: non-integer x = 0.468696
#> Warning: non-integer x = 0.746230
#> Warning: non-integer x = 4.113219
#> Warning: non-integer x = 1.177539
#> Warning: non-integer x = 0.718770
#> Warning: non-integer x = 1.310927
#> Warning: non-integer x = 1.879159
#> Warning: non-integer x = 3.716983
#> Warning: non-integer x = 3.336029
#> Warning: non-integer x = 1.583876
#> Warning: non-integer x = 1.800928
#> Warning: non-integer x = 0.282430
#> Warning: non-integer x = 0.042771
#> Warning: non-integer x = 2.084823
#> Warning: non-integer x = 1.162401
#> Warning: non-integer x = 1.962786
#> Warning: non-integer x = 0.915187
#> Warning: non-integer x = 1.846446
#> Warning: non-integer x = 2.486013
#> Warning: non-integer x = 0.572713
#> Warning: non-integer x = 2.953910
#> Warning: non-integer x = 1.398357
#> Warning: non-integer x = 1.605247
#> Warning: non-integer x = 1.167348
#> Warning: non-integer x = 0.362298
#> Warning: non-integer x = 2.075489
#> Warning: non-integer x = 1.422941
#> Warning: non-integer x = 2.980898
#> Warning: non-integer x = 0.297920
#> Warning: non-integer x = 1.836923
#> Warning: non-integer x = 1.871164
#> Warning: non-integer x = 3.072513
#> Warning: non-integer x = 0.335511
#> Warning: non-integer x = 1.767985
#> Warning: non-integer x = 1.146560
#> Warning: non-integer x = 1.819590
#> Warning: non-integer x = 3.112656
#> Warning: non-integer x = 1.179667
#> Warning: non-integer x = 2.213951
#> Warning: non-integer x = 3.379497
#> Warning: non-integer x = 0.193272
#> Warning: non-integer x = 2.815508
#> Warning: non-integer x = 2.525285
#> Warning: non-integer x = 1.547222
#> Warning: non-integer x = 3.204603
#> Warning: non-integer x = 0.129122
#> Warning: non-integer x = 0.908435
#> Warning: non-integer x = 1.409245
#> Warning: non-integer x = 3.389339
#> Warning: non-integer x = 1.443344
#> Warning: non-integer x = 2.455130
#> Warning: non-integer x = 1.619623
#> Warning: non-integer x = 0.703164
#> Warning: non-integer x = 1.270440
#> Warning: non-integer x = 2.412726
#> Warning: non-integer x = 2.031445
#> Warning: non-integer x = 1.112961
#> Warning: non-integer x = 0.690221
#> Warning: non-integer x = 2.552590
#> Warning: non-integer x = 1.275326
#> Warning: non-integer x = 0.643690
#> Warning: non-integer x = 2.018724
#> Warning: non-integer x = 3.456108
#> Warning: non-integer x = 3.497720
#> Warning: non-integer x = 2.257560
#> Warning: non-integer x = 1.232301
#> Warning: non-integer x = 2.863187
#> Warning: non-integer x = 3.351531
#> Warning: non-integer x = 0.132967
#> Warning: non-integer x = 0.281124
#> Warning: non-integer x = 0.324480
#> Warning: non-integer x = 1.019051
#> Warning: non-integer x = 1.791886
#> Warning: non-integer x = 3.438568
#> Warning: non-integer x = 1.054107
#> Warning: non-integer x = 2.850501
#> Warning: non-integer x = 3.149954
#> Warning: non-integer x = 1.057344
#> Warning: non-integer x = 1.130176
#> Warning: non-integer x = 4.407844
#> Warning: non-integer x = 1.619460
#> Warning: non-integer x = 0.616005
#> Warning: non-integer x = 2.565091
#> Warning: non-integer x = 0.329835
#> Warning: non-integer x = 2.383632
#> Warning: non-integer x = 0.486229
#> Warning: non-integer x = 0.386903
#> Warning: non-integer x = 0.910345
#> Warning: non-integer x = 3.635087
#> Warning: non-integer x = 0.284453
#> Warning: non-integer x = 0.959916
#> Warning: non-integer x = 1.832407
#> Warning: non-integer x = 0.467305
#> Warning: non-integer x = 2.219001
#> Warning: non-integer x = 0.042582
#> Warning: non-integer x = 2.425312
#> Warning: non-integer x = 0.290860
#> Warning: non-integer x = 2.570245
#> Warning: non-integer x = 3.042200
#> Warning: non-integer x = 0.424334
#> Warning: non-integer x = 1.349049
#> Warning: non-integer x = 0.622900
#> Warning: non-integer x = 1.437024
#> Warning: non-integer x = 1.643583
#> Warning: non-integer x = 1.174816
#> Warning: non-integer x = 2.801130
#> Warning: non-integer x = 3.284552
#> Warning: non-integer x = 1.102687
#> Warning: non-integer x = 0.485978
#> Warning: non-integer x = 0.311075
#> Warning: non-integer x = 2.267685
#> Warning: non-integer x = 0.408508
#> Warning: non-integer x = 0.281320
#> Warning: non-integer x = 0.246261
#> Warning: non-integer x = 2.149389
#> Warning: non-integer x = 0.802298
#> Warning: non-integer x = 1.208692
#> Warning: non-integer x = 2.097582
#> Warning: non-integer x = 2.406341
#> Warning: non-integer x = 0.040181
#> Warning: non-integer x = 0.576019
#> Warning: non-integer x = 1.703515
#> Warning: non-integer x = 0.068686
#> Warning: non-integer x = 0.291651
#> Warning: non-integer x = 0.219418
#> Warning: non-integer x = 0.233892
#> Warning: non-integer x = 0.878685
#> Warning: non-integer x = 1.795621
#> Warning: non-integer x = 1.898298
#> Warning: non-integer x = 0.762000
#> Warning: non-integer x = 0.065351
#> Warning: non-integer x = 1.759057
#> Warning: non-integer x = 2.324489
#> Warning: non-integer x = 2.642625
#> Warning: non-integer x = 0.300467
#> Warning: non-integer x = 2.479089
#> Warning: non-integer x = 3.905472
#> Warning: non-integer x = 2.836118
#> Warning: non-integer x = 2.188008
#> Warning: non-integer x = 2.100501
#> Warning: non-integer x = 1.757274
#> Warning: non-integer x = 1.562895
#> Warning: non-integer x = 0.742178
#> Warning: non-integer x = 2.725244
#> Warning: non-integer x = 3.313010
#> Warning: non-integer x = 1.082297
#> Warning: non-integer x = 0.790651
#> Warning: non-integer x = 2.090867
#> Warning: non-integer x = 1.350405
#> Warning: non-integer x = 1.034454
#> Warning: non-integer x = 1.371006
#> Warning: non-integer x = 0.880131
#> Warning: non-integer x = 2.600839
#> Warning: non-integer x = 3.444393
#> Warning: non-integer x = 2.402353
#> Warning: non-integer x = 2.845998
#> Warning: non-integer x = 0.155057
#> Warning: non-integer x = 1.320439
#> Warning: non-integer x = 0.001893
#> Warning: non-integer x = 2.025030
#> Warning: non-integer x = 0.151057
#> Warning: non-integer x = 1.717117
#> Warning: non-integer x = 1.680895
#> Warning: non-integer x = 1.558937
#> Warning: non-integer x = 2.857245
#> Warning: non-integer x = 1.431752
#> Warning: non-integer x = 0.592822
#> Warning: non-integer x = 1.052197
#> Warning: non-integer x = 0.981918
#> Warning: non-integer x = 1.412707
#> Warning: non-integer x = 1.664752
#> Warning: non-integer x = 1.001940
#> Warning: non-integer x = 0.283541
#> Warning: non-integer x = 0.453699
#> Warning: non-integer x = 3.038802
#> Warning: non-integer x = 0.264518
#> Warning: non-integer x = 0.123533
#> Warning: non-integer x = 0.141169
#> Warning: non-integer x = 0.295950
#> Warning: non-integer x = 2.316683
#> Warning: non-integer x = 2.898998
#> Warning: non-integer x = 1.775880
#> Warning: non-integer x = 1.977040
#> Warning: non-integer x = 0.085718
#> Warning: non-integer x = 2.977054
#> Warning: non-integer x = 3.296945
#> Warning: non-integer x = 0.684995
#> Warning: non-integer x = 0.080803
#> Warning: non-integer x = 0.297810
#> Warning: non-integer x = 0.808894
#> Warning: non-integer x = 1.051056
#> Warning: non-integer x = 0.855416
#> Warning: non-integer x = 2.975965
#> Warning: non-integer x = 2.438856
#> Warning: non-integer x = 0.369893
#> Warning: non-integer x = 0.977837
#> Warning: non-integer x = 2.107984
#> Warning: non-integer x = 0.489449
#> Warning: non-integer x = 0.044967
#> Warning: non-integer x = 1.127640
#> Warning: non-integer x = 1.518171
#> Warning: non-integer x = 1.729531
#> Warning: non-integer x = 2.044084
#> Warning: non-integer x = 0.982614
#> Warning: non-integer x = 0.627398
#> Warning: non-integer x = 1.155012
#> Warning: non-integer x = 2.382229
#> Warning: non-integer x = 1.274674
#> Warning: non-integer x = 0.473537
#> Warning: non-integer x = 2.319153
#> Warning: non-integer x = 0.228578
#> Warning: non-integer x = 1.624466
#> Warning: non-integer x = 1.441352
#> Warning: non-integer x = 1.925183
#> Warning: non-integer x = 0.546600
#> Warning: non-integer x = 0.943901
#> Warning: non-integer x = 1.838153
#> Warning: non-integer x = 2.444314
#> Warning: non-integer x = 0.738212
#> Warning: non-integer x = 1.852339
#> Warning: non-integer x = 0.859953
#> Warning: non-integer x = 0.490704
#> Warning: non-integer x = 2.198652
#> Warning: non-integer x = 0.708738
#> Warning: non-integer x = 1.022798
#> Warning: non-integer x = 0.042249
#> Warning: non-integer x = 1.557723
#> Warning: non-integer x = 1.514807
#> Warning: non-integer x = 2.989839
#> Warning: non-integer x = 2.094868
#> Warning: non-integer x = 2.429432
#> Warning: non-integer x = 0.785080
#> Warning: non-integer x = 0.108071
#> Warning: non-integer x = 1.022106
#> Warning: non-integer x = 0.489924
#> Warning: non-integer x = 2.237583
#> Warning: non-integer x = 3.062932
#> Warning: non-integer x = 0.092810
#> Warning: non-integer x = 0.344742
#> Warning: non-integer x = 1.885379
#> Warning: non-integer x = 0.487263
#> Warning: non-integer x = 0.000444
#> Warning: non-integer x = 0.797000
#> Warning: non-integer x = 1.506096
#> Warning: non-integer x = 3.022515
#> Warning: non-integer x = 2.536841
#> Warning: non-integer x = 3.835260
#> Warning: non-integer x = 3.179182
#> Warning: non-integer x = 1.614452
#> Warning: non-integer x = 0.686594
#> Warning: non-integer x = 2.223676
#> Warning: non-integer x = 0.496803
#> Warning: non-integer x = 1.470832
#> Warning: non-integer x = 0.781206
#> Warning: non-integer x = 0.642223
#> Warning: non-integer x = 1.119458
#> Warning: non-integer x = 0.124709
#> Warning: non-integer x = 2.680964
#> Warning: non-integer x = 1.436911
#> Warning: non-integer x = 1.184427
#> Warning: non-integer x = 0.749991
#> Warning: non-integer x = 0.717586
#> Warning: non-integer x = 0.318968
#> Warning: non-integer x = 0.870492
#> Warning: non-integer x = 2.751772
#> Warning: non-integer x = 1.034334
#> Warning: non-integer x = 0.162770
#> Warning: non-integer x = 0.461192
#> Warning: non-integer x = 2.524137
#> Warning: non-integer x = 2.091604
#> Warning: non-integer x = 0.561036
#> Warning: non-integer x = 1.763800
#> Warning: non-integer x = 1.853742
#> Warning: non-integer x = 2.771797
#> Warning: non-integer x = 0.577242
#> Warning: non-integer x = 1.608877
#> Warning: non-integer x = 0.180071
#> Warning: non-integer x = 0.851889
#> Warning: non-integer x = 2.420869
#> Warning: non-integer x = 2.748042
#> Warning: non-integer x = 1.272463
#> Warning: non-integer x = 0.931753
#> Warning: non-integer x = 1.108947
#> Warning: non-integer x = 0.906952
#> Warning: non-integer x = 0.690747
#> Warning: non-integer x = 2.037989
#> Warning: non-integer x = 0.764349
#> Warning: non-integer x = 1.472625
#> Warning: non-integer x = 3.295740
#> Warning: non-integer x = 0.908996
#> Warning: non-integer x = 1.541637
#> Warning: non-integer x = 0.970055
#> Warning: non-integer x = 1.880875
#> Warning: non-integer x = 2.972984
#> Warning: non-integer x = 2.070526
#> Warning: non-integer x = 0.414774
#> Warning: non-integer x = 1.975331
#> Warning: non-integer x = 0.839196
#> Warning: non-integer x = 3.265683
#> Warning: non-integer x = 0.224023
#> Warning: non-integer x = 0.815901
#> Warning: non-integer x = 0.181993
#> Warning: non-integer x = 0.196363
#> Warning: non-integer x = 0.026572
#> Warning: non-integer x = 1.432101
#> Warning: non-integer x = 0.171537
#> Warning: non-integer x = 3.228549
#> Warning: non-integer x = 3.659902
#> Warning: non-integer x = 0.173251
#> Warning: non-integer x = 0.697608
#> Warning: non-integer x = 0.699016
#> Warning: non-integer x = 0.097598
#> Warning: non-integer x = 1.831567
#> Warning: non-integer x = 0.326722
#> Warning: non-integer x = 0.793975
#> Warning: non-integer x = 2.441953
#> Warning: non-integer x = 3.278417
#> Warning: non-integer x = 1.065901
#> Warning: non-integer x = 0.600573
#> Warning: non-integer x = 0.158666
#> Warning: non-integer x = 2.308609
#> Warning: non-integer x = 0.611721
#> Warning: non-integer x = 0.285412
#> Warning: non-integer x = 0.198854
#> Warning: non-integer x = 0.318844
#> Warning: non-integer x = 2.927519
#> Warning: non-integer x = 1.603399
#> Warning: non-integer x = 1.030490
#> Warning: non-integer x = 2.390438
#> Warning: non-integer x = 3.845859
#> Warning: non-integer x = 1.736911
#> Warning: non-integer x = 2.439856
#> Warning: non-integer x = 0.540180
#> Warning: non-integer x = 0.583227
#> Warning: non-integer x = 1.632596
#> Warning: non-integer x = 0.533096
#> Warning: non-integer x = 0.328692
#> Warning: non-integer x = 0.206176
#> Warning: non-integer x = 0.172047
#> Warning: non-integer x = 1.775635
#> Warning: non-integer x = 0.930498
#> Warning: non-integer x = 0.936858
#> Warning: non-integer x = 0.736175
#> Warning: non-integer x = 0.701531
#> Warning: non-integer x = 0.222868
#> Warning: non-integer x = 2.835169
#> Warning: non-integer x = 3.180041
#> Warning: non-integer x = 1.509924
#> Warning: non-integer x = 2.489877
#> Warning: non-integer x = 0.765188
#> Warning: non-integer x = 0.151168
#> Warning: non-integer x = 2.933566
#> Warning: non-integer x = 2.997885
#> Warning: non-integer x = 1.974540
#> Warning: non-integer x = 1.628270
#> Warning: non-integer x = 2.155003
#> Warning: non-integer x = 0.041034
#> Warning: non-integer x = 2.837989
#> Warning: non-integer x = 0.451607
#> Warning: non-integer x = 1.399865
#> Warning: non-integer x = 1.124025
#> Warning: non-integer x = 1.321088
#> Warning: non-integer x = 3.230987
#> Warning: non-integer x = 3.271648
#> Warning: non-integer x = 0.674391
#> Warning: non-integer x = 2.962253
#> Warning: non-integer x = 2.309283
#> Warning: non-integer x = 0.611602
#> Warning: non-integer x = 2.329262
#> Warning: non-integer x = 2.208016
#> Warning: non-integer x = 0.603542
#> Warning: non-integer x = 1.332210
#> Warning: non-integer x = 1.396350
#> Warning: non-integer x = 0.669222
#> Warning: non-integer x = 0.559407
#> Warning: non-integer x = 0.517787
#> Warning: non-integer x = 0.085585
#> Warning: non-integer x = 0.879113
#> Warning: non-integer x = 0.797876
#> Warning: non-integer x = 4.385649
#> Warning: non-integer x = 1.146167
#> Warning: non-integer x = 2.311017
#> Warning: non-integer x = 0.275069
#> Warning: non-integer x = 1.307194
#> Warning: non-integer x = 0.804965
#> Warning: non-integer x = 1.914517
#> Warning: non-integer x = 3.668045
#> Warning: non-integer x = 1.702570
#> Warning: non-integer x = 2.005429
#> Warning: non-integer x = 3.084367
#> Warning: non-integer x = 0.406963
#> Warning: non-integer x = 4.659923
#> Warning: non-integer x = 0.931214
#> Warning: non-integer x = 2.165503
#> Warning: non-integer x = 1.198809
#> Warning: non-integer x = 0.221362
#> Warning: non-integer x = 1.339280
#> Warning: non-integer x = 0.422327
#> Warning: non-integer x = 1.235218
#> Warning: non-integer x = 2.449770
#> Warning: non-integer x = 0.532331
#> Warning: non-integer x = 2.945007
#> Warning: non-integer x = 0.477235
#> Warning: non-integer x = 1.057805
#> Warning: non-integer x = 2.628849
#> Warning: non-integer x = 0.157113
#> Warning: non-integer x = 4.634477
#> Warning: non-integer x = 0.081118
#> Warning: non-integer x = 1.163951
#> Warning: non-integer x = 1.415516
#> Warning: non-integer x = 1.531555
#> Warning: non-integer x = 0.547205
#> Warning: non-integer x = 0.775174
#> Warning: non-integer x = 0.718060
#> Warning: non-integer x = 0.865082
#> Warning: non-integer x = 1.769067
#> Warning: non-integer x = 1.541773
#> Warning: non-integer x = 1.337124
#> Warning: non-integer x = 2.019864
#> Warning: non-integer x = 0.306951
#> Warning: non-integer x = 3.276195
#> Warning: non-integer x = 0.155819
#> Warning: non-integer x = 1.295687
#> Warning: non-integer x = 1.393933
#> Warning: non-integer x = 1.549776
#> Warning: non-integer x = 0.745001
#> Warning: non-integer x = 4.283372
#> Warning: non-integer x = 0.952641
#> Warning: non-integer x = 0.504833
#> Warning: non-integer x = 0.007553
#> Warning: non-integer x = 0.805943
#> Warning: non-integer x = 0.816985
#> Warning: non-integer x = 0.120369
#> Warning: non-integer x = 1.705297
#> Warning: non-integer x = 4.364361
#> Warning: non-integer x = 2.636773
#> Warning: non-integer x = 0.108169
#> Warning: non-integer x = 1.598025
#> Warning: non-integer x = 0.905936
#> Warning: non-integer x = 1.322457
#> Warning: non-integer x = 0.380121
#> Warning: non-integer x = 1.734312
#> Warning: non-integer x = 2.562275
#> Warning: non-integer x = 0.513992
#> Warning: non-integer x = 1.375927
#> Warning: non-integer x = 0.510691
#> Warning: non-integer x = 2.874219
#> Warning: non-integer x = 2.889267
#> Warning: non-integer x = 2.628719
#> Warning: non-integer x = 0.007089
#> Warning: non-integer x = 1.161751
#> Warning: non-integer x = 2.643100
#> Warning: non-integer x = 1.610300
#> Warning: non-integer x = 3.633184
#> Warning: non-integer x = 2.314730
#> Warning: non-integer x = 1.158742
#> Warning: non-integer x = 1.739085
#> Warning: non-integer x = 3.171998
#> Warning: non-integer x = 2.212527
#> Warning: non-integer x = 0.471506
#> Warning: non-integer x = 2.672965
#> Warning: non-integer x = 1.582493
#> Warning: non-integer x = 0.622182
#> Warning: non-integer x = 0.704296
#> Warning: non-integer x = 0.731997
#> Warning: non-integer x = 1.260750
#> Warning: non-integer x = 1.051965
#> Warning: non-integer x = 2.749164
#> Warning: non-integer x = 0.354530
#> Warning: non-integer x = 0.668599
#> Warning: non-integer x = 1.279069
#> Warning: non-integer x = 1.799099
#> Warning: non-integer x = 1.897489
#> Warning: non-integer x = 2.367138
#> Warning: non-integer x = 1.081728
#> Warning: non-integer x = 1.907051
#> Warning: non-integer x = 2.546654
#> Warning: non-integer x = 0.160304
#> Warning: non-integer x = 0.212602
#> Warning: non-integer x = 1.898704
#> Warning: non-integer x = 1.800428
#> Warning: non-integer x = 0.428432
#> Warning: non-integer x = 0.397133
#> Warning: non-integer x = 0.536876
#> Warning: non-integer x = 1.126732
#> Warning: non-integer x = 1.535016
#> Warning: non-integer x = 1.376154
#> Warning: non-integer x = 0.829387
#> Warning: non-integer x = 1.074022
#> Warning: non-integer x = 2.658557
#> Warning: non-integer x = 0.306580
#> Warning: non-integer x = 0.717957
#> Warning: non-integer x = 1.528481
#> Warning: non-integer x = 1.301953
#> Warning: non-integer x = 1.837221
#> Warning: non-integer x = 1.382397
#> Warning: non-integer x = 1.173760
#> Warning: non-integer x = 0.697304
#> Warning: non-integer x = 0.200145
#> Warning: non-integer x = 0.364499
#> Warning: non-integer x = 1.537058
#> Warning: non-integer x = 1.266625
#> Warning: non-integer x = 1.994348
#> Warning: non-integer x = 2.086790
#> Warning: non-integer x = 1.588641
#> Warning: non-integer x = 3.346574
#> Warning: non-integer x = 1.163143
#> Warning: non-integer x = 0.457142
#> Warning: non-integer x = 1.529548
#> Warning: non-integer x = 0.882865
#> Warning: non-integer x = 1.829669
#> Warning: non-integer x = 1.646647
#> Warning: non-integer x = 3.031428
#> Warning: non-integer x = 2.797670
#> Warning: non-integer x = 1.117692
#> Warning: non-integer x = 0.400194
#> Warning: non-integer x = 2.363595
#> Warning: non-integer x = 2.425656
#> Warning: non-integer x = 0.828173
#> Warning: non-integer x = 1.438368
#> Warning: non-integer x = 0.394729
#> Warning: non-integer x = 3.460171
#> Warning: non-integer x = 1.495356
#> Warning: non-integer x = 2.489380
#> Warning: non-integer x = 1.453390
#> Warning: non-integer x = 0.452715
#> Warning: non-integer x = 2.148826
#> Warning: non-integer x = 1.601037
#> Warning: non-integer x = 1.505209
#> Warning: non-integer x = 1.361998
#> Warning: non-integer x = 0.471994
#> Warning: non-integer x = 1.413326
#> Warning: non-integer x = 1.974681
#> Warning: non-integer x = 2.182181
#> Warning: non-integer x = 2.701590
#> Warning: non-integer x = 0.894756
#> Warning: non-integer x = 0.216206
#> Warning: non-integer x = 0.149984
#> Warning: non-integer x = 1.919069
#> Warning: non-integer x = 1.467452
#> Warning: non-integer x = 0.417027
#> Warning: non-integer x = 1.607096
#> Warning: non-integer x = 1.863795
#> Warning: non-integer x = 1.417293
#> Warning: non-integer x = 1.690940
#> Warning: non-integer x = 0.323408
#> Warning: non-integer x = 0.738383
#> Warning: non-integer x = 0.420241
#> Warning: non-integer x = 2.029016
#> Warning: non-integer x = 0.357895
#> Warning: non-integer x = 0.989086
#> Warning: non-integer x = 1.028627
#> Warning: non-integer x = 0.113911
#> Warning: non-integer x = 0.392217
#> Warning: non-integer x = 1.540240
#> Warning: non-integer x = 1.316552
#> Warning: non-integer x = 1.249818
#> Warning: non-integer x = 0.925233
#> Warning: non-integer x = 0.608096
#> Warning: non-integer x = 1.250267
#> Warning: non-integer x = 1.495776
#> Warning: non-integer x = 1.663851
#> Warning: non-integer x = 1.367946
#> Warning: non-integer x = 1.910977
#> Warning: non-integer x = 3.982046
#> Warning: non-integer x = 2.201492
#> Warning: non-integer x = 2.167125
#> Warning: non-integer x = 0.079149
#> Warning: non-integer x = 1.696822
#> Warning: non-integer x = 0.564183
#> Warning: non-integer x = 2.777558
#> Warning: non-integer x = 3.210070
#> Warning: non-integer x = 2.360042
#> Warning: non-integer x = 1.025890
#> Warning: non-integer x = 2.516380
#> Warning: non-integer x = 1.589615
#> Warning: non-integer x = 2.164415
#> Warning: non-integer x = 1.330722
#> Warning: non-integer x = 2.995358
#> Warning: non-integer x = 1.786859
#> Warning: non-integer x = 3.582962
#> Warning: non-integer x = 0.664735
#> Warning: non-integer x = 0.652669
#> Warning: non-integer x = 0.275776
#> Warning: non-integer x = 2.153036
#> Warning: non-integer x = 1.822523
#> Warning: non-integer x = 1.546823
#> Warning: non-integer x = 1.881873
#> Warning: non-integer x = 1.341860
#> Warning: non-integer x = 0.216397
#> Warning: non-integer x = 0.736956
#> Warning: non-integer x = 3.017033
#> Warning: non-integer x = 0.695233
#> Warning: non-integer x = 1.244955
#> Warning: non-integer x = 2.478839
#> Warning: non-integer x = 1.342009
#> Warning: non-integer x = 1.901848
#> Warning: non-integer x = 1.588296
#> Warning: non-integer x = 0.660130
#> Warning: non-integer x = 1.513840
#> Warning: non-integer x = 0.788583
#> Warning: non-integer x = 3.055420
#> Warning: non-integer x = 0.003765
#> Warning: non-integer x = 0.413193
#> Warning: non-integer x = 2.080798
#> Warning: non-integer x = 0.269783
#> Warning: non-integer x = 0.672224
#> Warning: non-integer x = 0.782571
#> Warning: non-integer x = 1.268077
#> Warning: non-integer x = 2.025955
#> Warning: non-integer x = 1.090809
#> Warning: non-integer x = 2.246498
#> Warning: non-integer x = 2.177385
#> Warning: non-integer x = 3.150664
#> Warning: non-integer x = 1.167336
#> Warning: non-integer x = 0.445551
#> Warning: non-integer x = 1.024321
#> Warning: non-integer x = 3.512333
#> Warning: non-integer x = 0.098465
#> Warning: non-integer x = 1.048796
#> Warning: non-integer x = 3.130498
#> Warning: non-integer x = 0.957127
#> Warning: non-integer x = 0.733306
#> Warning: non-integer x = 0.676773
#> Warning: non-integer x = 1.039675
#> Warning: non-integer x = 0.313741
#> Warning: non-integer x = 0.262528
#> Warning: non-integer x = 0.732172
#> Warning: non-integer x = 2.387966
#> Warning: non-integer x = 2.284098
#> Warning: non-integer x = 3.353307
#> Warning: non-integer x = 0.065123
#> Warning: non-integer x = 2.162335
#> Warning: non-integer x = 1.290106
#> Warning: non-integer x = 2.180195
#> Warning: non-integer x = 0.130425
#> Warning: non-integer x = 1.196264
#> Warning: non-integer x = 1.305597
#> Warning: non-integer x = 1.149688
#> Warning: non-integer x = 2.565947
#> Warning: non-integer x = 0.737023
#> Warning: non-integer x = 1.675451
#> Warning: non-integer x = 1.867108
#> Warning: non-integer x = 0.400270
#> Warning: non-integer x = 1.351672
#> Warning: non-integer x = 1.185843
#> Warning: non-integer x = 1.982900
#> Warning: non-integer x = 1.423276
#> Warning: non-integer x = 1.870036
#> Warning: non-integer x = 3.549977
#> Warning: non-integer x = 1.831029
#> Warning: non-integer x = 1.953085
#> Warning: non-integer x = 1.574430
#> Warning: non-integer x = 0.519418
#> Warning: non-integer x = 0.732086
#> Warning: non-integer x = 0.356808
#> Warning: non-integer x = 0.046109
#> Warning: non-integer x = 1.339571
#> Warning: non-integer x = 3.250498
#> Warning: non-integer x = 1.742935
#> Warning: non-integer x = 1.395859
#> Warning: non-integer x = 1.449611
#> Warning: non-integer x = 2.337975
#> Warning: non-integer x = 1.404309
#> Warning: non-integer x = 0.236918
#> Warning: non-integer x = 0.660329
#> Warning: non-integer x = 1.432248
#> Warning: non-integer x = 0.190106
#> Warning: non-integer x = 2.074351
#> Warning: non-integer x = 0.291751
#> Warning: non-integer x = 3.197834
#> Warning: non-integer x = 0.075656
#> Warning: non-integer x = 2.727137
#> Warning: non-integer x = 1.144903
#> Warning: non-integer x = 2.521737
#> Warning: non-integer x = 0.883311
#> Warning: non-integer x = 1.952138
#> Warning: non-integer x = 1.991611
#> Warning: non-integer x = 1.121597
#> Warning: non-integer x = 1.228814
#> Warning: non-integer x = 0.621994
#> Warning: non-integer x = 0.217405
#> Warning: non-integer x = 0.653273
#> Warning: non-integer x = 0.722954
#> Warning: non-integer x = 1.222107
#> Warning: non-integer x = 0.296415
#> Warning: non-integer x = 1.911552
#> Warning: non-integer x = 0.437701
#> Warning: non-integer x = 0.091739
#> Warning: non-integer x = 1.956423
#> Warning: non-integer x = 0.992051
#> Warning: non-integer x = 3.274569
#> Warning: non-integer x = 0.930668
#> Warning: non-integer x = 0.620646
#> Warning: non-integer x = 0.995631
#> Warning: non-integer x = 1.202229
#> Warning: non-integer x = 0.808921
#> Warning: non-integer x = 1.569577
#> Warning: non-integer x = 0.135844
#> Warning: non-integer x = 0.725509
#> Warning: non-integer x = 1.020261
#> Warning: non-integer x = 0.633003
#> Warning: non-integer x = 2.111319
#> Warning: non-integer x = 0.397923
#> Warning: non-integer x = 0.448148
#> Warning: non-integer x = 0.172955
#> Warning: non-integer x = 1.890628
#> Warning: non-integer x = 1.039367
#> Warning: non-integer x = 3.105165
#> Warning: non-integer x = 0.261808
#> Warning: non-integer x = 1.295746
#> Warning: non-integer x = 2.266094
#> Warning: non-integer x = 0.938475
#> Warning: non-integer x = 0.126023
#> Warning: non-integer x = 3.482405