Estimate the Individual Treatment Effect using Poisson Regression

Estimates the Individual Treatment Effect using Poisson Regression given a response vector, a treatment vector, and a features matrix.

Usage

estimate_ite_poisson(y, z, X, X_names, include_offset, offset_name)

Arguments

y

the observed response vector

z

the treatment vector

X

the features matrix

X_names

the names of the covariates

include_offset

whether or not to include an offset when estimating the ITE, for poisson only

offset_name

the name of the offset, if it is to be included

Value

a vector of ITE estimates

Examples

dataset <- generate_cre_dataset(n = 1000, rho = 0, n_rules = 2, p = 10,
                                effect_size = 2, binary = FALSE)

# Initialize parameters
y <- abs(dataset[["y"]])
z <- dataset[["z"]]
X <- as.data.frame(dataset[["X"]])
X_names <- names(as.data.frame(X))
include_offset <- FALSE
offset_name <- NA

ite_list <- estimate_ite_poisson(y, z, X, X_names, include_offset, offset_name)
#> Warning: non-integer x = 2.173708
#> Warning: non-integer x = 2.800476
#> Warning: non-integer x = 1.080897
#> Warning: non-integer x = 1.933369
#> Warning: non-integer x = 0.395976
#> Warning: non-integer x = 0.097398
#> Warning: non-integer x = 0.916127
#> Warning: non-integer x = 1.122045
#> Warning: non-integer x = 3.247042
#> Warning: non-integer x = 2.658243
#> Warning: non-integer x = 2.416787
#> Warning: non-integer x = 1.330870
#> Warning: non-integer x = 1.597202
#> Warning: non-integer x = 1.057442
#> Warning: non-integer x = 0.438258
#> Warning: non-integer x = 0.702046
#> Warning: non-integer x = 1.196545
#> Warning: non-integer x = 0.002151
#> Warning: non-integer x = 2.361779
#> Warning: non-integer x = 0.757939
#> Warning: non-integer x = 1.202156
#> Warning: non-integer x = 2.455119
#> Warning: non-integer x = 1.425788
#> Warning: non-integer x = 3.110803
#> Warning: non-integer x = 0.239567
#> Warning: non-integer x = 2.068703
#> Warning: non-integer x = 2.383228
#> Warning: non-integer x = 2.539904
#> Warning: non-integer x = 1.308000
#> Warning: non-integer x = 1.200906
#> Warning: non-integer x = 2.552221
#> Warning: non-integer x = 3.684952
#> Warning: non-integer x = 0.345111
#> Warning: non-integer x = 0.771591
#> Warning: non-integer x = 0.895662
#> Warning: non-integer x = 1.631051
#> Warning: non-integer x = 0.230977
#> Warning: non-integer x = 1.843840
#> Warning: non-integer x = 1.742333
#> Warning: non-integer x = 0.052270
#> Warning: non-integer x = 3.090873
#> Warning: non-integer x = 3.738999
#> Warning: non-integer x = 1.147590
#> Warning: non-integer x = 1.863836
#> Warning: non-integer x = 2.221142
#> Warning: non-integer x = 1.745189
#> Warning: non-integer x = 0.718651
#> Warning: non-integer x = 0.726197
#> Warning: non-integer x = 0.681893
#> Warning: non-integer x = 2.336969
#> Warning: non-integer x = 1.951563
#> Warning: non-integer x = 1.766905
#> Warning: non-integer x = 2.183034
#> Warning: non-integer x = 0.048585
#> Warning: non-integer x = 0.655307
#> Warning: non-integer x = 1.462462
#> Warning: non-integer x = 1.794486
#> Warning: non-integer x = 1.021063
#> Warning: non-integer x = 1.241412
#> Warning: non-integer x = 0.694886
#> Warning: non-integer x = 3.544603
#> Warning: non-integer x = 0.035032
#> Warning: non-integer x = 3.362518
#> Warning: non-integer x = 2.313857
#> Warning: non-integer x = 4.146210
#> Warning: non-integer x = 2.633564
#> Warning: non-integer x = 2.610216
#> Warning: non-integer x = 0.771676
#> Warning: non-integer x = 1.967124
#> Warning: non-integer x = 1.378566
#> Warning: non-integer x = 4.531625
#> Warning: non-integer x = 0.712299
#> Warning: non-integer x = 2.496403
#> Warning: non-integer x = 1.260448
#> Warning: non-integer x = 0.568627
#> Warning: non-integer x = 2.265954
#> Warning: non-integer x = 0.778613
#> Warning: non-integer x = 0.064648
#> Warning: non-integer x = 0.862267
#> Warning: non-integer x = 2.984508
#> Warning: non-integer x = 0.264782
#> Warning: non-integer x = 0.611730
#> Warning: non-integer x = 0.195392
#> Warning: non-integer x = 1.201128
#> Warning: non-integer x = 0.067833
#> Warning: non-integer x = 2.038757
#> Warning: non-integer x = 0.773889
#> Warning: non-integer x = 0.548430
#> Warning: non-integer x = 2.041769
#> Warning: non-integer x = 0.393439
#> Warning: non-integer x = 2.895722
#> Warning: non-integer x = 0.039661
#> Warning: non-integer x = 2.016359
#> Warning: non-integer x = 0.906129
#> Warning: non-integer x = 1.063147
#> Warning: non-integer x = 1.502783
#> Warning: non-integer x = 2.733059
#> Warning: non-integer x = 2.545521
#> Warning: non-integer x = 1.338517
#> Warning: non-integer x = 1.840151
#> Warning: non-integer x = 1.837413
#> Warning: non-integer x = 1.511063
#> Warning: non-integer x = 2.269855
#> Warning: non-integer x = 2.190528
#> Warning: non-integer x = 2.536109
#> Warning: non-integer x = 0.707777
#> Warning: non-integer x = 2.195833
#> Warning: non-integer x = 0.250646
#> Warning: non-integer x = 1.151897
#> Warning: non-integer x = 0.051399
#> Warning: non-integer x = 1.305876
#> Warning: non-integer x = 2.966822
#> Warning: non-integer x = 2.341414
#> Warning: non-integer x = 0.636300
#> Warning: non-integer x = 2.098870
#> Warning: non-integer x = 0.813504
#> Warning: non-integer x = 0.882357
#> Warning: non-integer x = 4.171374
#> Warning: non-integer x = 1.168277
#> Warning: non-integer x = 1.084840
#> Warning: non-integer x = 2.270247
#> Warning: non-integer x = 1.792608
#> Warning: non-integer x = 3.172192
#> Warning: non-integer x = 0.649853
#> Warning: non-integer x = 3.390568
#> Warning: non-integer x = 4.173591
#> Warning: non-integer x = 0.457127
#> Warning: non-integer x = 0.845255
#> Warning: non-integer x = 2.682744
#> Warning: non-integer x = 1.256687
#> Warning: non-integer x = 0.081490
#> Warning: non-integer x = 3.764421
#> Warning: non-integer x = 3.496635
#> Warning: non-integer x = 1.820377
#> Warning: non-integer x = 0.746853
#> Warning: non-integer x = 0.584239
#> Warning: non-integer x = 2.376828
#> Warning: non-integer x = 2.349496
#> Warning: non-integer x = 2.211861
#> Warning: non-integer x = 0.423199
#> Warning: non-integer x = 1.153597
#> Warning: non-integer x = 2.245152
#> Warning: non-integer x = 4.106855
#> Warning: non-integer x = 1.783829
#> Warning: non-integer x = 0.738296
#> Warning: non-integer x = 3.258763
#> Warning: non-integer x = 1.066131
#> Warning: non-integer x = 0.004991
#> Warning: non-integer x = 2.297825
#> Warning: non-integer x = 5.467817
#> Warning: non-integer x = 0.653962
#> Warning: non-integer x = 3.401440
#> Warning: non-integer x = 0.157349
#> Warning: non-integer x = 1.335316
#> Warning: non-integer x = 1.981769
#> Warning: non-integer x = 0.442540
#> Warning: non-integer x = 0.186485
#> Warning: non-integer x = 2.189216
#> Warning: non-integer x = 1.954824
#> Warning: non-integer x = 0.853000
#> Warning: non-integer x = 1.161910
#> Warning: non-integer x = 1.537570
#> Warning: non-integer x = 3.143613
#> Warning: non-integer x = 4.112108
#> Warning: non-integer x = 1.916484
#> Warning: non-integer x = 2.342804
#> Warning: non-integer x = 1.664416
#> Warning: non-integer x = 2.519071
#> Warning: non-integer x = 0.484880
#> Warning: non-integer x = 3.004233
#> Warning: non-integer x = 3.104748
#> Warning: non-integer x = 1.455284
#> Warning: non-integer x = 1.587799
#> Warning: non-integer x = 0.799776
#> Warning: non-integer x = 2.662177
#> Warning: non-integer x = 0.631525
#> Warning: non-integer x = 0.922945
#> Warning: non-integer x = 0.263677
#> Warning: non-integer x = 1.299202
#> Warning: non-integer x = 1.086124
#> Warning: non-integer x = 1.529896
#> Warning: non-integer x = 1.547426
#> Warning: non-integer x = 1.471806
#> Warning: non-integer x = 0.067869
#> Warning: non-integer x = 1.863436
#> Warning: non-integer x = 0.137734
#> Warning: non-integer x = 0.450985
#> Warning: non-integer x = 0.321061
#> Warning: non-integer x = 3.225952
#> Warning: non-integer x = 0.216557
#> Warning: non-integer x = 2.603418
#> Warning: non-integer x = 0.342257
#> Warning: non-integer x = 0.210485
#> Warning: non-integer x = 2.281780
#> Warning: non-integer x = 2.197231
#> Warning: non-integer x = 0.390155
#> Warning: non-integer x = 2.038457
#> Warning: non-integer x = 3.858728
#> Warning: non-integer x = 1.315646
#> Warning: non-integer x = 3.266549
#> Warning: non-integer x = 2.207678
#> Warning: non-integer x = 1.417304
#> Warning: non-integer x = 0.705732
#> Warning: non-integer x = 0.481601
#> Warning: non-integer x = 3.155053
#> Warning: non-integer x = 1.678274
#> Warning: non-integer x = 3.806015
#> Warning: non-integer x = 2.738873
#> Warning: non-integer x = 0.717383
#> Warning: non-integer x = 1.321041
#> Warning: non-integer x = 2.603309
#> Warning: non-integer x = 3.259868
#> Warning: non-integer x = 2.527383
#> Warning: non-integer x = 0.284545
#> Warning: non-integer x = 3.557442
#> Warning: non-integer x = 0.169860
#> Warning: non-integer x = 2.218986
#> Warning: non-integer x = 2.538048
#> Warning: non-integer x = 0.598475
#> Warning: non-integer x = 3.395727
#> Warning: non-integer x = 1.870898
#> Warning: non-integer x = 0.975537
#> Warning: non-integer x = 2.939096
#> Warning: non-integer x = 1.464213
#> Warning: non-integer x = 0.688474
#> Warning: non-integer x = 1.187032
#> Warning: non-integer x = 0.703652
#> Warning: non-integer x = 1.458494
#> Warning: non-integer x = 2.799473
#> Warning: non-integer x = 0.943318
#> Warning: non-integer x = 0.550693
#> Warning: non-integer x = 2.105022
#> Warning: non-integer x = 1.889063
#> Warning: non-integer x = 0.184582
#> Warning: non-integer x = 2.886287
#> Warning: non-integer x = 3.402797
#> Warning: non-integer x = 0.990790
#> Warning: non-integer x = 0.656618
#> Warning: non-integer x = 1.132630
#> Warning: non-integer x = 2.296388
#> Warning: non-integer x = 0.206581
#> Warning: non-integer x = 1.575164
#> Warning: non-integer x = 2.480869
#> Warning: non-integer x = 0.800496
#> Warning: non-integer x = 0.453687
#> Warning: non-integer x = 2.367935
#> Warning: non-integer x = 0.452668
#> Warning: non-integer x = 1.967122
#> Warning: non-integer x = 3.247613
#> Warning: non-integer x = 4.779607
#> Warning: non-integer x = 3.094423
#> Warning: non-integer x = 0.377012
#> Warning: non-integer x = 2.229610
#> Warning: non-integer x = 2.724672
#> Warning: non-integer x = 0.119096
#> Warning: non-integer x = 1.006222
#> Warning: non-integer x = 2.840985
#> Warning: non-integer x = 1.718873
#> Warning: non-integer x = 2.274233
#> Warning: non-integer x = 0.267868
#> Warning: non-integer x = 2.621052
#> Warning: non-integer x = 3.213916
#> Warning: non-integer x = 3.078060
#> Warning: non-integer x = 0.253174
#> Warning: non-integer x = 2.155817
#> Warning: non-integer x = 1.599004
#> Warning: non-integer x = 1.032320
#> Warning: non-integer x = 2.219183
#> Warning: non-integer x = 0.770325
#> Warning: non-integer x = 3.811291
#> Warning: non-integer x = 1.802845
#> Warning: non-integer x = 3.210699
#> Warning: non-integer x = 2.032833
#> Warning: non-integer x = 0.129642
#> Warning: non-integer x = 2.102925
#> Warning: non-integer x = 1.708024
#> Warning: non-integer x = 1.213502
#> Warning: non-integer x = 2.783418
#> Warning: non-integer x = 0.028577
#> Warning: non-integer x = 3.209003
#> Warning: non-integer x = 1.091103
#> Warning: non-integer x = 0.674911
#> Warning: non-integer x = 0.397088
#> Warning: non-integer x = 2.342579
#> Warning: non-integer x = 3.584228
#> Warning: non-integer x = 3.291820
#> Warning: non-integer x = 1.922941
#> Warning: non-integer x = 0.371337
#> Warning: non-integer x = 2.996600
#> Warning: non-integer x = 2.001990
#> Warning: non-integer x = 2.642184
#> Warning: non-integer x = 0.585162
#> Warning: non-integer x = 0.155350
#> Warning: non-integer x = 0.754467
#> Warning: non-integer x = 2.831041
#> Warning: non-integer x = 1.261704
#> Warning: non-integer x = 2.352399
#> Warning: non-integer x = 0.292519
#> Warning: non-integer x = 2.078565
#> Warning: non-integer x = 2.618924
#> Warning: non-integer x = 1.001485
#> Warning: non-integer x = 1.155563
#> Warning: non-integer x = 0.889864
#> Warning: non-integer x = 2.193760
#> Warning: non-integer x = 2.129257
#> Warning: non-integer x = 1.398533
#> Warning: non-integer x = 3.622577
#> Warning: non-integer x = 1.503020
#> Warning: non-integer x = 1.383068
#> Warning: non-integer x = 2.643785
#> Warning: non-integer x = 2.572952
#> Warning: non-integer x = 3.727254
#> Warning: non-integer x = 1.312868
#> Warning: non-integer x = 2.262227
#> Warning: non-integer x = 2.137432
#> Warning: non-integer x = 2.988166
#> Warning: non-integer x = 1.308767
#> Warning: non-integer x = 2.554239
#> Warning: non-integer x = 1.844926
#> Warning: non-integer x = 2.463261
#> Warning: non-integer x = 2.085403
#> Warning: non-integer x = 0.700318
#> Warning: non-integer x = 2.865784
#> Warning: non-integer x = 3.177730
#> Warning: non-integer x = 2.089259
#> Warning: non-integer x = 0.042788
#> Warning: non-integer x = 1.354723
#> Warning: non-integer x = 3.066455
#> Warning: non-integer x = 0.225872
#> Warning: non-integer x = 0.459535
#> Warning: non-integer x = 0.691858
#> Warning: non-integer x = 1.480165
#> Warning: non-integer x = 0.429894
#> Warning: non-integer x = 2.258614
#> Warning: non-integer x = 2.037200
#> Warning: non-integer x = 1.550300
#> Warning: non-integer x = 0.791911
#> Warning: non-integer x = 2.486809
#> Warning: non-integer x = 0.317781
#> Warning: non-integer x = 1.049592
#> Warning: non-integer x = 0.388361
#> Warning: non-integer x = 3.022345
#> Warning: non-integer x = 2.115637
#> Warning: non-integer x = 0.562256
#> Warning: non-integer x = 1.771028
#> Warning: non-integer x = 3.034857
#> Warning: non-integer x = 1.982887
#> Warning: non-integer x = 3.222783
#> Warning: non-integer x = 2.622123
#> Warning: non-integer x = 1.217017
#> Warning: non-integer x = 0.567590
#> Warning: non-integer x = 0.048883
#> Warning: non-integer x = 4.076044
#> Warning: non-integer x = 2.547771
#> Warning: non-integer x = 1.947699
#> Warning: non-integer x = 3.487235
#> Warning: non-integer x = 1.509962
#> Warning: non-integer x = 2.481612
#> Warning: non-integer x = 1.601078
#> Warning: non-integer x = 2.905173
#> Warning: non-integer x = 2.591894
#> Warning: non-integer x = 4.115277
#> Warning: non-integer x = 2.543138
#> Warning: non-integer x = 2.396517
#> Warning: non-integer x = 2.713380
#> Warning: non-integer x = 0.011581
#> Warning: non-integer x = 0.961800
#> Warning: non-integer x = 0.698421
#> Warning: non-integer x = 3.217695
#> Warning: non-integer x = 0.880037
#> Warning: non-integer x = 3.178583
#> Warning: non-integer x = 1.323673
#> Warning: non-integer x = 1.276984
#> Warning: non-integer x = 2.294332
#> Warning: non-integer x = 0.672063
#> Warning: non-integer x = 3.258593
#> Warning: non-integer x = 0.109132
#> Warning: non-integer x = 0.293072
#> Warning: non-integer x = 1.240872
#> Warning: non-integer x = 0.045941
#> Warning: non-integer x = 0.094530
#> Warning: non-integer x = 1.009084
#> Warning: non-integer x = 1.215771
#> Warning: non-integer x = 3.114490
#> Warning: non-integer x = 0.903251
#> Warning: non-integer x = 0.231566
#> Warning: non-integer x = 0.061395
#> Warning: non-integer x = 0.884251
#> Warning: non-integer x = 3.914610
#> Warning: non-integer x = 3.564882
#> Warning: non-integer x = 0.147413
#> Warning: non-integer x = 1.229713
#> Warning: non-integer x = 1.477917
#> Warning: non-integer x = 1.805870
#> Warning: non-integer x = 0.166905
#> Warning: non-integer x = 3.068833
#> Warning: non-integer x = 1.922027
#> Warning: non-integer x = 2.113568
#> Warning: non-integer x = 0.637161
#> Warning: non-integer x = 0.104398
#> Warning: non-integer x = 2.429260
#> Warning: non-integer x = 0.296487
#> Warning: non-integer x = 2.650311
#> Warning: non-integer x = 0.041144
#> Warning: non-integer x = 2.861281
#> Warning: non-integer x = 0.227506
#> Warning: non-integer x = 1.118353
#> Warning: non-integer x = 1.818004
#> Warning: non-integer x = 1.696249
#> Warning: non-integer x = 0.095148
#> Warning: non-integer x = 2.463107
#> Warning: non-integer x = 1.696544
#> Warning: non-integer x = 0.114131
#> Warning: non-integer x = 2.583449
#> Warning: non-integer x = 2.234581
#> Warning: non-integer x = 3.615865
#> Warning: non-integer x = 1.945187
#> Warning: non-integer x = 3.500858
#> Warning: non-integer x = 3.620748
#> Warning: non-integer x = 1.778180
#> Warning: non-integer x = 1.394391
#> Warning: non-integer x = 0.958490
#> Warning: non-integer x = 2.604801
#> Warning: non-integer x = 1.352642
#> Warning: non-integer x = 1.113233
#> Warning: non-integer x = 2.915983
#> Warning: non-integer x = 0.105330
#> Warning: non-integer x = 1.213622
#> Warning: non-integer x = 0.437396
#> Warning: non-integer x = 0.184653
#> Warning: non-integer x = 0.312588
#> Warning: non-integer x = 0.735832
#> Warning: non-integer x = 0.529824
#> Warning: non-integer x = 0.712994
#> Warning: non-integer x = 2.672610
#> Warning: non-integer x = 1.394836
#> Warning: non-integer x = 0.197774
#> Warning: non-integer x = 0.272774
#> Warning: non-integer x = 2.482441
#> Warning: non-integer x = 1.117349
#> Warning: non-integer x = 0.480590
#> Warning: non-integer x = 2.606141
#> Warning: non-integer x = 0.826906
#> Warning: non-integer x = 4.646208
#> Warning: non-integer x = 1.029009
#> Warning: non-integer x = 0.820691
#> Warning: non-integer x = 0.110376
#> Warning: non-integer x = 2.464498
#> Warning: non-integer x = 0.205304
#> Warning: non-integer x = 0.407338
#> Warning: non-integer x = 0.141816
#> Warning: non-integer x = 0.313686
#> Warning: non-integer x = 0.664003
#> Warning: non-integer x = 1.377908
#> Warning: non-integer x = 1.295142
#> Warning: non-integer x = 1.340325
#> Warning: non-integer x = 0.293463
#> Warning: non-integer x = 1.597125
#> Warning: non-integer x = 0.815715
#> Warning: non-integer x = 2.194257
#> Warning: non-integer x = 4.559196
#> Warning: non-integer x = 3.296533
#> Warning: non-integer x = 1.812634
#> Warning: non-integer x = 0.912280
#> Warning: non-integer x = 0.331993
#> Warning: non-integer x = 0.177912
#> Warning: non-integer x = 3.037428
#> Warning: non-integer x = 0.210895
#> Warning: non-integer x = 1.935259
#> Warning: non-integer x = 1.239704
#> Warning: non-integer x = 0.451415
#> Warning: non-integer x = 2.917009
#> Warning: non-integer x = 4.171267
#> Warning: non-integer x = 0.109591
#> Warning: non-integer x = 2.184582
#> Warning: non-integer x = 0.796898
#> Warning: non-integer x = 3.055116
#> Warning: non-integer x = 0.136388
#> Warning: non-integer x = 0.944940
#> Warning: non-integer x = 0.509552
#> Warning: non-integer x = 2.325289
#> Warning: non-integer x = 1.116761
#> Warning: non-integer x = 1.867721
#> Warning: non-integer x = 0.827370
#> Warning: non-integer x = 0.171566
#> Warning: non-integer x = 0.299247
#> Warning: non-integer x = 1.545900
#> Warning: non-integer x = 2.071376
#> Warning: non-integer x = 2.022888
#> Warning: non-integer x = 3.071331
#> Warning: non-integer x = 0.118839
#> Warning: non-integer x = 1.224646
#> Warning: non-integer x = 0.470384
#> Warning: non-integer x = 2.362905
#> Warning: non-integer x = 1.015982
#> Warning: non-integer x = 1.339492
#> Warning: non-integer x = 1.386561
#> Warning: non-integer x = 2.990445
#> Warning: non-integer x = 1.706584
#> Warning: non-integer x = 0.042911
#> Warning: non-integer x = 0.876394
#> Warning: non-integer x = 1.932852
#> Warning: non-integer x = 0.661612
#> Warning: non-integer x = 3.091090
#> Warning: non-integer x = 1.481948
#> Warning: non-integer x = 1.999400
#> Warning: non-integer x = 0.481237
#> Warning: non-integer x = 2.566287
#> Warning: non-integer x = 0.397760
#> Warning: non-integer x = 2.163154
#> Warning: non-integer x = 1.604753
#> Warning: non-integer x = 0.289316
#> Warning: non-integer x = 2.551129
#> Warning: non-integer x = 2.109661
#> Warning: non-integer x = 0.488146
#> Warning: non-integer x = 1.058883
#> Warning: non-integer x = 2.834464
#> Warning: non-integer x = 2.109725
#> Warning: non-integer x = 1.510705
#> Warning: non-integer x = 2.176451
#> Warning: non-integer x = 1.791082
#> Warning: non-integer x = 0.222877
#> Warning: non-integer x = 2.423380
#> Warning: non-integer x = 2.267586
#> Warning: non-integer x = 0.250863
#> Warning: non-integer x = 2.055926
#> Warning: non-integer x = 2.586520
#> Warning: non-integer x = 0.228690
#> Warning: non-integer x = 0.570818
#> Warning: non-integer x = 0.701251
#> Warning: non-integer x = 0.245453
#> Warning: non-integer x = 2.764482
#> Warning: non-integer x = 1.077953
#> Warning: non-integer x = 0.374639
#> Warning: non-integer x = 1.165447
#> Warning: non-integer x = 0.243482
#> Warning: non-integer x = 0.498521
#> Warning: non-integer x = 0.365618
#> Warning: non-integer x = 1.402298
#> Warning: non-integer x = 1.651543
#> Warning: non-integer x = 1.112400
#> Warning: non-integer x = 0.965469
#> Warning: non-integer x = 2.689185
#> Warning: non-integer x = 0.179059
#> Warning: non-integer x = 0.739417
#> Warning: non-integer x = 2.494531
#> Warning: non-integer x = 1.343637
#> Warning: non-integer x = 0.305959
#> Warning: non-integer x = 1.057811
#> Warning: non-integer x = 1.034535
#> Warning: non-integer x = 0.999564
#> Warning: non-integer x = 1.239310
#> Warning: non-integer x = 2.161419
#> Warning: non-integer x = 0.016445
#> Warning: non-integer x = 0.605469
#> Warning: non-integer x = 1.843453
#> Warning: non-integer x = 0.835352
#> Warning: non-integer x = 1.483964
#> Warning: non-integer x = 0.466545
#> Warning: non-integer x = 0.637639
#> Warning: non-integer x = 2.324097
#> Warning: non-integer x = 2.366587
#> Warning: non-integer x = 1.114879
#> Warning: non-integer x = 2.081361
#> Warning: non-integer x = 1.997964
#> Warning: non-integer x = 1.239689
#> Warning: non-integer x = 3.458046
#> Warning: non-integer x = 1.373411
#> Warning: non-integer x = 1.539369
#> Warning: non-integer x = 1.374551
#> Warning: non-integer x = 0.593023
#> Warning: non-integer x = 0.100683
#> Warning: non-integer x = 2.752785
#> Warning: non-integer x = 1.907664
#> Warning: non-integer x = 0.467948
#> Warning: non-integer x = 1.281738
#> Warning: non-integer x = 1.493169
#> Warning: non-integer x = 2.479791
#> Warning: non-integer x = 0.035580
#> Warning: non-integer x = 0.444799
#> Warning: non-integer x = 1.529720
#> Warning: non-integer x = 1.249413
#> Warning: non-integer x = 1.782576
#> Warning: non-integer x = 2.093951
#> Warning: non-integer x = 4.225114
#> Warning: non-integer x = 0.313240
#> Warning: non-integer x = 0.070772
#> Warning: non-integer x = 0.289078
#> Warning: non-integer x = 0.495228
#> Warning: non-integer x = 2.122997
#> Warning: non-integer x = 3.198690
#> Warning: non-integer x = 2.033020
#> Warning: non-integer x = 3.915830
#> Warning: non-integer x = 0.496859
#> Warning: non-integer x = 3.580055
#> Warning: non-integer x = 0.600810
#> Warning: non-integer x = 1.844253
#> Warning: non-integer x = 3.882762
#> Warning: non-integer x = 0.821070
#> Warning: non-integer x = 5.677803
#> Warning: non-integer x = 2.954826
#> Warning: non-integer x = 1.925046
#> Warning: non-integer x = 1.079080
#> Warning: non-integer x = 0.355926
#> Warning: non-integer x = 0.430090
#> Warning: non-integer x = 0.758368
#> Warning: non-integer x = 2.468389
#> Warning: non-integer x = 0.447723
#> Warning: non-integer x = 0.962232
#> Warning: non-integer x = 1.149966
#> Warning: non-integer x = 0.158773
#> Warning: non-integer x = 0.353694
#> Warning: non-integer x = 2.097777
#> Warning: non-integer x = 1.689868
#> Warning: non-integer x = 0.844348
#> Warning: non-integer x = 0.312547
#> Warning: non-integer x = 2.688079
#> Warning: non-integer x = 1.089823
#> Warning: non-integer x = 0.636693
#> Warning: non-integer x = 2.371757
#> Warning: non-integer x = 3.516708
#> Warning: non-integer x = 0.296939
#> Warning: non-integer x = 0.676258
#> Warning: non-integer x = 0.796185
#> Warning: non-integer x = 0.420837
#> Warning: non-integer x = 3.022377
#> Warning: non-integer x = 2.561331
#> Warning: non-integer x = 1.350371
#> Warning: non-integer x = 1.009960
#> Warning: non-integer x = 1.257787
#> Warning: non-integer x = 3.544449
#> Warning: non-integer x = 1.171805
#> Warning: non-integer x = 0.759800
#> Warning: non-integer x = 1.649054
#> Warning: non-integer x = 1.168666
#> Warning: non-integer x = 1.115906
#> Warning: non-integer x = 1.876065
#> Warning: non-integer x = 0.897626
#> Warning: non-integer x = 4.187030
#> Warning: non-integer x = 0.076849
#> Warning: non-integer x = 0.207095
#> Warning: non-integer x = 1.251786
#> Warning: non-integer x = 2.867762
#> Warning: non-integer x = 0.451960
#> Warning: non-integer x = 0.494467
#> Warning: non-integer x = 1.305547
#> Warning: non-integer x = 0.545978
#> Warning: non-integer x = 0.314020
#> Warning: non-integer x = 0.024033
#> Warning: non-integer x = 1.075811
#> Warning: non-integer x = 0.723687
#> Warning: non-integer x = 0.097015
#> Warning: non-integer x = 0.541591
#> Warning: non-integer x = 1.291317
#> Warning: non-integer x = 0.714557
#> Warning: non-integer x = 0.526121
#> Warning: non-integer x = 1.622810
#> Warning: non-integer x = 1.714730
#> Warning: non-integer x = 1.199971
#> Warning: non-integer x = 0.092109
#> Warning: non-integer x = 3.604448
#> Warning: non-integer x = 1.410965
#> Warning: non-integer x = 2.180226
#> Warning: non-integer x = 1.885329
#> Warning: non-integer x = 1.310656
#> Warning: non-integer x = 3.570970
#> Warning: non-integer x = 2.254845
#> Warning: non-integer x = 1.496076
#> Warning: non-integer x = 0.301919
#> Warning: non-integer x = 0.664659
#> Warning: non-integer x = 2.304024
#> Warning: non-integer x = 1.206239
#> Warning: non-integer x = 2.127507
#> Warning: non-integer x = 1.372151
#> Warning: non-integer x = 3.312394
#> Warning: non-integer x = 1.550543
#> Warning: non-integer x = 1.295170
#> Warning: non-integer x = 0.485318
#> Warning: non-integer x = 2.795049
#> Warning: non-integer x = 0.520383
#> Warning: non-integer x = 1.551195
#> Warning: non-integer x = 0.716813
#> Warning: non-integer x = 0.287106
#> Warning: non-integer x = 0.672414
#> Warning: non-integer x = 0.864530
#> Warning: non-integer x = 1.081201
#> Warning: non-integer x = 0.402853
#> Warning: non-integer x = 0.312943
#> Warning: non-integer x = 0.259662
#> Warning: non-integer x = 1.548980
#> Warning: non-integer x = 0.585448
#> Warning: non-integer x = 0.886876
#> Warning: non-integer x = 1.454519
#> Warning: non-integer x = 1.279811
#> Warning: non-integer x = 2.783894
#> Warning: non-integer x = 0.301418
#> Warning: non-integer x = 2.694482
#> Warning: non-integer x = 0.735561
#> Warning: non-integer x = 0.697028
#> Warning: non-integer x = 3.284918
#> Warning: non-integer x = 0.391339
#> Warning: non-integer x = 0.992591
#> Warning: non-integer x = 0.067250
#> Warning: non-integer x = 0.061427
#> Warning: non-integer x = 0.160772
#> Warning: non-integer x = 1.267558
#> Warning: non-integer x = 0.905113
#> Warning: non-integer x = 1.936560
#> Warning: non-integer x = 1.020561
#> Warning: non-integer x = 1.200781
#> Warning: non-integer x = 2.055405
#> Warning: non-integer x = 1.019565
#> Warning: non-integer x = 0.277668
#> Warning: non-integer x = 0.831285
#> Warning: non-integer x = 0.893250
#> Warning: non-integer x = 2.429435
#> Warning: non-integer x = 1.849251
#> Warning: non-integer x = 0.859750
#> Warning: non-integer x = 2.460006
#> Warning: non-integer x = 1.402236
#> Warning: non-integer x = 2.594962
#> Warning: non-integer x = 0.568427
#> Warning: non-integer x = 0.805484
#> Warning: non-integer x = 1.067306
#> Warning: non-integer x = 0.995596
#> Warning: non-integer x = 1.502569
#> Warning: non-integer x = 2.979364
#> Warning: non-integer x = 0.390432
#> Warning: non-integer x = 2.571589
#> Warning: non-integer x = 0.430563
#> Warning: non-integer x = 0.445582
#> Warning: non-integer x = 0.708362
#> Warning: non-integer x = 2.414585
#> Warning: non-integer x = 2.830407
#> Warning: non-integer x = 1.547397
#> Warning: non-integer x = 0.014808
#> Warning: non-integer x = 0.405635
#> Warning: non-integer x = 1.579314
#> Warning: non-integer x = 0.342618
#> Warning: non-integer x = 0.522428
#> Warning: non-integer x = 1.509158
#> Warning: non-integer x = 1.271280
#> Warning: non-integer x = 0.041316
#> Warning: non-integer x = 0.214423
#> Warning: non-integer x = 2.254358
#> Warning: non-integer x = 0.824453
#> Warning: non-integer x = 2.477433
#> Warning: non-integer x = 2.024874
#> Warning: non-integer x = 0.288021
#> Warning: non-integer x = 2.121730
#> Warning: non-integer x = 2.043928
#> Warning: non-integer x = 2.058932
#> Warning: non-integer x = 4.597365
#> Warning: non-integer x = 0.229746
#> Warning: non-integer x = 0.619463
#> Warning: non-integer x = 1.609558
#> Warning: non-integer x = 0.864574
#> Warning: non-integer x = 0.152986
#> Warning: non-integer x = 0.408511
#> Warning: non-integer x = 0.410037
#> Warning: non-integer x = 0.741169
#> Warning: non-integer x = 1.853019
#> Warning: non-integer x = 1.099917
#> Warning: non-integer x = 2.487987
#> Warning: non-integer x = 1.351064
#> Warning: non-integer x = 0.554934
#> Warning: non-integer x = 4.007433
#> Warning: non-integer x = 0.380581
#> Warning: non-integer x = 1.549738
#> Warning: non-integer x = 0.850990
#> Warning: non-integer x = 1.951504
#> Warning: non-integer x = 3.157395
#> Warning: non-integer x = 0.300760
#> Warning: non-integer x = 0.840529
#> Warning: non-integer x = 0.287707
#> Warning: non-integer x = 1.065513
#> Warning: non-integer x = 1.476169
#> Warning: non-integer x = 1.464793
#> Warning: non-integer x = 2.215295
#> Warning: non-integer x = 0.630239
#> Warning: non-integer x = 1.597100
#> Warning: non-integer x = 0.993385
#> Warning: non-integer x = 1.740696
#> Warning: non-integer x = 0.722686
#> Warning: non-integer x = 0.350228
#> Warning: non-integer x = 0.768572
#> Warning: non-integer x = 0.800233
#> Warning: non-integer x = 0.459641
#> Warning: non-integer x = 1.618047
#> Warning: non-integer x = 2.423986
#> Warning: non-integer x = 0.589134
#> Warning: non-integer x = 1.190602
#> Warning: non-integer x = 0.497655
#> Warning: non-integer x = 2.276579
#> Warning: non-integer x = 1.583026
#> Warning: non-integer x = 1.636032
#> Warning: non-integer x = 0.661341
#> Warning: non-integer x = 0.968769
#> Warning: non-integer x = 0.225470
#> Warning: non-integer x = 2.514336
#> Warning: non-integer x = 0.072197
#> Warning: non-integer x = 0.336733
#> Warning: non-integer x = 0.557342
#> Warning: non-integer x = 4.614561
#> Warning: non-integer x = 0.030069
#> Warning: non-integer x = 1.904081
#> Warning: non-integer x = 0.944542
#> Warning: non-integer x = 0.008727
#> Warning: non-integer x = 3.047273
#> Warning: non-integer x = 1.740940
#> Warning: non-integer x = 2.899885
#> Warning: non-integer x = 3.936534
#> Warning: non-integer x = 3.415226
#> Warning: non-integer x = 0.708103
#> Warning: non-integer x = 1.332320
#> Warning: non-integer x = 1.491141
#> Warning: non-integer x = 0.418512
#> Warning: non-integer x = 1.134544
#> Warning: non-integer x = 3.713602
#> Warning: non-integer x = 1.215939
#> Warning: non-integer x = 2.663854
#> Warning: non-integer x = 1.120221
#> Warning: non-integer x = 3.084411
#> Warning: non-integer x = 0.690180
#> Warning: non-integer x = 2.562943
#> Warning: non-integer x = 2.512901
#> Warning: non-integer x = 1.329495
#> Warning: non-integer x = 3.247169
#> Warning: non-integer x = 0.817255
#> Warning: non-integer x = 1.287389
#> Warning: non-integer x = 2.239422
#> Warning: non-integer x = 0.281911
#> Warning: non-integer x = 1.110790
#> Warning: non-integer x = 0.904068
#> Warning: non-integer x = 1.491738
#> Warning: non-integer x = 0.091312
#> Warning: non-integer x = 0.639635
#> Warning: non-integer x = 1.321382
#> Warning: non-integer x = 0.261795
#> Warning: non-integer x = 1.098956
#> Warning: non-integer x = 1.259777
#> Warning: non-integer x = 0.631490
#> Warning: non-integer x = 2.067462
#> Warning: non-integer x = 0.469206
#> Warning: non-integer x = 1.194417
#> Warning: non-integer x = 0.579339
#> Warning: non-integer x = 1.904655
#> Warning: non-integer x = 1.506025
#> Warning: non-integer x = 1.793369
#> Warning: non-integer x = 1.040737
#> Warning: non-integer x = 2.152436
#> Warning: non-integer x = 0.038391
#> Warning: non-integer x = 1.465573
#> Warning: non-integer x = 2.747902
#> Warning: non-integer x = 1.361206
#> Warning: non-integer x = 1.790275
#> Warning: non-integer x = 1.713487
#> Warning: non-integer x = 1.795946
#> Warning: non-integer x = 0.878236
#> Warning: non-integer x = 1.435478
#> Warning: non-integer x = 0.278900
#> Warning: non-integer x = 1.547324
#> Warning: non-integer x = 0.236877
#> Warning: non-integer x = 1.515097
#> Warning: non-integer x = 0.988017
#> Warning: non-integer x = 0.088315
#> Warning: non-integer x = 2.642285
#> Warning: non-integer x = 1.061822
#> Warning: non-integer x = 0.043463
#> Warning: non-integer x = 0.966734
#> Warning: non-integer x = 3.311342
#> Warning: non-integer x = 1.886943
#> Warning: non-integer x = 0.003259
#> Warning: non-integer x = 0.539066
#> Warning: non-integer x = 0.739922
#> Warning: non-integer x = 1.555017
#> Warning: non-integer x = 2.014428
#> Warning: non-integer x = 1.116894
#> Warning: non-integer x = 2.924992
#> Warning: non-integer x = 0.556300
#> Warning: non-integer x = 0.931601
#> Warning: non-integer x = 1.063429
#> Warning: non-integer x = 0.175555
#> Warning: non-integer x = 1.619080
#> Warning: non-integer x = 1.267038
#> Warning: non-integer x = 0.057756
#> Warning: non-integer x = 2.516196
#> Warning: non-integer x = 1.070312
#> Warning: non-integer x = 1.347869
#> Warning: non-integer x = 3.050514
#> Warning: non-integer x = 0.740944
#> Warning: non-integer x = 0.419147
#> Warning: non-integer x = 0.089739
#> Warning: non-integer x = 0.241089
#> Warning: non-integer x = 0.202427
#> Warning: non-integer x = 2.546706
#> Warning: non-integer x = 0.683910
#> Warning: non-integer x = 3.302893
#> Warning: non-integer x = 2.659439
#> Warning: non-integer x = 2.350468
#> Warning: non-integer x = 1.357947
#> Warning: non-integer x = 0.546625
#> Warning: non-integer x = 0.197431
#> Warning: non-integer x = 1.185177
#> Warning: non-integer x = 2.589240
#> Warning: non-integer x = 2.173818
#> Warning: non-integer x = 0.221989
#> Warning: non-integer x = 3.431575
#> Warning: non-integer x = 0.271554
#> Warning: non-integer x = 1.339786
#> Warning: non-integer x = 0.675139
#> Warning: non-integer x = 0.796324
#> Warning: non-integer x = 1.429501
#> Warning: non-integer x = 3.276520
#> Warning: non-integer x = 0.084324
#> Warning: non-integer x = 0.367441
#> Warning: non-integer x = 0.284713
#> Warning: non-integer x = 2.137395
#> Warning: non-integer x = 2.042014
#> Warning: non-integer x = 0.220150
#> Warning: non-integer x = 0.244300
#> Warning: non-integer x = 1.256571
#> Warning: non-integer x = 0.205115
#> Warning: non-integer x = 0.078763
#> Warning: non-integer x = 1.362165
#> Warning: non-integer x = 2.865703
#> Warning: non-integer x = 0.295331
#> Warning: non-integer x = 3.153115
#> Warning: non-integer x = 0.927631
#> Warning: non-integer x = 0.864682
#> Warning: non-integer x = 2.922313
#> Warning: non-integer x = 1.987256
#> Warning: non-integer x = 2.476396
#> Warning: non-integer x = 0.652857
#> Warning: non-integer x = 0.333989
#> Warning: non-integer x = 1.816718
#> Warning: non-integer x = 2.321351
#> Warning: non-integer x = 0.648113
#> Warning: non-integer x = 2.739283
#> Warning: non-integer x = 2.590231
#> Warning: non-integer x = 0.020810
#> Warning: non-integer x = 1.668228
#> Warning: non-integer x = 2.003885
#> Warning: non-integer x = 0.571829
#> Warning: non-integer x = 0.193745
#> Warning: non-integer x = 1.789613
#> Warning: non-integer x = 1.304449
#> Warning: non-integer x = 0.352365
#> Warning: non-integer x = 1.950697
#> Warning: non-integer x = 1.055859
#> Warning: non-integer x = 0.554786
#> Warning: non-integer x = 0.127974
#> Warning: non-integer x = 1.727542
#> Warning: non-integer x = 0.788628
#> Warning: non-integer x = 3.931880
#> Warning: non-integer x = 1.000906
#> Warning: non-integer x = 0.745148
#> Warning: non-integer x = 0.346071
#> Warning: non-integer x = 1.225115
#> Warning: non-integer x = 0.363977
#> Warning: non-integer x = 2.460117
#> Warning: non-integer x = 2.160205
#> Warning: non-integer x = 3.356058
#> Warning: non-integer x = 0.204047
#> Warning: non-integer x = 0.585659
#> Warning: non-integer x = 2.239643
#> Warning: non-integer x = 2.181102
#> Warning: non-integer x = 1.852444
#> Warning: non-integer x = 3.903030
#> Warning: non-integer x = 2.348858
#> Warning: non-integer x = 0.397736
#> Warning: non-integer x = 0.783994
#> Warning: non-integer x = 0.470932
#> Warning: non-integer x = 0.705602
#> Warning: non-integer x = 0.774070
#> Warning: non-integer x = 1.293444
#> Warning: non-integer x = 1.550977
#> Warning: non-integer x = 2.299881
#> Warning: non-integer x = 1.438714
#> Warning: non-integer x = 0.435669
#> Warning: non-integer x = 2.814818
#> Warning: non-integer x = 1.458443
#> Warning: non-integer x = 1.196407
#> Warning: non-integer x = 0.288167
#> Warning: non-integer x = 0.634714
#> Warning: non-integer x = 0.196278
#> Warning: non-integer x = 3.122801
#> Warning: non-integer x = 0.697386
#> Warning: non-integer x = 1.206271
#> Warning: non-integer x = 1.551877
#> Warning: non-integer x = 1.218780
#> Warning: non-integer x = 1.484920
#> Warning: non-integer x = 0.833153
#> Warning: non-integer x = 0.005324
#> Warning: non-integer x = 1.635172
#> Warning: non-integer x = 1.596226
#> Warning: non-integer x = 2.615184
#> Warning: non-integer x = 1.302142
#> Warning: non-integer x = 0.084644
#> Warning: non-integer x = 2.945906