Estimate the Individual Treatment Effect using Poisson Regression

Estimates the Individual Treatment Effect using Poisson Regression given a response vector, a treatment vector, and a features matrix.

Usage

estimate_ite_poisson(y, z, X, X_names, include_offset, offset_name)

Arguments

y

the observed response vector

z

the treatment vector

X

the features matrix

X_names

the names of the covariates

include_offset

whether or not to include an offset when estimating the ITE, for poisson only

offset_name

the name of the offset, if it is to be included

Value

a vector of ITE estimates

Examples

dataset <- generate_cre_dataset(n = 1000, rho = 0, n_rules = 2, p = 10,
                                effect_size = 2, binary = FALSE)

# Initialize parameters
y <- abs(dataset[["y"]])
z <- dataset[["z"]]
X <- as.data.frame(dataset[["X"]])
X_names <- names(as.data.frame(X))
include_offset <- FALSE
offset_name <- NA

ite_list <- estimate_ite_poisson(y, z, X, X_names, include_offset, offset_name)
#> Warning: non-integer x = 0.175512
#> Warning: non-integer x = 0.133029
#> Warning: non-integer x = 1.901907
#> Warning: non-integer x = 1.226758
#> Warning: non-integer x = 4.314261
#> Warning: non-integer x = 1.775886
#> Warning: non-integer x = 1.994380
#> Warning: non-integer x = 2.534269
#> Warning: non-integer x = 1.211469
#> Warning: non-integer x = 2.858785
#> Warning: non-integer x = 0.574633
#> Warning: non-integer x = 2.815928
#> Warning: non-integer x = 1.255566
#> Warning: non-integer x = 1.401583
#> Warning: non-integer x = 0.595123
#> Warning: non-integer x = 0.506012
#> Warning: non-integer x = 0.226638
#> Warning: non-integer x = 3.802055
#> Warning: non-integer x = 0.663873
#> Warning: non-integer x = 2.915202
#> Warning: non-integer x = 1.740831
#> Warning: non-integer x = 2.265770
#> Warning: non-integer x = 0.602739
#> Warning: non-integer x = 1.317158
#> Warning: non-integer x = 0.992152
#> Warning: non-integer x = 0.889729
#> Warning: non-integer x = 3.398839
#> Warning: non-integer x = 0.603395
#> Warning: non-integer x = 3.620659
#> Warning: non-integer x = 0.837350
#> Warning: non-integer x = 0.078959
#> Warning: non-integer x = 2.022605
#> Warning: non-integer x = 2.575098
#> Warning: non-integer x = 0.025293
#> Warning: non-integer x = 3.092319
#> Warning: non-integer x = 1.488734
#> Warning: non-integer x = 1.391984
#> Warning: non-integer x = 1.433283
#> Warning: non-integer x = 0.137518
#> Warning: non-integer x = 1.320645
#> Warning: non-integer x = 0.863643
#> Warning: non-integer x = 2.096155
#> Warning: non-integer x = 0.270028
#> Warning: non-integer x = 0.355274
#> Warning: non-integer x = 0.247738
#> Warning: non-integer x = 1.348610
#> Warning: non-integer x = 3.124692
#> Warning: non-integer x = 0.318510
#> Warning: non-integer x = 3.160596
#> Warning: non-integer x = 3.048983
#> Warning: non-integer x = 0.117592
#> Warning: non-integer x = 1.799832
#> Warning: non-integer x = 0.813439
#> Warning: non-integer x = 1.089505
#> Warning: non-integer x = 1.789013
#> Warning: non-integer x = 0.774056
#> Warning: non-integer x = 0.422628
#> Warning: non-integer x = 1.669748
#> Warning: non-integer x = 1.136517
#> Warning: non-integer x = 1.733030
#> Warning: non-integer x = 0.076787
#> Warning: non-integer x = 1.124138
#> Warning: non-integer x = 0.985825
#> Warning: non-integer x = 0.545816
#> Warning: non-integer x = 0.833306
#> Warning: non-integer x = 2.354109
#> Warning: non-integer x = 2.820230
#> Warning: non-integer x = 1.323070
#> Warning: non-integer x = 2.366633
#> Warning: non-integer x = 2.564188
#> Warning: non-integer x = 1.749210
#> Warning: non-integer x = 1.880565
#> Warning: non-integer x = 0.423673
#> Warning: non-integer x = 0.017096
#> Warning: non-integer x = 0.446834
#> Warning: non-integer x = 0.772893
#> Warning: non-integer x = 0.478922
#> Warning: non-integer x = 1.301064
#> Warning: non-integer x = 1.943323
#> Warning: non-integer x = 5.077556
#> Warning: non-integer x = 0.883843
#> Warning: non-integer x = 1.293895
#> Warning: non-integer x = 2.051538
#> Warning: non-integer x = 4.067383
#> Warning: non-integer x = 1.331148
#> Warning: non-integer x = 1.058387
#> Warning: non-integer x = 0.881137
#> Warning: non-integer x = 2.643920
#> Warning: non-integer x = 1.978167
#> Warning: non-integer x = 0.160725
#> Warning: non-integer x = 2.795553
#> Warning: non-integer x = 0.134934
#> Warning: non-integer x = 1.172175
#> Warning: non-integer x = 0.166204
#> Warning: non-integer x = 2.336992
#> Warning: non-integer x = 2.012806
#> Warning: non-integer x = 1.675742
#> Warning: non-integer x = 0.400582
#> Warning: non-integer x = 1.929820
#> Warning: non-integer x = 0.216953
#> Warning: non-integer x = 2.587421
#> Warning: non-integer x = 0.427161
#> Warning: non-integer x = 2.795450
#> Warning: non-integer x = 1.263660
#> Warning: non-integer x = 0.407968
#> Warning: non-integer x = 0.882293
#> Warning: non-integer x = 0.792623
#> Warning: non-integer x = 1.055855
#> Warning: non-integer x = 0.967931
#> Warning: non-integer x = 0.248254
#> Warning: non-integer x = 1.992710
#> Warning: non-integer x = 0.675680
#> Warning: non-integer x = 1.477965
#> Warning: non-integer x = 0.882437
#> Warning: non-integer x = 1.625485
#> Warning: non-integer x = 0.910743
#> Warning: non-integer x = 3.027582
#> Warning: non-integer x = 1.881514
#> Warning: non-integer x = 3.393320
#> Warning: non-integer x = 3.367293
#> Warning: non-integer x = 3.628088
#> Warning: non-integer x = 3.719241
#> Warning: non-integer x = 2.634760
#> Warning: non-integer x = 2.435689
#> Warning: non-integer x = 0.393625
#> Warning: non-integer x = 2.639648
#> Warning: non-integer x = 0.228091
#> Warning: non-integer x = 0.940335
#> Warning: non-integer x = 0.777871
#> Warning: non-integer x = 2.553025
#> Warning: non-integer x = 0.138365
#> Warning: non-integer x = 3.210265
#> Warning: non-integer x = 0.869397
#> Warning: non-integer x = 2.367189
#> Warning: non-integer x = 3.639530
#> Warning: non-integer x = 1.685204
#> Warning: non-integer x = 3.240087
#> Warning: non-integer x = 4.670400
#> Warning: non-integer x = 2.335501
#> Warning: non-integer x = 4.030827
#> Warning: non-integer x = 2.619820
#> Warning: non-integer x = 0.463021
#> Warning: non-integer x = 0.530456
#> Warning: non-integer x = 5.049901
#> Warning: non-integer x = 0.286762
#> Warning: non-integer x = 0.983264
#> Warning: non-integer x = 1.770676
#> Warning: non-integer x = 1.559555
#> Warning: non-integer x = 1.179405
#> Warning: non-integer x = 0.791797
#> Warning: non-integer x = 1.060161
#> Warning: non-integer x = 0.080216
#> Warning: non-integer x = 2.855823
#> Warning: non-integer x = 1.630773
#> Warning: non-integer x = 0.022977
#> Warning: non-integer x = 1.507970
#> Warning: non-integer x = 1.338077
#> Warning: non-integer x = 3.112351
#> Warning: non-integer x = 3.417553
#> Warning: non-integer x = 1.257743
#> Warning: non-integer x = 1.327397
#> Warning: non-integer x = 2.790318
#> Warning: non-integer x = 0.018000
#> Warning: non-integer x = 1.437155
#> Warning: non-integer x = 0.650060
#> Warning: non-integer x = 1.934322
#> Warning: non-integer x = 0.542731
#> Warning: non-integer x = 0.615043
#> Warning: non-integer x = 0.011112
#> Warning: non-integer x = 3.084240
#> Warning: non-integer x = 2.385218
#> Warning: non-integer x = 1.945547
#> Warning: non-integer x = 1.197051
#> Warning: non-integer x = 0.197577
#> Warning: non-integer x = 1.920905
#> Warning: non-integer x = 2.120485
#> Warning: non-integer x = 0.406062
#> Warning: non-integer x = 3.607891
#> Warning: non-integer x = 2.567236
#> Warning: non-integer x = 3.551324
#> Warning: non-integer x = 0.516063
#> Warning: non-integer x = 2.802041
#> Warning: non-integer x = 3.176201
#> Warning: non-integer x = 0.191357
#> Warning: non-integer x = 1.137763
#> Warning: non-integer x = 0.939236
#> Warning: non-integer x = 1.048685
#> Warning: non-integer x = 1.269085
#> Warning: non-integer x = 3.724227
#> Warning: non-integer x = 3.099306
#> Warning: non-integer x = 0.353950
#> Warning: non-integer x = 2.394766
#> Warning: non-integer x = 3.902179
#> Warning: non-integer x = 2.867220
#> Warning: non-integer x = 2.492585
#> Warning: non-integer x = 1.938107
#> Warning: non-integer x = 0.426522
#> Warning: non-integer x = 2.958040
#> Warning: non-integer x = 3.494602
#> Warning: non-integer x = 0.215813
#> Warning: non-integer x = 1.218571
#> Warning: non-integer x = 3.426614
#> Warning: non-integer x = 1.189336
#> Warning: non-integer x = 0.261250
#> Warning: non-integer x = 1.670564
#> Warning: non-integer x = 1.139409
#> Warning: non-integer x = 2.951534
#> Warning: non-integer x = 0.860051
#> Warning: non-integer x = 1.141526
#> Warning: non-integer x = 2.028760
#> Warning: non-integer x = 2.366202
#> Warning: non-integer x = 0.420403
#> Warning: non-integer x = 2.700336
#> Warning: non-integer x = 4.195342
#> Warning: non-integer x = 0.084862
#> Warning: non-integer x = 3.203688
#> Warning: non-integer x = 1.262094
#> Warning: non-integer x = 1.471538
#> Warning: non-integer x = 4.490123
#> Warning: non-integer x = 0.228992
#> Warning: non-integer x = 1.606038
#> Warning: non-integer x = 0.191268
#> Warning: non-integer x = 5.131642
#> Warning: non-integer x = 0.437962
#> Warning: non-integer x = 2.650357
#> Warning: non-integer x = 1.650499
#> Warning: non-integer x = 2.505464
#> Warning: non-integer x = 1.852102
#> Warning: non-integer x = 1.732339
#> Warning: non-integer x = 2.660613
#> Warning: non-integer x = 0.063396
#> Warning: non-integer x = 2.103903
#> Warning: non-integer x = 0.106599
#> Warning: non-integer x = 1.506436
#> Warning: non-integer x = 0.137594
#> Warning: non-integer x = 2.648464
#> Warning: non-integer x = 1.966046
#> Warning: non-integer x = 2.125734
#> Warning: non-integer x = 0.745345
#> Warning: non-integer x = 0.810086
#> Warning: non-integer x = 4.772204
#> Warning: non-integer x = 2.129181
#> Warning: non-integer x = 5.911523
#> Warning: non-integer x = 1.988987
#> Warning: non-integer x = 2.298028
#> Warning: non-integer x = 1.985758
#> Warning: non-integer x = 1.856190
#> Warning: non-integer x = 2.713904
#> Warning: non-integer x = 0.227052
#> Warning: non-integer x = 0.319768
#> Warning: non-integer x = 3.789148
#> Warning: non-integer x = 0.831907
#> Warning: non-integer x = 0.196099
#> Warning: non-integer x = 3.254448
#> Warning: non-integer x = 3.479009
#> Warning: non-integer x = 3.382266
#> Warning: non-integer x = 2.122582
#> Warning: non-integer x = 1.119352
#> Warning: non-integer x = 3.631210
#> Warning: non-integer x = 0.836976
#> Warning: non-integer x = 2.531300
#> Warning: non-integer x = 3.244262
#> Warning: non-integer x = 0.594983
#> Warning: non-integer x = 2.249652
#> Warning: non-integer x = 1.457837
#> Warning: non-integer x = 3.438888
#> Warning: non-integer x = 0.912620
#> Warning: non-integer x = 1.602210
#> Warning: non-integer x = 2.288434
#> Warning: non-integer x = 0.845702
#> Warning: non-integer x = 1.325919
#> Warning: non-integer x = 3.302703
#> Warning: non-integer x = 1.153948
#> Warning: non-integer x = 2.711384
#> Warning: non-integer x = 0.652622
#> Warning: non-integer x = 0.390550
#> Warning: non-integer x = 1.507664
#> Warning: non-integer x = 0.161239
#> Warning: non-integer x = 1.564548
#> Warning: non-integer x = 3.417911
#> Warning: non-integer x = 1.897219
#> Warning: non-integer x = 1.689454
#> Warning: non-integer x = 0.053634
#> Warning: non-integer x = 0.993366
#> Warning: non-integer x = 2.869105
#> Warning: non-integer x = 0.970033
#> Warning: non-integer x = 2.802919
#> Warning: non-integer x = 2.012884
#> Warning: non-integer x = 0.028978
#> Warning: non-integer x = 0.556532
#> Warning: non-integer x = 0.375812
#> Warning: non-integer x = 0.490629
#> Warning: non-integer x = 2.782013
#> Warning: non-integer x = 2.505173
#> Warning: non-integer x = 2.010006
#> Warning: non-integer x = 2.084103
#> Warning: non-integer x = 1.545852
#> Warning: non-integer x = 0.529644
#> Warning: non-integer x = 2.781166
#> Warning: non-integer x = 1.602098
#> Warning: non-integer x = 2.265726
#> Warning: non-integer x = 1.701848
#> Warning: non-integer x = 0.682797
#> Warning: non-integer x = 1.233708
#> Warning: non-integer x = 2.624367
#> Warning: non-integer x = 2.633309
#> Warning: non-integer x = 1.456290
#> Warning: non-integer x = 2.913339
#> Warning: non-integer x = 1.302329
#> Warning: non-integer x = 3.479322
#> Warning: non-integer x = 0.053725
#> Warning: non-integer x = 0.866361
#> Warning: non-integer x = 0.319885
#> Warning: non-integer x = 1.859206
#> Warning: non-integer x = 2.855756
#> Warning: non-integer x = 0.361548
#> Warning: non-integer x = 3.077846
#> Warning: non-integer x = 1.539129
#> Warning: non-integer x = 3.626478
#> Warning: non-integer x = 0.988946
#> Warning: non-integer x = 0.585704
#> Warning: non-integer x = 2.313958
#> Warning: non-integer x = 0.919631
#> Warning: non-integer x = 3.948185
#> Warning: non-integer x = 1.166273
#> Warning: non-integer x = 1.674237
#> Warning: non-integer x = 3.950569
#> Warning: non-integer x = 3.033202
#> Warning: non-integer x = 3.551294
#> Warning: non-integer x = 1.721274
#> Warning: non-integer x = 0.074159
#> Warning: non-integer x = 3.917546
#> Warning: non-integer x = 0.942845
#> Warning: non-integer x = 3.249282
#> Warning: non-integer x = 2.649623
#> Warning: non-integer x = 1.617432
#> Warning: non-integer x = 0.943084
#> Warning: non-integer x = 1.534454
#> Warning: non-integer x = 0.790055
#> Warning: non-integer x = 0.365632
#> Warning: non-integer x = 0.568841
#> Warning: non-integer x = 0.631042
#> Warning: non-integer x = 3.847634
#> Warning: non-integer x = 3.731291
#> Warning: non-integer x = 2.202290
#> Warning: non-integer x = 0.252288
#> Warning: non-integer x = 0.792156
#> Warning: non-integer x = 1.831333
#> Warning: non-integer x = 2.878096
#> Warning: non-integer x = 0.028583
#> Warning: non-integer x = 1.973434
#> Warning: non-integer x = 3.496422
#> Warning: non-integer x = 2.084923
#> Warning: non-integer x = 1.316678
#> Warning: non-integer x = 2.704571
#> Warning: non-integer x = 2.035995
#> Warning: non-integer x = 2.397755
#> Warning: non-integer x = 1.846507
#> Warning: non-integer x = 4.108440
#> Warning: non-integer x = 1.073360
#> Warning: non-integer x = 1.056908
#> Warning: non-integer x = 0.173348
#> Warning: non-integer x = 0.111592
#> Warning: non-integer x = 3.438757
#> Warning: non-integer x = 3.749549
#> Warning: non-integer x = 2.959089
#> Warning: non-integer x = 1.343019
#> Warning: non-integer x = 1.264960
#> Warning: non-integer x = 1.565965
#> Warning: non-integer x = 1.111641
#> Warning: non-integer x = 1.919599
#> Warning: non-integer x = 2.406642
#> Warning: non-integer x = 1.305539
#> Warning: non-integer x = 2.854257
#> Warning: non-integer x = 4.620972
#> Warning: non-integer x = 0.190850
#> Warning: non-integer x = 0.006556
#> Warning: non-integer x = 0.430921
#> Warning: non-integer x = 2.424066
#> Warning: non-integer x = 0.648051
#> Warning: non-integer x = 0.394903
#> Warning: non-integer x = 0.490428
#> Warning: non-integer x = 0.660076
#> Warning: non-integer x = 0.643602
#> Warning: non-integer x = 3.332360
#> Warning: non-integer x = 0.746748
#> Warning: non-integer x = 3.493425
#> Warning: non-integer x = 2.011750
#> Warning: non-integer x = 1.062968
#> Warning: non-integer x = 1.041452
#> Warning: non-integer x = 2.633145
#> Warning: non-integer x = 1.668394
#> Warning: non-integer x = 1.130029
#> Warning: non-integer x = 1.433188
#> Warning: non-integer x = 3.499416
#> Warning: non-integer x = 1.019835
#> Warning: non-integer x = 0.643734
#> Warning: non-integer x = 1.029089
#> Warning: non-integer x = 2.734559
#> Warning: non-integer x = 1.132430
#> Warning: non-integer x = 2.488751
#> Warning: non-integer x = 0.295572
#> Warning: non-integer x = 1.983326
#> Warning: non-integer x = 1.124016
#> Warning: non-integer x = 0.139320
#> Warning: non-integer x = 3.061716
#> Warning: non-integer x = 0.563539
#> Warning: non-integer x = 0.224183
#> Warning: non-integer x = 1.596216
#> Warning: non-integer x = 2.047000
#> Warning: non-integer x = 2.076026
#> Warning: non-integer x = 3.427378
#> Warning: non-integer x = 0.276982
#> Warning: non-integer x = 0.945020
#> Warning: non-integer x = 2.482737
#> Warning: non-integer x = 2.403164
#> Warning: non-integer x = 2.098380
#> Warning: non-integer x = 1.255098
#> Warning: non-integer x = 3.360078
#> Warning: non-integer x = 0.827513
#> Warning: non-integer x = 1.003288
#> Warning: non-integer x = 0.904053
#> Warning: non-integer x = 1.370788
#> Warning: non-integer x = 0.690967
#> Warning: non-integer x = 0.500023
#> Warning: non-integer x = 0.150457
#> Warning: non-integer x = 2.100612
#> Warning: non-integer x = 0.107250
#> Warning: non-integer x = 2.986743
#> Warning: non-integer x = 1.819897
#> Warning: non-integer x = 2.954012
#> Warning: non-integer x = 0.353382
#> Warning: non-integer x = 1.420870
#> Warning: non-integer x = 1.965417
#> Warning: non-integer x = 1.557600
#> Warning: non-integer x = 2.008025
#> Warning: non-integer x = 4.126724
#> Warning: non-integer x = 1.236225
#> Warning: non-integer x = 0.723042
#> Warning: non-integer x = 1.445957
#> Warning: non-integer x = 2.780378
#> Warning: non-integer x = 2.144138
#> Warning: non-integer x = 2.094601
#> Warning: non-integer x = 2.036871
#> Warning: non-integer x = 1.228961
#> Warning: non-integer x = 1.243108
#> Warning: non-integer x = 1.650388
#> Warning: non-integer x = 0.125398
#> Warning: non-integer x = 0.150506
#> Warning: non-integer x = 4.263305
#> Warning: non-integer x = 2.376810
#> Warning: non-integer x = 1.118008
#> Warning: non-integer x = 1.481665
#> Warning: non-integer x = 0.574610
#> Warning: non-integer x = 1.907216
#> Warning: non-integer x = 1.395284
#> Warning: non-integer x = 0.290553
#> Warning: non-integer x = 2.035763
#> Warning: non-integer x = 0.871948
#> Warning: non-integer x = 2.126007
#> Warning: non-integer x = 2.170816
#> Warning: non-integer x = 0.894923
#> Warning: non-integer x = 2.482027
#> Warning: non-integer x = 3.037056
#> Warning: non-integer x = 0.204827
#> Warning: non-integer x = 0.148179
#> Warning: non-integer x = 2.758699
#> Warning: non-integer x = 0.238443
#> Warning: non-integer x = 2.539839
#> Warning: non-integer x = 0.115540
#> Warning: non-integer x = 1.094852
#> Warning: non-integer x = 1.149376
#> Warning: non-integer x = 1.319986
#> Warning: non-integer x = 0.596255
#> Warning: non-integer x = 0.241480
#> Warning: non-integer x = 0.914241
#> Warning: non-integer x = 0.003848
#> Warning: non-integer x = 2.029055
#> Warning: non-integer x = 2.281729
#> Warning: non-integer x = 1.574156
#> Warning: non-integer x = 0.762058
#> Warning: non-integer x = 0.577460
#> Warning: non-integer x = 0.468860
#> Warning: non-integer x = 1.234397
#> Warning: non-integer x = 3.202792
#> Warning: non-integer x = 1.559782
#> Warning: non-integer x = 5.257248
#> Warning: non-integer x = 0.603996
#> Warning: non-integer x = 3.045209
#> Warning: non-integer x = 2.129580
#> Warning: non-integer x = 1.918585
#> Warning: non-integer x = 0.402296
#> Warning: non-integer x = 0.054409
#> Warning: non-integer x = 0.201424
#> Warning: non-integer x = 2.581235
#> Warning: non-integer x = 1.745840
#> Warning: non-integer x = 2.171295
#> Warning: non-integer x = 0.055831
#> Warning: non-integer x = 0.381680
#> Warning: non-integer x = 0.064339
#> Warning: non-integer x = 2.343445
#> Warning: non-integer x = 1.208433
#> Warning: non-integer x = 2.089363
#> Warning: non-integer x = 3.075925
#> Warning: non-integer x = 0.689578
#> Warning: non-integer x = 1.173438
#> Warning: non-integer x = 2.274335
#> Warning: non-integer x = 0.971421
#> Warning: non-integer x = 1.574174
#> Warning: non-integer x = 0.326803
#> Warning: non-integer x = 0.070206
#> Warning: non-integer x = 0.459925
#> Warning: non-integer x = 1.422923
#> Warning: non-integer x = 1.626170
#> Warning: non-integer x = 1.043904
#> Warning: non-integer x = 2.089644
#> Warning: non-integer x = 0.693284
#> Warning: non-integer x = 1.766820
#> Warning: non-integer x = 0.072311
#> Warning: non-integer x = 2.228040
#> Warning: non-integer x = 2.006879
#> Warning: non-integer x = 0.946628
#> Warning: non-integer x = 1.833376
#> Warning: non-integer x = 1.446759
#> Warning: non-integer x = 1.346860
#> Warning: non-integer x = 1.717792
#> Warning: non-integer x = 0.159639
#> Warning: non-integer x = 1.690238
#> Warning: non-integer x = 0.102139
#> Warning: non-integer x = 2.134269
#> Warning: non-integer x = 0.241887
#> Warning: non-integer x = 1.499085
#> Warning: non-integer x = 0.588615
#> Warning: non-integer x = 1.042702
#> Warning: non-integer x = 1.781718
#> Warning: non-integer x = 2.147242
#> Warning: non-integer x = 2.100632
#> Warning: non-integer x = 1.900295
#> Warning: non-integer x = 0.729688
#> Warning: non-integer x = 0.717022
#> Warning: non-integer x = 0.583932
#> Warning: non-integer x = 2.167473
#> Warning: non-integer x = 0.276015
#> Warning: non-integer x = 1.788824
#> Warning: non-integer x = 0.158631
#> Warning: non-integer x = 1.585191
#> Warning: non-integer x = 0.946087
#> Warning: non-integer x = 2.867072
#> Warning: non-integer x = 0.908348
#> Warning: non-integer x = 0.448787
#> Warning: non-integer x = 0.750410
#> Warning: non-integer x = 1.815585
#> Warning: non-integer x = 2.141799
#> Warning: non-integer x = 4.645345
#> Warning: non-integer x = 0.420774
#> Warning: non-integer x = 0.174070
#> Warning: non-integer x = 0.532204
#> Warning: non-integer x = 4.055354
#> Warning: non-integer x = 0.976000
#> Warning: non-integer x = 1.725340
#> Warning: non-integer x = 2.305219
#> Warning: non-integer x = 1.403758
#> Warning: non-integer x = 1.765778
#> Warning: non-integer x = 2.804092
#> Warning: non-integer x = 0.631249
#> Warning: non-integer x = 0.446788
#> Warning: non-integer x = 1.327430
#> Warning: non-integer x = 0.915604
#> Warning: non-integer x = 1.005456
#> Warning: non-integer x = 1.682382
#> Warning: non-integer x = 2.133741
#> Warning: non-integer x = 0.386705
#> Warning: non-integer x = 1.339587
#> Warning: non-integer x = 1.151457
#> Warning: non-integer x = 3.173308
#> Warning: non-integer x = 2.190908
#> Warning: non-integer x = 1.630440
#> Warning: non-integer x = 0.893896
#> Warning: non-integer x = 1.550546
#> Warning: non-integer x = 0.135513
#> Warning: non-integer x = 3.109886
#> Warning: non-integer x = 0.181733
#> Warning: non-integer x = 2.254085
#> Warning: non-integer x = 0.497604
#> Warning: non-integer x = 2.738449
#> Warning: non-integer x = 0.395698
#> Warning: non-integer x = 0.911650
#> Warning: non-integer x = 0.108125
#> Warning: non-integer x = 2.996734
#> Warning: non-integer x = 1.855366
#> Warning: non-integer x = 0.596298
#> Warning: non-integer x = 1.594994
#> Warning: non-integer x = 2.639849
#> Warning: non-integer x = 2.612031
#> Warning: non-integer x = 0.988816
#> Warning: non-integer x = 1.054764
#> Warning: non-integer x = 2.230330
#> Warning: non-integer x = 1.141866
#> Warning: non-integer x = 2.182430
#> Warning: non-integer x = 1.107260
#> Warning: non-integer x = 2.308737
#> Warning: non-integer x = 0.501994
#> Warning: non-integer x = 0.283168
#> Warning: non-integer x = 1.371597
#> Warning: non-integer x = 2.410196
#> Warning: non-integer x = 0.049449
#> Warning: non-integer x = 0.216871
#> Warning: non-integer x = 1.268439
#> Warning: non-integer x = 1.530453
#> Warning: non-integer x = 2.592463
#> Warning: non-integer x = 2.163751
#> Warning: non-integer x = 0.790024
#> Warning: non-integer x = 0.559712
#> Warning: non-integer x = 0.482290
#> Warning: non-integer x = 1.112456
#> Warning: non-integer x = 0.462063
#> Warning: non-integer x = 0.097179
#> Warning: non-integer x = 0.109403
#> Warning: non-integer x = 0.915785
#> Warning: non-integer x = 1.776138
#> Warning: non-integer x = 1.215954
#> Warning: non-integer x = 0.690504
#> Warning: non-integer x = 2.357387
#> Warning: non-integer x = 0.361157
#> Warning: non-integer x = 0.574936
#> Warning: non-integer x = 0.325889
#> Warning: non-integer x = 0.438210
#> Warning: non-integer x = 2.455100
#> Warning: non-integer x = 0.788932
#> Warning: non-integer x = 2.114647
#> Warning: non-integer x = 1.576438
#> Warning: non-integer x = 0.726915
#> Warning: non-integer x = 1.254978
#> Warning: non-integer x = 1.386729
#> Warning: non-integer x = 0.204997
#> Warning: non-integer x = 0.459273
#> Warning: non-integer x = 2.991690
#> Warning: non-integer x = 0.669587
#> Warning: non-integer x = 2.246800
#> Warning: non-integer x = 2.962383
#> Warning: non-integer x = 1.224591
#> Warning: non-integer x = 4.148183
#> Warning: non-integer x = 0.343824
#> Warning: non-integer x = 0.373380
#> Warning: non-integer x = 1.411453
#> Warning: non-integer x = 1.733880
#> Warning: non-integer x = 4.486444
#> Warning: non-integer x = 0.649315
#> Warning: non-integer x = 1.859237
#> Warning: non-integer x = 1.744794
#> Warning: non-integer x = 1.023359
#> Warning: non-integer x = 2.392206
#> Warning: non-integer x = 1.366059
#> Warning: non-integer x = 1.657406
#> Warning: non-integer x = 0.287523
#> Warning: non-integer x = 1.665629
#> Warning: non-integer x = 3.314711
#> Warning: non-integer x = 0.382466
#> Warning: non-integer x = 0.949405
#> Warning: non-integer x = 1.385666
#> Warning: non-integer x = 0.666743
#> Warning: non-integer x = 1.560140
#> Warning: non-integer x = 1.279030
#> Warning: non-integer x = 0.696733
#> Warning: non-integer x = 2.409776
#> Warning: non-integer x = 0.363178
#> Warning: non-integer x = 0.604810
#> Warning: non-integer x = 0.227346
#> Warning: non-integer x = 1.145751
#> Warning: non-integer x = 3.017276
#> Warning: non-integer x = 1.594331
#> Warning: non-integer x = 0.732195
#> Warning: non-integer x = 0.314508
#> Warning: non-integer x = 1.788125
#> Warning: non-integer x = 0.667764
#> Warning: non-integer x = 2.590830
#> Warning: non-integer x = 0.676616
#> Warning: non-integer x = 0.774107
#> Warning: non-integer x = 1.432190
#> Warning: non-integer x = 2.449886
#> Warning: non-integer x = 1.118475
#> Warning: non-integer x = 1.562259
#> Warning: non-integer x = 0.857220
#> Warning: non-integer x = 1.951822
#> Warning: non-integer x = 3.459265
#> Warning: non-integer x = 3.095314
#> Warning: non-integer x = 0.392289
#> Warning: non-integer x = 0.939551
#> Warning: non-integer x = 0.615461
#> Warning: non-integer x = 0.487041
#> Warning: non-integer x = 0.905177
#> Warning: non-integer x = 0.728024
#> Warning: non-integer x = 1.191790
#> Warning: non-integer x = 1.651050
#> Warning: non-integer x = 0.780523
#> Warning: non-integer x = 0.943954
#> Warning: non-integer x = 1.190601
#> Warning: non-integer x = 1.091146
#> Warning: non-integer x = 1.156776
#> Warning: non-integer x = 0.646251
#> Warning: non-integer x = 1.127208
#> Warning: non-integer x = 0.869271
#> Warning: non-integer x = 1.344103
#> Warning: non-integer x = 0.079461
#> Warning: non-integer x = 0.527641
#> Warning: non-integer x = 0.795330
#> Warning: non-integer x = 0.632274
#> Warning: non-integer x = 1.506164
#> Warning: non-integer x = 0.328471
#> Warning: non-integer x = 0.928112
#> Warning: non-integer x = 2.562325
#> Warning: non-integer x = 1.833297
#> Warning: non-integer x = 1.263035
#> Warning: non-integer x = 1.920935
#> Warning: non-integer x = 2.451954
#> Warning: non-integer x = 2.716450
#> Warning: non-integer x = 0.048445
#> Warning: non-integer x = 0.272433
#> Warning: non-integer x = 1.991638
#> Warning: non-integer x = 0.399927
#> Warning: non-integer x = 0.372415
#> Warning: non-integer x = 2.791670
#> Warning: non-integer x = 0.026898
#> Warning: non-integer x = 3.155968
#> Warning: non-integer x = 1.981696
#> Warning: non-integer x = 2.608990
#> Warning: non-integer x = 0.138344
#> Warning: non-integer x = 1.189958
#> Warning: non-integer x = 0.563677
#> Warning: non-integer x = 2.300091
#> Warning: non-integer x = 1.574668
#> Warning: non-integer x = 1.815988
#> Warning: non-integer x = 1.047997
#> Warning: non-integer x = 0.865031
#> Warning: non-integer x = 1.487404
#> Warning: non-integer x = 0.665289
#> Warning: non-integer x = 1.532450
#> Warning: non-integer x = 0.009591
#> Warning: non-integer x = 1.270716
#> Warning: non-integer x = 1.198320
#> Warning: non-integer x = 2.230588
#> Warning: non-integer x = 1.664910
#> Warning: non-integer x = 2.267870
#> Warning: non-integer x = 0.717481
#> Warning: non-integer x = 1.086395
#> Warning: non-integer x = 0.560412
#> Warning: non-integer x = 3.157263
#> Warning: non-integer x = 0.619656
#> Warning: non-integer x = 0.182535
#> Warning: non-integer x = 0.710459
#> Warning: non-integer x = 1.032679
#> Warning: non-integer x = 2.446381
#> Warning: non-integer x = 1.064125
#> Warning: non-integer x = 1.896983
#> Warning: non-integer x = 0.098592
#> Warning: non-integer x = 1.101768
#> Warning: non-integer x = 0.328410
#> Warning: non-integer x = 1.120611
#> Warning: non-integer x = 0.479526
#> Warning: non-integer x = 2.562446
#> Warning: non-integer x = 1.322032
#> Warning: non-integer x = 0.087328
#> Warning: non-integer x = 2.537643
#> Warning: non-integer x = 1.118248
#> Warning: non-integer x = 2.620006
#> Warning: non-integer x = 0.588435
#> Warning: non-integer x = 0.626813
#> Warning: non-integer x = 1.682844
#> Warning: non-integer x = 0.597646
#> Warning: non-integer x = 1.069385
#> Warning: non-integer x = 0.109369
#> Warning: non-integer x = 1.556514
#> Warning: non-integer x = 0.659882
#> Warning: non-integer x = 1.941397
#> Warning: non-integer x = 1.119960
#> Warning: non-integer x = 3.069513
#> Warning: non-integer x = 1.390502
#> Warning: non-integer x = 0.562669
#> Warning: non-integer x = 3.221992
#> Warning: non-integer x = 1.737148
#> Warning: non-integer x = 0.249969
#> Warning: non-integer x = 1.506614
#> Warning: non-integer x = 0.512038
#> Warning: non-integer x = 1.298425
#> Warning: non-integer x = 0.012027
#> Warning: non-integer x = 0.055995
#> Warning: non-integer x = 1.626048
#> Warning: non-integer x = 1.765296
#> Warning: non-integer x = 0.566843
#> Warning: non-integer x = 1.795031
#> Warning: non-integer x = 2.862501
#> Warning: non-integer x = 0.187786
#> Warning: non-integer x = 0.297795
#> Warning: non-integer x = 2.228554
#> Warning: non-integer x = 0.522203
#> Warning: non-integer x = 1.738018
#> Warning: non-integer x = 2.088390
#> Warning: non-integer x = 0.046393
#> Warning: non-integer x = 1.508353
#> Warning: non-integer x = 0.355617
#> Warning: non-integer x = 2.805097
#> Warning: non-integer x = 1.539645
#> Warning: non-integer x = 1.004601
#> Warning: non-integer x = 0.202821
#> Warning: non-integer x = 0.752635
#> Warning: non-integer x = 0.369305
#> Warning: non-integer x = 1.613077
#> Warning: non-integer x = 3.378457
#> Warning: non-integer x = 0.766802
#> Warning: non-integer x = 0.999495
#> Warning: non-integer x = 1.695557
#> Warning: non-integer x = 0.633537
#> Warning: non-integer x = 0.851088
#> Warning: non-integer x = 0.226135
#> Warning: non-integer x = 0.267665
#> Warning: non-integer x = 2.898263
#> Warning: non-integer x = 0.318055
#> Warning: non-integer x = 3.490253
#> Warning: non-integer x = 0.149503
#> Warning: non-integer x = 0.754783
#> Warning: non-integer x = 1.466642
#> Warning: non-integer x = 1.122062
#> Warning: non-integer x = 0.646658
#> Warning: non-integer x = 0.521367
#> Warning: non-integer x = 0.330009
#> Warning: non-integer x = 1.944209
#> Warning: non-integer x = 2.178994
#> Warning: non-integer x = 1.654037
#> Warning: non-integer x = 0.617051
#> Warning: non-integer x = 1.262375
#> Warning: non-integer x = 1.025125
#> Warning: non-integer x = 0.947020
#> Warning: non-integer x = 1.703455
#> Warning: non-integer x = 1.629886
#> Warning: non-integer x = 1.062239
#> Warning: non-integer x = 1.696799
#> Warning: non-integer x = 1.843511
#> Warning: non-integer x = 1.946373
#> Warning: non-integer x = 0.974275
#> Warning: non-integer x = 0.383907
#> Warning: non-integer x = 2.625029
#> Warning: non-integer x = 1.999332
#> Warning: non-integer x = 2.567718
#> Warning: non-integer x = 2.168494
#> Warning: non-integer x = 2.326779
#> Warning: non-integer x = 2.535477
#> Warning: non-integer x = 0.250950
#> Warning: non-integer x = 1.780447
#> Warning: non-integer x = 1.028938
#> Warning: non-integer x = 1.772286
#> Warning: non-integer x = 0.300154
#> Warning: non-integer x = 1.158293
#> Warning: non-integer x = 1.150026
#> Warning: non-integer x = 2.025538
#> Warning: non-integer x = 2.641500
#> Warning: non-integer x = 2.811350
#> Warning: non-integer x = 0.110554
#> Warning: non-integer x = 0.907039
#> Warning: non-integer x = 0.086117
#> Warning: non-integer x = 0.267792
#> Warning: non-integer x = 0.095132
#> Warning: non-integer x = 0.770269
#> Warning: non-integer x = 1.334045
#> Warning: non-integer x = 1.053571
#> Warning: non-integer x = 0.776330
#> Warning: non-integer x = 1.937181
#> Warning: non-integer x = 0.503973
#> Warning: non-integer x = 1.818752
#> Warning: non-integer x = 3.303737
#> Warning: non-integer x = 2.398305
#> Warning: non-integer x = 1.040079
#> Warning: non-integer x = 1.268211
#> Warning: non-integer x = 1.177220
#> Warning: non-integer x = 1.787042
#> Warning: non-integer x = 0.346939
#> Warning: non-integer x = 1.617805
#> Warning: non-integer x = 1.201106
#> Warning: non-integer x = 1.220315
#> Warning: non-integer x = 0.768076
#> Warning: non-integer x = 1.572339
#> Warning: non-integer x = 2.343242
#> Warning: non-integer x = 1.187671
#> Warning: non-integer x = 0.182626
#> Warning: non-integer x = 0.813455
#> Warning: non-integer x = 2.046070
#> Warning: non-integer x = 1.112057
#> Warning: non-integer x = 3.089564
#> Warning: non-integer x = 1.163338
#> Warning: non-integer x = 0.854381
#> Warning: non-integer x = 1.577361
#> Warning: non-integer x = 1.915039
#> Warning: non-integer x = 0.403174
#> Warning: non-integer x = 0.357402
#> Warning: non-integer x = 2.527101
#> Warning: non-integer x = 0.683566
#> Warning: non-integer x = 1.999973
#> Warning: non-integer x = 0.690475
#> Warning: non-integer x = 3.025501
#> Warning: non-integer x = 0.383880
#> Warning: non-integer x = 1.967384
#> Warning: non-integer x = 0.444272
#> Warning: non-integer x = 1.034921
#> Warning: non-integer x = 0.246822
#> Warning: non-integer x = 0.070593
#> Warning: non-integer x = 1.223720
#> Warning: non-integer x = 0.368910
#> Warning: non-integer x = 0.409553
#> Warning: non-integer x = 3.771856
#> Warning: non-integer x = 0.714882
#> Warning: non-integer x = 1.089232
#> Warning: non-integer x = 0.608521
#> Warning: non-integer x = 2.909095
#> Warning: non-integer x = 2.544858
#> Warning: non-integer x = 3.802301
#> Warning: non-integer x = 0.349157
#> Warning: non-integer x = 3.056925
#> Warning: non-integer x = 0.380541
#> Warning: non-integer x = 0.779466
#> Warning: non-integer x = 0.171833
#> Warning: non-integer x = 3.458172
#> Warning: non-integer x = 0.047184
#> Warning: non-integer x = 1.577260
#> Warning: non-integer x = 3.567189
#> Warning: non-integer x = 1.024277
#> Warning: non-integer x = 2.280384
#> Warning: non-integer x = 0.994133
#> Warning: non-integer x = 0.202515
#> Warning: non-integer x = 1.491823
#> Warning: non-integer x = 1.077487
#> Warning: non-integer x = 2.050959
#> Warning: non-integer x = 0.509926
#> Warning: non-integer x = 0.043348
#> Warning: non-integer x = 0.485222
#> Warning: non-integer x = 2.329246
#> Warning: non-integer x = 1.016805
#> Warning: non-integer x = 0.271477
#> Warning: non-integer x = 1.746880
#> Warning: non-integer x = 0.542683
#> Warning: non-integer x = 0.116509
#> Warning: non-integer x = 2.884260
#> Warning: non-integer x = 0.025671
#> Warning: non-integer x = 0.786695
#> Warning: non-integer x = 1.595626
#> Warning: non-integer x = 0.163223
#> Warning: non-integer x = 0.480860
#> Warning: non-integer x = 0.760070
#> Warning: non-integer x = 1.787195
#> Warning: non-integer x = 0.757874
#> Warning: non-integer x = 1.103751
#> Warning: non-integer x = 0.748391
#> Warning: non-integer x = 2.414660
#> Warning: non-integer x = 0.811375
#> Warning: non-integer x = 0.344388
#> Warning: non-integer x = 0.124635
#> Warning: non-integer x = 0.509214
#> Warning: non-integer x = 2.788439
#> Warning: non-integer x = 0.944958
#> Warning: non-integer x = 0.830676
#> Warning: non-integer x = 3.782399
#> Warning: non-integer x = 1.138699
#> Warning: non-integer x = 0.528293
#> Warning: non-integer x = 1.127430
#> Warning: non-integer x = 0.308487
#> Warning: non-integer x = 1.093721
#> Warning: non-integer x = 1.479530
#> Warning: non-integer x = 1.813268
#> Warning: non-integer x = 1.387828
#> Warning: non-integer x = 0.896277
#> Warning: non-integer x = 0.325687
#> Warning: non-integer x = 1.435432
#> Warning: non-integer x = 0.800625
#> Warning: non-integer x = 0.087676
#> Warning: non-integer x = 1.071352
#> Warning: non-integer x = 1.044131
#> Warning: non-integer x = 1.991322
#> Warning: non-integer x = 0.953577
#> Warning: non-integer x = 0.270697
#> Warning: non-integer x = 0.443195
#> Warning: non-integer x = 0.831672
#> Warning: non-integer x = 0.257994
#> Warning: non-integer x = 3.485818
#> Warning: non-integer x = 4.275861
#> Warning: non-integer x = 1.345119
#> Warning: non-integer x = 0.690380
#> Warning: non-integer x = 1.481365
#> Warning: non-integer x = 0.341976
#> Warning: non-integer x = 1.906183
#> Warning: non-integer x = 1.805369
#> Warning: non-integer x = 2.857953
#> Warning: non-integer x = 0.475569
#> Warning: non-integer x = 1.786206
#> Warning: non-integer x = 1.950395
#> Warning: non-integer x = 2.109863
#> Warning: non-integer x = 0.307865
#> Warning: non-integer x = 1.270560
#> Warning: non-integer x = 1.785453
#> Warning: non-integer x = 0.837892
#> Warning: non-integer x = 0.626085
#> Warning: non-integer x = 2.258224
#> Warning: non-integer x = 0.904575