Estimate the Individual Treatment Effect using Poisson Regression

Estimates the Individual Treatment Effect using Poisson Regression given a response vector, a treatment vector, and a features matrix.

estimate_ite_poisson(y, z, X, X_names, include_offset, offset_name)

Arguments

y
the observed response vector

z
the treatment vector

X
the features matrix

X_names
the names of the covariates

include_offset
whether or not to include an offset when estimating the ITE, for poisson only

offset_name
the name of the offset, if it is to be included

Value

a vector of ITE estimates

Examples

dataset <- generate_cre_dataset(n = 1000, rho = 0, n_rules = 2, p = 10,
                                effect_size = 2, binary = FALSE)

# Initialize parameters
y <- abs(dataset[["y"]])
z <- dataset[["z"]]
X <- as.data.frame(dataset[["X"]])
X_names <- names(as.data.frame(X))
include_offset <- FALSE
offset_name <- NA

ite_list <- estimate_ite_poisson(y, z, X, X_names, include_offset, offset_name)

#> Warning: non-integer x = 0.818411

#> Warning: non-integer x = 1.812466

#> Warning: non-integer x = 0.093313

#> Warning: non-integer x = 3.319376

#> Warning: non-integer x = 2.566639

#> Warning: non-integer x = 0.834272

#> Warning: non-integer x = 3.738203

#> Warning: non-integer x = 1.988186

#> Warning: non-integer x = 0.213560

#> Warning: non-integer x = 2.717304

#> Warning: non-integer x = 0.228226

#> Warning: non-integer x = 0.918080

#> Warning: non-integer x = 2.769608

#> Warning: non-integer x = 0.606496

#> Warning: non-integer x = 3.576929

#> Warning: non-integer x = 2.158823

#> Warning: non-integer x = 2.110007

#> Warning: non-integer x = 0.839163

#> Warning: non-integer x = 3.434614

#> Warning: non-integer x = 4.116503

#> Warning: non-integer x = 0.280322

#> Warning: non-integer x = 0.072475

#> Warning: non-integer x = 1.403409

#> Warning: non-integer x = 1.989214

#> Warning: non-integer x = 3.435644

#> Warning: non-integer x = 0.983370

#> Warning: non-integer x = 2.262143

#> Warning: non-integer x = 2.537533

#> Warning: non-integer x = 3.562237

#> Warning: non-integer x = 0.259621

#> Warning: non-integer x = 3.951872

#> Warning: non-integer x = 1.947020

#> Warning: non-integer x = 1.324924

#> Warning: non-integer x = 2.175724

#> Warning: non-integer x = 1.223495

#> Warning: non-integer x = 2.129889

#> Warning: non-integer x = 0.803949

#> Warning: non-integer x = 5.193881

#> Warning: non-integer x = 1.138707

#> Warning: non-integer x = 2.068021

#> Warning: non-integer x = 0.413811

#> Warning: non-integer x = 0.453726

#> Warning: non-integer x = 2.485264

#> Warning: non-integer x = 3.388139

#> Warning: non-integer x = 1.507596

#> Warning: non-integer x = 2.309589

#> Warning: non-integer x = 0.397696

#> Warning: non-integer x = 1.570106

#> Warning: non-integer x = 0.159514

#> Warning: non-integer x = 2.937083

#> Warning: non-integer x = 2.471248

#> Warning: non-integer x = 3.289057

#> Warning: non-integer x = 2.339581

#> Warning: non-integer x = 4.575198

#> Warning: non-integer x = 1.001921

#> Warning: non-integer x = 4.083683

#> Warning: non-integer x = 0.859258

#> Warning: non-integer x = 0.197943

#> Warning: non-integer x = 2.906376

#> Warning: non-integer x = 1.514963

#> Warning: non-integer x = 2.002341

#> Warning: non-integer x = 1.607759

#> Warning: non-integer x = 3.126246

#> Warning: non-integer x = 0.323460

#> Warning: non-integer x = 4.810284

#> Warning: non-integer x = 1.871289

#> Warning: non-integer x = 2.925773

#> Warning: non-integer x = 3.497463

#> Warning: non-integer x = 0.616539

#> Warning: non-integer x = 0.468132

#> Warning: non-integer x = 1.422529

#> Warning: non-integer x = 3.202016

#> Warning: non-integer x = 1.356831

#> Warning: non-integer x = 1.816067

#> Warning: non-integer x = 2.533606

#> Warning: non-integer x = 2.961672

#> Warning: non-integer x = 1.732505

#> Warning: non-integer x = 0.768884

#> Warning: non-integer x = 1.140347

#> Warning: non-integer x = 0.948110

#> Warning: non-integer x = 2.212917

#> Warning: non-integer x = 2.312109

#> Warning: non-integer x = 2.277845

#> Warning: non-integer x = 0.503051

#> Warning: non-integer x = 2.808169

#> Warning: non-integer x = 1.451886

#> Warning: non-integer x = 1.956129

#> Warning: non-integer x = 1.983395

#> Warning: non-integer x = 1.428236

#> Warning: non-integer x = 1.994187

#> Warning: non-integer x = 0.469340

#> Warning: non-integer x = 3.191063

#> Warning: non-integer x = 0.491003

#> Warning: non-integer x = 0.029013

#> Warning: non-integer x = 1.045782

#> Warning: non-integer x = 0.605396

#> Warning: non-integer x = 0.187501

#> Warning: non-integer x = 1.682868

#> Warning: non-integer x = 1.391128

#> Warning: non-integer x = 1.785368

#> Warning: non-integer x = 2.085631

#> Warning: non-integer x = 3.604904

#> Warning: non-integer x = 0.419589

#> Warning: non-integer x = 2.836422

#> Warning: non-integer x = 3.910611

#> Warning: non-integer x = 3.327742

#> Warning: non-integer x = 3.587452

#> Warning: non-integer x = 2.553737

#> Warning: non-integer x = 1.029961

#> Warning: non-integer x = 0.309123

#> Warning: non-integer x = 1.246233

#> Warning: non-integer x = 4.784148

#> Warning: non-integer x = 3.332936

#> Warning: non-integer x = 0.636984

#> Warning: non-integer x = 3.519101

#> Warning: non-integer x = 0.475809

#> Warning: non-integer x = 1.580263

#> Warning: non-integer x = 1.573953

#> Warning: non-integer x = 2.216663

#> Warning: non-integer x = 0.318501

#> Warning: non-integer x = 0.736040

#> Warning: non-integer x = 0.892893

#> Warning: non-integer x = 1.710014

#> Warning: non-integer x = 1.172077

#> Warning: non-integer x = 2.885347

#> Warning: non-integer x = 0.012991

#> Warning: non-integer x = 3.083811

#> Warning: non-integer x = 1.245382

#> Warning: non-integer x = 2.217484

#> Warning: non-integer x = 3.423426

#> Warning: non-integer x = 3.170595

#> Warning: non-integer x = 2.233223

#> Warning: non-integer x = 0.131132

#> Warning: non-integer x = 0.237590

#> Warning: non-integer x = 0.155961

#> Warning: non-integer x = 0.728880

#> Warning: non-integer x = 0.077219

#> Warning: non-integer x = 1.284575

#> Warning: non-integer x = 0.754372

#> Warning: non-integer x = 4.647799

#> Warning: non-integer x = 1.761853

#> Warning: non-integer x = 1.506507

#> Warning: non-integer x = 0.459397

#> Warning: non-integer x = 2.173641

#> Warning: non-integer x = 0.896577

#> Warning: non-integer x = 0.934926

#> Warning: non-integer x = 0.158511

#> Warning: non-integer x = 2.171720

#> Warning: non-integer x = 0.170975

#> Warning: non-integer x = 0.148599

#> Warning: non-integer x = 1.522332

#> Warning: non-integer x = 1.469062

#> Warning: non-integer x = 2.244809

#> Warning: non-integer x = 2.006615

#> Warning: non-integer x = 0.385973

#> Warning: non-integer x = 1.488713

#> Warning: non-integer x = 0.002711

#> Warning: non-integer x = 2.433818

#> Warning: non-integer x = 0.386286

#> Warning: non-integer x = 0.367739

#> Warning: non-integer x = 1.307309

#> Warning: non-integer x = 1.448248

#> Warning: non-integer x = 2.264570

#> Warning: non-integer x = 0.177854

#> Warning: non-integer x = 5.405956

#> Warning: non-integer x = 2.087977

#> Warning: non-integer x = 0.261648

#> Warning: non-integer x = 3.551993

#> Warning: non-integer x = 2.785348

#> Warning: non-integer x = 0.866881

#> Warning: non-integer x = 2.626071

#> Warning: non-integer x = 3.080203

#> Warning: non-integer x = 0.566482

#> Warning: non-integer x = 5.212566

#> Warning: non-integer x = 1.525649

#> Warning: non-integer x = 1.839953

#> Warning: non-integer x = 2.004865

#> Warning: non-integer x = 0.157148

#> Warning: non-integer x = 0.690596

#> Warning: non-integer x = 1.629547

#> Warning: non-integer x = 0.933308

#> Warning: non-integer x = 1.394362

#> Warning: non-integer x = 2.735029

#> Warning: non-integer x = 5.171125

#> Warning: non-integer x = 2.376218

#> Warning: non-integer x = 3.717227

#> Warning: non-integer x = 2.153490

#> Warning: non-integer x = 0.861104

#> Warning: non-integer x = 3.030326

#> Warning: non-integer x = 1.347695

#> Warning: non-integer x = 2.580659

#> Warning: non-integer x = 0.128977

#> Warning: non-integer x = 0.104650

#> Warning: non-integer x = 4.786783

#> Warning: non-integer x = 2.902190

#> Warning: non-integer x = 1.456386

#> Warning: non-integer x = 2.404385

#> Warning: non-integer x = 1.231140

#> Warning: non-integer x = 3.545347

#> Warning: non-integer x = 0.340856

#> Warning: non-integer x = 3.734082

#> Warning: non-integer x = 0.694342

#> Warning: non-integer x = 0.053060

#> Warning: non-integer x = 4.316786

#> Warning: non-integer x = 3.286881

#> Warning: non-integer x = 1.986769

#> Warning: non-integer x = 2.031459

#> Warning: non-integer x = 2.334733

#> Warning: non-integer x = 2.745043

#> Warning: non-integer x = 3.647433

#> Warning: non-integer x = 0.811549

#> Warning: non-integer x = 1.625091

#> Warning: non-integer x = 2.592330

#> Warning: non-integer x = 1.986787

#> Warning: non-integer x = 0.212307

#> Warning: non-integer x = 2.564241

#> Warning: non-integer x = 1.383548

#> Warning: non-integer x = 0.927896

#> Warning: non-integer x = 0.348933

#> Warning: non-integer x = 3.442917

#> Warning: non-integer x = 1.339018

#> Warning: non-integer x = 1.715827

#> Warning: non-integer x = 0.874511

#> Warning: non-integer x = 1.600601

#> Warning: non-integer x = 1.206879

#> Warning: non-integer x = 3.242256

#> Warning: non-integer x = 2.807917

#> Warning: non-integer x = 2.137505

#> Warning: non-integer x = 2.995029

#> Warning: non-integer x = 0.815391

#> Warning: non-integer x = 1.493979

#> Warning: non-integer x = 3.124958

#> Warning: non-integer x = 2.419797

#> Warning: non-integer x = 0.118601

#> Warning: non-integer x = 1.246329

#> Warning: non-integer x = 0.409860

#> Warning: non-integer x = 2.573984

#> Warning: non-integer x = 1.559795

#> Warning: non-integer x = 2.353325

#> Warning: non-integer x = 0.259734

#> Warning: non-integer x = 0.795927

#> Warning: non-integer x = 0.265461

#> Warning: non-integer x = 2.383701

#> Warning: non-integer x = 3.097023

#> Warning: non-integer x = 0.785164

#> Warning: non-integer x = 1.014707

#> Warning: non-integer x = 0.406071

#> Warning: non-integer x = 0.129883

#> Warning: non-integer x = 3.728631

#> Warning: non-integer x = 2.817090

#> Warning: non-integer x = 1.392950

#> Warning: non-integer x = 0.925181

#> Warning: non-integer x = 0.749369

#> Warning: non-integer x = 0.294196

#> Warning: non-integer x = 1.404482

#> Warning: non-integer x = 1.938280

#> Warning: non-integer x = 1.036347

#> Warning: non-integer x = 1.126986

#> Warning: non-integer x = 2.382690

#> Warning: non-integer x = 0.367686

#> Warning: non-integer x = 2.231953

#> Warning: non-integer x = 2.248497

#> Warning: non-integer x = 1.534295

#> Warning: non-integer x = 1.215637

#> Warning: non-integer x = 0.477665

#> Warning: non-integer x = 2.463973

#> Warning: non-integer x = 0.451808

#> Warning: non-integer x = 0.875981

#> Warning: non-integer x = 0.681863

#> Warning: non-integer x = 1.107367

#> Warning: non-integer x = 2.014463

#> Warning: non-integer x = 0.537305

#> Warning: non-integer x = 2.696728

#> Warning: non-integer x = 3.985693

#> Warning: non-integer x = 0.447997

#> Warning: non-integer x = 0.807898

#> Warning: non-integer x = 2.092313

#> Warning: non-integer x = 1.375810

#> Warning: non-integer x = 1.436025

#> Warning: non-integer x = 0.957740

#> Warning: non-integer x = 0.058908

#> Warning: non-integer x = 2.217981

#> Warning: non-integer x = 2.203586

#> Warning: non-integer x = 3.737975

#> Warning: non-integer x = 1.623179

#> Warning: non-integer x = 3.040426

#> Warning: non-integer x = 0.178260

#> Warning: non-integer x = 4.042507

#> Warning: non-integer x = 1.494262

#> Warning: non-integer x = 0.428297

#> Warning: non-integer x = 0.100153

#> Warning: non-integer x = 1.182496

#> Warning: non-integer x = 0.497527

#> Warning: non-integer x = 1.344388

#> Warning: non-integer x = 3.095897

#> Warning: non-integer x = 1.402336

#> Warning: non-integer x = 0.306643

#> Warning: non-integer x = 0.870917

#> Warning: non-integer x = 0.968391

#> Warning: non-integer x = 1.271602

#> Warning: non-integer x = 2.289963

#> Warning: non-integer x = 0.588482

#> Warning: non-integer x = 2.580785

#> Warning: non-integer x = 0.930694

#> Warning: non-integer x = 1.883646

#> Warning: non-integer x = 1.131157

#> Warning: non-integer x = 3.680368

#> Warning: non-integer x = 0.946079

#> Warning: non-integer x = 2.091903

#> Warning: non-integer x = 0.246024

#> Warning: non-integer x = 1.636454

#> Warning: non-integer x = 0.278529

#> Warning: non-integer x = 1.494308

#> Warning: non-integer x = 4.266889

#> Warning: non-integer x = 3.072262

#> Warning: non-integer x = 0.040138

#> Warning: non-integer x = 0.644444

#> Warning: non-integer x = 2.100080

#> Warning: non-integer x = 0.620024

#> Warning: non-integer x = 1.371585

#> Warning: non-integer x = 1.313662

#> Warning: non-integer x = 1.332009

#> Warning: non-integer x = 1.518654

#> Warning: non-integer x = 0.123272

#> Warning: non-integer x = 1.678836

#> Warning: non-integer x = 1.956883

#> Warning: non-integer x = 0.668926

#> Warning: non-integer x = 1.566374

#> Warning: non-integer x = 0.669178

#> Warning: non-integer x = 1.095853

#> Warning: non-integer x = 1.900194

#> Warning: non-integer x = 1.191659

#> Warning: non-integer x = 1.719460

#> Warning: non-integer x = 2.739960

#> Warning: non-integer x = 0.763239

#> Warning: non-integer x = 2.188662

#> Warning: non-integer x = 0.566342

#> Warning: non-integer x = 1.075696

#> Warning: non-integer x = 2.104756

#> Warning: non-integer x = 0.574034

#> Warning: non-integer x = 2.901400

#> Warning: non-integer x = 1.799097

#> Warning: non-integer x = 1.072438

#> Warning: non-integer x = 1.384271

#> Warning: non-integer x = 2.628068

#> Warning: non-integer x = 1.178426

#> Warning: non-integer x = 0.809479

#> Warning: non-integer x = 0.902285

#> Warning: non-integer x = 2.000026

#> Warning: non-integer x = 0.991824

#> Warning: non-integer x = 0.091407

#> Warning: non-integer x = 2.290414

#> Warning: non-integer x = 2.643510

#> Warning: non-integer x = 0.946409

#> Warning: non-integer x = 1.901979

#> Warning: non-integer x = 1.759421

#> Warning: non-integer x = 2.893742

#> Warning: non-integer x = 0.099488

#> Warning: non-integer x = 2.933754

#> Warning: non-integer x = 0.893968

#> Warning: non-integer x = 1.707662

#> Warning: non-integer x = 1.847478

#> Warning: non-integer x = 2.928385

#> Warning: non-integer x = 0.680550

#> Warning: non-integer x = 1.986613

#> Warning: non-integer x = 2.659176

#> Warning: non-integer x = 0.835028

#> Warning: non-integer x = 1.976201

#> Warning: non-integer x = 0.967545

#> Warning: non-integer x = 1.743336

#> Warning: non-integer x = 1.950089

#> Warning: non-integer x = 3.635587

#> Warning: non-integer x = 0.631259

#> Warning: non-integer x = 1.362588

#> Warning: non-integer x = 1.967790

#> Warning: non-integer x = 0.482858

#> Warning: non-integer x = 3.155974

#> Warning: non-integer x = 4.131053

#> Warning: non-integer x = 1.882116

#> Warning: non-integer x = 0.050796

#> Warning: non-integer x = 0.475913

#> Warning: non-integer x = 2.636971

#> Warning: non-integer x = 0.165023

#> Warning: non-integer x = 3.268283

#> Warning: non-integer x = 2.618927

#> Warning: non-integer x = 2.219648

#> Warning: non-integer x = 2.111568

#> Warning: non-integer x = 5.648446

#> Warning: non-integer x = 3.725920

#> Warning: non-integer x = 0.228135

#> Warning: non-integer x = 0.289274

#> Warning: non-integer x = 1.330396

#> Warning: non-integer x = 3.445762

#> Warning: non-integer x = 2.369284

#> Warning: non-integer x = 0.783291

#> Warning: non-integer x = 0.873786

#> Warning: non-integer x = 1.168445

#> Warning: non-integer x = 1.502323

#> Warning: non-integer x = 0.239659

#> Warning: non-integer x = 0.764696

#> Warning: non-integer x = 1.019066

#> Warning: non-integer x = 1.413955

#> Warning: non-integer x = 1.158933

#> Warning: non-integer x = 1.491028

#> Warning: non-integer x = 0.070195

#> Warning: non-integer x = 2.529016

#> Warning: non-integer x = 2.661320

#> Warning: non-integer x = 1.099671

#> Warning: non-integer x = 0.488092

#> Warning: non-integer x = 0.101518

#> Warning: non-integer x = 3.959552

#> Warning: non-integer x = 3.946046

#> Warning: non-integer x = 0.450629

#> Warning: non-integer x = 1.227220

#> Warning: non-integer x = 1.903444

#> Warning: non-integer x = 0.993048

#> Warning: non-integer x = 0.617566

#> Warning: non-integer x = 1.336255

#> Warning: non-integer x = 2.520035

#> Warning: non-integer x = 1.085875

#> Warning: non-integer x = 0.923327

#> Warning: non-integer x = 2.036317

#> Warning: non-integer x = 0.343354

#> Warning: non-integer x = 0.801509

#> Warning: non-integer x = 1.181924

#> Warning: non-integer x = 1.779169

#> Warning: non-integer x = 1.277586

#> Warning: non-integer x = 0.042996

#> Warning: non-integer x = 1.295355

#> Warning: non-integer x = 0.665131

#> Warning: non-integer x = 1.420199

#> Warning: non-integer x = 1.360091

#> Warning: non-integer x = 3.799971

#> Warning: non-integer x = 1.585857

#> Warning: non-integer x = 2.186576

#> Warning: non-integer x = 1.590556

#> Warning: non-integer x = 2.930662

#> Warning: non-integer x = 0.811693

#> Warning: non-integer x = 0.931873

#> Warning: non-integer x = 0.486540

#> Warning: non-integer x = 0.567580

#> Warning: non-integer x = 4.165963

#> Warning: non-integer x = 0.364305

#> Warning: non-integer x = 0.486094

#> Warning: non-integer x = 2.026248

#> Warning: non-integer x = 0.264454

#> Warning: non-integer x = 1.176868

#> Warning: non-integer x = 2.327538

#> Warning: non-integer x = 0.974856

#> Warning: non-integer x = 1.567146

#> Warning: non-integer x = 0.329921

#> Warning: non-integer x = 1.713692

#> Warning: non-integer x = 0.924334

#> Warning: non-integer x = 3.387119

#> Warning: non-integer x = 1.914012

#> Warning: non-integer x = 0.415590

#> Warning: non-integer x = 0.778317

#> Warning: non-integer x = 1.402058

#> Warning: non-integer x = 3.218841

#> Warning: non-integer x = 1.313503

#> Warning: non-integer x = 0.024518

#> Warning: non-integer x = 1.822870

#> Warning: non-integer x = 1.508866

#> Warning: non-integer x = 2.747339

#> Warning: non-integer x = 1.982240

#> Warning: non-integer x = 1.280968

#> Warning: non-integer x = 1.223141

#> Warning: non-integer x = 0.629107

#> Warning: non-integer x = 2.644838

#> Warning: non-integer x = 0.744792

#> Warning: non-integer x = 0.094435

#> Warning: non-integer x = 2.729195

#> Warning: non-integer x = 1.579148

#> Warning: non-integer x = 1.983847

#> Warning: non-integer x = 1.572654

#> Warning: non-integer x = 2.108159

#> Warning: non-integer x = 2.691125

#> Warning: non-integer x = 0.088165

#> Warning: non-integer x = 0.287382

#> Warning: non-integer x = 0.170295

#> Warning: non-integer x = 1.009560

#> Warning: non-integer x = 2.912709

#> Warning: non-integer x = 0.738831

#> Warning: non-integer x = 0.460421

#> Warning: non-integer x = 0.540920

#> Warning: non-integer x = 1.210192

#> Warning: non-integer x = 2.485592

#> Warning: non-integer x = 2.490673

#> Warning: non-integer x = 1.125574

#> Warning: non-integer x = 1.144731

#> Warning: non-integer x = 1.229116

#> Warning: non-integer x = 0.169907

#> Warning: non-integer x = 1.243375

#> Warning: non-integer x = 1.622521

#> Warning: non-integer x = 0.795642

#> Warning: non-integer x = 2.136736

#> Warning: non-integer x = 0.634266

#> Warning: non-integer x = 0.707852

#> Warning: non-integer x = 1.540443

#> Warning: non-integer x = 1.148280

#> Warning: non-integer x = 0.827622

#> Warning: non-integer x = 2.304657

#> Warning: non-integer x = 3.264440

#> Warning: non-integer x = 1.713514

#> Warning: non-integer x = 0.283667

#> Warning: non-integer x = 1.747616

#> Warning: non-integer x = 0.033026

#> Warning: non-integer x = 1.336183

#> Warning: non-integer x = 2.405803

#> Warning: non-integer x = 4.083574

#> Warning: non-integer x = 0.166280

#> Warning: non-integer x = 0.053704

#> Warning: non-integer x = 1.448754

#> Warning: non-integer x = 0.096037

#> Warning: non-integer x = 0.136531

#> Warning: non-integer x = 3.740619

#> Warning: non-integer x = 0.618124

#> Warning: non-integer x = 0.499890

#> Warning: non-integer x = 0.210873

#> Warning: non-integer x = 0.042380

#> Warning: non-integer x = 0.883513

#> Warning: non-integer x = 0.662839

#> Warning: non-integer x = 0.209529

#> Warning: non-integer x = 1.987023

#> Warning: non-integer x = 0.076758

#> Warning: non-integer x = 2.149083

#> Warning: non-integer x = 1.747312

#> Warning: non-integer x = 0.645205

#> Warning: non-integer x = 0.394241

#> Warning: non-integer x = 0.223192

#> Warning: non-integer x = 1.209597

#> Warning: non-integer x = 0.014558

#> Warning: non-integer x = 0.781350

#> Warning: non-integer x = 2.115608

#> Warning: non-integer x = 4.148996

#> Warning: non-integer x = 0.372046

#> Warning: non-integer x = 1.920380

#> Warning: non-integer x = 1.947659

#> Warning: non-integer x = 0.193745

#> Warning: non-integer x = 0.437194

#> Warning: non-integer x = 1.018226

#> Warning: non-integer x = 0.878045

#> Warning: non-integer x = 2.287573

#> Warning: non-integer x = 1.613622

#> Warning: non-integer x = 1.015760

#> Warning: non-integer x = 0.436244

#> Warning: non-integer x = 2.036475

#> Warning: non-integer x = 1.008310

#> Warning: non-integer x = 0.248033

#> Warning: non-integer x = 0.894542

#> Warning: non-integer x = 2.102576

#> Warning: non-integer x = 0.965429

#> Warning: non-integer x = 1.232523

#> Warning: non-integer x = 0.191362

#> Warning: non-integer x = 1.254916

#> Warning: non-integer x = 2.593798

#> Warning: non-integer x = 0.825646

#> Warning: non-integer x = 3.392939

#> Warning: non-integer x = 1.577834

#> Warning: non-integer x = 0.411049

#> Warning: non-integer x = 1.478330

#> Warning: non-integer x = 1.652995

#> Warning: non-integer x = 0.182452

#> Warning: non-integer x = 1.117300

#> Warning: non-integer x = 0.594301

#> Warning: non-integer x = 1.398979

#> Warning: non-integer x = 0.100920

#> Warning: non-integer x = 0.201755

#> Warning: non-integer x = 0.455776

#> Warning: non-integer x = 0.332393

#> Warning: non-integer x = 1.394444

#> Warning: non-integer x = 3.006147

#> Warning: non-integer x = 0.052041

#> Warning: non-integer x = 3.090936

#> Warning: non-integer x = 0.269152

#> Warning: non-integer x = 0.638910

#> Warning: non-integer x = 0.143672

#> Warning: non-integer x = 0.467527

#> Warning: non-integer x = 1.759486

#> Warning: non-integer x = 3.243840

#> Warning: non-integer x = 2.123519

#> Warning: non-integer x = 0.998043

#> Warning: non-integer x = 1.179563

#> Warning: non-integer x = 0.488796

#> Warning: non-integer x = 0.705676

#> Warning: non-integer x = 2.124922

#> Warning: non-integer x = 2.799886

#> Warning: non-integer x = 1.328074

#> Warning: non-integer x = 1.120476

#> Warning: non-integer x = 1.352024

#> Warning: non-integer x = 2.824767

#> Warning: non-integer x = 1.065667

#> Warning: non-integer x = 0.186417

#> Warning: non-integer x = 1.983955

#> Warning: non-integer x = 0.282081

#> Warning: non-integer x = 1.056540

#> Warning: non-integer x = 0.741936

#> Warning: non-integer x = 0.047195

#> Warning: non-integer x = 2.336004

#> Warning: non-integer x = 2.766805

#> Warning: non-integer x = 1.596709

#> Warning: non-integer x = 0.887666

#> Warning: non-integer x = 0.103593

#> Warning: non-integer x = 1.845049

#> Warning: non-integer x = 3.768878

#> Warning: non-integer x = 3.242232

#> Warning: non-integer x = 1.613517

#> Warning: non-integer x = 1.310305

#> Warning: non-integer x = 0.088783

#> Warning: non-integer x = 0.196972

#> Warning: non-integer x = 1.249991

#> Warning: non-integer x = 3.600979

#> Warning: non-integer x = 0.319185

#> Warning: non-integer x = 0.579418

#> Warning: non-integer x = 0.267039

#> Warning: non-integer x = 2.903085

#> Warning: non-integer x = 0.777604

#> Warning: non-integer x = 2.328078

#> Warning: non-integer x = 0.778374

#> Warning: non-integer x = 1.764152

#> Warning: non-integer x = 0.241061

#> Warning: non-integer x = 1.633807

#> Warning: non-integer x = 1.006910

#> Warning: non-integer x = 3.314313

#> Warning: non-integer x = 0.710286

#> Warning: non-integer x = 1.631958

#> Warning: non-integer x = 0.221402

#> Warning: non-integer x = 3.539317

#> Warning: non-integer x = 1.099713

#> Warning: non-integer x = 1.962311

#> Warning: non-integer x = 1.823583

#> Warning: non-integer x = 0.318492

#> Warning: non-integer x = 0.867670

#> Warning: non-integer x = 1.008995

#> Warning: non-integer x = 1.904121

#> Warning: non-integer x = 4.093894

#> Warning: non-integer x = 0.182856

#> Warning: non-integer x = 0.691117

#> Warning: non-integer x = 0.139835

#> Warning: non-integer x = 0.352716

#> Warning: non-integer x = 0.587142

#> Warning: non-integer x = 3.009195

#> Warning: non-integer x = 0.825640

#> Warning: non-integer x = 3.064561

#> Warning: non-integer x = 1.687083

#> Warning: non-integer x = 1.721901

#> Warning: non-integer x = 1.301296

#> Warning: non-integer x = 0.819467

#> Warning: non-integer x = 0.221742

#> Warning: non-integer x = 1.034083

#> Warning: non-integer x = 1.278964

#> Warning: non-integer x = 3.149996

#> Warning: non-integer x = 1.067526

#> Warning: non-integer x = 0.230215

#> Warning: non-integer x = 1.825990

#> Warning: non-integer x = 1.325862

#> Warning: non-integer x = 0.819403

#> Warning: non-integer x = 2.020212

#> Warning: non-integer x = 1.968235

#> Warning: non-integer x = 0.230114

#> Warning: non-integer x = 1.765793

#> Warning: non-integer x = 0.563495

#> Warning: non-integer x = 0.817071

#> Warning: non-integer x = 0.419169

#> Warning: non-integer x = 1.329041

#> Warning: non-integer x = 1.328575

#> Warning: non-integer x = 1.836611

#> Warning: non-integer x = 1.140884

#> Warning: non-integer x = 2.126594

#> Warning: non-integer x = 0.063729

#> Warning: non-integer x = 3.428706

#> Warning: non-integer x = 0.186551

#> Warning: non-integer x = 2.502612

#> Warning: non-integer x = 1.055395

#> Warning: non-integer x = 1.892156

#> Warning: non-integer x = 0.160496

#> Warning: non-integer x = 1.751127

#> Warning: non-integer x = 2.039298

#> Warning: non-integer x = 1.081608

#> Warning: non-integer x = 0.126100

#> Warning: non-integer x = 1.595230

#> Warning: non-integer x = 0.112592

#> Warning: non-integer x = 1.662349

#> Warning: non-integer x = 0.364953

#> Warning: non-integer x = 0.874428

#> Warning: non-integer x = 0.129530

#> Warning: non-integer x = 1.533942

#> Warning: non-integer x = 1.810578

#> Warning: non-integer x = 1.235023

#> Warning: non-integer x = 0.352355

#> Warning: non-integer x = 2.040136

#> Warning: non-integer x = 0.561133

#> Warning: non-integer x = 1.232909

#> Warning: non-integer x = 2.474716

#> Warning: non-integer x = 2.238003

#> Warning: non-integer x = 0.909769

#> Warning: non-integer x = 0.276159

#> Warning: non-integer x = 1.576888

#> Warning: non-integer x = 0.607964

#> Warning: non-integer x = 0.132952

#> Warning: non-integer x = 0.706656

#> Warning: non-integer x = 0.368351

#> Warning: non-integer x = 0.556747

#> Warning: non-integer x = 0.403746

#> Warning: non-integer x = 1.472129

#> Warning: non-integer x = 3.230659

#> Warning: non-integer x = 0.150438

#> Warning: non-integer x = 0.552585

#> Warning: non-integer x = 1.041337

#> Warning: non-integer x = 1.196300

#> Warning: non-integer x = 1.420933

#> Warning: non-integer x = 0.741957

#> Warning: non-integer x = 0.176232

#> Warning: non-integer x = 0.199594

#> Warning: non-integer x = 1.533962

#> Warning: non-integer x = 2.094521

#> Warning: non-integer x = 0.269743

#> Warning: non-integer x = 0.135063

#> Warning: non-integer x = 0.933440

#> Warning: non-integer x = 2.648486

#> Warning: non-integer x = 2.166206

#> Warning: non-integer x = 0.586044

#> Warning: non-integer x = 0.049697

#> Warning: non-integer x = 1.499679

#> Warning: non-integer x = 0.070945

#> Warning: non-integer x = 0.706144

#> Warning: non-integer x = 3.025893

#> Warning: non-integer x = 1.120259

#> Warning: non-integer x = 1.155408

#> Warning: non-integer x = 1.758466

#> Warning: non-integer x = 0.969044

#> Warning: non-integer x = 1.701569

#> Warning: non-integer x = 0.628273

#> Warning: non-integer x = 3.438035

#> Warning: non-integer x = 0.232582

#> Warning: non-integer x = 0.517629

#> Warning: non-integer x = 2.443759

#> Warning: non-integer x = 2.413125

#> Warning: non-integer x = 2.506373

#> Warning: non-integer x = 0.934643

#> Warning: non-integer x = 1.861005

#> Warning: non-integer x = 1.576325

#> Warning: non-integer x = 1.123077

#> Warning: non-integer x = 1.720012

#> Warning: non-integer x = 1.192903

#> Warning: non-integer x = 1.364428

#> Warning: non-integer x = 0.226580

#> Warning: non-integer x = 0.983597

#> Warning: non-integer x = 0.637187

#> Warning: non-integer x = 0.387888

#> Warning: non-integer x = 0.884746

#> Warning: non-integer x = 2.828744

#> Warning: non-integer x = 1.287937

#> Warning: non-integer x = 1.056086

#> Warning: non-integer x = 1.596901

#> Warning: non-integer x = 1.254015

#> Warning: non-integer x = 1.058041

#> Warning: non-integer x = 1.165366

#> Warning: non-integer x = 0.057616

#> Warning: non-integer x = 1.831208

#> Warning: non-integer x = 2.494697

#> Warning: non-integer x = 3.091456

#> Warning: non-integer x = 0.746792

#> Warning: non-integer x = 2.009915

#> Warning: non-integer x = 0.528603

#> Warning: non-integer x = 1.560342

#> Warning: non-integer x = 0.054281

#> Warning: non-integer x = 0.410442

#> Warning: non-integer x = 1.566680

#> Warning: non-integer x = 3.527673

#> Warning: non-integer x = 1.786741

#> Warning: non-integer x = 0.504596

#> Warning: non-integer x = 0.410822

#> Warning: non-integer x = 0.601499

#> Warning: non-integer x = 0.799642

#> Warning: non-integer x = 0.091207

#> Warning: non-integer x = 0.094159

#> Warning: non-integer x = 0.286635

#> Warning: non-integer x = 0.469376

#> Warning: non-integer x = 4.076853

#> Warning: non-integer x = 1.208077

#> Warning: non-integer x = 0.553251

#> Warning: non-integer x = 1.430918

#> Warning: non-integer x = 0.173720

#> Warning: non-integer x = 0.462526

#> Warning: non-integer x = 1.283685

#> Warning: non-integer x = 3.203828

#> Warning: non-integer x = 0.207268

#> Warning: non-integer x = 1.911052

#> Warning: non-integer x = 2.918421

#> Warning: non-integer x = 2.445711

#> Warning: non-integer x = 0.958991

#> Warning: non-integer x = 2.654028

#> Warning: non-integer x = 0.937294

#> Warning: non-integer x = 0.406029

#> Warning: non-integer x = 1.031416

#> Warning: non-integer x = 1.504218

#> Warning: non-integer x = 0.992211

#> Warning: non-integer x = 0.622556

#> Warning: non-integer x = 0.576722

#> Warning: non-integer x = 2.076208

#> Warning: non-integer x = 0.011630

#> Warning: non-integer x = 1.940295

#> Warning: non-integer x = 0.712764

#> Warning: non-integer x = 0.027014

#> Warning: non-integer x = 3.288621

#> Warning: non-integer x = 0.133314

#> Warning: non-integer x = 0.898345

#> Warning: non-integer x = 0.782004

#> Warning: non-integer x = 1.281433

#> Warning: non-integer x = 0.359330

#> Warning: non-integer x = 0.312095

#> Warning: non-integer x = 1.793499

#> Warning: non-integer x = 1.960729

#> Warning: non-integer x = 2.247019

#> Warning: non-integer x = 2.447370

#> Warning: non-integer x = 1.803529

#> Warning: non-integer x = 1.419586

#> Warning: non-integer x = 2.556674

#> Warning: non-integer x = 0.564117

#> Warning: non-integer x = 1.090079

#> Warning: non-integer x = 3.430775

#> Warning: non-integer x = 0.028151

#> Warning: non-integer x = 1.887705

#> Warning: non-integer x = 1.087087

#> Warning: non-integer x = 1.965088

#> Warning: non-integer x = 1.776766

#> Warning: non-integer x = 0.510715

#> Warning: non-integer x = 1.775173

#> Warning: non-integer x = 3.283432

#> Warning: non-integer x = 0.024354

#> Warning: non-integer x = 0.268145

#> Warning: non-integer x = 2.210055

#> Warning: non-integer x = 0.620023

#> Warning: non-integer x = 1.395033

#> Warning: non-integer x = 1.623551

#> Warning: non-integer x = 0.051608

#> Warning: non-integer x = 1.670311

#> Warning: non-integer x = 2.291136

#> Warning: non-integer x = 0.961891

#> Warning: non-integer x = 0.708302

#> Warning: non-integer x = 0.682729

#> Warning: non-integer x = 0.137483

#> Warning: non-integer x = 1.104567

#> Warning: non-integer x = 0.404114

#> Warning: non-integer x = 0.769919

#> Warning: non-integer x = 4.893263

#> Warning: non-integer x = 0.922172

#> Warning: non-integer x = 2.007588

#> Warning: non-integer x = 2.841098

#> Warning: non-integer x = 0.364077

#> Warning: non-integer x = 1.879078

#> Warning: non-integer x = 1.637088

#> Warning: non-integer x = 2.259341

#> Warning: non-integer x = 0.207675

#> Warning: non-integer x = 2.129697

#> Warning: non-integer x = 0.197037

#> Warning: non-integer x = 3.025178

#> Warning: non-integer x = 3.204677

#> Warning: non-integer x = 1.267912

#> Warning: non-integer x = 1.150780

#> Warning: non-integer x = 1.537757

#> Warning: non-integer x = 1.815234

#> Warning: non-integer x = 1.318929

#> Warning: non-integer x = 2.257734

#> Warning: non-integer x = 1.519121

#> Warning: non-integer x = 1.270142

#> Warning: non-integer x = 0.738072

#> Warning: non-integer x = 1.186279

#> Warning: non-integer x = 1.057469

#> Warning: non-integer x = 2.170747

#> Warning: non-integer x = 0.053913

#> Warning: non-integer x = 0.411499

#> Warning: non-integer x = 0.174855

#> Warning: non-integer x = 1.043116

#> Warning: non-integer x = 0.390184

#> Warning: non-integer x = 0.396541

#> Warning: non-integer x = 1.040215

#> Warning: non-integer x = 1.266014

#> Warning: non-integer x = 2.298150

#> Warning: non-integer x = 0.184565

#> Warning: non-integer x = 1.209996

#> Warning: non-integer x = 3.212165

#> Warning: non-integer x = 0.630524

#> Warning: non-integer x = 2.502447

#> Warning: non-integer x = 0.475698

#> Warning: non-integer x = 0.168472

#> Warning: non-integer x = 1.344841

#> Warning: non-integer x = 1.363502

#> Warning: non-integer x = 0.857601

#> Warning: non-integer x = 1.727083

#> Warning: non-integer x = 0.501919

#> Warning: non-integer x = 2.522346

#> Warning: non-integer x = 0.080090

#> Warning: non-integer x = 1.930318

#> Warning: non-integer x = 0.845334

#> Warning: non-integer x = 2.005817

#> Warning: non-integer x = 1.397512

#> Warning: non-integer x = 0.378944

#> Warning: non-integer x = 2.322020

#> Warning: non-integer x = 0.589464

#> Warning: non-integer x = 0.744682

#> Warning: non-integer x = 1.106757

#> Warning: non-integer x = 0.122207

#> Warning: non-integer x = 1.200181

#> Warning: non-integer x = 2.493262

#> Warning: non-integer x = 2.844099

#> Warning: non-integer x = 1.520730

#> Warning: non-integer x = 2.518703

#> Warning: non-integer x = 2.273870

#> Warning: non-integer x = 0.233377

#> Warning: non-integer x = 2.120307

#> Warning: non-integer x = 0.556383

#> Warning: non-integer x = 0.085673

#> Warning: non-integer x = 1.107403

#> Warning: non-integer x = 2.530241

#> Warning: non-integer x = 2.131957

#> Warning: non-integer x = 0.468478

#> Warning: non-integer x = 0.602216

#> Warning: non-integer x = 1.620098

#> Warning: non-integer x = 2.580568

#> Warning: non-integer x = 0.641274

#> Warning: non-integer x = 1.444714

#> Warning: non-integer x = 1.445468

#> Warning: non-integer x = 0.180519

#> Warning: non-integer x = 2.027202

#> Warning: non-integer x = 0.229199

#> Warning: non-integer x = 1.234674

#> Warning: non-integer x = 0.744987

#> Warning: non-integer x = 2.110930

#> Warning: non-integer x = 1.258714

#> Warning: non-integer x = 0.752385

#> Warning: non-integer x = 2.947695

#> Warning: non-integer x = 2.078563

#> Warning: non-integer x = 0.284561

#> Warning: non-integer x = 0.503402

#> Warning: non-integer x = 0.427246

#> Warning: non-integer x = 1.515218

#> Warning: non-integer x = 2.600015

#> Warning: non-integer x = 2.395862

#> Warning: non-integer x = 0.883733

#> Warning: non-integer x = 1.390390

#> Warning: non-integer x = 1.079918

#> Warning: non-integer x = 0.002538

#> Warning: non-integer x = 1.591603

#> Warning: non-integer x = 1.982484

#> Warning: non-integer x = 0.618214

#> Warning: non-integer x = 0.861790

#> Warning: non-integer x = 1.280738

#> Warning: non-integer x = 1.838625

#> Warning: non-integer x = 0.753723

#> Warning: non-integer x = 0.574344

#> Warning: non-integer x = 2.341038

#> Warning: non-integer x = 0.940671

#> Warning: non-integer x = 1.169576

#> Warning: non-integer x = 0.342522

#> Warning: non-integer x = 1.921574

#> Warning: non-integer x = 0.476322

#> Warning: non-integer x = 0.189307

#> Warning: non-integer x = 0.356697

#> Warning: non-integer x = 0.639351

#> Warning: non-integer x = 2.016889

#> Warning: non-integer x = 3.056377

#> Warning: non-integer x = 1.588569

#> Warning: non-integer x = 1.703205

#> Warning: non-integer x = 0.512691

#> Warning: non-integer x = 1.321745

#> Warning: non-integer x = 0.181450

#> Warning: non-integer x = 2.953239

#> Warning: non-integer x = 1.272760

#> Warning: non-integer x = 1.679721

#> Warning: non-integer x = 2.259121

#> Warning: non-integer x = 0.800963

#> Warning: non-integer x = 1.188832

#> Warning: non-integer x = 1.442783

#> Warning: non-integer x = 1.013092

#> Warning: non-integer x = 2.027226

#> Warning: non-integer x = 1.941113

#> Warning: non-integer x = 1.051876

#> Warning: non-integer x = 0.338022

#> Warning: non-integer x = 0.701308

#> Warning: non-integer x = 1.664996

#> Warning: non-integer x = 0.937757

#> Warning: non-integer x = 1.466884

#> Warning: non-integer x = 1.415257

#> Warning: non-integer x = 1.396350

#> Warning: non-integer x = 2.211278

#> Warning: non-integer x = 1.847682

#> Warning: non-integer x = 2.507013

#> Warning: non-integer x = 1.252816

#> Warning: non-integer x = 1.520093

#> Warning: non-integer x = 1.268894

#> Warning: non-integer x = 3.347721

#> Warning: non-integer x = 0.115034

#> Warning: non-integer x = 0.698884

#> Warning: non-integer x = 0.232340

#> Warning: non-integer x = 0.450263

#> Warning: non-integer x = 0.080601

#> Warning: non-integer x = 0.209157

#> Warning: non-integer x = 1.349066