Estimate the Individual Treatment Effect using Poisson Regression

Estimates the Individual Treatment Effect using Poisson Regression given a response vector, a treatment vector, and a features matrix.

Usage

estimate_ite_poisson(y, z, X, X_names, include_offset, offset_name)

Arguments

y

the observed response vector

z

the treatment vector

X

the features matrix

X_names

the names of the covariates

include_offset

whether or not to include an offset when estimating the ITE, for poisson only

offset_name

the name of the offset, if it is to be included

Value

a vector of ITE estimates

Examples

dataset <- generate_cre_dataset(n = 1000, rho = 0, n_rules = 2, p = 10,
                                effect_size = 2, binary = FALSE)

# Initialize parameters
y <- abs(dataset[["y"]])
z <- dataset[["z"]]
X <- as.data.frame(dataset[["X"]])
X_names <- names(as.data.frame(X))
include_offset <- FALSE
offset_name <- NA

ite_list <- estimate_ite_poisson(y, z, X, X_names, include_offset, offset_name)
#> Warning: non-integer x = 1.521391
#> Warning: non-integer x = 0.887520
#> Warning: non-integer x = 0.118603
#> Warning: non-integer x = 3.447522
#> Warning: non-integer x = 0.643607
#> Warning: non-integer x = 3.072708
#> Warning: non-integer x = 3.279828
#> Warning: non-integer x = 0.094717
#> Warning: non-integer x = 1.024306
#> Warning: non-integer x = 1.831892
#> Warning: non-integer x = 2.116683
#> Warning: non-integer x = 0.763714
#> Warning: non-integer x = 0.730335
#> Warning: non-integer x = 1.677696
#> Warning: non-integer x = 1.471220
#> Warning: non-integer x = 4.784593
#> Warning: non-integer x = 0.234718
#> Warning: non-integer x = 1.575203
#> Warning: non-integer x = 0.579517
#> Warning: non-integer x = 0.016716
#> Warning: non-integer x = 1.342998
#> Warning: non-integer x = 0.200177
#> Warning: non-integer x = 2.205801
#> Warning: non-integer x = 1.064939
#> Warning: non-integer x = 0.261190
#> Warning: non-integer x = 0.157443
#> Warning: non-integer x = 1.333612
#> Warning: non-integer x = 2.461386
#> Warning: non-integer x = 0.987924
#> Warning: non-integer x = 2.913448
#> Warning: non-integer x = 2.057663
#> Warning: non-integer x = 0.769247
#> Warning: non-integer x = 1.172829
#> Warning: non-integer x = 1.126717
#> Warning: non-integer x = 3.595793
#> Warning: non-integer x = 2.252845
#> Warning: non-integer x = 4.194279
#> Warning: non-integer x = 2.934685
#> Warning: non-integer x = 3.270636
#> Warning: non-integer x = 0.570661
#> Warning: non-integer x = 0.415214
#> Warning: non-integer x = 2.847561
#> Warning: non-integer x = 0.078071
#> Warning: non-integer x = 0.002729
#> Warning: non-integer x = 1.786731
#> Warning: non-integer x = 0.900704
#> Warning: non-integer x = 1.244516
#> Warning: non-integer x = 3.663364
#> Warning: non-integer x = 2.519357
#> Warning: non-integer x = 1.855593
#> Warning: non-integer x = 0.547438
#> Warning: non-integer x = 3.549047
#> Warning: non-integer x = 1.879606
#> Warning: non-integer x = 0.418758
#> Warning: non-integer x = 2.720253
#> Warning: non-integer x = 3.696839
#> Warning: non-integer x = 0.614387
#> Warning: non-integer x = 2.270087
#> Warning: non-integer x = 2.578183
#> Warning: non-integer x = 0.219405
#> Warning: non-integer x = 2.254693
#> Warning: non-integer x = 2.885753
#> Warning: non-integer x = 1.413586
#> Warning: non-integer x = 2.487464
#> Warning: non-integer x = 2.212693
#> Warning: non-integer x = 1.552117
#> Warning: non-integer x = 1.647964
#> Warning: non-integer x = 0.508101
#> Warning: non-integer x = 0.967827
#> Warning: non-integer x = 2.336333
#> Warning: non-integer x = 0.711940
#> Warning: non-integer x = 1.919570
#> Warning: non-integer x = 0.818235
#> Warning: non-integer x = 1.820296
#> Warning: non-integer x = 0.471532
#> Warning: non-integer x = 2.021203
#> Warning: non-integer x = 0.309245
#> Warning: non-integer x = 1.417963
#> Warning: non-integer x = 1.964457
#> Warning: non-integer x = 0.206801
#> Warning: non-integer x = 1.114272
#> Warning: non-integer x = 1.333486
#> Warning: non-integer x = 2.201631
#> Warning: non-integer x = 2.867347
#> Warning: non-integer x = 2.154182
#> Warning: non-integer x = 0.088164
#> Warning: non-integer x = 1.349268
#> Warning: non-integer x = 2.919684
#> Warning: non-integer x = 1.453151
#> Warning: non-integer x = 0.652049
#> Warning: non-integer x = 0.465531
#> Warning: non-integer x = 0.583446
#> Warning: non-integer x = 0.753023
#> Warning: non-integer x = 3.971154
#> Warning: non-integer x = 1.373538
#> Warning: non-integer x = 0.224884
#> Warning: non-integer x = 3.260218
#> Warning: non-integer x = 0.001405
#> Warning: non-integer x = 0.869122
#> Warning: non-integer x = 1.403816
#> Warning: non-integer x = 0.895140
#> Warning: non-integer x = 0.687424
#> Warning: non-integer x = 5.034925
#> Warning: non-integer x = 0.612140
#> Warning: non-integer x = 1.815095
#> Warning: non-integer x = 2.496564
#> Warning: non-integer x = 2.478355
#> Warning: non-integer x = 0.241107
#> Warning: non-integer x = 0.939613
#> Warning: non-integer x = 3.267630
#> Warning: non-integer x = 0.667266
#> Warning: non-integer x = 0.900395
#> Warning: non-integer x = 0.119327
#> Warning: non-integer x = 2.273827
#> Warning: non-integer x = 0.126896
#> Warning: non-integer x = 3.253782
#> Warning: non-integer x = 2.588690
#> Warning: non-integer x = 2.312740
#> Warning: non-integer x = 0.574131
#> Warning: non-integer x = 0.939884
#> Warning: non-integer x = 0.343119
#> Warning: non-integer x = 0.870277
#> Warning: non-integer x = 0.816800
#> Warning: non-integer x = 2.403661
#> Warning: non-integer x = 1.071557
#> Warning: non-integer x = 0.703814
#> Warning: non-integer x = 0.560999
#> Warning: non-integer x = 0.076803
#> Warning: non-integer x = 0.393715
#> Warning: non-integer x = 0.197013
#> Warning: non-integer x = 2.048914
#> Warning: non-integer x = 2.755750
#> Warning: non-integer x = 1.879879
#> Warning: non-integer x = 1.302147
#> Warning: non-integer x = 1.685524
#> Warning: non-integer x = 1.836056
#> Warning: non-integer x = 0.214412
#> Warning: non-integer x = 1.307795
#> Warning: non-integer x = 2.772234
#> Warning: non-integer x = 3.036184
#> Warning: non-integer x = 2.816661
#> Warning: non-integer x = 2.123276
#> Warning: non-integer x = 1.160231
#> Warning: non-integer x = 2.091827
#> Warning: non-integer x = 1.292360
#> Warning: non-integer x = 0.495049
#> Warning: non-integer x = 0.522854
#> Warning: non-integer x = 0.421675
#> Warning: non-integer x = 2.599178
#> Warning: non-integer x = 1.912398
#> Warning: non-integer x = 4.129065
#> Warning: non-integer x = 0.681513
#> Warning: non-integer x = 0.853617
#> Warning: non-integer x = 3.744633
#> Warning: non-integer x = 2.047358
#> Warning: non-integer x = 1.092838
#> Warning: non-integer x = 1.317335
#> Warning: non-integer x = 0.239344
#> Warning: non-integer x = 1.546109
#> Warning: non-integer x = 0.124793
#> Warning: non-integer x = 4.004322
#> Warning: non-integer x = 0.685581
#> Warning: non-integer x = 1.249984
#> Warning: non-integer x = 2.079124
#> Warning: non-integer x = 1.639791
#> Warning: non-integer x = 1.880618
#> Warning: non-integer x = 3.867602
#> Warning: non-integer x = 0.408851
#> Warning: non-integer x = 1.026043
#> Warning: non-integer x = 1.222126
#> Warning: non-integer x = 0.893998
#> Warning: non-integer x = 1.064758
#> Warning: non-integer x = 0.180274
#> Warning: non-integer x = 0.345589
#> Warning: non-integer x = 2.552004
#> Warning: non-integer x = 0.838891
#> Warning: non-integer x = 1.628493
#> Warning: non-integer x = 3.003575
#> Warning: non-integer x = 1.686670
#> Warning: non-integer x = 1.831552
#> Warning: non-integer x = 2.613996
#> Warning: non-integer x = 1.302893
#> Warning: non-integer x = 0.660047
#> Warning: non-integer x = 1.312863
#> Warning: non-integer x = 0.574186
#> Warning: non-integer x = 1.714737
#> Warning: non-integer x = 2.500728
#> Warning: non-integer x = 0.642138
#> Warning: non-integer x = 3.089848
#> Warning: non-integer x = 2.397783
#> Warning: non-integer x = 1.101568
#> Warning: non-integer x = 1.816474
#> Warning: non-integer x = 1.530360
#> Warning: non-integer x = 1.823582
#> Warning: non-integer x = 1.539291
#> Warning: non-integer x = 1.284594
#> Warning: non-integer x = 0.426103
#> Warning: non-integer x = 0.399487
#> Warning: non-integer x = 2.599112
#> Warning: non-integer x = 1.827115
#> Warning: non-integer x = 1.307653
#> Warning: non-integer x = 1.552154
#> Warning: non-integer x = 0.264187
#> Warning: non-integer x = 2.464493
#> Warning: non-integer x = 1.740795
#> Warning: non-integer x = 0.573488
#> Warning: non-integer x = 0.561701
#> Warning: non-integer x = 2.545229
#> Warning: non-integer x = 1.017058
#> Warning: non-integer x = 3.433352
#> Warning: non-integer x = 1.326932
#> Warning: non-integer x = 4.232189
#> Warning: non-integer x = 0.466384
#> Warning: non-integer x = 1.198313
#> Warning: non-integer x = 0.254555
#> Warning: non-integer x = 2.369962
#> Warning: non-integer x = 0.217308
#> Warning: non-integer x = 2.531994
#> Warning: non-integer x = 2.875115
#> Warning: non-integer x = 2.336513
#> Warning: non-integer x = 0.116403
#> Warning: non-integer x = 1.109314
#> Warning: non-integer x = 1.329924
#> Warning: non-integer x = 0.701796
#> Warning: non-integer x = 0.889244
#> Warning: non-integer x = 0.978745
#> Warning: non-integer x = 1.708570
#> Warning: non-integer x = 1.070207
#> Warning: non-integer x = 1.671850
#> Warning: non-integer x = 0.379201
#> Warning: non-integer x = 1.132989
#> Warning: non-integer x = 0.342209
#> Warning: non-integer x = 1.421542
#> Warning: non-integer x = 2.119304
#> Warning: non-integer x = 0.020980
#> Warning: non-integer x = 0.337150
#> Warning: non-integer x = 0.186876
#> Warning: non-integer x = 1.944986
#> Warning: non-integer x = 0.424786
#> Warning: non-integer x = 0.112181
#> Warning: non-integer x = 3.898343
#> Warning: non-integer x = 3.655722
#> Warning: non-integer x = 1.201856
#> Warning: non-integer x = 2.445791
#> Warning: non-integer x = 0.868835
#> Warning: non-integer x = 0.175991
#> Warning: non-integer x = 3.353654
#> Warning: non-integer x = 2.958878
#> Warning: non-integer x = 1.272619
#> Warning: non-integer x = 0.307839
#> Warning: non-integer x = 1.212218
#> Warning: non-integer x = 1.963892
#> Warning: non-integer x = 0.502867
#> Warning: non-integer x = 0.905902
#> Warning: non-integer x = 2.875958
#> Warning: non-integer x = 0.949947
#> Warning: non-integer x = 3.148458
#> Warning: non-integer x = 3.313544
#> Warning: non-integer x = 2.314087
#> Warning: non-integer x = 0.476922
#> Warning: non-integer x = 2.026370
#> Warning: non-integer x = 1.585118
#> Warning: non-integer x = 1.359125
#> Warning: non-integer x = 1.711689
#> Warning: non-integer x = 0.487398
#> Warning: non-integer x = 1.376373
#> Warning: non-integer x = 0.144560
#> Warning: non-integer x = 1.613309
#> Warning: non-integer x = 0.647235
#> Warning: non-integer x = 0.873105
#> Warning: non-integer x = 3.855223
#> Warning: non-integer x = 1.570902
#> Warning: non-integer x = 2.862409
#> Warning: non-integer x = 1.812670
#> Warning: non-integer x = 0.826584
#> Warning: non-integer x = 0.561583
#> Warning: non-integer x = 2.523404
#> Warning: non-integer x = 1.445734
#> Warning: non-integer x = 1.399921
#> Warning: non-integer x = 2.252998
#> Warning: non-integer x = 0.131159
#> Warning: non-integer x = 1.395672
#> Warning: non-integer x = 1.315703
#> Warning: non-integer x = 0.463567
#> Warning: non-integer x = 2.306462
#> Warning: non-integer x = 0.184494
#> Warning: non-integer x = 4.246356
#> Warning: non-integer x = 2.284224
#> Warning: non-integer x = 1.374805
#> Warning: non-integer x = 0.479770
#> Warning: non-integer x = 3.339901
#> Warning: non-integer x = 1.528037
#> Warning: non-integer x = 0.311025
#> Warning: non-integer x = 0.224380
#> Warning: non-integer x = 0.320874
#> Warning: non-integer x = 0.773484
#> Warning: non-integer x = 2.182501
#> Warning: non-integer x = 1.108114
#> Warning: non-integer x = 1.164922
#> Warning: non-integer x = 1.167676
#> Warning: non-integer x = 0.091024
#> Warning: non-integer x = 2.237876
#> Warning: non-integer x = 3.503926
#> Warning: non-integer x = 1.312189
#> Warning: non-integer x = 0.096876
#> Warning: non-integer x = 0.740827
#> Warning: non-integer x = 2.028546
#> Warning: non-integer x = 1.618845
#> Warning: non-integer x = 2.684608
#> Warning: non-integer x = 2.814935
#> Warning: non-integer x = 1.694056
#> Warning: non-integer x = 0.672757
#> Warning: non-integer x = 0.188906
#> Warning: non-integer x = 0.448159
#> Warning: non-integer x = 0.490490
#> Warning: non-integer x = 2.616240
#> Warning: non-integer x = 3.239044
#> Warning: non-integer x = 3.218259
#> Warning: non-integer x = 0.995465
#> Warning: non-integer x = 2.914015
#> Warning: non-integer x = 2.319365
#> Warning: non-integer x = 1.571850
#> Warning: non-integer x = 0.819973
#> Warning: non-integer x = 2.153649
#> Warning: non-integer x = 0.773575
#> Warning: non-integer x = 0.732250
#> Warning: non-integer x = 1.309471
#> Warning: non-integer x = 3.565259
#> Warning: non-integer x = 0.473599
#> Warning: non-integer x = 1.097766
#> Warning: non-integer x = 2.257278
#> Warning: non-integer x = 0.246207
#> Warning: non-integer x = 1.039096
#> Warning: non-integer x = 0.877678
#> Warning: non-integer x = 2.235145
#> Warning: non-integer x = 0.815294
#> Warning: non-integer x = 1.184399
#> Warning: non-integer x = 2.058785
#> Warning: non-integer x = 0.475869
#> Warning: non-integer x = 3.057209
#> Warning: non-integer x = 0.444778
#> Warning: non-integer x = 2.663750
#> Warning: non-integer x = 0.236354
#> Warning: non-integer x = 2.315761
#> Warning: non-integer x = 3.691991
#> Warning: non-integer x = 2.042489
#> Warning: non-integer x = 2.511934
#> Warning: non-integer x = 1.801652
#> Warning: non-integer x = 1.740382
#> Warning: non-integer x = 1.887464
#> Warning: non-integer x = 2.257369
#> Warning: non-integer x = 0.641686
#> Warning: non-integer x = 0.775078
#> Warning: non-integer x = 0.485705
#> Warning: non-integer x = 2.206771
#> Warning: non-integer x = 0.717034
#> Warning: non-integer x = 1.026831
#> Warning: non-integer x = 2.966274
#> Warning: non-integer x = 0.160142
#> Warning: non-integer x = 1.345468
#> Warning: non-integer x = 3.371311
#> Warning: non-integer x = 3.545503
#> Warning: non-integer x = 3.169472
#> Warning: non-integer x = 0.822342
#> Warning: non-integer x = 1.460823
#> Warning: non-integer x = 0.455864
#> Warning: non-integer x = 0.281913
#> Warning: non-integer x = 1.209990
#> Warning: non-integer x = 1.452953
#> Warning: non-integer x = 0.677527
#> Warning: non-integer x = 2.504128
#> Warning: non-integer x = 1.170357
#> Warning: non-integer x = 1.206990
#> Warning: non-integer x = 0.139730
#> Warning: non-integer x = 0.259304
#> Warning: non-integer x = 2.982960
#> Warning: non-integer x = 1.159062
#> Warning: non-integer x = 2.670251
#> Warning: non-integer x = 2.153629
#> Warning: non-integer x = 1.090423
#> Warning: non-integer x = 0.336488
#> Warning: non-integer x = 2.534312
#> Warning: non-integer x = 1.465307
#> Warning: non-integer x = 1.828415
#> Warning: non-integer x = 2.034897
#> Warning: non-integer x = 2.293374
#> Warning: non-integer x = 0.071821
#> Warning: non-integer x = 0.946136
#> Warning: non-integer x = 0.621395
#> Warning: non-integer x = 0.098351
#> Warning: non-integer x = 2.957690
#> Warning: non-integer x = 1.671626
#> Warning: non-integer x = 0.074067
#> Warning: non-integer x = 0.940406
#> Warning: non-integer x = 2.619544
#> Warning: non-integer x = 0.752952
#> Warning: non-integer x = 0.947340
#> Warning: non-integer x = 1.397202
#> Warning: non-integer x = 3.201111
#> Warning: non-integer x = 2.530023
#> Warning: non-integer x = 0.853698
#> Warning: non-integer x = 0.103287
#> Warning: non-integer x = 1.248586
#> Warning: non-integer x = 0.801670
#> Warning: non-integer x = 1.392470
#> Warning: non-integer x = 1.585398
#> Warning: non-integer x = 4.717833
#> Warning: non-integer x = 0.446766
#> Warning: non-integer x = 1.024374
#> Warning: non-integer x = 1.913715
#> Warning: non-integer x = 0.090262
#> Warning: non-integer x = 0.030660
#> Warning: non-integer x = 0.312592
#> Warning: non-integer x = 1.667274
#> Warning: non-integer x = 0.575956
#> Warning: non-integer x = 3.315860
#> Warning: non-integer x = 1.036147
#> Warning: non-integer x = 1.495338
#> Warning: non-integer x = 0.171595
#> Warning: non-integer x = 0.187992
#> Warning: non-integer x = 1.454239
#> Warning: non-integer x = 2.439561
#> Warning: non-integer x = 2.150087
#> Warning: non-integer x = 2.508992
#> Warning: non-integer x = 1.401021
#> Warning: non-integer x = 1.903772
#> Warning: non-integer x = 0.721154
#> Warning: non-integer x = 0.328607
#> Warning: non-integer x = 1.650544
#> Warning: non-integer x = 1.018610
#> Warning: non-integer x = 1.355172
#> Warning: non-integer x = 0.831146
#> Warning: non-integer x = 0.041587
#> Warning: non-integer x = 3.497807
#> Warning: non-integer x = 0.391151
#> Warning: non-integer x = 2.047021
#> Warning: non-integer x = 0.191762
#> Warning: non-integer x = 1.049118
#> Warning: non-integer x = 1.913931
#> Warning: non-integer x = 0.144950
#> Warning: non-integer x = 1.396132
#> Warning: non-integer x = 0.965283
#> Warning: non-integer x = 0.016745
#> Warning: non-integer x = 0.772033
#> Warning: non-integer x = 0.326837
#> Warning: non-integer x = 0.790828
#> Warning: non-integer x = 1.087742
#> Warning: non-integer x = 2.856512
#> Warning: non-integer x = 1.062409
#> Warning: non-integer x = 0.238571
#> Warning: non-integer x = 2.943939
#> Warning: non-integer x = 1.532391
#> Warning: non-integer x = 1.138945
#> Warning: non-integer x = 1.691602
#> Warning: non-integer x = 0.491995
#> Warning: non-integer x = 2.558831
#> Warning: non-integer x = 1.441681
#> Warning: non-integer x = 0.495036
#> Warning: non-integer x = 0.386791
#> Warning: non-integer x = 2.745629
#> Warning: non-integer x = 1.635628
#> Warning: non-integer x = 0.817493
#> Warning: non-integer x = 0.384472
#> Warning: non-integer x = 1.605519
#> Warning: non-integer x = 1.121056
#> Warning: non-integer x = 1.853555
#> Warning: non-integer x = 1.360035
#> Warning: non-integer x = 0.820903
#> Warning: non-integer x = 0.777940
#> Warning: non-integer x = 0.608443
#> Warning: non-integer x = 2.488749
#> Warning: non-integer x = 1.714441
#> Warning: non-integer x = 0.562917
#> Warning: non-integer x = 1.638451
#> Warning: non-integer x = 2.329638
#> Warning: non-integer x = 0.153642
#> Warning: non-integer x = 1.002635
#> Warning: non-integer x = 4.138626
#> Warning: non-integer x = 0.164232
#> Warning: non-integer x = 1.666851
#> Warning: non-integer x = 2.824156
#> Warning: non-integer x = 1.678732
#> Warning: non-integer x = 2.626520
#> Warning: non-integer x = 2.103914
#> Warning: non-integer x = 1.240175
#> Warning: non-integer x = 0.112431
#> Warning: non-integer x = 2.198803
#> Warning: non-integer x = 0.981681
#> Warning: non-integer x = 2.212143
#> Warning: non-integer x = 1.704907
#> Warning: non-integer x = 0.862861
#> Warning: non-integer x = 2.737372
#> Warning: non-integer x = 2.329113
#> Warning: non-integer x = 0.778462
#> Warning: non-integer x = 1.550559
#> Warning: non-integer x = 1.611504
#> Warning: non-integer x = 1.871838
#> Warning: non-integer x = 1.108818
#> Warning: non-integer x = 0.697359
#> Warning: non-integer x = 1.622173
#> Warning: non-integer x = 1.215600
#> Warning: non-integer x = 0.033662
#> Warning: non-integer x = 0.045931
#> Warning: non-integer x = 2.525894
#> Warning: non-integer x = 0.428878
#> Warning: non-integer x = 1.083054
#> Warning: non-integer x = 0.214985
#> Warning: non-integer x = 3.061140
#> Warning: non-integer x = 0.091858
#> Warning: non-integer x = 1.941849
#> Warning: non-integer x = 0.383474
#> Warning: non-integer x = 2.478862
#> Warning: non-integer x = 1.827821
#> Warning: non-integer x = 1.182632
#> Warning: non-integer x = 0.423702
#> Warning: non-integer x = 1.593210
#> Warning: non-integer x = 0.639830
#> Warning: non-integer x = 2.013986
#> Warning: non-integer x = 0.421167
#> Warning: non-integer x = 1.545852
#> Warning: non-integer x = 0.570734
#> Warning: non-integer x = 1.440062
#> Warning: non-integer x = 2.172035
#> Warning: non-integer x = 1.680553
#> Warning: non-integer x = 0.376128
#> Warning: non-integer x = 0.881360
#> Warning: non-integer x = 0.861259
#> Warning: non-integer x = 0.834333
#> Warning: non-integer x = 0.440069
#> Warning: non-integer x = 2.780081
#> Warning: non-integer x = 1.625967
#> Warning: non-integer x = 0.269587
#> Warning: non-integer x = 0.362290
#> Warning: non-integer x = 1.858228
#> Warning: non-integer x = 0.365095
#> Warning: non-integer x = 0.581741
#> Warning: non-integer x = 0.302837
#> Warning: non-integer x = 1.287876
#> Warning: non-integer x = 1.186492
#> Warning: non-integer x = 0.978069
#> Warning: non-integer x = 1.052398
#> Warning: non-integer x = 1.183242
#> Warning: non-integer x = 2.672755
#> Warning: non-integer x = 0.007271
#> Warning: non-integer x = 0.509963
#> Warning: non-integer x = 1.222826
#> Warning: non-integer x = 0.784968
#> Warning: non-integer x = 1.315805
#> Warning: non-integer x = 0.108770
#> Warning: non-integer x = 0.197121
#> Warning: non-integer x = 0.613634
#> Warning: non-integer x = 1.664812
#> Warning: non-integer x = 2.271223
#> Warning: non-integer x = 2.717116
#> Warning: non-integer x = 0.645394
#> Warning: non-integer x = 1.512021
#> Warning: non-integer x = 2.348657
#> Warning: non-integer x = 3.626765
#> Warning: non-integer x = 2.680998
#> Warning: non-integer x = 0.311894
#> Warning: non-integer x = 1.617437
#> Warning: non-integer x = 1.014720
#> Warning: non-integer x = 1.548167
#> Warning: non-integer x = 0.253260
#> Warning: non-integer x = 0.584125
#> Warning: non-integer x = 4.409865
#> Warning: non-integer x = 1.702741
#> Warning: non-integer x = 0.460848
#> Warning: non-integer x = 2.502054
#> Warning: non-integer x = 1.451856
#> Warning: non-integer x = 0.833336
#> Warning: non-integer x = 0.052506
#> Warning: non-integer x = 1.468280
#> Warning: non-integer x = 3.126024
#> Warning: non-integer x = 2.070193
#> Warning: non-integer x = 2.140353
#> Warning: non-integer x = 0.642901
#> Warning: non-integer x = 0.927361
#> Warning: non-integer x = 0.246876
#> Warning: non-integer x = 0.868700
#> Warning: non-integer x = 2.110950
#> Warning: non-integer x = 1.607074
#> Warning: non-integer x = 1.659619
#> Warning: non-integer x = 1.421704
#> Warning: non-integer x = 1.749802
#> Warning: non-integer x = 3.408456
#> Warning: non-integer x = 1.906526
#> Warning: non-integer x = 2.661618
#> Warning: non-integer x = 1.743832
#> Warning: non-integer x = 0.108898
#> Warning: non-integer x = 2.009999
#> Warning: non-integer x = 0.921294
#> Warning: non-integer x = 0.661317
#> Warning: non-integer x = 0.731966
#> Warning: non-integer x = 2.476961
#> Warning: non-integer x = 0.838377
#> Warning: non-integer x = 1.860162
#> Warning: non-integer x = 1.779590
#> Warning: non-integer x = 0.461457
#> Warning: non-integer x = 1.033026
#> Warning: non-integer x = 1.125074
#> Warning: non-integer x = 0.279839
#> Warning: non-integer x = 0.929117
#> Warning: non-integer x = 0.788654
#> Warning: non-integer x = 2.051428
#> Warning: non-integer x = 0.890404
#> Warning: non-integer x = 3.130247
#> Warning: non-integer x = 2.280143
#> Warning: non-integer x = 0.611591
#> Warning: non-integer x = 0.811506
#> Warning: non-integer x = 0.401571
#> Warning: non-integer x = 0.503970
#> Warning: non-integer x = 2.408663
#> Warning: non-integer x = 0.725579
#> Warning: non-integer x = 0.820431
#> Warning: non-integer x = 0.302874
#> Warning: non-integer x = 0.761120
#> Warning: non-integer x = 3.336875
#> Warning: non-integer x = 1.310002
#> Warning: non-integer x = 2.999562
#> Warning: non-integer x = 0.845005
#> Warning: non-integer x = 2.739514
#> Warning: non-integer x = 0.279311
#> Warning: non-integer x = 1.817692
#> Warning: non-integer x = 1.048733
#> Warning: non-integer x = 2.441738
#> Warning: non-integer x = 2.556105
#> Warning: non-integer x = 2.059943
#> Warning: non-integer x = 2.241341
#> Warning: non-integer x = 1.583860
#> Warning: non-integer x = 0.933733
#> Warning: non-integer x = 2.306874
#> Warning: non-integer x = 2.759385
#> Warning: non-integer x = 1.233588
#> Warning: non-integer x = 3.646629
#> Warning: non-integer x = 1.331660
#> Warning: non-integer x = 2.477980
#> Warning: non-integer x = 0.198068
#> Warning: non-integer x = 1.126863
#> Warning: non-integer x = 2.683696
#> Warning: non-integer x = 1.753995
#> Warning: non-integer x = 1.080207
#> Warning: non-integer x = 1.272346
#> Warning: non-integer x = 0.667995
#> Warning: non-integer x = 1.918827
#> Warning: non-integer x = 0.626934
#> Warning: non-integer x = 3.187733
#> Warning: non-integer x = 2.441507
#> Warning: non-integer x = 2.262283
#> Warning: non-integer x = 1.819952
#> Warning: non-integer x = 1.775303
#> Warning: non-integer x = 0.509117
#> Warning: non-integer x = 1.762886
#> Warning: non-integer x = 0.853188
#> Warning: non-integer x = 0.539720
#> Warning: non-integer x = 0.967656
#> Warning: non-integer x = 1.427542
#> Warning: non-integer x = 1.892995
#> Warning: non-integer x = 1.621686
#> Warning: non-integer x = 1.277493
#> Warning: non-integer x = 1.374193
#> Warning: non-integer x = 1.477322
#> Warning: non-integer x = 2.729319
#> Warning: non-integer x = 0.356799
#> Warning: non-integer x = 0.237991
#> Warning: non-integer x = 1.397416
#> Warning: non-integer x = 0.362517
#> Warning: non-integer x = 1.155786
#> Warning: non-integer x = 1.626675
#> Warning: non-integer x = 1.571215
#> Warning: non-integer x = 0.131352
#> Warning: non-integer x = 0.367965
#> Warning: non-integer x = 3.342558
#> Warning: non-integer x = 1.979226
#> Warning: non-integer x = 2.107959
#> Warning: non-integer x = 0.671366
#> Warning: non-integer x = 0.385737
#> Warning: non-integer x = 2.534293
#> Warning: non-integer x = 0.899019
#> Warning: non-integer x = 1.396105
#> Warning: non-integer x = 2.721236
#> Warning: non-integer x = 1.059625
#> Warning: non-integer x = 2.352696
#> Warning: non-integer x = 1.390529
#> Warning: non-integer x = 0.383898
#> Warning: non-integer x = 1.304462
#> Warning: non-integer x = 0.061076
#> Warning: non-integer x = 2.287517
#> Warning: non-integer x = 1.156450
#> Warning: non-integer x = 0.751499
#> Warning: non-integer x = 1.596402
#> Warning: non-integer x = 1.714359
#> Warning: non-integer x = 0.865194
#> Warning: non-integer x = 1.764031
#> Warning: non-integer x = 0.364238
#> Warning: non-integer x = 2.922840
#> Warning: non-integer x = 1.334281
#> Warning: non-integer x = 0.479273
#> Warning: non-integer x = 2.069389
#> Warning: non-integer x = 0.384723
#> Warning: non-integer x = 0.575095
#> Warning: non-integer x = 0.487503
#> Warning: non-integer x = 0.607252
#> Warning: non-integer x = 2.041908
#> Warning: non-integer x = 1.242493
#> Warning: non-integer x = 0.764751
#> Warning: non-integer x = 1.799862
#> Warning: non-integer x = 0.511605
#> Warning: non-integer x = 0.956867
#> Warning: non-integer x = 2.031956
#> Warning: non-integer x = 0.423911
#> Warning: non-integer x = 1.097233
#> Warning: non-integer x = 1.645181
#> Warning: non-integer x = 1.051479
#> Warning: non-integer x = 1.370275
#> Warning: non-integer x = 0.186539
#> Warning: non-integer x = 2.212732
#> Warning: non-integer x = 1.149576
#> Warning: non-integer x = 0.346775
#> Warning: non-integer x = 1.947587
#> Warning: non-integer x = 2.731989
#> Warning: non-integer x = 0.412855
#> Warning: non-integer x = 1.069964
#> Warning: non-integer x = 0.242143
#> Warning: non-integer x = 0.955912
#> Warning: non-integer x = 0.277765
#> Warning: non-integer x = 0.160659
#> Warning: non-integer x = 0.742388
#> Warning: non-integer x = 0.077142
#> Warning: non-integer x = 0.960911
#> Warning: non-integer x = 0.755217
#> Warning: non-integer x = 2.157717
#> Warning: non-integer x = 0.341709
#> Warning: non-integer x = 1.522462
#> Warning: non-integer x = 3.571466
#> Warning: non-integer x = 0.384969
#> Warning: non-integer x = 2.972990
#> Warning: non-integer x = 0.492637
#> Warning: non-integer x = 0.825933
#> Warning: non-integer x = 1.883706
#> Warning: non-integer x = 2.447081
#> Warning: non-integer x = 1.121285
#> Warning: non-integer x = 3.038001
#> Warning: non-integer x = 1.240061
#> Warning: non-integer x = 2.226268
#> Warning: non-integer x = 0.269986
#> Warning: non-integer x = 1.977449
#> Warning: non-integer x = 0.915897
#> Warning: non-integer x = 0.396322
#> Warning: non-integer x = 2.841229
#> Warning: non-integer x = 1.984866
#> Warning: non-integer x = 0.157533
#> Warning: non-integer x = 0.424704
#> Warning: non-integer x = 3.078116
#> Warning: non-integer x = 1.408289
#> Warning: non-integer x = 1.711799
#> Warning: non-integer x = 1.882598
#> Warning: non-integer x = 1.262230
#> Warning: non-integer x = 0.154624
#> Warning: non-integer x = 1.589509
#> Warning: non-integer x = 0.318276
#> Warning: non-integer x = 3.129971
#> Warning: non-integer x = 1.557985
#> Warning: non-integer x = 0.522088
#> Warning: non-integer x = 0.609402
#> Warning: non-integer x = 0.790154
#> Warning: non-integer x = 1.886869
#> Warning: non-integer x = 0.086540
#> Warning: non-integer x = 0.767631
#> Warning: non-integer x = 0.689129
#> Warning: non-integer x = 0.308448
#> Warning: non-integer x = 3.302778
#> Warning: non-integer x = 0.897996
#> Warning: non-integer x = 1.188951
#> Warning: non-integer x = 0.468221
#> Warning: non-integer x = 2.007874
#> Warning: non-integer x = 0.077602
#> Warning: non-integer x = 0.704848
#> Warning: non-integer x = 1.151393
#> Warning: non-integer x = 2.441479
#> Warning: non-integer x = 0.220809
#> Warning: non-integer x = 0.308947
#> Warning: non-integer x = 2.195087
#> Warning: non-integer x = 3.469527
#> Warning: non-integer x = 0.181481
#> Warning: non-integer x = 0.029063
#> Warning: non-integer x = 0.013276
#> Warning: non-integer x = 1.115209
#> Warning: non-integer x = 0.714560
#> Warning: non-integer x = 1.319516
#> Warning: non-integer x = 1.247366
#> Warning: non-integer x = 0.017833
#> Warning: non-integer x = 0.940903
#> Warning: non-integer x = 0.342345
#> Warning: non-integer x = 2.471441
#> Warning: non-integer x = 0.807630
#> Warning: non-integer x = 0.673523
#> Warning: non-integer x = 1.869119
#> Warning: non-integer x = 1.116123
#> Warning: non-integer x = 1.064037
#> Warning: non-integer x = 2.849496
#> Warning: non-integer x = 1.176910
#> Warning: non-integer x = 0.087852
#> Warning: non-integer x = 0.795176
#> Warning: non-integer x = 2.783284
#> Warning: non-integer x = 0.400724
#> Warning: non-integer x = 2.342831
#> Warning: non-integer x = 0.072201
#> Warning: non-integer x = 0.505805
#> Warning: non-integer x = 0.004215
#> Warning: non-integer x = 1.008843
#> Warning: non-integer x = 0.135727
#> Warning: non-integer x = 1.766337
#> Warning: non-integer x = 1.275022
#> Warning: non-integer x = 1.926245
#> Warning: non-integer x = 0.617269
#> Warning: non-integer x = 0.787647
#> Warning: non-integer x = 0.403132
#> Warning: non-integer x = 1.087645
#> Warning: non-integer x = 1.166149
#> Warning: non-integer x = 0.187793
#> Warning: non-integer x = 0.428166
#> Warning: non-integer x = 0.454756
#> Warning: non-integer x = 0.115786
#> Warning: non-integer x = 1.650907
#> Warning: non-integer x = 1.521209
#> Warning: non-integer x = 1.942132
#> Warning: non-integer x = 0.806184
#> Warning: non-integer x = 0.411159
#> Warning: non-integer x = 0.053887
#> Warning: non-integer x = 0.131650
#> Warning: non-integer x = 2.468316
#> Warning: non-integer x = 0.579091
#> Warning: non-integer x = 0.204895
#> Warning: non-integer x = 0.801982
#> Warning: non-integer x = 3.151103
#> Warning: non-integer x = 2.145739
#> Warning: non-integer x = 0.032394
#> Warning: non-integer x = 0.074398
#> Warning: non-integer x = 1.402153
#> Warning: non-integer x = 1.203837
#> Warning: non-integer x = 2.139788
#> Warning: non-integer x = 0.201358
#> Warning: non-integer x = 1.958530
#> Warning: non-integer x = 3.777587
#> Warning: non-integer x = 1.148436
#> Warning: non-integer x = 1.257105
#> Warning: non-integer x = 0.897863
#> Warning: non-integer x = 1.334264
#> Warning: non-integer x = 1.476558
#> Warning: non-integer x = 1.250898
#> Warning: non-integer x = 2.022725
#> Warning: non-integer x = 0.977313
#> Warning: non-integer x = 0.431814
#> Warning: non-integer x = 1.019253
#> Warning: non-integer x = 0.327752
#> Warning: non-integer x = 1.641585
#> Warning: non-integer x = 1.657739
#> Warning: non-integer x = 0.651481
#> Warning: non-integer x = 0.499031
#> Warning: non-integer x = 0.702893
#> Warning: non-integer x = 0.963705
#> Warning: non-integer x = 3.640712
#> Warning: non-integer x = 2.122544
#> Warning: non-integer x = 0.148078
#> Warning: non-integer x = 1.874536
#> Warning: non-integer x = 0.538796
#> Warning: non-integer x = 0.441277
#> Warning: non-integer x = 2.193670
#> Warning: non-integer x = 1.114936
#> Warning: non-integer x = 0.356034
#> Warning: non-integer x = 0.383622
#> Warning: non-integer x = 1.102133
#> Warning: non-integer x = 1.687291
#> Warning: non-integer x = 1.121082
#> Warning: non-integer x = 0.475489
#> Warning: non-integer x = 0.561569
#> Warning: non-integer x = 1.549104
#> Warning: non-integer x = 3.242242
#> Warning: non-integer x = 1.257762
#> Warning: non-integer x = 0.261828
#> Warning: non-integer x = 1.697942
#> Warning: non-integer x = 1.503167
#> Warning: non-integer x = 0.722303
#> Warning: non-integer x = 3.527690
#> Warning: non-integer x = 1.913522
#> Warning: non-integer x = 1.179808
#> Warning: non-integer x = 0.453867
#> Warning: non-integer x = 1.667568
#> Warning: non-integer x = 1.093568
#> Warning: non-integer x = 0.330112
#> Warning: non-integer x = 0.943712
#> Warning: non-integer x = 2.109060
#> Warning: non-integer x = 3.140045
#> Warning: non-integer x = 1.335630
#> Warning: non-integer x = 0.474852
#> Warning: non-integer x = 0.265829
#> Warning: non-integer x = 1.533582
#> Warning: non-integer x = 0.113867
#> Warning: non-integer x = 0.236193
#> Warning: non-integer x = 1.271228
#> Warning: non-integer x = 0.864030
#> Warning: non-integer x = 0.065587
#> Warning: non-integer x = 0.216091
#> Warning: non-integer x = 0.823970
#> Warning: non-integer x = 0.801096
#> Warning: non-integer x = 2.528915
#> Warning: non-integer x = 1.814121
#> Warning: non-integer x = 0.193317
#> Warning: non-integer x = 2.876536
#> Warning: non-integer x = 0.474258
#> Warning: non-integer x = 1.125779
#> Warning: non-integer x = 0.855415
#> Warning: non-integer x = 0.218863
#> Warning: non-integer x = 2.130197
#> Warning: non-integer x = 2.337149
#> Warning: non-integer x = 1.402341
#> Warning: non-integer x = 0.299132
#> Warning: non-integer x = 1.026315
#> Warning: non-integer x = 1.356187
#> Warning: non-integer x = 2.792491
#> Warning: non-integer x = 1.110354
#> Warning: non-integer x = 1.228096
#> Warning: non-integer x = 2.275235
#> Warning: non-integer x = 2.322216
#> Warning: non-integer x = 0.617695
#> Warning: non-integer x = 2.413996
#> Warning: non-integer x = 0.834567
#> Warning: non-integer x = 0.320030
#> Warning: non-integer x = 2.332053
#> Warning: non-integer x = 0.588241
#> Warning: non-integer x = 3.962789
#> Warning: non-integer x = 2.335065
#> Warning: non-integer x = 0.290624
#> Warning: non-integer x = 1.407680
#> Warning: non-integer x = 1.320990
#> Warning: non-integer x = 1.135612
#> Warning: non-integer x = 1.504903
#> Warning: non-integer x = 2.563531
#> Warning: non-integer x = 3.621231
#> Warning: non-integer x = 2.033009
#> Warning: non-integer x = 0.772309
#> Warning: non-integer x = 0.374898
#> Warning: non-integer x = 1.773742
#> Warning: non-integer x = 1.712649
#> Warning: non-integer x = 2.627405
#> Warning: non-integer x = 0.062975
#> Warning: non-integer x = 0.206337
#> Warning: non-integer x = 1.420954
#> Warning: non-integer x = 1.968419
#> Warning: non-integer x = 0.992695
#> Warning: non-integer x = 0.173834
#> Warning: non-integer x = 1.462366
#> Warning: non-integer x = 0.643105
#> Warning: non-integer x = 0.490535
#> Warning: non-integer x = 0.867849
#> Warning: non-integer x = 0.042690
#> Warning: non-integer x = 0.518683
#> Warning: non-integer x = 1.419685
#> Warning: non-integer x = 1.331956
#> Warning: non-integer x = 1.168467
#> Warning: non-integer x = 3.117666
#> Warning: non-integer x = 2.843952
#> Warning: non-integer x = 2.885756
#> Warning: non-integer x = 0.007630
#> Warning: non-integer x = 1.661278
#> Warning: non-integer x = 1.216706
#> Warning: non-integer x = 1.510672
#> Warning: non-integer x = 1.567926
#> Warning: non-integer x = 2.795431
#> Warning: non-integer x = 1.301066
#> Warning: non-integer x = 0.066334
#> Warning: non-integer x = 3.400326
#> Warning: non-integer x = 0.994187
#> Warning: non-integer x = 1.452218
#> Warning: non-integer x = 0.750702
#> Warning: non-integer x = 1.817621
#> Warning: non-integer x = 1.757643
#> Warning: non-integer x = 1.998397
#> Warning: non-integer x = 0.669555
#> Warning: non-integer x = 0.937740
#> Warning: non-integer x = 2.814152
#> Warning: non-integer x = 0.253497
#> Warning: non-integer x = 3.127857
#> Warning: non-integer x = 0.567473
#> Warning: non-integer x = 2.189085
#> Warning: non-integer x = 0.501505
#> Warning: non-integer x = 0.197165
#> Warning: non-integer x = 2.527452
#> Warning: non-integer x = 1.281083
#> Warning: non-integer x = 1.667021
#> Warning: non-integer x = 0.353159
#> Warning: non-integer x = 1.584775
#> Warning: non-integer x = 1.859993
#> Warning: non-integer x = 1.776090
#> Warning: non-integer x = 1.551579
#> Warning: non-integer x = 3.090297
#> Warning: non-integer x = 0.983235
#> Warning: non-integer x = 1.111238
#> Warning: non-integer x = 2.587701