Estimate the Individual Treatment Effect using Poisson Regression

Estimates the Individual Treatment Effect using Poisson Regression given a response vector, a treatment vector, and a features matrix.

Usage

estimate_ite_poisson(y, z, X, X_names, include_offset, offset_name)

Arguments

y

the observed response vector

z

the treatment vector

X

the features matrix

X_names

the names of the covariates

include_offset

whether or not to include an offset when estimating the ITE, for poisson only

offset_name

the name of the offset, if it is to be included

Value

a vector of ITE estimates

Examples

dataset <- generate_cre_dataset(n = 1000, rho = 0, n_rules = 2, p = 10,
                                effect_size = 2, binary = FALSE)

# Initialize parameters
y <- abs(dataset[["y"]])
z <- dataset[["z"]]
X <- as.data.frame(dataset[["X"]])
X_names <- names(as.data.frame(X))
include_offset <- FALSE
offset_name <- NA

ite_list <- estimate_ite_poisson(y, z, X, X_names, include_offset, offset_name)
#> Warning: non-integer x = 0.456879
#> Warning: non-integer x = 1.489806
#> Warning: non-integer x = 0.494746
#> Warning: non-integer x = 4.890329
#> Warning: non-integer x = 3.824089
#> Warning: non-integer x = 0.923054
#> Warning: non-integer x = 1.125050
#> Warning: non-integer x = 0.811831
#> Warning: non-integer x = 3.429299
#> Warning: non-integer x = 0.099516
#> Warning: non-integer x = 2.482716
#> Warning: non-integer x = 2.306013
#> Warning: non-integer x = 2.332381
#> Warning: non-integer x = 2.886826
#> Warning: non-integer x = 0.471568
#> Warning: non-integer x = 2.907377
#> Warning: non-integer x = 1.493900
#> Warning: non-integer x = 0.203277
#> Warning: non-integer x = 0.540013
#> Warning: non-integer x = 1.704954
#> Warning: non-integer x = 1.755060
#> Warning: non-integer x = 2.046890
#> Warning: non-integer x = 1.037612
#> Warning: non-integer x = 2.959313
#> Warning: non-integer x = 2.512865
#> Warning: non-integer x = 0.123419
#> Warning: non-integer x = 0.545066
#> Warning: non-integer x = 1.386223
#> Warning: non-integer x = 2.914829
#> Warning: non-integer x = 1.064236
#> Warning: non-integer x = 4.495462
#> Warning: non-integer x = 2.029214
#> Warning: non-integer x = 0.766872
#> Warning: non-integer x = 1.325937
#> Warning: non-integer x = 1.826203
#> Warning: non-integer x = 1.135489
#> Warning: non-integer x = 2.682334
#> Warning: non-integer x = 0.175049
#> Warning: non-integer x = 2.167225
#> Warning: non-integer x = 2.387175
#> Warning: non-integer x = 1.619745
#> Warning: non-integer x = 1.616440
#> Warning: non-integer x = 1.704252
#> Warning: non-integer x = 1.596871
#> Warning: non-integer x = 1.632049
#> Warning: non-integer x = 2.100764
#> Warning: non-integer x = 0.719695
#> Warning: non-integer x = 2.862873
#> Warning: non-integer x = 0.011514
#> Warning: non-integer x = 0.182400
#> Warning: non-integer x = 0.205090
#> Warning: non-integer x = 4.436501
#> Warning: non-integer x = 2.683465
#> Warning: non-integer x = 1.292553
#> Warning: non-integer x = 0.330294
#> Warning: non-integer x = 1.216663
#> Warning: non-integer x = 1.876704
#> Warning: non-integer x = 1.386567
#> Warning: non-integer x = 2.330928
#> Warning: non-integer x = 1.412703
#> Warning: non-integer x = 2.303606
#> Warning: non-integer x = 3.677253
#> Warning: non-integer x = 1.744782
#> Warning: non-integer x = 0.449380
#> Warning: non-integer x = 1.469390
#> Warning: non-integer x = 2.641266
#> Warning: non-integer x = 1.614350
#> Warning: non-integer x = 0.103002
#> Warning: non-integer x = 0.109639
#> Warning: non-integer x = 3.059717
#> Warning: non-integer x = 1.610822
#> Warning: non-integer x = 1.857188
#> Warning: non-integer x = 1.598204
#> Warning: non-integer x = 0.144650
#> Warning: non-integer x = 0.196685
#> Warning: non-integer x = 0.127937
#> Warning: non-integer x = 2.924841
#> Warning: non-integer x = 2.713614
#> Warning: non-integer x = 0.170375
#> Warning: non-integer x = 0.565443
#> Warning: non-integer x = 4.440424
#> Warning: non-integer x = 0.106728
#> Warning: non-integer x = 0.961376
#> Warning: non-integer x = 2.800082
#> Warning: non-integer x = 4.014813
#> Warning: non-integer x = 2.431742
#> Warning: non-integer x = 1.431030
#> Warning: non-integer x = 3.023100
#> Warning: non-integer x = 0.107851
#> Warning: non-integer x = 2.461724
#> Warning: non-integer x = 1.007244
#> Warning: non-integer x = 0.226180
#> Warning: non-integer x = 1.700853
#> Warning: non-integer x = 2.261105
#> Warning: non-integer x = 2.128226
#> Warning: non-integer x = 1.000592
#> Warning: non-integer x = 0.797414
#> Warning: non-integer x = 1.759851
#> Warning: non-integer x = 0.424798
#> Warning: non-integer x = 0.204778
#> Warning: non-integer x = 0.598120
#> Warning: non-integer x = 3.492223
#> Warning: non-integer x = 2.850677
#> Warning: non-integer x = 0.127777
#> Warning: non-integer x = 2.189885
#> Warning: non-integer x = 0.542404
#> Warning: non-integer x = 2.127339
#> Warning: non-integer x = 0.190935
#> Warning: non-integer x = 5.009747
#> Warning: non-integer x = 0.275064
#> Warning: non-integer x = 2.841297
#> Warning: non-integer x = 2.065564
#> Warning: non-integer x = 3.600884
#> Warning: non-integer x = 2.925648
#> Warning: non-integer x = 0.742861
#> Warning: non-integer x = 0.295042
#> Warning: non-integer x = 4.565977
#> Warning: non-integer x = 1.671789
#> Warning: non-integer x = 2.644703
#> Warning: non-integer x = 1.408451
#> Warning: non-integer x = 0.851468
#> Warning: non-integer x = 1.794322
#> Warning: non-integer x = 3.851071
#> Warning: non-integer x = 1.877141
#> Warning: non-integer x = 0.425472
#> Warning: non-integer x = 1.689146
#> Warning: non-integer x = 2.080257
#> Warning: non-integer x = 6.265866
#> Warning: non-integer x = 2.373686
#> Warning: non-integer x = 2.732018
#> Warning: non-integer x = 3.568149
#> Warning: non-integer x = 1.378464
#> Warning: non-integer x = 1.760666
#> Warning: non-integer x = 2.246144
#> Warning: non-integer x = 0.199937
#> Warning: non-integer x = 4.095424
#> Warning: non-integer x = 1.326759
#> Warning: non-integer x = 0.318765
#> Warning: non-integer x = 0.096914
#> Warning: non-integer x = 0.124523
#> Warning: non-integer x = 0.619778
#> Warning: non-integer x = 1.328075
#> Warning: non-integer x = 0.950078
#> Warning: non-integer x = 2.410076
#> Warning: non-integer x = 1.100609
#> Warning: non-integer x = 0.656719
#> Warning: non-integer x = 0.841012
#> Warning: non-integer x = 2.246684
#> Warning: non-integer x = 1.112348
#> Warning: non-integer x = 1.523656
#> Warning: non-integer x = 2.514551
#> Warning: non-integer x = 0.337387
#> Warning: non-integer x = 1.936857
#> Warning: non-integer x = 0.131064
#> Warning: non-integer x = 0.041231
#> Warning: non-integer x = 0.047863
#> Warning: non-integer x = 0.394289
#> Warning: non-integer x = 0.910715
#> Warning: non-integer x = 0.892059
#> Warning: non-integer x = 1.304704
#> Warning: non-integer x = 0.964091
#> Warning: non-integer x = 2.622917
#> Warning: non-integer x = 3.999929
#> Warning: non-integer x = 0.064835
#> Warning: non-integer x = 0.974333
#> Warning: non-integer x = 2.561134
#> Warning: non-integer x = 1.144034
#> Warning: non-integer x = 2.000034
#> Warning: non-integer x = 1.208943
#> Warning: non-integer x = 3.176010
#> Warning: non-integer x = 2.702670
#> Warning: non-integer x = 0.545826
#> Warning: non-integer x = 4.934110
#> Warning: non-integer x = 1.881093
#> Warning: non-integer x = 0.122587
#> Warning: non-integer x = 1.519245
#> Warning: non-integer x = 3.192798
#> Warning: non-integer x = 1.068438
#> Warning: non-integer x = 0.071991
#> Warning: non-integer x = 2.000716
#> Warning: non-integer x = 1.409635
#> Warning: non-integer x = 2.840260
#> Warning: non-integer x = 0.468068
#> Warning: non-integer x = 1.642840
#> Warning: non-integer x = 2.051297
#> Warning: non-integer x = 2.772703
#> Warning: non-integer x = 0.173354
#> Warning: non-integer x = 1.012343
#> Warning: non-integer x = 2.760291
#> Warning: non-integer x = 0.470948
#> Warning: non-integer x = 1.277116
#> Warning: non-integer x = 1.944623
#> Warning: non-integer x = 1.029186
#> Warning: non-integer x = 4.632567
#> Warning: non-integer x = 0.045431
#> Warning: non-integer x = 2.106154
#> Warning: non-integer x = 0.408853
#> Warning: non-integer x = 0.794557
#> Warning: non-integer x = 2.450835
#> Warning: non-integer x = 1.109195
#> Warning: non-integer x = 0.325941
#> Warning: non-integer x = 3.240039
#> Warning: non-integer x = 1.151159
#> Warning: non-integer x = 0.376259
#> Warning: non-integer x = 2.258486
#> Warning: non-integer x = 0.166633
#> Warning: non-integer x = 2.198402
#> Warning: non-integer x = 0.458715
#> Warning: non-integer x = 0.295415
#> Warning: non-integer x = 3.978164
#> Warning: non-integer x = 3.375800
#> Warning: non-integer x = 1.775300
#> Warning: non-integer x = 0.916317
#> Warning: non-integer x = 1.501444
#> Warning: non-integer x = 0.812185
#> Warning: non-integer x = 0.077247
#> Warning: non-integer x = 2.006817
#> Warning: non-integer x = 0.616757
#> Warning: non-integer x = 3.699260
#> Warning: non-integer x = 0.816640
#> Warning: non-integer x = 1.989990
#> Warning: non-integer x = 2.989602
#> Warning: non-integer x = 0.055087
#> Warning: non-integer x = 0.565388
#> Warning: non-integer x = 1.326825
#> Warning: non-integer x = 1.001388
#> Warning: non-integer x = 0.299212
#> Warning: non-integer x = 0.990918
#> Warning: non-integer x = 0.367408
#> Warning: non-integer x = 1.523318
#> Warning: non-integer x = 1.251771
#> Warning: non-integer x = 2.487677
#> Warning: non-integer x = 1.461112
#> Warning: non-integer x = 0.504659
#> Warning: non-integer x = 0.541125
#> Warning: non-integer x = 2.772193
#> Warning: non-integer x = 3.322805
#> Warning: non-integer x = 0.747494
#> Warning: non-integer x = 0.137457
#> Warning: non-integer x = 2.180577
#> Warning: non-integer x = 1.794995
#> Warning: non-integer x = 1.711083
#> Warning: non-integer x = 0.998313
#> Warning: non-integer x = 0.016065
#> Warning: non-integer x = 0.959359
#> Warning: non-integer x = 0.032843
#> Warning: non-integer x = 0.666065
#> Warning: non-integer x = 3.886710
#> Warning: non-integer x = 2.817100
#> Warning: non-integer x = 0.151447
#> Warning: non-integer x = 3.414278
#> Warning: non-integer x = 0.275101
#> Warning: non-integer x = 0.946006
#> Warning: non-integer x = 2.450531
#> Warning: non-integer x = 0.262588
#> Warning: non-integer x = 2.593528
#> Warning: non-integer x = 2.870965
#> Warning: non-integer x = 2.188924
#> Warning: non-integer x = 2.659050
#> Warning: non-integer x = 2.076031
#> Warning: non-integer x = 0.614710
#> Warning: non-integer x = 0.553920
#> Warning: non-integer x = 3.320861
#> Warning: non-integer x = 1.366018
#> Warning: non-integer x = 1.038784
#> Warning: non-integer x = 1.016998
#> Warning: non-integer x = 0.398743
#> Warning: non-integer x = 0.290915
#> Warning: non-integer x = 0.161900
#> Warning: non-integer x = 2.843518
#> Warning: non-integer x = 1.575470
#> Warning: non-integer x = 3.001489
#> Warning: non-integer x = 1.116987
#> Warning: non-integer x = 2.986001
#> Warning: non-integer x = 2.286529
#> Warning: non-integer x = 1.280394
#> Warning: non-integer x = 3.209503
#> Warning: non-integer x = 1.286048
#> Warning: non-integer x = 3.397908
#> Warning: non-integer x = 3.375759
#> Warning: non-integer x = 1.615787
#> Warning: non-integer x = 2.001331
#> Warning: non-integer x = 0.623483
#> Warning: non-integer x = 0.737266
#> Warning: non-integer x = 1.160683
#> Warning: non-integer x = 0.010917
#> Warning: non-integer x = 3.504973
#> Warning: non-integer x = 1.758615
#> Warning: non-integer x = 1.828217
#> Warning: non-integer x = 1.080315
#> Warning: non-integer x = 0.916608
#> Warning: non-integer x = 4.118376
#> Warning: non-integer x = 3.406253
#> Warning: non-integer x = 2.761822
#> Warning: non-integer x = 2.807245
#> Warning: non-integer x = 1.351555
#> Warning: non-integer x = 0.779415
#> Warning: non-integer x = 3.219844
#> Warning: non-integer x = 0.225459
#> Warning: non-integer x = 0.202643
#> Warning: non-integer x = 0.884533
#> Warning: non-integer x = 2.550737
#> Warning: non-integer x = 3.589312
#> Warning: non-integer x = 0.535413
#> Warning: non-integer x = 1.312815
#> Warning: non-integer x = 3.039608
#> Warning: non-integer x = 1.979485
#> Warning: non-integer x = 1.767032
#> Warning: non-integer x = 2.311045
#> Warning: non-integer x = 1.173351
#> Warning: non-integer x = 5.234573
#> Warning: non-integer x = 3.425252
#> Warning: non-integer x = 0.083605
#> Warning: non-integer x = 1.864542
#> Warning: non-integer x = 1.254379
#> Warning: non-integer x = 1.050037
#> Warning: non-integer x = 0.635252
#> Warning: non-integer x = 2.955772
#> Warning: non-integer x = 0.944550
#> Warning: non-integer x = 2.546368
#> Warning: non-integer x = 1.803563
#> Warning: non-integer x = 0.843370
#> Warning: non-integer x = 0.182222
#> Warning: non-integer x = 1.816178
#> Warning: non-integer x = 2.071535
#> Warning: non-integer x = 1.325818
#> Warning: non-integer x = 0.445072
#> Warning: non-integer x = 0.553156
#> Warning: non-integer x = 2.257373
#> Warning: non-integer x = 0.586474
#> Warning: non-integer x = 1.954639
#> Warning: non-integer x = 3.957715
#> Warning: non-integer x = 0.514834
#> Warning: non-integer x = 2.069502
#> Warning: non-integer x = 0.489252
#> Warning: non-integer x = 0.543091
#> Warning: non-integer x = 1.792968
#> Warning: non-integer x = 2.120558
#> Warning: non-integer x = 0.382705
#> Warning: non-integer x = 0.992130
#> Warning: non-integer x = 0.176020
#> Warning: non-integer x = 0.424063
#> Warning: non-integer x = 0.577458
#> Warning: non-integer x = 3.628702
#> Warning: non-integer x = 0.267032
#> Warning: non-integer x = 2.723623
#> Warning: non-integer x = 0.838577
#> Warning: non-integer x = 2.350011
#> Warning: non-integer x = 0.097154
#> Warning: non-integer x = 0.769140
#> Warning: non-integer x = 4.141423
#> Warning: non-integer x = 0.565416
#> Warning: non-integer x = 2.661020
#> Warning: non-integer x = 0.348056
#> Warning: non-integer x = 1.667472
#> Warning: non-integer x = 0.444861
#> Warning: non-integer x = 1.425654
#> Warning: non-integer x = 2.028028
#> Warning: non-integer x = 0.426802
#> Warning: non-integer x = 0.919953
#> Warning: non-integer x = 1.114300
#> Warning: non-integer x = 0.951767
#> Warning: non-integer x = 2.366279
#> Warning: non-integer x = 2.420164
#> Warning: non-integer x = 0.517115
#> Warning: non-integer x = 1.788779
#> Warning: non-integer x = 2.053977
#> Warning: non-integer x = 0.320759
#> Warning: non-integer x = 0.022443
#> Warning: non-integer x = 1.547866
#> Warning: non-integer x = 2.515674
#> Warning: non-integer x = 2.664522
#> Warning: non-integer x = 2.213775
#> Warning: non-integer x = 0.275693
#> Warning: non-integer x = 0.577308
#> Warning: non-integer x = 1.409185
#> Warning: non-integer x = 0.381002
#> Warning: non-integer x = 0.507496
#> Warning: non-integer x = 1.676832
#> Warning: non-integer x = 0.999006
#> Warning: non-integer x = 3.301411
#> Warning: non-integer x = 1.656980
#> Warning: non-integer x = 1.228705
#> Warning: non-integer x = 0.549769
#> Warning: non-integer x = 1.379380
#> Warning: non-integer x = 0.021918
#> Warning: non-integer x = 1.578871
#> Warning: non-integer x = 4.385952
#> Warning: non-integer x = 1.702054
#> Warning: non-integer x = 1.854289
#> Warning: non-integer x = 1.908763
#> Warning: non-integer x = 1.252006
#> Warning: non-integer x = 0.825495
#> Warning: non-integer x = 1.209879
#> Warning: non-integer x = 1.449071
#> Warning: non-integer x = 1.019531
#> Warning: non-integer x = 0.924272
#> Warning: non-integer x = 0.443122
#> Warning: non-integer x = 1.996520
#> Warning: non-integer x = 0.012910
#> Warning: non-integer x = 1.861917
#> Warning: non-integer x = 2.468322
#> Warning: non-integer x = 2.627107
#> Warning: non-integer x = 0.746761
#> Warning: non-integer x = 0.591409
#> Warning: non-integer x = 0.035564
#> Warning: non-integer x = 0.383182
#> Warning: non-integer x = 0.941788
#> Warning: non-integer x = 1.409384
#> Warning: non-integer x = 1.523333
#> Warning: non-integer x = 0.682273
#> Warning: non-integer x = 2.074485
#> Warning: non-integer x = 2.273705
#> Warning: non-integer x = 2.480292
#> Warning: non-integer x = 1.596709
#> Warning: non-integer x = 1.720799
#> Warning: non-integer x = 1.635220
#> Warning: non-integer x = 1.015768
#> Warning: non-integer x = 0.416883
#> Warning: non-integer x = 1.262516
#> Warning: non-integer x = 2.767332
#> Warning: non-integer x = 1.333975
#> Warning: non-integer x = 0.579387
#> Warning: non-integer x = 0.578771
#> Warning: non-integer x = 2.446087
#> Warning: non-integer x = 2.613331
#> Warning: non-integer x = 3.600107
#> Warning: non-integer x = 3.852013
#> Warning: non-integer x = 1.564280
#> Warning: non-integer x = 1.059522
#> Warning: non-integer x = 1.397172
#> Warning: non-integer x = 0.379764
#> Warning: non-integer x = 1.287250
#> Warning: non-integer x = 1.279613
#> Warning: non-integer x = 1.854842
#> Warning: non-integer x = 1.820735
#> Warning: non-integer x = 0.244728
#> Warning: non-integer x = 2.572538
#> Warning: non-integer x = 1.158816
#> Warning: non-integer x = 1.243178
#> Warning: non-integer x = 0.101984
#> Warning: non-integer x = 1.996464
#> Warning: non-integer x = 1.449763
#> Warning: non-integer x = 0.546417
#> Warning: non-integer x = 0.109844
#> Warning: non-integer x = 2.208273
#> Warning: non-integer x = 0.130270
#> Warning: non-integer x = 1.871423
#> Warning: non-integer x = 0.088549
#> Warning: non-integer x = 1.470905
#> Warning: non-integer x = 2.604925
#> Warning: non-integer x = 0.067088
#> Warning: non-integer x = 0.968690
#> Warning: non-integer x = 0.320439
#> Warning: non-integer x = 0.598921
#> Warning: non-integer x = 3.374386
#> Warning: non-integer x = 0.770341
#> Warning: non-integer x = 1.464991
#> Warning: non-integer x = 2.084049
#> Warning: non-integer x = 0.105897
#> Warning: non-integer x = 0.281049
#> Warning: non-integer x = 0.595356
#> Warning: non-integer x = 1.242873
#> Warning: non-integer x = 2.227059
#> Warning: non-integer x = 1.949817
#> Warning: non-integer x = 1.324747
#> Warning: non-integer x = 0.319531
#> Warning: non-integer x = 1.101648
#> Warning: non-integer x = 1.512652
#> Warning: non-integer x = 0.366639
#> Warning: non-integer x = 1.309302
#> Warning: non-integer x = 0.524582
#> Warning: non-integer x = 1.265929
#> Warning: non-integer x = 2.897336
#> Warning: non-integer x = 0.652635
#> Warning: non-integer x = 3.194360
#> Warning: non-integer x = 0.227537
#> Warning: non-integer x = 1.807809
#> Warning: non-integer x = 0.240155
#> Warning: non-integer x = 2.826700
#> Warning: non-integer x = 1.718813
#> Warning: non-integer x = 4.284997
#> Warning: non-integer x = 5.001056
#> Warning: non-integer x = 2.433671
#> Warning: non-integer x = 1.968146
#> Warning: non-integer x = 0.661762
#> Warning: non-integer x = 2.021908
#> Warning: non-integer x = 0.298369
#> Warning: non-integer x = 1.410491
#> Warning: non-integer x = 2.027573
#> Warning: non-integer x = 0.309552
#> Warning: non-integer x = 1.221145
#> Warning: non-integer x = 3.750890
#> Warning: non-integer x = 1.643894
#> Warning: non-integer x = 0.523361
#> Warning: non-integer x = 2.442453
#> Warning: non-integer x = 1.360447
#> Warning: non-integer x = 0.819094
#> Warning: non-integer x = 2.417349
#> Warning: non-integer x = 1.127275
#> Warning: non-integer x = 1.142338
#> Warning: non-integer x = 0.226893
#> Warning: non-integer x = 0.353288
#> Warning: non-integer x = 0.784317
#> Warning: non-integer x = 0.314783
#> Warning: non-integer x = 1.308981
#> Warning: non-integer x = 0.461345
#> Warning: non-integer x = 1.476880
#> Warning: non-integer x = 1.926122
#> Warning: non-integer x = 0.821690
#> Warning: non-integer x = 2.993643
#> Warning: non-integer x = 1.376629
#> Warning: non-integer x = 1.072351
#> Warning: non-integer x = 0.173409
#> Warning: non-integer x = 1.242119
#> Warning: non-integer x = 2.959232
#> Warning: non-integer x = 2.216966
#> Warning: non-integer x = 0.759965
#> Warning: non-integer x = 1.500075
#> Warning: non-integer x = 1.153939
#> Warning: non-integer x = 1.524055
#> Warning: non-integer x = 1.970696
#> Warning: non-integer x = 1.894332
#> Warning: non-integer x = 0.249007
#> Warning: non-integer x = 0.767340
#> Warning: non-integer x = 1.733524
#> Warning: non-integer x = 0.324790
#> Warning: non-integer x = 0.502991
#> Warning: non-integer x = 1.454277
#> Warning: non-integer x = 1.448365
#> Warning: non-integer x = 2.545469
#> Warning: non-integer x = 0.257480
#> Warning: non-integer x = 1.891544
#> Warning: non-integer x = 0.413969
#> Warning: non-integer x = 1.883880
#> Warning: non-integer x = 2.244689
#> Warning: non-integer x = 1.919268
#> Warning: non-integer x = 2.724864
#> Warning: non-integer x = 2.145090
#> Warning: non-integer x = 1.675283
#> Warning: non-integer x = 0.488487
#> Warning: non-integer x = 0.426692
#> Warning: non-integer x = 0.138765
#> Warning: non-integer x = 1.753951
#> Warning: non-integer x = 1.244979
#> Warning: non-integer x = 2.291991
#> Warning: non-integer x = 1.402444
#> Warning: non-integer x = 2.838155
#> Warning: non-integer x = 2.036859
#> Warning: non-integer x = 0.248537
#> Warning: non-integer x = 0.914066
#> Warning: non-integer x = 1.476236
#> Warning: non-integer x = 2.302677
#> Warning: non-integer x = 1.166803
#> Warning: non-integer x = 1.871569
#> Warning: non-integer x = 1.452495
#> Warning: non-integer x = 1.730742
#> Warning: non-integer x = 2.945447
#> Warning: non-integer x = 2.032431
#> Warning: non-integer x = 1.836459
#> Warning: non-integer x = 1.137156
#> Warning: non-integer x = 1.869974
#> Warning: non-integer x = 1.067893
#> Warning: non-integer x = 2.450323
#> Warning: non-integer x = 0.871976
#> Warning: non-integer x = 0.900346
#> Warning: non-integer x = 1.770269
#> Warning: non-integer x = 0.656113
#> Warning: non-integer x = 1.221407
#> Warning: non-integer x = 2.260538
#> Warning: non-integer x = 3.817044
#> Warning: non-integer x = 1.574219
#> Warning: non-integer x = 0.983520
#> Warning: non-integer x = 2.346362
#> Warning: non-integer x = 1.642309
#> Warning: non-integer x = 1.199176
#> Warning: non-integer x = 0.695913
#> Warning: non-integer x = 0.351320
#> Warning: non-integer x = 2.636195
#> Warning: non-integer x = 1.721613
#> Warning: non-integer x = 1.967323
#> Warning: non-integer x = 0.447730
#> Warning: non-integer x = 2.696112
#> Warning: non-integer x = 0.604953
#> Warning: non-integer x = 2.828914
#> Warning: non-integer x = 2.055688
#> Warning: non-integer x = 0.086021
#> Warning: non-integer x = 3.833843
#> Warning: non-integer x = 2.044475
#> Warning: non-integer x = 1.835364
#> Warning: non-integer x = 0.892182
#> Warning: non-integer x = 3.175368
#> Warning: non-integer x = 1.958256
#> Warning: non-integer x = 0.519601
#> Warning: non-integer x = 0.001990
#> Warning: non-integer x = 0.601535
#> Warning: non-integer x = 1.055889
#> Warning: non-integer x = 0.948783
#> Warning: non-integer x = 1.003551
#> Warning: non-integer x = 3.277086
#> Warning: non-integer x = 0.722833
#> Warning: non-integer x = 0.024719
#> Warning: non-integer x = 1.225925
#> Warning: non-integer x = 0.879255
#> Warning: non-integer x = 0.965262
#> Warning: non-integer x = 0.779769
#> Warning: non-integer x = 0.401402
#> Warning: non-integer x = 1.702031
#> Warning: non-integer x = 0.414464
#> Warning: non-integer x = 0.407794
#> Warning: non-integer x = 0.964611
#> Warning: non-integer x = 0.495077
#> Warning: non-integer x = 2.706271
#> Warning: non-integer x = 2.267814
#> Warning: non-integer x = 1.598252
#> Warning: non-integer x = 0.223104
#> Warning: non-integer x = 1.345797
#> Warning: non-integer x = 0.624041
#> Warning: non-integer x = 2.349828
#> Warning: non-integer x = 0.124004
#> Warning: non-integer x = 0.511794
#> Warning: non-integer x = 2.206786
#> Warning: non-integer x = 0.183042
#> Warning: non-integer x = 3.138495
#> Warning: non-integer x = 0.772278
#> Warning: non-integer x = 2.390848
#> Warning: non-integer x = 2.854088
#> Warning: non-integer x = 1.822923
#> Warning: non-integer x = 0.550871
#> Warning: non-integer x = 1.041601
#> Warning: non-integer x = 0.717388
#> Warning: non-integer x = 2.956995
#> Warning: non-integer x = 1.296914
#> Warning: non-integer x = 1.071571
#> Warning: non-integer x = 0.197058
#> Warning: non-integer x = 0.635956
#> Warning: non-integer x = 1.021969
#> Warning: non-integer x = 0.587228
#> Warning: non-integer x = 1.766768
#> Warning: non-integer x = 0.318167
#> Warning: non-integer x = 0.758000
#> Warning: non-integer x = 1.752928
#> Warning: non-integer x = 1.103167
#> Warning: non-integer x = 0.153544
#> Warning: non-integer x = 2.323783
#> Warning: non-integer x = 0.026547
#> Warning: non-integer x = 0.682138
#> Warning: non-integer x = 2.017458
#> Warning: non-integer x = 1.300385
#> Warning: non-integer x = 2.402430
#> Warning: non-integer x = 1.883285
#> Warning: non-integer x = 0.327929
#> Warning: non-integer x = 2.182575
#> Warning: non-integer x = 2.657218
#> Warning: non-integer x = 1.146871
#> Warning: non-integer x = 0.304370
#> Warning: non-integer x = 1.915115
#> Warning: non-integer x = 2.227329
#> Warning: non-integer x = 4.250994
#> Warning: non-integer x = 3.184593
#> Warning: non-integer x = 0.979358
#> Warning: non-integer x = 0.919113
#> Warning: non-integer x = 0.470215
#> Warning: non-integer x = 1.023009
#> Warning: non-integer x = 0.067865
#> Warning: non-integer x = 3.040907
#> Warning: non-integer x = 2.088902
#> Warning: non-integer x = 2.788344
#> Warning: non-integer x = 1.410479
#> Warning: non-integer x = 1.994314
#> Warning: non-integer x = 1.315601
#> Warning: non-integer x = 0.485030
#> Warning: non-integer x = 0.832385
#> Warning: non-integer x = 1.962781
#> Warning: non-integer x = 1.451084
#> Warning: non-integer x = 1.499598
#> Warning: non-integer x = 1.504095
#> Warning: non-integer x = 0.282879
#> Warning: non-integer x = 2.604948
#> Warning: non-integer x = 0.224484
#> Warning: non-integer x = 0.744755
#> Warning: non-integer x = 1.378051
#> Warning: non-integer x = 2.003479
#> Warning: non-integer x = 0.643786
#> Warning: non-integer x = 1.746697
#> Warning: non-integer x = 2.367189
#> Warning: non-integer x = 0.766633
#> Warning: non-integer x = 4.091512
#> Warning: non-integer x = 0.585645
#> Warning: non-integer x = 2.813151
#> Warning: non-integer x = 2.528060
#> Warning: non-integer x = 2.998427
#> Warning: non-integer x = 1.181311
#> Warning: non-integer x = 0.501690
#> Warning: non-integer x = 4.078846
#> Warning: non-integer x = 2.978303
#> Warning: non-integer x = 2.988262
#> Warning: non-integer x = 0.344241
#> Warning: non-integer x = 0.995538
#> Warning: non-integer x = 1.466635
#> Warning: non-integer x = 0.676346
#> Warning: non-integer x = 1.978385
#> Warning: non-integer x = 0.395289
#> Warning: non-integer x = 0.725009
#> Warning: non-integer x = 1.156258
#> Warning: non-integer x = 1.900077
#> Warning: non-integer x = 0.364502
#> Warning: non-integer x = 0.486086
#> Warning: non-integer x = 0.537055
#> Warning: non-integer x = 0.003369
#> Warning: non-integer x = 3.047889
#> Warning: non-integer x = 2.037167
#> Warning: non-integer x = 2.337452
#> Warning: non-integer x = 1.568618
#> Warning: non-integer x = 2.182014
#> Warning: non-integer x = 2.346370
#> Warning: non-integer x = 1.704094
#> Warning: non-integer x = 2.172504
#> Warning: non-integer x = 0.806177
#> Warning: non-integer x = 0.955252
#> Warning: non-integer x = 3.112184
#> Warning: non-integer x = 1.947151
#> Warning: non-integer x = 0.631770
#> Warning: non-integer x = 2.537779
#> Warning: non-integer x = 1.414806
#> Warning: non-integer x = 4.757060
#> Warning: non-integer x = 0.245966
#> Warning: non-integer x = 0.851356
#> Warning: non-integer x = 2.010523
#> Warning: non-integer x = 0.885310
#> Warning: non-integer x = 0.580681
#> Warning: non-integer x = 1.220759
#> Warning: non-integer x = 0.550740
#> Warning: non-integer x = 1.345942
#> Warning: non-integer x = 0.382504
#> Warning: non-integer x = 1.343464
#> Warning: non-integer x = 1.880290
#> Warning: non-integer x = 0.327245
#> Warning: non-integer x = 3.181543
#> Warning: non-integer x = 0.287520
#> Warning: non-integer x = 0.689405
#> Warning: non-integer x = 1.351593
#> Warning: non-integer x = 1.819214
#> Warning: non-integer x = 2.337922
#> Warning: non-integer x = 0.188426
#> Warning: non-integer x = 2.667866
#> Warning: non-integer x = 0.386377
#> Warning: non-integer x = 2.135640
#> Warning: non-integer x = 2.885382
#> Warning: non-integer x = 0.955238
#> Warning: non-integer x = 1.349946
#> Warning: non-integer x = 1.399225
#> Warning: non-integer x = 1.445966
#> Warning: non-integer x = 0.312702
#> Warning: non-integer x = 0.239059
#> Warning: non-integer x = 0.249314
#> Warning: non-integer x = 0.879136
#> Warning: non-integer x = 1.383241
#> Warning: non-integer x = 0.937790
#> Warning: non-integer x = 0.731037
#> Warning: non-integer x = 1.625088
#> Warning: non-integer x = 0.575407
#> Warning: non-integer x = 3.302949
#> Warning: non-integer x = 0.491590
#> Warning: non-integer x = 0.285410
#> Warning: non-integer x = 0.545288
#> Warning: non-integer x = 0.562770
#> Warning: non-integer x = 1.445716
#> Warning: non-integer x = 1.901283
#> Warning: non-integer x = 1.779935
#> Warning: non-integer x = 0.647670
#> Warning: non-integer x = 3.506677
#> Warning: non-integer x = 0.611767
#> Warning: non-integer x = 2.060648
#> Warning: non-integer x = 3.601516
#> Warning: non-integer x = 0.330660
#> Warning: non-integer x = 2.462394
#> Warning: non-integer x = 0.378861
#> Warning: non-integer x = 3.008570
#> Warning: non-integer x = 0.815282
#> Warning: non-integer x = 0.816712
#> Warning: non-integer x = 1.815672
#> Warning: non-integer x = 2.923131
#> Warning: non-integer x = 0.306089
#> Warning: non-integer x = 0.006346
#> Warning: non-integer x = 2.259331
#> Warning: non-integer x = 2.708752
#> Warning: non-integer x = 1.696709
#> Warning: non-integer x = 0.925292
#> Warning: non-integer x = 1.352696
#> Warning: non-integer x = 1.559085
#> Warning: non-integer x = 0.259343
#> Warning: non-integer x = 1.406333
#> Warning: non-integer x = 0.533209
#> Warning: non-integer x = 1.421824
#> Warning: non-integer x = 0.278340
#> Warning: non-integer x = 1.582091
#> Warning: non-integer x = 0.621753
#> Warning: non-integer x = 4.785320
#> Warning: non-integer x = 1.789249
#> Warning: non-integer x = 0.731190
#> Warning: non-integer x = 0.323029
#> Warning: non-integer x = 1.809411
#> Warning: non-integer x = 1.031129
#> Warning: non-integer x = 2.441765
#> Warning: non-integer x = 0.194868
#> Warning: non-integer x = 0.304272
#> Warning: non-integer x = 0.810346
#> Warning: non-integer x = 1.937120
#> Warning: non-integer x = 0.976055
#> Warning: non-integer x = 2.384959
#> Warning: non-integer x = 0.184575
#> Warning: non-integer x = 2.356938
#> Warning: non-integer x = 1.214633
#> Warning: non-integer x = 1.471342
#> Warning: non-integer x = 1.273560
#> Warning: non-integer x = 2.253483
#> Warning: non-integer x = 0.585855
#> Warning: non-integer x = 0.030872
#> Warning: non-integer x = 0.451004
#> Warning: non-integer x = 0.552187
#> Warning: non-integer x = 0.015937
#> Warning: non-integer x = 1.423041
#> Warning: non-integer x = 1.048241
#> Warning: non-integer x = 1.474147
#> Warning: non-integer x = 1.704650
#> Warning: non-integer x = 2.925005
#> Warning: non-integer x = 0.264448
#> Warning: non-integer x = 1.115750
#> Warning: non-integer x = 0.940845
#> Warning: non-integer x = 0.597708
#> Warning: non-integer x = 4.723147
#> Warning: non-integer x = 1.930907
#> Warning: non-integer x = 0.503259
#> Warning: non-integer x = 1.569458
#> Warning: non-integer x = 1.269536
#> Warning: non-integer x = 0.784335
#> Warning: non-integer x = 0.245706
#> Warning: non-integer x = 1.105439
#> Warning: non-integer x = 3.441664
#> Warning: non-integer x = 2.619341
#> Warning: non-integer x = 1.709384
#> Warning: non-integer x = 1.927079
#> Warning: non-integer x = 0.473718
#> Warning: non-integer x = 2.693384
#> Warning: non-integer x = 0.677801
#> Warning: non-integer x = 0.183923
#> Warning: non-integer x = 0.515570
#> Warning: non-integer x = 3.656991
#> Warning: non-integer x = 0.352643
#> Warning: non-integer x = 1.145581
#> Warning: non-integer x = 2.132227
#> Warning: non-integer x = 5.101162
#> Warning: non-integer x = 2.839431
#> Warning: non-integer x = 1.334730
#> Warning: non-integer x = 0.316076
#> Warning: non-integer x = 0.176550
#> Warning: non-integer x = 0.927807
#> Warning: non-integer x = 1.742921
#> Warning: non-integer x = 1.805402
#> Warning: non-integer x = 1.469092
#> Warning: non-integer x = 0.626378
#> Warning: non-integer x = 0.918420
#> Warning: non-integer x = 0.044503
#> Warning: non-integer x = 0.005014
#> Warning: non-integer x = 1.771645
#> Warning: non-integer x = 0.035966
#> Warning: non-integer x = 1.390437
#> Warning: non-integer x = 2.569272
#> Warning: non-integer x = 0.704422
#> Warning: non-integer x = 0.856608
#> Warning: non-integer x = 0.824395
#> Warning: non-integer x = 0.331647
#> Warning: non-integer x = 2.218014
#> Warning: non-integer x = 2.198655
#> Warning: non-integer x = 0.378821
#> Warning: non-integer x = 3.483857
#> Warning: non-integer x = 2.793648
#> Warning: non-integer x = 2.486722
#> Warning: non-integer x = 1.927543
#> Warning: non-integer x = 2.175280
#> Warning: non-integer x = 1.191953
#> Warning: non-integer x = 1.221544
#> Warning: non-integer x = 2.413666
#> Warning: non-integer x = 1.234094
#> Warning: non-integer x = 2.463109
#> Warning: non-integer x = 1.678532
#> Warning: non-integer x = 0.454584
#> Warning: non-integer x = 0.193769
#> Warning: non-integer x = 0.555798
#> Warning: non-integer x = 1.579461
#> Warning: non-integer x = 1.712413
#> Warning: non-integer x = 0.733009
#> Warning: non-integer x = 4.193157
#> Warning: non-integer x = 0.098505
#> Warning: non-integer x = 1.386585
#> Warning: non-integer x = 1.448696
#> Warning: non-integer x = 0.746390
#> Warning: non-integer x = 1.553723
#> Warning: non-integer x = 2.733091
#> Warning: non-integer x = 2.595828
#> Warning: non-integer x = 0.216209
#> Warning: non-integer x = 1.052513
#> Warning: non-integer x = 0.843564
#> Warning: non-integer x = 0.887884
#> Warning: non-integer x = 0.101143
#> Warning: non-integer x = 2.529758
#> Warning: non-integer x = 1.220518
#> Warning: non-integer x = 0.506884
#> Warning: non-integer x = 0.455984
#> Warning: non-integer x = 1.311727
#> Warning: non-integer x = 1.689543
#> Warning: non-integer x = 1.097677
#> Warning: non-integer x = 0.092537
#> Warning: non-integer x = 1.029762
#> Warning: non-integer x = 0.213387
#> Warning: non-integer x = 0.979223
#> Warning: non-integer x = 1.990135
#> Warning: non-integer x = 0.550101
#> Warning: non-integer x = 1.639858
#> Warning: non-integer x = 1.037837
#> Warning: non-integer x = 0.149129
#> Warning: non-integer x = 0.119297
#> Warning: non-integer x = 4.001085
#> Warning: non-integer x = 0.561223
#> Warning: non-integer x = 1.684477
#> Warning: non-integer x = 0.253397
#> Warning: non-integer x = 0.410384
#> Warning: non-integer x = 1.851233
#> Warning: non-integer x = 0.600299
#> Warning: non-integer x = 1.746652
#> Warning: non-integer x = 1.316776
#> Warning: non-integer x = 0.829245
#> Warning: non-integer x = 2.289124
#> Warning: non-integer x = 0.693850
#> Warning: non-integer x = 1.347549
#> Warning: non-integer x = 1.420709
#> Warning: non-integer x = 0.805444
#> Warning: non-integer x = 1.676210
#> Warning: non-integer x = 1.656711
#> Warning: non-integer x = 0.856057
#> Warning: non-integer x = 1.267421
#> Warning: non-integer x = 1.612985
#> Warning: non-integer x = 0.914959
#> Warning: non-integer x = 0.890895
#> Warning: non-integer x = 2.427841
#> Warning: non-integer x = 0.150082
#> Warning: non-integer x = 2.921377
#> Warning: non-integer x = 0.399018
#> Warning: non-integer x = 2.112732
#> Warning: non-integer x = 2.898758
#> Warning: non-integer x = 1.830217
#> Warning: non-integer x = 0.675479
#> Warning: non-integer x = 0.715014
#> Warning: non-integer x = 0.175464
#> Warning: non-integer x = 1.264781
#> Warning: non-integer x = 0.715933
#> Warning: non-integer x = 0.742346
#> Warning: non-integer x = 0.249929
#> Warning: non-integer x = 2.372067
#> Warning: non-integer x = 0.981963
#> Warning: non-integer x = 1.246378
#> Warning: non-integer x = 1.240815
#> Warning: non-integer x = 2.444982
#> Warning: non-integer x = 0.601497
#> Warning: non-integer x = 0.811107
#> Warning: non-integer x = 0.638239
#> Warning: non-integer x = 0.948679
#> Warning: non-integer x = 3.438324
#> Warning: non-integer x = 1.558160
#> Warning: non-integer x = 0.781643
#> Warning: non-integer x = 0.573539
#> Warning: non-integer x = 0.279430
#> Warning: non-integer x = 1.596987
#> Warning: non-integer x = 0.910828
#> Warning: non-integer x = 1.053047
#> Warning: non-integer x = 0.050308
#> Warning: non-integer x = 2.173955
#> Warning: non-integer x = 1.265028
#> Warning: non-integer x = 1.916871
#> Warning: non-integer x = 0.052966
#> Warning: non-integer x = 1.589578
#> Warning: non-integer x = 0.919551
#> Warning: non-integer x = 1.763766
#> Warning: non-integer x = 1.060563
#> Warning: non-integer x = 0.588908
#> Warning: non-integer x = 1.679980
#> Warning: non-integer x = 1.572749
#> Warning: non-integer x = 4.742095
#> Warning: non-integer x = 1.426677
#> Warning: non-integer x = 1.124568
#> Warning: non-integer x = 3.206657
#> Warning: non-integer x = 0.264522
#> Warning: non-integer x = 0.064834
#> Warning: non-integer x = 3.513006
#> Warning: non-integer x = 0.458741
#> Warning: non-integer x = 0.266589
#> Warning: non-integer x = 1.238921
#> Warning: non-integer x = 1.818396
#> Warning: non-integer x = 0.804205