Estimate the Individual Treatment Effect using Poisson Regression

Estimates the Individual Treatment Effect using Poisson Regression given a response vector, a treatment vector, and a features matrix.

Usage

estimate_ite_poisson(y, z, X, X_names, include_offset, offset_name)

Arguments

y

the observed response vector

z

the treatment vector

X

the features matrix

X_names

the names of the covariates

include_offset

whether or not to include an offset when estimating the ITE, for poisson only

offset_name

the name of the offset, if it is to be included

Value

a vector of ITE estimates

Examples

dataset <- generate_cre_dataset(n = 1000, rho = 0, n_rules = 2, p = 10,
                                effect_size = 2, binary = FALSE)

# Initialize parameters
y <- abs(dataset[["y"]])
z <- dataset[["z"]]
X <- as.data.frame(dataset[["X"]])
X_names <- names(as.data.frame(X))
include_offset <- FALSE
offset_name <- NA

ite_list <- estimate_ite_poisson(y, z, X, X_names, include_offset, offset_name)
#> Warning: non-integer x = 1.203200
#> Warning: non-integer x = 1.060025
#> Warning: non-integer x = 2.221068
#> Warning: non-integer x = 0.010544
#> Warning: non-integer x = 1.343772
#> Warning: non-integer x = 2.566239
#> Warning: non-integer x = 2.274212
#> Warning: non-integer x = 1.051541
#> Warning: non-integer x = 0.057646
#> Warning: non-integer x = 2.652076
#> Warning: non-integer x = 2.169305
#> Warning: non-integer x = 2.241127
#> Warning: non-integer x = 3.922574
#> Warning: non-integer x = 1.777353
#> Warning: non-integer x = 0.429118
#> Warning: non-integer x = 2.660026
#> Warning: non-integer x = 2.654840
#> Warning: non-integer x = 2.105018
#> Warning: non-integer x = 1.205038
#> Warning: non-integer x = 3.555361
#> Warning: non-integer x = 2.329127
#> Warning: non-integer x = 4.070174
#> Warning: non-integer x = 1.180560
#> Warning: non-integer x = 0.921051
#> Warning: non-integer x = 2.534916
#> Warning: non-integer x = 2.266219
#> Warning: non-integer x = 1.713514
#> Warning: non-integer x = 2.914396
#> Warning: non-integer x = 1.670354
#> Warning: non-integer x = 2.991951
#> Warning: non-integer x = 0.457259
#> Warning: non-integer x = 1.339909
#> Warning: non-integer x = 0.287856
#> Warning: non-integer x = 1.355332
#> Warning: non-integer x = 1.293831
#> Warning: non-integer x = 1.373549
#> Warning: non-integer x = 1.141722
#> Warning: non-integer x = 3.553815
#> Warning: non-integer x = 0.597721
#> Warning: non-integer x = 3.602387
#> Warning: non-integer x = 1.200555
#> Warning: non-integer x = 1.526040
#> Warning: non-integer x = 1.543350
#> Warning: non-integer x = 0.827432
#> Warning: non-integer x = 2.895076
#> Warning: non-integer x = 2.686957
#> Warning: non-integer x = 0.058675
#> Warning: non-integer x = 2.827496
#> Warning: non-integer x = 1.719273
#> Warning: non-integer x = 2.689621
#> Warning: non-integer x = 1.668129
#> Warning: non-integer x = 0.058314
#> Warning: non-integer x = 3.108681
#> Warning: non-integer x = 1.132864
#> Warning: non-integer x = 0.096446
#> Warning: non-integer x = 0.117911
#> Warning: non-integer x = 2.600460
#> Warning: non-integer x = 1.767565
#> Warning: non-integer x = 2.077696
#> Warning: non-integer x = 0.334606
#> Warning: non-integer x = 1.842451
#> Warning: non-integer x = 0.545489
#> Warning: non-integer x = 0.918288
#> Warning: non-integer x = 0.455011
#> Warning: non-integer x = 2.328799
#> Warning: non-integer x = 2.039847
#> Warning: non-integer x = 0.898405
#> Warning: non-integer x = 1.600880
#> Warning: non-integer x = 2.095047
#> Warning: non-integer x = 1.573914
#> Warning: non-integer x = 2.223244
#> Warning: non-integer x = 0.290297
#> Warning: non-integer x = 2.909089
#> Warning: non-integer x = 0.964340
#> Warning: non-integer x = 1.798632
#> Warning: non-integer x = 3.032812
#> Warning: non-integer x = 1.909518
#> Warning: non-integer x = 1.709495
#> Warning: non-integer x = 1.176705
#> Warning: non-integer x = 2.830682
#> Warning: non-integer x = 3.653322
#> Warning: non-integer x = 0.006737
#> Warning: non-integer x = 0.153031
#> Warning: non-integer x = 1.016400
#> Warning: non-integer x = 0.699320
#> Warning: non-integer x = 0.660282
#> Warning: non-integer x = 0.195908
#> Warning: non-integer x = 2.874999
#> Warning: non-integer x = 4.244913
#> Warning: non-integer x = 1.989948
#> Warning: non-integer x = 1.127152
#> Warning: non-integer x = 1.318783
#> Warning: non-integer x = 1.608415
#> Warning: non-integer x = 0.287856
#> Warning: non-integer x = 3.471688
#> Warning: non-integer x = 1.816511
#> Warning: non-integer x = 2.186838
#> Warning: non-integer x = 0.081141
#> Warning: non-integer x = 1.925399
#> Warning: non-integer x = 0.067744
#> Warning: non-integer x = 3.924985
#> Warning: non-integer x = 1.938056
#> Warning: non-integer x = 0.511946
#> Warning: non-integer x = 1.501335
#> Warning: non-integer x = 3.178552
#> Warning: non-integer x = 2.592805
#> Warning: non-integer x = 1.150899
#> Warning: non-integer x = 2.764021
#> Warning: non-integer x = 0.140508
#> Warning: non-integer x = 0.990179
#> Warning: non-integer x = 1.693459
#> Warning: non-integer x = 2.812403
#> Warning: non-integer x = 2.561952
#> Warning: non-integer x = 3.183937
#> Warning: non-integer x = 4.438757
#> Warning: non-integer x = 1.280840
#> Warning: non-integer x = 2.573064
#> Warning: non-integer x = 0.904246
#> Warning: non-integer x = 2.256646
#> Warning: non-integer x = 1.234988
#> Warning: non-integer x = 0.529480
#> Warning: non-integer x = 0.777965
#> Warning: non-integer x = 1.584813
#> Warning: non-integer x = 0.274823
#> Warning: non-integer x = 3.063512
#> Warning: non-integer x = 1.557375
#> Warning: non-integer x = 0.132558
#> Warning: non-integer x = 0.931566
#> Warning: non-integer x = 1.444881
#> Warning: non-integer x = 4.615722
#> Warning: non-integer x = 3.663849
#> Warning: non-integer x = 2.080721
#> Warning: non-integer x = 2.952225
#> Warning: non-integer x = 0.823851
#> Warning: non-integer x = 4.632623
#> Warning: non-integer x = 0.487366
#> Warning: non-integer x = 0.733813
#> Warning: non-integer x = 4.830545
#> Warning: non-integer x = 0.949901
#> Warning: non-integer x = 0.967448
#> Warning: non-integer x = 1.526855
#> Warning: non-integer x = 1.870532
#> Warning: non-integer x = 2.236068
#> Warning: non-integer x = 2.733521
#> Warning: non-integer x = 2.265446
#> Warning: non-integer x = 1.336140
#> Warning: non-integer x = 0.628268
#> Warning: non-integer x = 3.741654
#> Warning: non-integer x = 3.657271
#> Warning: non-integer x = 4.212549
#> Warning: non-integer x = 2.468958
#> Warning: non-integer x = 0.531330
#> Warning: non-integer x = 1.953413
#> Warning: non-integer x = 1.122121
#> Warning: non-integer x = 2.208454
#> Warning: non-integer x = 0.417560
#> Warning: non-integer x = 0.803176
#> Warning: non-integer x = 1.259319
#> Warning: non-integer x = 2.252453
#> Warning: non-integer x = 2.333105
#> Warning: non-integer x = 2.270347
#> Warning: non-integer x = 0.633496
#> Warning: non-integer x = 2.008301
#> Warning: non-integer x = 4.394123
#> Warning: non-integer x = 0.259343
#> Warning: non-integer x = 0.318310
#> Warning: non-integer x = 2.190508
#> Warning: non-integer x = 3.519007
#> Warning: non-integer x = 1.757076
#> Warning: non-integer x = 3.659031
#> Warning: non-integer x = 2.292280
#> Warning: non-integer x = 0.874586
#> Warning: non-integer x = 1.215709
#> Warning: non-integer x = 3.121385
#> Warning: non-integer x = 1.135719
#> Warning: non-integer x = 3.088536
#> Warning: non-integer x = 3.399834
#> Warning: non-integer x = 1.901080
#> Warning: non-integer x = 1.598350
#> Warning: non-integer x = 3.235217
#> Warning: non-integer x = 4.393599
#> Warning: non-integer x = 0.610645
#> Warning: non-integer x = 2.314885
#> Warning: non-integer x = 1.295186
#> Warning: non-integer x = 1.831402
#> Warning: non-integer x = 1.512270
#> Warning: non-integer x = 0.886396
#> Warning: non-integer x = 0.101703
#> Warning: non-integer x = 2.148089
#> Warning: non-integer x = 2.140756
#> Warning: non-integer x = 0.331524
#> Warning: non-integer x = 1.463878
#> Warning: non-integer x = 0.230380
#> Warning: non-integer x = 0.924018
#> Warning: non-integer x = 1.506030
#> Warning: non-integer x = 0.951760
#> Warning: non-integer x = 2.063752
#> Warning: non-integer x = 0.137523
#> Warning: non-integer x = 1.411520
#> Warning: non-integer x = 2.769824
#> Warning: non-integer x = 3.120841
#> Warning: non-integer x = 2.070847
#> Warning: non-integer x = 1.150507
#> Warning: non-integer x = 2.220488
#> Warning: non-integer x = 0.429530
#> Warning: non-integer x = 0.690781
#> Warning: non-integer x = 1.429596
#> Warning: non-integer x = 3.011828
#> Warning: non-integer x = 0.069403
#> Warning: non-integer x = 4.343293
#> Warning: non-integer x = 1.208744
#> Warning: non-integer x = 1.306876
#> Warning: non-integer x = 2.717693
#> Warning: non-integer x = 2.833695
#> Warning: non-integer x = 1.948196
#> Warning: non-integer x = 1.086389
#> Warning: non-integer x = 2.144888
#> Warning: non-integer x = 0.584533
#> Warning: non-integer x = 2.540776
#> Warning: non-integer x = 2.755538
#> Warning: non-integer x = 1.975356
#> Warning: non-integer x = 2.512708
#> Warning: non-integer x = 0.338753
#> Warning: non-integer x = 1.464311
#> Warning: non-integer x = 0.512161
#> Warning: non-integer x = 1.479556
#> Warning: non-integer x = 3.484887
#> Warning: non-integer x = 3.415471
#> Warning: non-integer x = 4.084744
#> Warning: non-integer x = 1.669757
#> Warning: non-integer x = 0.936222
#> Warning: non-integer x = 3.779834
#> Warning: non-integer x = 3.550901
#> Warning: non-integer x = 2.388699
#> Warning: non-integer x = 1.424609
#> Warning: non-integer x = 1.369870
#> Warning: non-integer x = 1.055591
#> Warning: non-integer x = 0.333405
#> Warning: non-integer x = 3.939029
#> Warning: non-integer x = 3.082025
#> Warning: non-integer x = 1.332775
#> Warning: non-integer x = 2.186219
#> Warning: non-integer x = 1.674260
#> Warning: non-integer x = 1.896900
#> Warning: non-integer x = 3.869760
#> Warning: non-integer x = 0.566115
#> Warning: non-integer x = 1.989503
#> Warning: non-integer x = 1.119329
#> Warning: non-integer x = 0.493743
#> Warning: non-integer x = 3.484720
#> Warning: non-integer x = 1.575448
#> Warning: non-integer x = 3.745001
#> Warning: non-integer x = 1.367662
#> Warning: non-integer x = 0.326533
#> Warning: non-integer x = 3.878119
#> Warning: non-integer x = 0.937231
#> Warning: non-integer x = 2.523356
#> Warning: non-integer x = 0.950404
#> Warning: non-integer x = 0.337694
#> Warning: non-integer x = 3.620818
#> Warning: non-integer x = 0.599822
#> Warning: non-integer x = 2.217087
#> Warning: non-integer x = 0.864733
#> Warning: non-integer x = 2.354076
#> Warning: non-integer x = 0.789097
#> Warning: non-integer x = 3.685898
#> Warning: non-integer x = 1.897235
#> Warning: non-integer x = 3.219528
#> Warning: non-integer x = 0.431890
#> Warning: non-integer x = 2.558337
#> Warning: non-integer x = 2.975927
#> Warning: non-integer x = 1.805620
#> Warning: non-integer x = 2.375844
#> Warning: non-integer x = 3.317730
#> Warning: non-integer x = 0.682236
#> Warning: non-integer x = 3.360268
#> Warning: non-integer x = 1.390644
#> Warning: non-integer x = 0.818823
#> Warning: non-integer x = 3.239666
#> Warning: non-integer x = 0.781097
#> Warning: non-integer x = 3.124404
#> Warning: non-integer x = 2.737897
#> Warning: non-integer x = 1.680125
#> Warning: non-integer x = 1.364839
#> Warning: non-integer x = 1.269804
#> Warning: non-integer x = 0.816142
#> Warning: non-integer x = 2.322842
#> Warning: non-integer x = 1.389942
#> Warning: non-integer x = 1.822096
#> Warning: non-integer x = 0.482650
#> Warning: non-integer x = 2.049636
#> Warning: non-integer x = 0.959973
#> Warning: non-integer x = 2.191739
#> Warning: non-integer x = 0.992110
#> Warning: non-integer x = 0.261821
#> Warning: non-integer x = 1.017979
#> Warning: non-integer x = 1.946199
#> Warning: non-integer x = 0.318371
#> Warning: non-integer x = 1.463086
#> Warning: non-integer x = 2.224083
#> Warning: non-integer x = 0.771000
#> Warning: non-integer x = 2.570592
#> Warning: non-integer x = 0.272395
#> Warning: non-integer x = 2.435950
#> Warning: non-integer x = 1.456807
#> Warning: non-integer x = 1.452810
#> Warning: non-integer x = 0.318982
#> Warning: non-integer x = 3.645491
#> Warning: non-integer x = 0.120424
#> Warning: non-integer x = 0.221885
#> Warning: non-integer x = 0.622386
#> Warning: non-integer x = 0.301874
#> Warning: non-integer x = 2.451202
#> Warning: non-integer x = 2.618331
#> Warning: non-integer x = 1.173081
#> Warning: non-integer x = 0.868503
#> Warning: non-integer x = 1.313057
#> Warning: non-integer x = 0.785772
#> Warning: non-integer x = 1.694640
#> Warning: non-integer x = 2.220910
#> Warning: non-integer x = 1.244232
#> Warning: non-integer x = 2.300848
#> Warning: non-integer x = 0.249237
#> Warning: non-integer x = 1.111871
#> Warning: non-integer x = 0.277993
#> Warning: non-integer x = 3.908579
#> Warning: non-integer x = 2.735020
#> Warning: non-integer x = 2.668567
#> Warning: non-integer x = 0.078497
#> Warning: non-integer x = 0.838814
#> Warning: non-integer x = 0.983603
#> Warning: non-integer x = 1.150015
#> Warning: non-integer x = 1.216898
#> Warning: non-integer x = 1.363011
#> Warning: non-integer x = 1.935864
#> Warning: non-integer x = 1.122981
#> Warning: non-integer x = 0.387803
#> Warning: non-integer x = 2.030208
#> Warning: non-integer x = 1.217705
#> Warning: non-integer x = 1.258575
#> Warning: non-integer x = 1.908486
#> Warning: non-integer x = 0.323055
#> Warning: non-integer x = 0.675316
#> Warning: non-integer x = 0.701031
#> Warning: non-integer x = 1.841256
#> Warning: non-integer x = 2.823300
#> Warning: non-integer x = 2.706434
#> Warning: non-integer x = 0.504418
#> Warning: non-integer x = 1.160464
#> Warning: non-integer x = 2.360843
#> Warning: non-integer x = 2.058480
#> Warning: non-integer x = 1.768415
#> Warning: non-integer x = 3.201344
#> Warning: non-integer x = 1.140094
#> Warning: non-integer x = 1.735926
#> Warning: non-integer x = 2.215544
#> Warning: non-integer x = 0.523418
#> Warning: non-integer x = 5.189598
#> Warning: non-integer x = 1.787809
#> Warning: non-integer x = 1.375076
#> Warning: non-integer x = 2.727333
#> Warning: non-integer x = 0.778326
#> Warning: non-integer x = 0.810078
#> Warning: non-integer x = 0.630905
#> Warning: non-integer x = 3.025078
#> Warning: non-integer x = 2.988009
#> Warning: non-integer x = 2.695434
#> Warning: non-integer x = 1.887105
#> Warning: non-integer x = 4.297333
#> Warning: non-integer x = 1.170024
#> Warning: non-integer x = 0.188818
#> Warning: non-integer x = 0.062051
#> Warning: non-integer x = 1.241360
#> Warning: non-integer x = 1.899325
#> Warning: non-integer x = 0.791287
#> Warning: non-integer x = 3.053027
#> Warning: non-integer x = 1.637371
#> Warning: non-integer x = 1.100099
#> Warning: non-integer x = 2.371111
#> Warning: non-integer x = 1.150047
#> Warning: non-integer x = 1.434753
#> Warning: non-integer x = 1.620187
#> Warning: non-integer x = 0.946757
#> Warning: non-integer x = 0.069160
#> Warning: non-integer x = 1.206214
#> Warning: non-integer x = 2.375997
#> Warning: non-integer x = 0.988367
#> Warning: non-integer x = 0.741261
#> Warning: non-integer x = 0.670191
#> Warning: non-integer x = 4.128515
#> Warning: non-integer x = 0.744128
#> Warning: non-integer x = 0.528487
#> Warning: non-integer x = 1.070597
#> Warning: non-integer x = 0.342030
#> Warning: non-integer x = 0.790995
#> Warning: non-integer x = 2.228829
#> Warning: non-integer x = 0.904835
#> Warning: non-integer x = 1.271459
#> Warning: non-integer x = 2.164832
#> Warning: non-integer x = 0.464579
#> Warning: non-integer x = 0.707543
#> Warning: non-integer x = 0.727258
#> Warning: non-integer x = 2.215320
#> Warning: non-integer x = 2.004558
#> Warning: non-integer x = 0.019433
#> Warning: non-integer x = 1.687042
#> Warning: non-integer x = 1.912462
#> Warning: non-integer x = 0.337633
#> Warning: non-integer x = 0.908933
#> Warning: non-integer x = 0.506506
#> Warning: non-integer x = 1.287480
#> Warning: non-integer x = 2.211532
#> Warning: non-integer x = 1.096915
#> Warning: non-integer x = 0.611813
#> Warning: non-integer x = 2.063311
#> Warning: non-integer x = 1.499480
#> Warning: non-integer x = 1.284395
#> Warning: non-integer x = 0.998456
#> Warning: non-integer x = 0.654641
#> Warning: non-integer x = 2.435279
#> Warning: non-integer x = 0.414174
#> Warning: non-integer x = 0.918739
#> Warning: non-integer x = 0.975981
#> Warning: non-integer x = 0.124042
#> Warning: non-integer x = 1.763368
#> Warning: non-integer x = 2.156038
#> Warning: non-integer x = 1.434196
#> Warning: non-integer x = 0.840750
#> Warning: non-integer x = 1.862226
#> Warning: non-integer x = 0.525536
#> Warning: non-integer x = 0.527754
#> Warning: non-integer x = 3.410185
#> Warning: non-integer x = 0.319370
#> Warning: non-integer x = 0.222672
#> Warning: non-integer x = 1.374071
#> Warning: non-integer x = 1.364228
#> Warning: non-integer x = 2.058061
#> Warning: non-integer x = 0.675793
#> Warning: non-integer x = 0.647863
#> Warning: non-integer x = 0.401060
#> Warning: non-integer x = 0.380643
#> Warning: non-integer x = 0.012496
#> Warning: non-integer x = 1.841361
#> Warning: non-integer x = 1.499208
#> Warning: non-integer x = 3.972824
#> Warning: non-integer x = 2.725212
#> Warning: non-integer x = 0.500009
#> Warning: non-integer x = 1.239123
#> Warning: non-integer x = 0.114704
#> Warning: non-integer x = 0.030579
#> Warning: non-integer x = 1.706541
#> Warning: non-integer x = 0.712426
#> Warning: non-integer x = 1.071105
#> Warning: non-integer x = 0.106787
#> Warning: non-integer x = 0.877981
#> Warning: non-integer x = 0.640045
#> Warning: non-integer x = 0.249514
#> Warning: non-integer x = 1.764144
#> Warning: non-integer x = 2.456759
#> Warning: non-integer x = 2.321151
#> Warning: non-integer x = 1.063700
#> Warning: non-integer x = 0.881980
#> Warning: non-integer x = 4.231390
#> Warning: non-integer x = 0.445116
#> Warning: non-integer x = 4.228245
#> Warning: non-integer x = 0.855517
#> Warning: non-integer x = 1.121461
#> Warning: non-integer x = 1.032940
#> Warning: non-integer x = 2.394251
#> Warning: non-integer x = 1.174303
#> Warning: non-integer x = 0.580254
#> Warning: non-integer x = 1.385034
#> Warning: non-integer x = 2.686692
#> Warning: non-integer x = 2.276537
#> Warning: non-integer x = 1.636827
#> Warning: non-integer x = 1.478178
#> Warning: non-integer x = 0.193903
#> Warning: non-integer x = 0.748823
#> Warning: non-integer x = 0.570077
#> Warning: non-integer x = 3.894158
#> Warning: non-integer x = 0.495330
#> Warning: non-integer x = 2.101745
#> Warning: non-integer x = 0.448473
#> Warning: non-integer x = 1.087099
#> Warning: non-integer x = 0.776590
#> Warning: non-integer x = 2.318697
#> Warning: non-integer x = 0.420300
#> Warning: non-integer x = 0.220951
#> Warning: non-integer x = 0.836537
#> Warning: non-integer x = 0.870074
#> Warning: non-integer x = 1.112781
#> Warning: non-integer x = 0.807505
#> Warning: non-integer x = 0.283155
#> Warning: non-integer x = 1.360386
#> Warning: non-integer x = 2.802682
#> Warning: non-integer x = 1.924378
#> Warning: non-integer x = 2.633454
#> Warning: non-integer x = 1.759141
#> Warning: non-integer x = 1.976819
#> Warning: non-integer x = 0.347815
#> Warning: non-integer x = 0.189591
#> Warning: non-integer x = 0.812289
#> Warning: non-integer x = 1.133003
#> Warning: non-integer x = 3.268166
#> Warning: non-integer x = 0.139538
#> Warning: non-integer x = 0.929422
#> Warning: non-integer x = 0.365175
#> Warning: non-integer x = 0.334727
#> Warning: non-integer x = 0.255962
#> Warning: non-integer x = 1.872820
#> Warning: non-integer x = 1.036534
#> Warning: non-integer x = 1.409485
#> Warning: non-integer x = 2.185439
#> Warning: non-integer x = 0.958270
#> Warning: non-integer x = 1.995124
#> Warning: non-integer x = 2.717933
#> Warning: non-integer x = 2.056348
#> Warning: non-integer x = 0.158322
#> Warning: non-integer x = 0.206376
#> Warning: non-integer x = 1.688784
#> Warning: non-integer x = 0.849860
#> Warning: non-integer x = 0.808390
#> Warning: non-integer x = 0.551212
#> Warning: non-integer x = 1.098784
#> Warning: non-integer x = 0.279697
#> Warning: non-integer x = 1.508381
#> Warning: non-integer x = 0.713522
#> Warning: non-integer x = 0.515303
#> Warning: non-integer x = 2.200027
#> Warning: non-integer x = 2.462078
#> Warning: non-integer x = 0.352161
#> Warning: non-integer x = 4.462853
#> Warning: non-integer x = 0.421862
#> Warning: non-integer x = 0.478509
#> Warning: non-integer x = 2.320215
#> Warning: non-integer x = 0.152318
#> Warning: non-integer x = 0.771495
#> Warning: non-integer x = 0.042025
#> Warning: non-integer x = 4.727657
#> Warning: non-integer x = 1.741173
#> Warning: non-integer x = 1.318779
#> Warning: non-integer x = 0.419828
#> Warning: non-integer x = 0.092012
#> Warning: non-integer x = 1.450576
#> Warning: non-integer x = 0.125987
#> Warning: non-integer x = 0.269407
#> Warning: non-integer x = 0.394259
#> Warning: non-integer x = 4.567736
#> Warning: non-integer x = 2.651275
#> Warning: non-integer x = 0.944359
#> Warning: non-integer x = 0.036484
#> Warning: non-integer x = 1.356877
#> Warning: non-integer x = 3.181249
#> Warning: non-integer x = 0.314115
#> Warning: non-integer x = 0.443844
#> Warning: non-integer x = 1.137940
#> Warning: non-integer x = 3.798335
#> Warning: non-integer x = 0.470426
#> Warning: non-integer x = 0.427663
#> Warning: non-integer x = 0.449394
#> Warning: non-integer x = 0.408510
#> Warning: non-integer x = 1.691410
#> Warning: non-integer x = 1.844950
#> Warning: non-integer x = 1.735142
#> Warning: non-integer x = 1.741237
#> Warning: non-integer x = 0.356699
#> Warning: non-integer x = 0.293062
#> Warning: non-integer x = 0.868900
#> Warning: non-integer x = 0.736135
#> Warning: non-integer x = 1.078716
#> Warning: non-integer x = 3.564638
#> Warning: non-integer x = 0.261588
#> Warning: non-integer x = 0.079943
#> Warning: non-integer x = 3.077113
#> Warning: non-integer x = 1.861669
#> Warning: non-integer x = 1.119923
#> Warning: non-integer x = 0.515375
#> Warning: non-integer x = 0.059158
#> Warning: non-integer x = 0.127325
#> Warning: non-integer x = 1.336849
#> Warning: non-integer x = 1.449449
#> Warning: non-integer x = 1.108286
#> Warning: non-integer x = 0.829000
#> Warning: non-integer x = 1.190113
#> Warning: non-integer x = 1.193853
#> Warning: non-integer x = 1.531022
#> Warning: non-integer x = 1.737044
#> Warning: non-integer x = 3.047617
#> Warning: non-integer x = 0.975906
#> Warning: non-integer x = 0.873828
#> Warning: non-integer x = 1.761465
#> Warning: non-integer x = 1.234558
#> Warning: non-integer x = 2.099073
#> Warning: non-integer x = 0.534605
#> Warning: non-integer x = 1.402176
#> Warning: non-integer x = 1.591520
#> Warning: non-integer x = 3.594791
#> Warning: non-integer x = 0.877365
#> Warning: non-integer x = 2.343399
#> Warning: non-integer x = 1.384639
#> Warning: non-integer x = 0.325161
#> Warning: non-integer x = 0.811238
#> Warning: non-integer x = 0.990644
#> Warning: non-integer x = 0.705863
#> Warning: non-integer x = 0.723258
#> Warning: non-integer x = 1.060468
#> Warning: non-integer x = 0.455244
#> Warning: non-integer x = 0.273384
#> Warning: non-integer x = 2.839535
#> Warning: non-integer x = 1.991892
#> Warning: non-integer x = 0.492192
#> Warning: non-integer x = 0.364275
#> Warning: non-integer x = 0.315653
#> Warning: non-integer x = 1.742295
#> Warning: non-integer x = 2.938551
#> Warning: non-integer x = 1.541128
#> Warning: non-integer x = 0.607770
#> Warning: non-integer x = 2.964230
#> Warning: non-integer x = 0.119051
#> Warning: non-integer x = 0.028036
#> Warning: non-integer x = 4.798402
#> Warning: non-integer x = 0.587894
#> Warning: non-integer x = 1.444996
#> Warning: non-integer x = 1.241383
#> Warning: non-integer x = 0.730719
#> Warning: non-integer x = 3.836441
#> Warning: non-integer x = 3.311790
#> Warning: non-integer x = 2.264057
#> Warning: non-integer x = 1.195216
#> Warning: non-integer x = 0.424150
#> Warning: non-integer x = 2.147660
#> Warning: non-integer x = 0.903773
#> Warning: non-integer x = 1.167698
#> Warning: non-integer x = 0.121194
#> Warning: non-integer x = 0.869546
#> Warning: non-integer x = 0.860421
#> Warning: non-integer x = 2.278221
#> Warning: non-integer x = 1.939516
#> Warning: non-integer x = 1.131802
#> Warning: non-integer x = 1.106157
#> Warning: non-integer x = 0.560058
#> Warning: non-integer x = 1.652167
#> Warning: non-integer x = 1.033996
#> Warning: non-integer x = 0.635961
#> Warning: non-integer x = 3.362607
#> Warning: non-integer x = 0.841717
#> Warning: non-integer x = 1.260047
#> Warning: non-integer x = 0.895849
#> Warning: non-integer x = 0.901515
#> Warning: non-integer x = 0.063546
#> Warning: non-integer x = 0.425501
#> Warning: non-integer x = 0.204883
#> Warning: non-integer x = 2.229192
#> Warning: non-integer x = 2.071643
#> Warning: non-integer x = 0.040072
#> Warning: non-integer x = 3.379740
#> Warning: non-integer x = 2.753702
#> Warning: non-integer x = 1.630079
#> Warning: non-integer x = 0.826142
#> Warning: non-integer x = 0.030235
#> Warning: non-integer x = 2.335951
#> Warning: non-integer x = 1.496142
#> Warning: non-integer x = 1.311469
#> Warning: non-integer x = 0.931237
#> Warning: non-integer x = 1.335219
#> Warning: non-integer x = 0.692243
#> Warning: non-integer x = 2.292150
#> Warning: non-integer x = 1.433618
#> Warning: non-integer x = 2.255795
#> Warning: non-integer x = 2.786232
#> Warning: non-integer x = 1.713552
#> Warning: non-integer x = 0.624906
#> Warning: non-integer x = 0.523288
#> Warning: non-integer x = 0.269605
#> Warning: non-integer x = 1.163358
#> Warning: non-integer x = 1.177675
#> Warning: non-integer x = 2.464748
#> Warning: non-integer x = 0.125250
#> Warning: non-integer x = 5.354764
#> Warning: non-integer x = 2.494628
#> Warning: non-integer x = 0.232738
#> Warning: non-integer x = 0.448643
#> Warning: non-integer x = 0.492771
#> Warning: non-integer x = 2.116309
#> Warning: non-integer x = 1.550416
#> Warning: non-integer x = 3.092441
#> Warning: non-integer x = 1.474653
#> Warning: non-integer x = 1.474003
#> Warning: non-integer x = 1.331683
#> Warning: non-integer x = 1.230097
#> Warning: non-integer x = 0.723036
#> Warning: non-integer x = 2.430808
#> Warning: non-integer x = 0.888726
#> Warning: non-integer x = 0.257239
#> Warning: non-integer x = 0.819994
#> Warning: non-integer x = 0.568606
#> Warning: non-integer x = 0.232104
#> Warning: non-integer x = 0.289251
#> Warning: non-integer x = 1.094217
#> Warning: non-integer x = 3.301413
#> Warning: non-integer x = 0.953072
#> Warning: non-integer x = 1.484857
#> Warning: non-integer x = 1.195699
#> Warning: non-integer x = 0.876176
#> Warning: non-integer x = 1.079804
#> Warning: non-integer x = 0.275713
#> Warning: non-integer x = 1.302358
#> Warning: non-integer x = 2.771813
#> Warning: non-integer x = 2.420476
#> Warning: non-integer x = 1.854918
#> Warning: non-integer x = 0.002535
#> Warning: non-integer x = 0.604999
#> Warning: non-integer x = 1.700835
#> Warning: non-integer x = 2.654490
#> Warning: non-integer x = 1.623235
#> Warning: non-integer x = 0.085888
#> Warning: non-integer x = 1.651634
#> Warning: non-integer x = 0.365920
#> Warning: non-integer x = 1.661335
#> Warning: non-integer x = 0.187076
#> Warning: non-integer x = 0.109675
#> Warning: non-integer x = 1.959655
#> Warning: non-integer x = 1.513340
#> Warning: non-integer x = 0.829775
#> Warning: non-integer x = 1.237275
#> Warning: non-integer x = 1.147099
#> Warning: non-integer x = 0.487721
#> Warning: non-integer x = 0.039308
#> Warning: non-integer x = 0.543948
#> Warning: non-integer x = 0.122832
#> Warning: non-integer x = 1.326549
#> Warning: non-integer x = 1.246487
#> Warning: non-integer x = 2.031743
#> Warning: non-integer x = 1.584824
#> Warning: non-integer x = 0.305029
#> Warning: non-integer x = 0.052824
#> Warning: non-integer x = 1.487663
#> Warning: non-integer x = 0.687205
#> Warning: non-integer x = 1.091548
#> Warning: non-integer x = 0.674552
#> Warning: non-integer x = 1.506427
#> Warning: non-integer x = 2.525539
#> Warning: non-integer x = 0.095940
#> Warning: non-integer x = 1.549447
#> Warning: non-integer x = 1.271056
#> Warning: non-integer x = 1.510887
#> Warning: non-integer x = 0.036094
#> Warning: non-integer x = 3.037995
#> Warning: non-integer x = 1.921470
#> Warning: non-integer x = 0.301035
#> Warning: non-integer x = 0.233106
#> Warning: non-integer x = 3.255216
#> Warning: non-integer x = 2.344871
#> Warning: non-integer x = 0.862384
#> Warning: non-integer x = 0.950505
#> Warning: non-integer x = 0.022900
#> Warning: non-integer x = 2.268023
#> Warning: non-integer x = 1.372636
#> Warning: non-integer x = 1.698686
#> Warning: non-integer x = 0.151466
#> Warning: non-integer x = 0.253957
#> Warning: non-integer x = 2.049523
#> Warning: non-integer x = 2.457965
#> Warning: non-integer x = 2.743411
#> Warning: non-integer x = 1.983705
#> Warning: non-integer x = 0.066893
#> Warning: non-integer x = 0.337422
#> Warning: non-integer x = 1.166069
#> Warning: non-integer x = 1.391971
#> Warning: non-integer x = 1.682941
#> Warning: non-integer x = 0.613688
#> Warning: non-integer x = 2.290159
#> Warning: non-integer x = 1.310443
#> Warning: non-integer x = 1.792838
#> Warning: non-integer x = 0.261293
#> Warning: non-integer x = 0.930556
#> Warning: non-integer x = 0.269787
#> Warning: non-integer x = 2.275636
#> Warning: non-integer x = 3.120161
#> Warning: non-integer x = 1.022295
#> Warning: non-integer x = 0.493228
#> Warning: non-integer x = 0.141138
#> Warning: non-integer x = 2.083549
#> Warning: non-integer x = 1.671322
#> Warning: non-integer x = 2.532332
#> Warning: non-integer x = 0.225503
#> Warning: non-integer x = 1.002318
#> Warning: non-integer x = 0.389453
#> Warning: non-integer x = 2.000102
#> Warning: non-integer x = 1.667927
#> Warning: non-integer x = 2.602332
#> Warning: non-integer x = 0.652187
#> Warning: non-integer x = 0.884575
#> Warning: non-integer x = 0.900727
#> Warning: non-integer x = 1.791378
#> Warning: non-integer x = 2.410187
#> Warning: non-integer x = 0.623469
#> Warning: non-integer x = 0.733940
#> Warning: non-integer x = 0.263251
#> Warning: non-integer x = 2.343396
#> Warning: non-integer x = 2.924936
#> Warning: non-integer x = 3.853225
#> Warning: non-integer x = 0.446586
#> Warning: non-integer x = 0.207626
#> Warning: non-integer x = 2.332003
#> Warning: non-integer x = 3.245376
#> Warning: non-integer x = 2.203550
#> Warning: non-integer x = 2.769979
#> Warning: non-integer x = 1.925287
#> Warning: non-integer x = 0.697144
#> Warning: non-integer x = 1.228001
#> Warning: non-integer x = 0.445921
#> Warning: non-integer x = 0.323447
#> Warning: non-integer x = 1.374412
#> Warning: non-integer x = 2.542994
#> Warning: non-integer x = 0.635592
#> Warning: non-integer x = 0.362059
#> Warning: non-integer x = 1.873433
#> Warning: non-integer x = 1.273554
#> Warning: non-integer x = 1.294386
#> Warning: non-integer x = 1.828722
#> Warning: non-integer x = 0.437308
#> Warning: non-integer x = 0.285936
#> Warning: non-integer x = 0.883880
#> Warning: non-integer x = 1.147421
#> Warning: non-integer x = 1.313235
#> Warning: non-integer x = 2.309712
#> Warning: non-integer x = 1.744076
#> Warning: non-integer x = 1.242005
#> Warning: non-integer x = 0.198810
#> Warning: non-integer x = 1.808872
#> Warning: non-integer x = 1.310986
#> Warning: non-integer x = 0.055501
#> Warning: non-integer x = 1.914360
#> Warning: non-integer x = 1.673503
#> Warning: non-integer x = 2.292571
#> Warning: non-integer x = 2.308022
#> Warning: non-integer x = 0.316142
#> Warning: non-integer x = 1.481109
#> Warning: non-integer x = 1.519276
#> Warning: non-integer x = 2.878379
#> Warning: non-integer x = 2.064493
#> Warning: non-integer x = 1.112127
#> Warning: non-integer x = 1.553544
#> Warning: non-integer x = 1.215117
#> Warning: non-integer x = 1.247874
#> Warning: non-integer x = 0.503595
#> Warning: non-integer x = 0.737545
#> Warning: non-integer x = 0.296707
#> Warning: non-integer x = 1.560751
#> Warning: non-integer x = 3.018633
#> Warning: non-integer x = 0.096777
#> Warning: non-integer x = 1.991503
#> Warning: non-integer x = 1.191238
#> Warning: non-integer x = 0.665474
#> Warning: non-integer x = 1.759900
#> Warning: non-integer x = 1.362915
#> Warning: non-integer x = 0.949228
#> Warning: non-integer x = 0.540615
#> Warning: non-integer x = 1.059817
#> Warning: non-integer x = 2.430861
#> Warning: non-integer x = 1.668157
#> Warning: non-integer x = 2.603451
#> Warning: non-integer x = 2.859522
#> Warning: non-integer x = 2.351354
#> Warning: non-integer x = 2.619250
#> Warning: non-integer x = 0.921743
#> Warning: non-integer x = 0.822111
#> Warning: non-integer x = 0.335957
#> Warning: non-integer x = 1.325973
#> Warning: non-integer x = 3.100009
#> Warning: non-integer x = 2.299542
#> Warning: non-integer x = 2.432324
#> Warning: non-integer x = 0.122275
#> Warning: non-integer x = 1.443431
#> Warning: non-integer x = 2.567590
#> Warning: non-integer x = 2.080129
#> Warning: non-integer x = 0.186852
#> Warning: non-integer x = 0.410207
#> Warning: non-integer x = 0.194716
#> Warning: non-integer x = 2.977427
#> Warning: non-integer x = 2.148024
#> Warning: non-integer x = 0.978098
#> Warning: non-integer x = 0.394642
#> Warning: non-integer x = 1.249619
#> Warning: non-integer x = 0.384495
#> Warning: non-integer x = 2.491237
#> Warning: non-integer x = 0.341868
#> Warning: non-integer x = 1.105286
#> Warning: non-integer x = 2.422442
#> Warning: non-integer x = 1.415969
#> Warning: non-integer x = 0.274518
#> Warning: non-integer x = 1.137590
#> Warning: non-integer x = 0.809048
#> Warning: non-integer x = 0.869008
#> Warning: non-integer x = 0.832581
#> Warning: non-integer x = 0.186168
#> Warning: non-integer x = 1.495550
#> Warning: non-integer x = 0.352331
#> Warning: non-integer x = 0.685807
#> Warning: non-integer x = 0.607429
#> Warning: non-integer x = 1.226654
#> Warning: non-integer x = 1.584685
#> Warning: non-integer x = 2.483981
#> Warning: non-integer x = 0.624781
#> Warning: non-integer x = 2.605829
#> Warning: non-integer x = 1.662160
#> Warning: non-integer x = 0.816379
#> Warning: non-integer x = 1.879605
#> Warning: non-integer x = 2.682350
#> Warning: non-integer x = 2.876071
#> Warning: non-integer x = 1.979520
#> Warning: non-integer x = 1.537949
#> Warning: non-integer x = 0.249197
#> Warning: non-integer x = 0.983365
#> Warning: non-integer x = 1.931434
#> Warning: non-integer x = 0.632137
#> Warning: non-integer x = 0.616532
#> Warning: non-integer x = 0.823329
#> Warning: non-integer x = 0.762965
#> Warning: non-integer x = 0.768681
#> Warning: non-integer x = 1.070119
#> Warning: non-integer x = 0.659582
#> Warning: non-integer x = 1.640366
#> Warning: non-integer x = 1.673227
#> Warning: non-integer x = 0.913275
#> Warning: non-integer x = 1.490096
#> Warning: non-integer x = 1.948251
#> Warning: non-integer x = 0.343625
#> Warning: non-integer x = 1.126407
#> Warning: non-integer x = 0.873860
#> Warning: non-integer x = 1.727636
#> Warning: non-integer x = 1.977867
#> Warning: non-integer x = 0.313410
#> Warning: non-integer x = 0.471197
#> Warning: non-integer x = 0.551535
#> Warning: non-integer x = 1.509848
#> Warning: non-integer x = 0.629702
#> Warning: non-integer x = 0.348436
#> Warning: non-integer x = 1.970867
#> Warning: non-integer x = 0.088140
#> Warning: non-integer x = 2.973687
#> Warning: non-integer x = 2.156538
#> Warning: non-integer x = 1.796577
#> Warning: non-integer x = 1.775842
#> Warning: non-integer x = 0.633888
#> Warning: non-integer x = 0.889686
#> Warning: non-integer x = 0.468254
#> Warning: non-integer x = 3.148896
#> Warning: non-integer x = 0.493841
#> Warning: non-integer x = 1.676505
#> Warning: non-integer x = 0.357543
#> Warning: non-integer x = 3.380797
#> Warning: non-integer x = 1.271456
#> Warning: non-integer x = 0.000583
#> Warning: non-integer x = 0.415780
#> Warning: non-integer x = 0.290462
#> Warning: non-integer x = 1.655459
#> Warning: non-integer x = 1.445051
#> Warning: non-integer x = 0.084909
#> Warning: non-integer x = 2.666839
#> Warning: non-integer x = 2.209760
#> Warning: non-integer x = 2.424336
#> Warning: non-integer x = 0.093149
#> Warning: non-integer x = 1.137543
#> Warning: non-integer x = 1.598132
#> Warning: non-integer x = 0.304918
#> Warning: non-integer x = 1.010290
#> Warning: non-integer x = 1.211061
#> Warning: non-integer x = 0.554728
#> Warning: non-integer x = 1.045729
#> Warning: non-integer x = 1.295697
#> Warning: non-integer x = 1.026977
#> Warning: non-integer x = 2.651577
#> Warning: non-integer x = 0.192999
#> Warning: non-integer x = 1.367307
#> Warning: non-integer x = 1.708293
#> Warning: non-integer x = 3.860420
#> Warning: non-integer x = 1.831957
#> Warning: non-integer x = 0.848644
#> Warning: non-integer x = 0.294619
#> Warning: non-integer x = 0.512704
#> Warning: non-integer x = 3.298718
#> Warning: non-integer x = 2.020898
#> Warning: non-integer x = 2.310049
#> Warning: non-integer x = 2.372237
#> Warning: non-integer x = 2.467452
#> Warning: non-integer x = 3.350859
#> Warning: non-integer x = 1.959802
#> Warning: non-integer x = 3.267403
#> Warning: non-integer x = 0.109324
#> Warning: non-integer x = 0.964025
#> Warning: non-integer x = 0.115303
#> Warning: non-integer x = 2.288336
#> Warning: non-integer x = 1.159781
#> Warning: non-integer x = 1.612435
#> Warning: non-integer x = 1.403611
#> Warning: non-integer x = 3.139936
#> Warning: non-integer x = 2.448233
#> Warning: non-integer x = 1.514228