Estimate the Individual Treatment Effect using Poisson Regression

Estimates the Individual Treatment Effect using Poisson Regression given a response vector, a treatment vector, and a features matrix.

Usage

estimate_ite_poisson(y, z, X, X_names, include_offset, offset_name)

Arguments

y

the observed response vector

z

the treatment vector

X

the features matrix

X_names

the names of the covariates

include_offset

whether or not to include an offset when estimating the ITE, for poisson only

offset_name

the name of the offset, if it is to be included

Value

a vector of ITE estimates

Examples

dataset <- generate_cre_dataset(n = 1000, rho = 0, n_rules = 2, p = 10,
                                effect_size = 2, binary = FALSE)

# Initialize parameters
y <- abs(dataset[["y"]])
z <- dataset[["z"]]
X <- as.data.frame(dataset[["X"]])
X_names <- names(as.data.frame(X))
include_offset <- FALSE
offset_name <- NA

ite_list <- estimate_ite_poisson(y, z, X, X_names, include_offset, offset_name)
#> Warning: non-integer x = 2.963416
#> Warning: non-integer x = 3.262616
#> Warning: non-integer x = 0.619355
#> Warning: non-integer x = 0.023706
#> Warning: non-integer x = 2.807754
#> Warning: non-integer x = 1.064135
#> Warning: non-integer x = 1.079569
#> Warning: non-integer x = 0.600038
#> Warning: non-integer x = 1.990271
#> Warning: non-integer x = 1.530926
#> Warning: non-integer x = 3.855605
#> Warning: non-integer x = 0.990837
#> Warning: non-integer x = 0.971855
#> Warning: non-integer x = 0.607105
#> Warning: non-integer x = 0.842852
#> Warning: non-integer x = 1.829751
#> Warning: non-integer x = 2.945851
#> Warning: non-integer x = 3.037270
#> Warning: non-integer x = 0.900273
#> Warning: non-integer x = 4.261793
#> Warning: non-integer x = 2.465365
#> Warning: non-integer x = 1.028311
#> Warning: non-integer x = 0.676878
#> Warning: non-integer x = 3.578988
#> Warning: non-integer x = 2.038527
#> Warning: non-integer x = 1.321585
#> Warning: non-integer x = 5.007156
#> Warning: non-integer x = 1.072940
#> Warning: non-integer x = 1.235337
#> Warning: non-integer x = 1.496016
#> Warning: non-integer x = 2.671240
#> Warning: non-integer x = 0.365859
#> Warning: non-integer x = 1.351049
#> Warning: non-integer x = 1.883915
#> Warning: non-integer x = 0.204347
#> Warning: non-integer x = 0.616811
#> Warning: non-integer x = 0.311266
#> Warning: non-integer x = 0.980801
#> Warning: non-integer x = 0.249917
#> Warning: non-integer x = 2.379650
#> Warning: non-integer x = 2.792830
#> Warning: non-integer x = 2.592191
#> Warning: non-integer x = 1.503886
#> Warning: non-integer x = 0.734312
#> Warning: non-integer x = 0.868660
#> Warning: non-integer x = 2.891208
#> Warning: non-integer x = 1.426717
#> Warning: non-integer x = 1.659302
#> Warning: non-integer x = 1.638152
#> Warning: non-integer x = 1.301482
#> Warning: non-integer x = 1.378352
#> Warning: non-integer x = 2.181030
#> Warning: non-integer x = 0.528900
#> Warning: non-integer x = 0.201289
#> Warning: non-integer x = 4.760651
#> Warning: non-integer x = 0.108659
#> Warning: non-integer x = 1.628947
#> Warning: non-integer x = 0.469962
#> Warning: non-integer x = 0.548030
#> Warning: non-integer x = 1.246407
#> Warning: non-integer x = 0.953947
#> Warning: non-integer x = 1.393077
#> Warning: non-integer x = 1.589260
#> Warning: non-integer x = 0.852867
#> Warning: non-integer x = 1.673240
#> Warning: non-integer x = 2.642432
#> Warning: non-integer x = 2.224322
#> Warning: non-integer x = 1.291044
#> Warning: non-integer x = 2.020677
#> Warning: non-integer x = 0.795732
#> Warning: non-integer x = 0.974803
#> Warning: non-integer x = 3.174193
#> Warning: non-integer x = 1.469426
#> Warning: non-integer x = 2.240905
#> Warning: non-integer x = 1.782187
#> Warning: non-integer x = 1.654870
#> Warning: non-integer x = 1.717452
#> Warning: non-integer x = 0.331997
#> Warning: non-integer x = 0.110046
#> Warning: non-integer x = 5.314144
#> Warning: non-integer x = 2.770454
#> Warning: non-integer x = 3.161053
#> Warning: non-integer x = 1.792867
#> Warning: non-integer x = 1.126057
#> Warning: non-integer x = 4.033747
#> Warning: non-integer x = 3.160210
#> Warning: non-integer x = 0.769250
#> Warning: non-integer x = 0.219889
#> Warning: non-integer x = 2.700808
#> Warning: non-integer x = 3.224498
#> Warning: non-integer x = 3.275601
#> Warning: non-integer x = 0.148404
#> Warning: non-integer x = 1.337238
#> Warning: non-integer x = 0.486678
#> Warning: non-integer x = 1.326806
#> Warning: non-integer x = 3.055630
#> Warning: non-integer x = 3.004833
#> Warning: non-integer x = 0.785443
#> Warning: non-integer x = 1.535055
#> Warning: non-integer x = 3.824825
#> Warning: non-integer x = 0.398985
#> Warning: non-integer x = 1.119424
#> Warning: non-integer x = 2.854999
#> Warning: non-integer x = 3.379844
#> Warning: non-integer x = 2.837002
#> Warning: non-integer x = 0.465808
#> Warning: non-integer x = 1.899937
#> Warning: non-integer x = 2.143123
#> Warning: non-integer x = 0.534352
#> Warning: non-integer x = 2.861542
#> Warning: non-integer x = 1.191643
#> Warning: non-integer x = 0.287352
#> Warning: non-integer x = 2.057722
#> Warning: non-integer x = 2.198176
#> Warning: non-integer x = 1.701234
#> Warning: non-integer x = 0.840997
#> Warning: non-integer x = 1.945719
#> Warning: non-integer x = 0.642622
#> Warning: non-integer x = 0.964305
#> Warning: non-integer x = 2.760725
#> Warning: non-integer x = 0.199507
#> Warning: non-integer x = 3.745313
#> Warning: non-integer x = 4.018896
#> Warning: non-integer x = 1.082672
#> Warning: non-integer x = 0.181196
#> Warning: non-integer x = 1.432456
#> Warning: non-integer x = 0.753953
#> Warning: non-integer x = 0.710958
#> Warning: non-integer x = 2.554216
#> Warning: non-integer x = 1.929814
#> Warning: non-integer x = 2.043009
#> Warning: non-integer x = 2.290318
#> Warning: non-integer x = 1.619085
#> Warning: non-integer x = 0.118493
#> Warning: non-integer x = 0.489193
#> Warning: non-integer x = 1.368370
#> Warning: non-integer x = 0.626685
#> Warning: non-integer x = 3.687855
#> Warning: non-integer x = 2.607177
#> Warning: non-integer x = 0.926232
#> Warning: non-integer x = 3.306393
#> Warning: non-integer x = 3.048952
#> Warning: non-integer x = 0.950510
#> Warning: non-integer x = 0.996823
#> Warning: non-integer x = 0.551179
#> Warning: non-integer x = 5.473938
#> Warning: non-integer x = 2.568496
#> Warning: non-integer x = 0.936298
#> Warning: non-integer x = 0.952311
#> Warning: non-integer x = 4.255587
#> Warning: non-integer x = 0.957154
#> Warning: non-integer x = 2.683964
#> Warning: non-integer x = 2.386624
#> Warning: non-integer x = 3.301624
#> Warning: non-integer x = 0.034874
#> Warning: non-integer x = 0.016624
#> Warning: non-integer x = 0.109269
#> Warning: non-integer x = 1.820144
#> Warning: non-integer x = 2.348011
#> Warning: non-integer x = 0.823551
#> Warning: non-integer x = 0.916483
#> Warning: non-integer x = 3.884752
#> Warning: non-integer x = 2.948549
#> Warning: non-integer x = 1.350663
#> Warning: non-integer x = 1.928434
#> Warning: non-integer x = 1.386911
#> Warning: non-integer x = 1.658633
#> Warning: non-integer x = 0.784633
#> Warning: non-integer x = 0.139149
#> Warning: non-integer x = 5.015606
#> Warning: non-integer x = 3.397691
#> Warning: non-integer x = 0.887505
#> Warning: non-integer x = 0.715899
#> Warning: non-integer x = 0.895787
#> Warning: non-integer x = 0.959160
#> Warning: non-integer x = 0.694762
#> Warning: non-integer x = 1.523896
#> Warning: non-integer x = 1.717515
#> Warning: non-integer x = 4.683024
#> Warning: non-integer x = 0.166421
#> Warning: non-integer x = 1.378501
#> Warning: non-integer x = 0.050670
#> Warning: non-integer x = 1.120978
#> Warning: non-integer x = 3.920757
#> Warning: non-integer x = 4.163745
#> Warning: non-integer x = 2.082256
#> Warning: non-integer x = 0.357292
#> Warning: non-integer x = 3.117639
#> Warning: non-integer x = 0.139841
#> Warning: non-integer x = 1.541880
#> Warning: non-integer x = 0.381269
#> Warning: non-integer x = 0.841467
#> Warning: non-integer x = 0.341782
#> Warning: non-integer x = 2.670921
#> Warning: non-integer x = 0.634113
#> Warning: non-integer x = 1.458710
#> Warning: non-integer x = 2.424559
#> Warning: non-integer x = 1.090539
#> Warning: non-integer x = 2.054378
#> Warning: non-integer x = 0.140366
#> Warning: non-integer x = 1.296643
#> Warning: non-integer x = 0.407075
#> Warning: non-integer x = 0.843405
#> Warning: non-integer x = 2.067281
#> Warning: non-integer x = 1.954213
#> Warning: non-integer x = 0.207869
#> Warning: non-integer x = 3.114283
#> Warning: non-integer x = 0.717386
#> Warning: non-integer x = 0.570328
#> Warning: non-integer x = 0.588210
#> Warning: non-integer x = 2.897491
#> Warning: non-integer x = 1.008892
#> Warning: non-integer x = 0.986458
#> Warning: non-integer x = 1.088020
#> Warning: non-integer x = 4.512704
#> Warning: non-integer x = 1.413765
#> Warning: non-integer x = 0.148029
#> Warning: non-integer x = 2.973049
#> Warning: non-integer x = 1.425745
#> Warning: non-integer x = 0.250944
#> Warning: non-integer x = 2.446726
#> Warning: non-integer x = 2.733314
#> Warning: non-integer x = 0.593312
#> Warning: non-integer x = 0.896183
#> Warning: non-integer x = 2.606116
#> Warning: non-integer x = 0.833752
#> Warning: non-integer x = 1.201580
#> Warning: non-integer x = 0.946593
#> Warning: non-integer x = 2.030531
#> Warning: non-integer x = 0.669567
#> Warning: non-integer x = 3.532083
#> Warning: non-integer x = 2.997164
#> Warning: non-integer x = 4.131534
#> Warning: non-integer x = 2.606899
#> Warning: non-integer x = 2.110668
#> Warning: non-integer x = 1.211108
#> Warning: non-integer x = 0.678904
#> Warning: non-integer x = 1.498763
#> Warning: non-integer x = 1.327809
#> Warning: non-integer x = 1.523224
#> Warning: non-integer x = 2.650112
#> Warning: non-integer x = 0.545843
#> Warning: non-integer x = 2.424127
#> Warning: non-integer x = 0.851275
#> Warning: non-integer x = 0.536773
#> Warning: non-integer x = 1.653827
#> Warning: non-integer x = 3.298134
#> Warning: non-integer x = 1.583785
#> Warning: non-integer x = 2.462967
#> Warning: non-integer x = 0.706164
#> Warning: non-integer x = 0.185598
#> Warning: non-integer x = 0.687115
#> Warning: non-integer x = 1.981780
#> Warning: non-integer x = 0.257074
#> Warning: non-integer x = 0.891842
#> Warning: non-integer x = 2.673237
#> Warning: non-integer x = 0.837765
#> Warning: non-integer x = 0.416394
#> Warning: non-integer x = 2.673409
#> Warning: non-integer x = 2.413339
#> Warning: non-integer x = 0.353086
#> Warning: non-integer x = 0.437583
#> Warning: non-integer x = 0.928916
#> Warning: non-integer x = 1.353512
#> Warning: non-integer x = 1.276729
#> Warning: non-integer x = 1.418561
#> Warning: non-integer x = 0.009530
#> Warning: non-integer x = 2.257284
#> Warning: non-integer x = 0.879883
#> Warning: non-integer x = 1.283955
#> Warning: non-integer x = 1.228178
#> Warning: non-integer x = 3.701750
#> Warning: non-integer x = 3.584367
#> Warning: non-integer x = 3.294421
#> Warning: non-integer x = 1.296739
#> Warning: non-integer x = 1.600065
#> Warning: non-integer x = 4.464299
#> Warning: non-integer x = 0.535350
#> Warning: non-integer x = 0.458016
#> Warning: non-integer x = 0.251049
#> Warning: non-integer x = 0.504585
#> Warning: non-integer x = 1.238697
#> Warning: non-integer x = 1.046845
#> Warning: non-integer x = 6.147341
#> Warning: non-integer x = 1.991287
#> Warning: non-integer x = 0.964633
#> Warning: non-integer x = 0.873475
#> Warning: non-integer x = 2.897497
#> Warning: non-integer x = 5.567493
#> Warning: non-integer x = 0.019006
#> Warning: non-integer x = 2.090801
#> Warning: non-integer x = 1.914985
#> Warning: non-integer x = 1.291861
#> Warning: non-integer x = 1.121833
#> Warning: non-integer x = 2.647272
#> Warning: non-integer x = 1.638881
#> Warning: non-integer x = 0.150357
#> Warning: non-integer x = 1.386301
#> Warning: non-integer x = 1.381724
#> Warning: non-integer x = 1.377907
#> Warning: non-integer x = 2.718618
#> Warning: non-integer x = 0.813842
#> Warning: non-integer x = 0.159963
#> Warning: non-integer x = 1.827758
#> Warning: non-integer x = 1.642740
#> Warning: non-integer x = 0.131093
#> Warning: non-integer x = 0.267183
#> Warning: non-integer x = 1.399642
#> Warning: non-integer x = 2.690048
#> Warning: non-integer x = 1.135774
#> Warning: non-integer x = 0.281134
#> Warning: non-integer x = 1.144544
#> Warning: non-integer x = 1.387381
#> Warning: non-integer x = 0.169350
#> Warning: non-integer x = 2.333010
#> Warning: non-integer x = 0.582262
#> Warning: non-integer x = 1.560848
#> Warning: non-integer x = 0.950898
#> Warning: non-integer x = 2.121924
#> Warning: non-integer x = 1.564429
#> Warning: non-integer x = 1.816721
#> Warning: non-integer x = 1.577224
#> Warning: non-integer x = 2.762547
#> Warning: non-integer x = 1.187980
#> Warning: non-integer x = 3.018170
#> Warning: non-integer x = 2.914077
#> Warning: non-integer x = 0.578744
#> Warning: non-integer x = 3.619969
#> Warning: non-integer x = 0.734002
#> Warning: non-integer x = 0.934871
#> Warning: non-integer x = 3.640643
#> Warning: non-integer x = 0.034594
#> Warning: non-integer x = 2.059261
#> Warning: non-integer x = 1.821830
#> Warning: non-integer x = 1.409598
#> Warning: non-integer x = 0.906394
#> Warning: non-integer x = 2.144452
#> Warning: non-integer x = 2.796092
#> Warning: non-integer x = 0.854614
#> Warning: non-integer x = 2.968179
#> Warning: non-integer x = 2.765013
#> Warning: non-integer x = 2.889111
#> Warning: non-integer x = 0.440896
#> Warning: non-integer x = 1.130893
#> Warning: non-integer x = 0.092991
#> Warning: non-integer x = 1.576159
#> Warning: non-integer x = 1.811811
#> Warning: non-integer x = 0.094885
#> Warning: non-integer x = 3.772713
#> Warning: non-integer x = 0.397899
#> Warning: non-integer x = 1.080670
#> Warning: non-integer x = 1.840037
#> Warning: non-integer x = 0.898840
#> Warning: non-integer x = 2.675477
#> Warning: non-integer x = 1.062211
#> Warning: non-integer x = 1.746443
#> Warning: non-integer x = 2.070446
#> Warning: non-integer x = 1.388297
#> Warning: non-integer x = 0.019360
#> Warning: non-integer x = 1.127899
#> Warning: non-integer x = 1.924229
#> Warning: non-integer x = 3.181211
#> Warning: non-integer x = 0.442875
#> Warning: non-integer x = 1.610030
#> Warning: non-integer x = 0.265896
#> Warning: non-integer x = 1.204928
#> Warning: non-integer x = 1.263469
#> Warning: non-integer x = 0.977857
#> Warning: non-integer x = 2.221705
#> Warning: non-integer x = 0.573440
#> Warning: non-integer x = 2.401518
#> Warning: non-integer x = 1.092168
#> Warning: non-integer x = 0.211828
#> Warning: non-integer x = 1.631176
#> Warning: non-integer x = 1.410943
#> Warning: non-integer x = 2.921441
#> Warning: non-integer x = 0.783405
#> Warning: non-integer x = 4.766854
#> Warning: non-integer x = 1.103725
#> Warning: non-integer x = 0.014392
#> Warning: non-integer x = 0.205555
#> Warning: non-integer x = 0.717228
#> Warning: non-integer x = 1.276288
#> Warning: non-integer x = 3.021159
#> Warning: non-integer x = 0.643953
#> Warning: non-integer x = 1.055770
#> Warning: non-integer x = 0.119001
#> Warning: non-integer x = 1.109685
#> Warning: non-integer x = 0.240183
#> Warning: non-integer x = 3.971792
#> Warning: non-integer x = 0.893938
#> Warning: non-integer x = 3.200440
#> Warning: non-integer x = 0.065302
#> Warning: non-integer x = 0.139391
#> Warning: non-integer x = 2.605745
#> Warning: non-integer x = 3.871709
#> Warning: non-integer x = 3.294807
#> Warning: non-integer x = 1.918093
#> Warning: non-integer x = 0.093648
#> Warning: non-integer x = 1.081778
#> Warning: non-integer x = 0.157614
#> Warning: non-integer x = 1.738085
#> Warning: non-integer x = 0.315785
#> Warning: non-integer x = 0.303150
#> Warning: non-integer x = 1.374614
#> Warning: non-integer x = 1.965177
#> Warning: non-integer x = 1.777077
#> Warning: non-integer x = 2.421170
#> Warning: non-integer x = 1.109009
#> Warning: non-integer x = 1.387434
#> Warning: non-integer x = 2.947914
#> Warning: non-integer x = 2.542236
#> Warning: non-integer x = 0.852474
#> Warning: non-integer x = 2.163917
#> Warning: non-integer x = 2.835623
#> Warning: non-integer x = 1.123121
#> Warning: non-integer x = 3.279993
#> Warning: non-integer x = 3.227249
#> Warning: non-integer x = 0.856669
#> Warning: non-integer x = 0.089720
#> Warning: non-integer x = 0.353752
#> Warning: non-integer x = 1.231874
#> Warning: non-integer x = 1.801440
#> Warning: non-integer x = 3.701535
#> Warning: non-integer x = 3.424982
#> Warning: non-integer x = 1.512317
#> Warning: non-integer x = 1.207532
#> Warning: non-integer x = 1.205687
#> Warning: non-integer x = 1.361830
#> Warning: non-integer x = 1.698425
#> Warning: non-integer x = 1.094340
#> Warning: non-integer x = 0.502821
#> Warning: non-integer x = 3.113678
#> Warning: non-integer x = 1.846693
#> Warning: non-integer x = 0.904733
#> Warning: non-integer x = 2.281444
#> Warning: non-integer x = 0.581085
#> Warning: non-integer x = 0.462103
#> Warning: non-integer x = 2.134824
#> Warning: non-integer x = 1.864070
#> Warning: non-integer x = 1.207117
#> Warning: non-integer x = 1.734183
#> Warning: non-integer x = 0.408012
#> Warning: non-integer x = 2.647893
#> Warning: non-integer x = 1.822174
#> Warning: non-integer x = 1.798635
#> Warning: non-integer x = 1.910181
#> Warning: non-integer x = 1.056510
#> Warning: non-integer x = 0.474087
#> Warning: non-integer x = 2.400194
#> Warning: non-integer x = 1.223395
#> Warning: non-integer x = 0.981251
#> Warning: non-integer x = 3.551126
#> Warning: non-integer x = 0.432612
#> Warning: non-integer x = 0.755945
#> Warning: non-integer x = 0.388182
#> Warning: non-integer x = 0.167449
#> Warning: non-integer x = 2.496428
#> Warning: non-integer x = 0.368791
#> Warning: non-integer x = 0.629021
#> Warning: non-integer x = 0.337452
#> Warning: non-integer x = 0.416148
#> Warning: non-integer x = 1.311771
#> Warning: non-integer x = 2.470900
#> Warning: non-integer x = 0.223755
#> Warning: non-integer x = 1.086468
#> Warning: non-integer x = 0.911118
#> Warning: non-integer x = 0.037397
#> Warning: non-integer x = 2.587214
#> Warning: non-integer x = 1.461958
#> Warning: non-integer x = 1.499300
#> Warning: non-integer x = 0.633815
#> Warning: non-integer x = 1.624960
#> Warning: non-integer x = 1.379396
#> Warning: non-integer x = 2.044004
#> Warning: non-integer x = 0.533754
#> Warning: non-integer x = 2.540665
#> Warning: non-integer x = 0.108310
#> Warning: non-integer x = 1.416773
#> Warning: non-integer x = 1.598868
#> Warning: non-integer x = 2.198335
#> Warning: non-integer x = 0.701628
#> Warning: non-integer x = 0.331313
#> Warning: non-integer x = 0.075656
#> Warning: non-integer x = 0.974077
#> Warning: non-integer x = 2.087445
#> Warning: non-integer x = 3.483981
#> Warning: non-integer x = 0.500344
#> Warning: non-integer x = 2.937225
#> Warning: non-integer x = 2.318398
#> Warning: non-integer x = 1.635642
#> Warning: non-integer x = 0.545843
#> Warning: non-integer x = 0.574032
#> Warning: non-integer x = 0.081946
#> Warning: non-integer x = 0.956191
#> Warning: non-integer x = 2.970048
#> Warning: non-integer x = 0.443264
#> Warning: non-integer x = 0.948899
#> Warning: non-integer x = 0.068049
#> Warning: non-integer x = 0.149359
#> Warning: non-integer x = 0.544818
#> Warning: non-integer x = 1.231266
#> Warning: non-integer x = 0.107133
#> Warning: non-integer x = 0.621563
#> Warning: non-integer x = 0.139412
#> Warning: non-integer x = 1.782896
#> Warning: non-integer x = 1.563920
#> Warning: non-integer x = 0.211790
#> Warning: non-integer x = 1.399806
#> Warning: non-integer x = 2.966675
#> Warning: non-integer x = 1.277840
#> Warning: non-integer x = 2.961140
#> Warning: non-integer x = 1.078179
#> Warning: non-integer x = 0.511016
#> Warning: non-integer x = 1.452897
#> Warning: non-integer x = 1.440812
#> Warning: non-integer x = 0.654314
#> Warning: non-integer x = 0.504476
#> Warning: non-integer x = 0.142262
#> Warning: non-integer x = 1.635669
#> Warning: non-integer x = 1.032399
#> Warning: non-integer x = 3.316770
#> Warning: non-integer x = 0.396700
#> Warning: non-integer x = 2.403458
#> Warning: non-integer x = 0.039024
#> Warning: non-integer x = 0.385995
#> Warning: non-integer x = 2.120606
#> Warning: non-integer x = 0.527936
#> Warning: non-integer x = 1.637980
#> Warning: non-integer x = 1.420485
#> Warning: non-integer x = 1.127996
#> Warning: non-integer x = 1.111582
#> Warning: non-integer x = 1.124644
#> Warning: non-integer x = 1.153443
#> Warning: non-integer x = 0.629283
#> Warning: non-integer x = 1.832217
#> Warning: non-integer x = 0.128094
#> Warning: non-integer x = 0.380698
#> Warning: non-integer x = 1.570876
#> Warning: non-integer x = 0.654145
#> Warning: non-integer x = 2.204442
#> Warning: non-integer x = 3.306180
#> Warning: non-integer x = 0.015399
#> Warning: non-integer x = 0.316443
#> Warning: non-integer x = 2.946888
#> Warning: non-integer x = 0.851723
#> Warning: non-integer x = 0.709603
#> Warning: non-integer x = 1.130735
#> Warning: non-integer x = 0.230667
#> Warning: non-integer x = 0.249097
#> Warning: non-integer x = 0.211061
#> Warning: non-integer x = 1.546213
#> Warning: non-integer x = 1.512755
#> Warning: non-integer x = 0.674868
#> Warning: non-integer x = 0.942275
#> Warning: non-integer x = 0.943887
#> Warning: non-integer x = 0.693307
#> Warning: non-integer x = 2.751844
#> Warning: non-integer x = 0.106430
#> Warning: non-integer x = 0.888194
#> Warning: non-integer x = 1.661921
#> Warning: non-integer x = 0.580574
#> Warning: non-integer x = 0.500812
#> Warning: non-integer x = 2.442689
#> Warning: non-integer x = 0.571641
#> Warning: non-integer x = 0.963751
#> Warning: non-integer x = 0.173589
#> Warning: non-integer x = 0.177701
#> Warning: non-integer x = 0.700641
#> Warning: non-integer x = 1.293986
#> Warning: non-integer x = 0.759469
#> Warning: non-integer x = 0.952819
#> Warning: non-integer x = 1.290940
#> Warning: non-integer x = 1.568973
#> Warning: non-integer x = 0.104359
#> Warning: non-integer x = 2.808456
#> Warning: non-integer x = 2.514532
#> Warning: non-integer x = 1.096150
#> Warning: non-integer x = 1.768577
#> Warning: non-integer x = 1.797648
#> Warning: non-integer x = 1.030365
#> Warning: non-integer x = 1.386111
#> Warning: non-integer x = 1.852178
#> Warning: non-integer x = 1.966701
#> Warning: non-integer x = 1.890267
#> Warning: non-integer x = 0.974775
#> Warning: non-integer x = 0.135233
#> Warning: non-integer x = 0.321535
#> Warning: non-integer x = 1.152556
#> Warning: non-integer x = 3.999956
#> Warning: non-integer x = 1.269749
#> Warning: non-integer x = 2.258979
#> Warning: non-integer x = 1.952826
#> Warning: non-integer x = 1.519062
#> Warning: non-integer x = 3.268452
#> Warning: non-integer x = 1.198715
#> Warning: non-integer x = 1.118661
#> Warning: non-integer x = 2.061617
#> Warning: non-integer x = 0.549685
#> Warning: non-integer x = 0.248464
#> Warning: non-integer x = 0.362255
#> Warning: non-integer x = 0.137729
#> Warning: non-integer x = 0.088257
#> Warning: non-integer x = 2.036696
#> Warning: non-integer x = 1.612398
#> Warning: non-integer x = 1.105467
#> Warning: non-integer x = 1.716825
#> Warning: non-integer x = 0.811456
#> Warning: non-integer x = 0.561371
#> Warning: non-integer x = 0.667713
#> Warning: non-integer x = 0.968710
#> Warning: non-integer x = 1.584040
#> Warning: non-integer x = 0.175424
#> Warning: non-integer x = 1.320899
#> Warning: non-integer x = 0.092284
#> Warning: non-integer x = 0.253980
#> Warning: non-integer x = 1.244580
#> Warning: non-integer x = 1.530210
#> Warning: non-integer x = 0.462966
#> Warning: non-integer x = 2.139238
#> Warning: non-integer x = 1.989819
#> Warning: non-integer x = 0.231551
#> Warning: non-integer x = 2.070204
#> Warning: non-integer x = 0.015765
#> Warning: non-integer x = 2.140696
#> Warning: non-integer x = 1.939312
#> Warning: non-integer x = 1.053775
#> Warning: non-integer x = 1.074293
#> Warning: non-integer x = 1.459439
#> Warning: non-integer x = 1.697932
#> Warning: non-integer x = 0.237292
#> Warning: non-integer x = 1.602586
#> Warning: non-integer x = 1.740431
#> Warning: non-integer x = 1.684559
#> Warning: non-integer x = 1.427418
#> Warning: non-integer x = 0.315735
#> Warning: non-integer x = 1.196877
#> Warning: non-integer x = 0.483152
#> Warning: non-integer x = 0.010709
#> Warning: non-integer x = 1.299855
#> Warning: non-integer x = 0.659316
#> Warning: non-integer x = 0.391029
#> Warning: non-integer x = 1.873919
#> Warning: non-integer x = 1.533976
#> Warning: non-integer x = 1.222190
#> Warning: non-integer x = 1.875671
#> Warning: non-integer x = 0.360647
#> Warning: non-integer x = 0.582320
#> Warning: non-integer x = 0.152119
#> Warning: non-integer x = 0.777078
#> Warning: non-integer x = 1.157565
#> Warning: non-integer x = 0.294172
#> Warning: non-integer x = 0.826791
#> Warning: non-integer x = 0.234449
#> Warning: non-integer x = 1.456034
#> Warning: non-integer x = 2.155205
#> Warning: non-integer x = 0.235464
#> Warning: non-integer x = 1.583052
#> Warning: non-integer x = 1.917518
#> Warning: non-integer x = 1.353358
#> Warning: non-integer x = 1.125189
#> Warning: non-integer x = 0.377933
#> Warning: non-integer x = 2.903179
#> Warning: non-integer x = 0.430119
#> Warning: non-integer x = 3.464135
#> Warning: non-integer x = 2.041767
#> Warning: non-integer x = 1.822490
#> Warning: non-integer x = 2.123542
#> Warning: non-integer x = 0.935158
#> Warning: non-integer x = 1.166435
#> Warning: non-integer x = 1.789605
#> Warning: non-integer x = 1.415123
#> Warning: non-integer x = 2.383452
#> Warning: non-integer x = 3.260043
#> Warning: non-integer x = 1.098655
#> Warning: non-integer x = 2.016033
#> Warning: non-integer x = 2.543722
#> Warning: non-integer x = 0.554108
#> Warning: non-integer x = 4.581488
#> Warning: non-integer x = 0.619176
#> Warning: non-integer x = 3.154367
#> Warning: non-integer x = 0.588401
#> Warning: non-integer x = 0.674028
#> Warning: non-integer x = 0.646524
#> Warning: non-integer x = 1.797392
#> Warning: non-integer x = 3.246410
#> Warning: non-integer x = 0.120731
#> Warning: non-integer x = 0.942030
#> Warning: non-integer x = 1.792992
#> Warning: non-integer x = 0.566171
#> Warning: non-integer x = 2.057904
#> Warning: non-integer x = 1.635084
#> Warning: non-integer x = 1.809192
#> Warning: non-integer x = 0.285294
#> Warning: non-integer x = 0.295182
#> Warning: non-integer x = 0.208205
#> Warning: non-integer x = 2.317904
#> Warning: non-integer x = 0.384116
#> Warning: non-integer x = 1.005832
#> Warning: non-integer x = 1.639359
#> Warning: non-integer x = 0.588670
#> Warning: non-integer x = 2.095915
#> Warning: non-integer x = 0.178954
#> Warning: non-integer x = 1.148172
#> Warning: non-integer x = 0.474838
#> Warning: non-integer x = 0.363758
#> Warning: non-integer x = 1.566184
#> Warning: non-integer x = 0.245484
#> Warning: non-integer x = 1.993736
#> Warning: non-integer x = 2.180377
#> Warning: non-integer x = 2.190796
#> Warning: non-integer x = 2.004675
#> Warning: non-integer x = 0.372405
#> Warning: non-integer x = 1.201272
#> Warning: non-integer x = 0.077953
#> Warning: non-integer x = 0.598489
#> Warning: non-integer x = 1.496404
#> Warning: non-integer x = 0.973886
#> Warning: non-integer x = 2.110714
#> Warning: non-integer x = 0.201602
#> Warning: non-integer x = 0.574076
#> Warning: non-integer x = 0.046808
#> Warning: non-integer x = 1.321866
#> Warning: non-integer x = 0.114498
#> Warning: non-integer x = 0.419818
#> Warning: non-integer x = 0.169978
#> Warning: non-integer x = 1.405910
#> Warning: non-integer x = 2.956370
#> Warning: non-integer x = 0.477218
#> Warning: non-integer x = 0.559175
#> Warning: non-integer x = 1.262838
#> Warning: non-integer x = 1.386562
#> Warning: non-integer x = 4.003595
#> Warning: non-integer x = 0.547749
#> Warning: non-integer x = 0.387605
#> Warning: non-integer x = 4.461092
#> Warning: non-integer x = 0.344016
#> Warning: non-integer x = 3.938097
#> Warning: non-integer x = 3.566841
#> Warning: non-integer x = 1.083624
#> Warning: non-integer x = 0.924082
#> Warning: non-integer x = 0.779489
#> Warning: non-integer x = 0.911843
#> Warning: non-integer x = 1.951488
#> Warning: non-integer x = 0.524679
#> Warning: non-integer x = 1.504148
#> Warning: non-integer x = 1.139342
#> Warning: non-integer x = 2.815809
#> Warning: non-integer x = 0.108148
#> Warning: non-integer x = 0.271134
#> Warning: non-integer x = 1.414174
#> Warning: non-integer x = 2.675528
#> Warning: non-integer x = 0.677237
#> Warning: non-integer x = 4.339052
#> Warning: non-integer x = 2.113434
#> Warning: non-integer x = 1.527038
#> Warning: non-integer x = 1.162655
#> Warning: non-integer x = 1.370883
#> Warning: non-integer x = 1.200984
#> Warning: non-integer x = 2.219602
#> Warning: non-integer x = 0.596347
#> Warning: non-integer x = 2.660731
#> Warning: non-integer x = 1.025852
#> Warning: non-integer x = 1.344047
#> Warning: non-integer x = 0.173822
#> Warning: non-integer x = 1.772296
#> Warning: non-integer x = 1.259322
#> Warning: non-integer x = 0.028189
#> Warning: non-integer x = 0.484717
#> Warning: non-integer x = 0.970941
#> Warning: non-integer x = 1.670835
#> Warning: non-integer x = 0.664729
#> Warning: non-integer x = 1.807700
#> Warning: non-integer x = 1.668630
#> Warning: non-integer x = 2.815814
#> Warning: non-integer x = 1.851615
#> Warning: non-integer x = 0.319652
#> Warning: non-integer x = 1.348000
#> Warning: non-integer x = 1.414807
#> Warning: non-integer x = 2.615237
#> Warning: non-integer x = 1.082049
#> Warning: non-integer x = 0.539002
#> Warning: non-integer x = 1.229049
#> Warning: non-integer x = 1.031960
#> Warning: non-integer x = 1.470342
#> Warning: non-integer x = 3.515940
#> Warning: non-integer x = 1.222934
#> Warning: non-integer x = 1.080997
#> Warning: non-integer x = 0.002390
#> Warning: non-integer x = 0.296063
#> Warning: non-integer x = 0.969305
#> Warning: non-integer x = 1.714776
#> Warning: non-integer x = 0.478403
#> Warning: non-integer x = 0.396193
#> Warning: non-integer x = 1.734835
#> Warning: non-integer x = 1.019705
#> Warning: non-integer x = 2.299697
#> Warning: non-integer x = 1.243464
#> Warning: non-integer x = 1.820239
#> Warning: non-integer x = 1.544006
#> Warning: non-integer x = 0.096716
#> Warning: non-integer x = 0.917838
#> Warning: non-integer x = 2.313789
#> Warning: non-integer x = 0.108760
#> Warning: non-integer x = 0.158872
#> Warning: non-integer x = 0.167903
#> Warning: non-integer x = 0.699912
#> Warning: non-integer x = 0.007860
#> Warning: non-integer x = 1.677525
#> Warning: non-integer x = 0.439700
#> Warning: non-integer x = 3.792134
#> Warning: non-integer x = 2.646394
#> Warning: non-integer x = 0.758813
#> Warning: non-integer x = 1.843677
#> Warning: non-integer x = 2.580593
#> Warning: non-integer x = 1.021196
#> Warning: non-integer x = 1.446998
#> Warning: non-integer x = 0.081806
#> Warning: non-integer x = 0.893585
#> Warning: non-integer x = 0.640572
#> Warning: non-integer x = 0.497276
#> Warning: non-integer x = 1.634393
#> Warning: non-integer x = 2.310552
#> Warning: non-integer x = 2.158689
#> Warning: non-integer x = 2.357690
#> Warning: non-integer x = 2.293282
#> Warning: non-integer x = 0.617049
#> Warning: non-integer x = 3.276525
#> Warning: non-integer x = 1.471628
#> Warning: non-integer x = 0.347742
#> Warning: non-integer x = 1.907625
#> Warning: non-integer x = 1.287476
#> Warning: non-integer x = 0.724144
#> Warning: non-integer x = 2.323668
#> Warning: non-integer x = 1.519322
#> Warning: non-integer x = 0.322483
#> Warning: non-integer x = 1.641662
#> Warning: non-integer x = 0.299040
#> Warning: non-integer x = 3.070600
#> Warning: non-integer x = 0.926138
#> Warning: non-integer x = 0.681474
#> Warning: non-integer x = 1.039051
#> Warning: non-integer x = 0.480616
#> Warning: non-integer x = 1.330148
#> Warning: non-integer x = 1.883841
#> Warning: non-integer x = 0.464812
#> Warning: non-integer x = 0.624966
#> Warning: non-integer x = 0.063568
#> Warning: non-integer x = 0.081837
#> Warning: non-integer x = 1.594833
#> Warning: non-integer x = 1.570979
#> Warning: non-integer x = 2.708965
#> Warning: non-integer x = 1.764877
#> Warning: non-integer x = 0.535259
#> Warning: non-integer x = 1.508581
#> Warning: non-integer x = 3.392008
#> Warning: non-integer x = 1.538108
#> Warning: non-integer x = 3.248900
#> Warning: non-integer x = 2.060338
#> Warning: non-integer x = 0.046498
#> Warning: non-integer x = 1.827474
#> Warning: non-integer x = 0.699376
#> Warning: non-integer x = 0.613399
#> Warning: non-integer x = 0.790169
#> Warning: non-integer x = 0.114246
#> Warning: non-integer x = 1.397761
#> Warning: non-integer x = 0.463830
#> Warning: non-integer x = 0.908937
#> Warning: non-integer x = 1.492368
#> Warning: non-integer x = 2.075611
#> Warning: non-integer x = 0.075379
#> Warning: non-integer x = 1.901201
#> Warning: non-integer x = 0.888430
#> Warning: non-integer x = 1.791679
#> Warning: non-integer x = 2.895225
#> Warning: non-integer x = 0.049962
#> Warning: non-integer x = 3.226629
#> Warning: non-integer x = 0.335796
#> Warning: non-integer x = 1.668249
#> Warning: non-integer x = 0.024765
#> Warning: non-integer x = 0.681498
#> Warning: non-integer x = 0.327656
#> Warning: non-integer x = 0.157704
#> Warning: non-integer x = 0.637825
#> Warning: non-integer x = 0.788588
#> Warning: non-integer x = 2.403490
#> Warning: non-integer x = 1.985464
#> Warning: non-integer x = 1.045374
#> Warning: non-integer x = 1.723602
#> Warning: non-integer x = 0.077633
#> Warning: non-integer x = 1.318571
#> Warning: non-integer x = 0.120910
#> Warning: non-integer x = 2.993865
#> Warning: non-integer x = 0.329828
#> Warning: non-integer x = 1.526675
#> Warning: non-integer x = 0.468704
#> Warning: non-integer x = 1.357697
#> Warning: non-integer x = 1.666601
#> Warning: non-integer x = 0.561058
#> Warning: non-integer x = 0.554017
#> Warning: non-integer x = 3.741202
#> Warning: non-integer x = 0.851093
#> Warning: non-integer x = 2.360233
#> Warning: non-integer x = 1.493630
#> Warning: non-integer x = 0.331429
#> Warning: non-integer x = 1.586068
#> Warning: non-integer x = 0.968288
#> Warning: non-integer x = 0.248803
#> Warning: non-integer x = 2.178363
#> Warning: non-integer x = 0.964617
#> Warning: non-integer x = 0.606685
#> Warning: non-integer x = 0.829336
#> Warning: non-integer x = 1.096432
#> Warning: non-integer x = 0.191592
#> Warning: non-integer x = 2.703567
#> Warning: non-integer x = 1.546176
#> Warning: non-integer x = 1.526051
#> Warning: non-integer x = 1.308640
#> Warning: non-integer x = 0.521978
#> Warning: non-integer x = 0.294164
#> Warning: non-integer x = 1.483124
#> Warning: non-integer x = 1.032677
#> Warning: non-integer x = 1.303856
#> Warning: non-integer x = 0.768307
#> Warning: non-integer x = 0.765029
#> Warning: non-integer x = 2.446868
#> Warning: non-integer x = 0.133725
#> Warning: non-integer x = 0.302058
#> Warning: non-integer x = 2.515920
#> Warning: non-integer x = 0.457873
#> Warning: non-integer x = 2.455030
#> Warning: non-integer x = 2.524674
#> Warning: non-integer x = 1.942832
#> Warning: non-integer x = 1.578074
#> Warning: non-integer x = 0.681943
#> Warning: non-integer x = 2.110240
#> Warning: non-integer x = 2.567364
#> Warning: non-integer x = 0.598644
#> Warning: non-integer x = 0.857251
#> Warning: non-integer x = 1.303401
#> Warning: non-integer x = 3.410558
#> Warning: non-integer x = 2.313112
#> Warning: non-integer x = 0.272993
#> Warning: non-integer x = 0.350908
#> Warning: non-integer x = 0.203176
#> Warning: non-integer x = 1.658880
#> Warning: non-integer x = 0.377083
#> Warning: non-integer x = 0.816090
#> Warning: non-integer x = 1.296962
#> Warning: non-integer x = 1.640389
#> Warning: non-integer x = 0.889705
#> Warning: non-integer x = 0.895551
#> Warning: non-integer x = 1.081931
#> Warning: non-integer x = 0.908304
#> Warning: non-integer x = 1.024641
#> Warning: non-integer x = 0.573580
#> Warning: non-integer x = 0.035447
#> Warning: non-integer x = 5.590777
#> Warning: non-integer x = 0.926226
#> Warning: non-integer x = 1.377172
#> Warning: non-integer x = 1.530932
#> Warning: non-integer x = 0.993906
#> Warning: non-integer x = 0.844538
#> Warning: non-integer x = 1.050862
#> Warning: non-integer x = 2.273264
#> Warning: non-integer x = 0.258610
#> Warning: non-integer x = 3.030129
#> Warning: non-integer x = 1.546232
#> Warning: non-integer x = 3.604547
#> Warning: non-integer x = 0.984941
#> Warning: non-integer x = 0.173406
#> Warning: non-integer x = 1.230733
#> Warning: non-integer x = 0.405654
#> Warning: non-integer x = 1.307871
#> Warning: non-integer x = 1.937188
#> Warning: non-integer x = 2.257707
#> Warning: non-integer x = 2.364841
#> Warning: non-integer x = 2.474600
#> Warning: non-integer x = 2.551008
#> Warning: non-integer x = 0.712164
#> Warning: non-integer x = 0.555054
#> Warning: non-integer x = 0.722131
#> Warning: non-integer x = 1.772399
#> Warning: non-integer x = 1.214323
#> Warning: non-integer x = 0.362046
#> Warning: non-integer x = 3.282391
#> Warning: non-integer x = 3.142822
#> Warning: non-integer x = 0.051665
#> Warning: non-integer x = 0.969616
#> Warning: non-integer x = 0.547899
#> Warning: non-integer x = 0.694613
#> Warning: non-integer x = 1.705081
#> Warning: non-integer x = 2.094957
#> Warning: non-integer x = 2.940574
#> Warning: non-integer x = 0.719471
#> Warning: non-integer x = 1.350489
#> Warning: non-integer x = 0.426406
#> Warning: non-integer x = 1.800477
#> Warning: non-integer x = 0.402590
#> Warning: non-integer x = 1.331529