Estimate the Individual Treatment Effect using Poisson Regression

Estimates the Individual Treatment Effect using Poisson Regression given a response vector, a treatment vector, and a features matrix.

Usage

estimate_ite_poisson(y, z, X, X_names, include_offset, offset_name)

Arguments

y

the observed response vector

z

the treatment vector

X

the features matrix

X_names

the names of the covariates

include_offset

whether or not to include an offset when estimating the ITE, for poisson only

offset_name

the name of the offset, if it is to be included

Value

a vector of ITE estimates

Examples

dataset <- generate_cre_dataset(n = 1000, rho = 0, n_rules = 2, p = 10,
                                effect_size = 2, binary = FALSE)

# Initialize parameters
y <- abs(dataset[["y"]])
z <- dataset[["z"]]
X <- as.data.frame(dataset[["X"]])
X_names <- names(as.data.frame(X))
include_offset <- FALSE
offset_name <- NA

ite_list <- estimate_ite_poisson(y, z, X, X_names, include_offset, offset_name)
#> Warning: non-integer x = 4.216108
#> Warning: non-integer x = 2.856873
#> Warning: non-integer x = 0.211029
#> Warning: non-integer x = 0.953084
#> Warning: non-integer x = 0.131661
#> Warning: non-integer x = 0.378598
#> Warning: non-integer x = 2.279657
#> Warning: non-integer x = 2.073524
#> Warning: non-integer x = 3.326493
#> Warning: non-integer x = 3.226091
#> Warning: non-integer x = 1.780381
#> Warning: non-integer x = 1.436116
#> Warning: non-integer x = 0.224594
#> Warning: non-integer x = 2.172155
#> Warning: non-integer x = 1.481205
#> Warning: non-integer x = 1.198650
#> Warning: non-integer x = 3.567220
#> Warning: non-integer x = 1.112024
#> Warning: non-integer x = 2.779253
#> Warning: non-integer x = 1.910945
#> Warning: non-integer x = 3.599445
#> Warning: non-integer x = 1.766653
#> Warning: non-integer x = 0.650559
#> Warning: non-integer x = 1.784796
#> Warning: non-integer x = 1.762642
#> Warning: non-integer x = 2.168911
#> Warning: non-integer x = 4.799669
#> Warning: non-integer x = 2.389071
#> Warning: non-integer x = 4.632086
#> Warning: non-integer x = 0.410648
#> Warning: non-integer x = 2.277474
#> Warning: non-integer x = 0.392630
#> Warning: non-integer x = 0.953612
#> Warning: non-integer x = 1.542806
#> Warning: non-integer x = 0.570286
#> Warning: non-integer x = 2.148761
#> Warning: non-integer x = 2.708595
#> Warning: non-integer x = 2.532679
#> Warning: non-integer x = 1.372456
#> Warning: non-integer x = 3.784602
#> Warning: non-integer x = 0.149524
#> Warning: non-integer x = 1.927339
#> Warning: non-integer x = 0.935555
#> Warning: non-integer x = 1.399244
#> Warning: non-integer x = 2.490399
#> Warning: non-integer x = 2.894918
#> Warning: non-integer x = 0.387096
#> Warning: non-integer x = 0.559760
#> Warning: non-integer x = 0.689713
#> Warning: non-integer x = 0.990536
#> Warning: non-integer x = 0.649783
#> Warning: non-integer x = 0.518075
#> Warning: non-integer x = 0.722382
#> Warning: non-integer x = 1.216037
#> Warning: non-integer x = 3.450117
#> Warning: non-integer x = 0.843905
#> Warning: non-integer x = 0.467544
#> Warning: non-integer x = 3.373060
#> Warning: non-integer x = 1.379264
#> Warning: non-integer x = 1.108480
#> Warning: non-integer x = 0.970859
#> Warning: non-integer x = 2.330174
#> Warning: non-integer x = 0.250668
#> Warning: non-integer x = 0.761760
#> Warning: non-integer x = 0.353584
#> Warning: non-integer x = 2.912835
#> Warning: non-integer x = 0.902920
#> Warning: non-integer x = 0.393390
#> Warning: non-integer x = 3.617608
#> Warning: non-integer x = 0.329881
#> Warning: non-integer x = 1.963123
#> Warning: non-integer x = 3.294451
#> Warning: non-integer x = 2.619060
#> Warning: non-integer x = 0.166905
#> Warning: non-integer x = 0.108479
#> Warning: non-integer x = 2.611130
#> Warning: non-integer x = 3.123120
#> Warning: non-integer x = 2.653398
#> Warning: non-integer x = 0.063166
#> Warning: non-integer x = 0.178179
#> Warning: non-integer x = 2.735267
#> Warning: non-integer x = 2.408691
#> Warning: non-integer x = 1.063392
#> Warning: non-integer x = 0.118481
#> Warning: non-integer x = 1.072526
#> Warning: non-integer x = 1.928570
#> Warning: non-integer x = 1.167502
#> Warning: non-integer x = 2.407363
#> Warning: non-integer x = 0.567815
#> Warning: non-integer x = 0.326211
#> Warning: non-integer x = 1.661878
#> Warning: non-integer x = 3.760978
#> Warning: non-integer x = 1.047360
#> Warning: non-integer x = 3.091124
#> Warning: non-integer x = 2.985022
#> Warning: non-integer x = 2.044382
#> Warning: non-integer x = 0.668560
#> Warning: non-integer x = 0.671593
#> Warning: non-integer x = 1.397369
#> Warning: non-integer x = 2.711452
#> Warning: non-integer x = 2.212512
#> Warning: non-integer x = 3.540686
#> Warning: non-integer x = 0.359251
#> Warning: non-integer x = 1.426328
#> Warning: non-integer x = 2.551535
#> Warning: non-integer x = 0.353510
#> Warning: non-integer x = 2.885775
#> Warning: non-integer x = 0.097581
#> Warning: non-integer x = 1.973508
#> Warning: non-integer x = 0.678203
#> Warning: non-integer x = 3.280803
#> Warning: non-integer x = 0.783820
#> Warning: non-integer x = 0.802205
#> Warning: non-integer x = 0.337827
#> Warning: non-integer x = 0.167423
#> Warning: non-integer x = 3.753344
#> Warning: non-integer x = 0.336541
#> Warning: non-integer x = 3.561347
#> Warning: non-integer x = 2.616471
#> Warning: non-integer x = 3.359113
#> Warning: non-integer x = 1.752564
#> Warning: non-integer x = 0.293917
#> Warning: non-integer x = 2.085748
#> Warning: non-integer x = 3.050658
#> Warning: non-integer x = 1.091588
#> Warning: non-integer x = 1.444247
#> Warning: non-integer x = 2.759450
#> Warning: non-integer x = 2.514866
#> Warning: non-integer x = 3.181402
#> Warning: non-integer x = 0.896403
#> Warning: non-integer x = 1.357645
#> Warning: non-integer x = 0.782405
#> Warning: non-integer x = 1.241835
#> Warning: non-integer x = 0.432809
#> Warning: non-integer x = 3.679469
#> Warning: non-integer x = 2.608968
#> Warning: non-integer x = 1.066749
#> Warning: non-integer x = 1.066299
#> Warning: non-integer x = 0.879669
#> Warning: non-integer x = 1.795733
#> Warning: non-integer x = 2.449465
#> Warning: non-integer x = 1.398644
#> Warning: non-integer x = 0.883101
#> Warning: non-integer x = 3.913987
#> Warning: non-integer x = 3.103667
#> Warning: non-integer x = 2.506657
#> Warning: non-integer x = 3.086815
#> Warning: non-integer x = 1.068989
#> Warning: non-integer x = 1.485617
#> Warning: non-integer x = 1.491996
#> Warning: non-integer x = 0.563225
#> Warning: non-integer x = 0.075240
#> Warning: non-integer x = 0.689983
#> Warning: non-integer x = 1.186838
#> Warning: non-integer x = 2.805037
#> Warning: non-integer x = 2.795103
#> Warning: non-integer x = 1.840388
#> Warning: non-integer x = 1.940047
#> Warning: non-integer x = 1.409304
#> Warning: non-integer x = 0.210173
#> Warning: non-integer x = 0.915346
#> Warning: non-integer x = 4.010417
#> Warning: non-integer x = 0.978258
#> Warning: non-integer x = 2.461629
#> Warning: non-integer x = 1.056110
#> Warning: non-integer x = 0.257551
#> Warning: non-integer x = 0.612851
#> Warning: non-integer x = 0.882533
#> Warning: non-integer x = 1.253746
#> Warning: non-integer x = 2.919093
#> Warning: non-integer x = 0.779807
#> Warning: non-integer x = 0.181659
#> Warning: non-integer x = 0.586882
#> Warning: non-integer x = 1.815357
#> Warning: non-integer x = 1.680004
#> Warning: non-integer x = 2.594097
#> Warning: non-integer x = 2.093782
#> Warning: non-integer x = 1.315914
#> Warning: non-integer x = 1.297411
#> Warning: non-integer x = 2.163995
#> Warning: non-integer x = 0.302838
#> Warning: non-integer x = 1.962223
#> Warning: non-integer x = 3.500816
#> Warning: non-integer x = 2.173386
#> Warning: non-integer x = 2.067410
#> Warning: non-integer x = 0.058593
#> Warning: non-integer x = 0.254787
#> Warning: non-integer x = 0.181152
#> Warning: non-integer x = 2.502825
#> Warning: non-integer x = 0.137814
#> Warning: non-integer x = 2.158696
#> Warning: non-integer x = 1.477399
#> Warning: non-integer x = 2.384678
#> Warning: non-integer x = 1.062793
#> Warning: non-integer x = 2.472335
#> Warning: non-integer x = 2.181945
#> Warning: non-integer x = 4.996711
#> Warning: non-integer x = 1.737330
#> Warning: non-integer x = 1.481704
#> Warning: non-integer x = 2.069436
#> Warning: non-integer x = 2.933457
#> Warning: non-integer x = 2.157352
#> Warning: non-integer x = 1.933125
#> Warning: non-integer x = 0.768427
#> Warning: non-integer x = 1.169204
#> Warning: non-integer x = 0.567798
#> Warning: non-integer x = 0.136980
#> Warning: non-integer x = 1.530038
#> Warning: non-integer x = 1.362264
#> Warning: non-integer x = 2.029424
#> Warning: non-integer x = 2.185552
#> Warning: non-integer x = 4.478305
#> Warning: non-integer x = 0.683063
#> Warning: non-integer x = 2.284980
#> Warning: non-integer x = 0.455545
#> Warning: non-integer x = 1.986902
#> Warning: non-integer x = 1.953284
#> Warning: non-integer x = 3.083374
#> Warning: non-integer x = 1.602273
#> Warning: non-integer x = 1.460657
#> Warning: non-integer x = 1.322527
#> Warning: non-integer x = 0.520506
#> Warning: non-integer x = 0.496201
#> Warning: non-integer x = 4.264002
#> Warning: non-integer x = 3.096526
#> Warning: non-integer x = 3.772010
#> Warning: non-integer x = 1.249795
#> Warning: non-integer x = 4.564168
#> Warning: non-integer x = 3.960082
#> Warning: non-integer x = 1.324822
#> Warning: non-integer x = 0.755727
#> Warning: non-integer x = 1.207707
#> Warning: non-integer x = 3.726669
#> Warning: non-integer x = 1.325356
#> Warning: non-integer x = 2.371082
#> Warning: non-integer x = 0.551934
#> Warning: non-integer x = 0.310380
#> Warning: non-integer x = 4.639745
#> Warning: non-integer x = 1.060638
#> Warning: non-integer x = 1.328813
#> Warning: non-integer x = 0.591210
#> Warning: non-integer x = 3.606046
#> Warning: non-integer x = 1.738625
#> Warning: non-integer x = 4.977908
#> Warning: non-integer x = 2.343279
#> Warning: non-integer x = 1.485884
#> Warning: non-integer x = 1.464637
#> Warning: non-integer x = 3.147117
#> Warning: non-integer x = 0.085731
#> Warning: non-integer x = 3.848622
#> Warning: non-integer x = 3.125599
#> Warning: non-integer x = 4.402860
#> Warning: non-integer x = 3.377661
#> Warning: non-integer x = 0.510310
#> Warning: non-integer x = 2.581186
#> Warning: non-integer x = 1.212647
#> Warning: non-integer x = 0.698648
#> Warning: non-integer x = 2.776456
#> Warning: non-integer x = 1.369697
#> Warning: non-integer x = 2.740013
#> Warning: non-integer x = 1.296459
#> Warning: non-integer x = 1.325388
#> Warning: non-integer x = 3.359694
#> Warning: non-integer x = 0.219059
#> Warning: non-integer x = 0.279111
#> Warning: non-integer x = 2.412661
#> Warning: non-integer x = 2.312142
#> Warning: non-integer x = 3.765495
#> Warning: non-integer x = 1.434258
#> Warning: non-integer x = 1.221306
#> Warning: non-integer x = 0.876460
#> Warning: non-integer x = 3.563005
#> Warning: non-integer x = 0.148410
#> Warning: non-integer x = 1.999735
#> Warning: non-integer x = 3.436253
#> Warning: non-integer x = 0.136377
#> Warning: non-integer x = 0.115413
#> Warning: non-integer x = 2.736463
#> Warning: non-integer x = 1.005294
#> Warning: non-integer x = 1.721298
#> Warning: non-integer x = 1.254956
#> Warning: non-integer x = 0.971774
#> Warning: non-integer x = 3.711250
#> Warning: non-integer x = 0.007354
#> Warning: non-integer x = 2.981318
#> Warning: non-integer x = 4.007922
#> Warning: non-integer x = 3.073212
#> Warning: non-integer x = 5.189112
#> Warning: non-integer x = 0.187186
#> Warning: non-integer x = 0.603832
#> Warning: non-integer x = 2.883069
#> Warning: non-integer x = 2.042725
#> Warning: non-integer x = 1.220090
#> Warning: non-integer x = 0.256957
#> Warning: non-integer x = 0.417735
#> Warning: non-integer x = 0.117792
#> Warning: non-integer x = 1.930637
#> Warning: non-integer x = 0.294782
#> Warning: non-integer x = 0.556902
#> Warning: non-integer x = 1.921044
#> Warning: non-integer x = 1.222143
#> Warning: non-integer x = 1.073267
#> Warning: non-integer x = 0.965284
#> Warning: non-integer x = 3.267371
#> Warning: non-integer x = 2.164153
#> Warning: non-integer x = 0.005115
#> Warning: non-integer x = 4.679197
#> Warning: non-integer x = 2.677894
#> Warning: non-integer x = 0.000649
#> Warning: non-integer x = 0.950074
#> Warning: non-integer x = 1.733596
#> Warning: non-integer x = 0.940128
#> Warning: non-integer x = 0.674956
#> Warning: non-integer x = 2.030629
#> Warning: non-integer x = 0.448470
#> Warning: non-integer x = 0.481377
#> Warning: non-integer x = 0.613804
#> Warning: non-integer x = 0.366398
#> Warning: non-integer x = 1.290582
#> Warning: non-integer x = 2.721207
#> Warning: non-integer x = 2.005504
#> Warning: non-integer x = 0.396300
#> Warning: non-integer x = 1.254301
#> Warning: non-integer x = 1.432778
#> Warning: non-integer x = 2.513277
#> Warning: non-integer x = 2.116195
#> Warning: non-integer x = 1.560476
#> Warning: non-integer x = 2.487230
#> Warning: non-integer x = 0.154242
#> Warning: non-integer x = 3.580160
#> Warning: non-integer x = 0.880889
#> Warning: non-integer x = 3.787990
#> Warning: non-integer x = 0.451653
#> Warning: non-integer x = 1.367383
#> Warning: non-integer x = 1.122841
#> Warning: non-integer x = 0.989838
#> Warning: non-integer x = 1.731761
#> Warning: non-integer x = 2.662974
#> Warning: non-integer x = 3.679492
#> Warning: non-integer x = 2.832437
#> Warning: non-integer x = 1.992745
#> Warning: non-integer x = 2.756674
#> Warning: non-integer x = 0.298562
#> Warning: non-integer x = 2.384427
#> Warning: non-integer x = 2.839536
#> Warning: non-integer x = 3.070311
#> Warning: non-integer x = 0.642107
#> Warning: non-integer x = 4.433928
#> Warning: non-integer x = 0.887259
#> Warning: non-integer x = 2.920082
#> Warning: non-integer x = 0.264146
#> Warning: non-integer x = 2.791237
#> Warning: non-integer x = 0.647108
#> Warning: non-integer x = 1.820405
#> Warning: non-integer x = 1.042465
#> Warning: non-integer x = 2.304670
#> Warning: non-integer x = 0.993482
#> Warning: non-integer x = 2.517990
#> Warning: non-integer x = 1.192450
#> Warning: non-integer x = 2.579512
#> Warning: non-integer x = 0.545651
#> Warning: non-integer x = 1.744961
#> Warning: non-integer x = 0.468602
#> Warning: non-integer x = 2.802296
#> Warning: non-integer x = 2.827585
#> Warning: non-integer x = 2.920934
#> Warning: non-integer x = 1.954955
#> Warning: non-integer x = 1.796515
#> Warning: non-integer x = 3.792999
#> Warning: non-integer x = 0.042215
#> Warning: non-integer x = 2.768839
#> Warning: non-integer x = 1.561356
#> Warning: non-integer x = 1.142247
#> Warning: non-integer x = 1.123366
#> Warning: non-integer x = 3.619264
#> Warning: non-integer x = 0.763126
#> Warning: non-integer x = 2.910090
#> Warning: non-integer x = 2.425137
#> Warning: non-integer x = 2.262841
#> Warning: non-integer x = 1.329607
#> Warning: non-integer x = 0.856160
#> Warning: non-integer x = 0.767663
#> Warning: non-integer x = 0.669899
#> Warning: non-integer x = 2.791722
#> Warning: non-integer x = 1.120532
#> Warning: non-integer x = 0.078272
#> Warning: non-integer x = 2.941311
#> Warning: non-integer x = 0.580392
#> Warning: non-integer x = 0.626459
#> Warning: non-integer x = 1.121044
#> Warning: non-integer x = 2.697965
#> Warning: non-integer x = 1.759756
#> Warning: non-integer x = 2.564463
#> Warning: non-integer x = 0.755899
#> Warning: non-integer x = 0.437899
#> Warning: non-integer x = 1.683882
#> Warning: non-integer x = 0.508832
#> Warning: non-integer x = 3.859053
#> Warning: non-integer x = 1.129311
#> Warning: non-integer x = 3.275286
#> Warning: non-integer x = 1.788340
#> Warning: non-integer x = 2.969321
#> Warning: non-integer x = 2.942420
#> Warning: non-integer x = 2.078784
#> Warning: non-integer x = 2.895722
#> Warning: non-integer x = 2.763343
#> Warning: non-integer x = 1.317237
#> Warning: non-integer x = 1.365431
#> Warning: non-integer x = 1.092850
#> Warning: non-integer x = 3.414487
#> Warning: non-integer x = 2.222965
#> Warning: non-integer x = 2.056039
#> Warning: non-integer x = 0.741295
#> Warning: non-integer x = 1.421035
#> Warning: non-integer x = 0.234576
#> Warning: non-integer x = 1.470338
#> Warning: non-integer x = 0.298044
#> Warning: non-integer x = 0.654547
#> Warning: non-integer x = 2.031921
#> Warning: non-integer x = 2.091020
#> Warning: non-integer x = 2.995329
#> Warning: non-integer x = 4.014613
#> Warning: non-integer x = 0.497117
#> Warning: non-integer x = 2.251924
#> Warning: non-integer x = 0.071514
#> Warning: non-integer x = 2.890676
#> Warning: non-integer x = 3.604240
#> Warning: non-integer x = 0.667634
#> Warning: non-integer x = 1.005869
#> Warning: non-integer x = 0.439211
#> Warning: non-integer x = 2.056907
#> Warning: non-integer x = 1.630537
#> Warning: non-integer x = 4.185629
#> Warning: non-integer x = 2.488804
#> Warning: non-integer x = 0.409348
#> Warning: non-integer x = 0.374574
#> Warning: non-integer x = 1.342556
#> Warning: non-integer x = 0.534634
#> Warning: non-integer x = 3.064347
#> Warning: non-integer x = 0.597346
#> Warning: non-integer x = 2.504135
#> Warning: non-integer x = 1.247766
#> Warning: non-integer x = 0.598643
#> Warning: non-integer x = 1.126263
#> Warning: non-integer x = 2.385862
#> Warning: non-integer x = 0.070556
#> Warning: non-integer x = 0.720054
#> Warning: non-integer x = 1.258519
#> Warning: non-integer x = 0.781575
#> Warning: non-integer x = 1.971609
#> Warning: non-integer x = 2.084940
#> Warning: non-integer x = 0.521343
#> Warning: non-integer x = 0.757518
#> Warning: non-integer x = 1.802220
#> Warning: non-integer x = 0.560864
#> Warning: non-integer x = 0.594204
#> Warning: non-integer x = 0.073640
#> Warning: non-integer x = 1.304149
#> Warning: non-integer x = 0.587776
#> Warning: non-integer x = 1.714723
#> Warning: non-integer x = 1.987677
#> Warning: non-integer x = 3.745451
#> Warning: non-integer x = 2.431949
#> Warning: non-integer x = 1.432480
#> Warning: non-integer x = 0.363430
#> Warning: non-integer x = 1.826074
#> Warning: non-integer x = 1.476766
#> Warning: non-integer x = 2.035814
#> Warning: non-integer x = 1.289436
#> Warning: non-integer x = 2.395766
#> Warning: non-integer x = 1.427919
#> Warning: non-integer x = 0.970119
#> Warning: non-integer x = 3.063702
#> Warning: non-integer x = 0.641402
#> Warning: non-integer x = 0.463036
#> Warning: non-integer x = 2.555714
#> Warning: non-integer x = 0.440498
#> Warning: non-integer x = 0.985237
#> Warning: non-integer x = 2.008444
#> Warning: non-integer x = 0.613855
#> Warning: non-integer x = 0.365951
#> Warning: non-integer x = 1.077225
#> Warning: non-integer x = 1.808921
#> Warning: non-integer x = 0.041456
#> Warning: non-integer x = 0.733758
#> Warning: non-integer x = 1.135861
#> Warning: non-integer x = 0.928317
#> Warning: non-integer x = 1.274274
#> Warning: non-integer x = 2.320295
#> Warning: non-integer x = 1.692284
#> Warning: non-integer x = 2.279644
#> Warning: non-integer x = 1.402546
#> Warning: non-integer x = 1.466203
#> Warning: non-integer x = 1.529757
#> Warning: non-integer x = 0.158728
#> Warning: non-integer x = 0.577309
#> Warning: non-integer x = 1.670688
#> Warning: non-integer x = 2.170386
#> Warning: non-integer x = 0.953936
#> Warning: non-integer x = 0.964173
#> Warning: non-integer x = 2.131996
#> Warning: non-integer x = 3.643033
#> Warning: non-integer x = 0.646238
#> Warning: non-integer x = 0.754793
#> Warning: non-integer x = 0.486496
#> Warning: non-integer x = 1.175319
#> Warning: non-integer x = 0.639981
#> Warning: non-integer x = 3.826854
#> Warning: non-integer x = 1.601244
#> Warning: non-integer x = 1.752646
#> Warning: non-integer x = 1.279583
#> Warning: non-integer x = 3.813820
#> Warning: non-integer x = 2.498872
#> Warning: non-integer x = 3.418447
#> Warning: non-integer x = 0.561706
#> Warning: non-integer x = 2.003080
#> Warning: non-integer x = 2.062315
#> Warning: non-integer x = 0.280034
#> Warning: non-integer x = 1.125100
#> Warning: non-integer x = 1.268687
#> Warning: non-integer x = 2.323964
#> Warning: non-integer x = 0.372539
#> Warning: non-integer x = 1.836211
#> Warning: non-integer x = 0.698042
#> Warning: non-integer x = 1.048926
#> Warning: non-integer x = 0.759455
#> Warning: non-integer x = 0.717679
#> Warning: non-integer x = 2.708989
#> Warning: non-integer x = 2.034061
#> Warning: non-integer x = 3.378617
#> Warning: non-integer x = 1.953534
#> Warning: non-integer x = 2.413183
#> Warning: non-integer x = 2.730049
#> Warning: non-integer x = 0.575567
#> Warning: non-integer x = 2.327751
#> Warning: non-integer x = 0.506257
#> Warning: non-integer x = 0.562673
#> Warning: non-integer x = 1.013257
#> Warning: non-integer x = 0.657855
#> Warning: non-integer x = 0.386458
#> Warning: non-integer x = 2.401601
#> Warning: non-integer x = 0.690038
#> Warning: non-integer x = 0.662264
#> Warning: non-integer x = 1.420138
#> Warning: non-integer x = 0.296039
#> Warning: non-integer x = 0.091260
#> Warning: non-integer x = 2.218734
#> Warning: non-integer x = 0.947218
#> Warning: non-integer x = 0.237564
#> Warning: non-integer x = 2.489792
#> Warning: non-integer x = 0.666237
#> Warning: non-integer x = 0.746358
#> Warning: non-integer x = 0.082132
#> Warning: non-integer x = 2.402856
#> Warning: non-integer x = 0.040753
#> Warning: non-integer x = 2.691182
#> Warning: non-integer x = 2.678118
#> Warning: non-integer x = 0.492262
#> Warning: non-integer x = 2.510284
#> Warning: non-integer x = 0.820447
#> Warning: non-integer x = 2.586625
#> Warning: non-integer x = 1.189889
#> Warning: non-integer x = 0.338233
#> Warning: non-integer x = 3.664013
#> Warning: non-integer x = 0.134402
#> Warning: non-integer x = 0.853662
#> Warning: non-integer x = 1.751012
#> Warning: non-integer x = 0.678340
#> Warning: non-integer x = 1.716834
#> Warning: non-integer x = 1.546966
#> Warning: non-integer x = 1.874591
#> Warning: non-integer x = 1.153257
#> Warning: non-integer x = 0.639335
#> Warning: non-integer x = 1.915993
#> Warning: non-integer x = 2.022293
#> Warning: non-integer x = 2.558353
#> Warning: non-integer x = 1.621697
#> Warning: non-integer x = 1.344434
#> Warning: non-integer x = 0.191943
#> Warning: non-integer x = 0.391786
#> Warning: non-integer x = 1.502827
#> Warning: non-integer x = 3.415867
#> Warning: non-integer x = 0.838301
#> Warning: non-integer x = 2.351399
#> Warning: non-integer x = 2.500494
#> Warning: non-integer x = 0.254427
#> Warning: non-integer x = 0.103362
#> Warning: non-integer x = 2.523707
#> Warning: non-integer x = 0.374408
#> Warning: non-integer x = 0.010077
#> Warning: non-integer x = 1.293873
#> Warning: non-integer x = 2.946879
#> Warning: non-integer x = 1.752341
#> Warning: non-integer x = 1.404092
#> Warning: non-integer x = 1.379322
#> Warning: non-integer x = 2.330296
#> Warning: non-integer x = 4.415246
#> Warning: non-integer x = 0.604861
#> Warning: non-integer x = 0.423182
#> Warning: non-integer x = 2.198017
#> Warning: non-integer x = 0.301006
#> Warning: non-integer x = 0.389876
#> Warning: non-integer x = 0.761592
#> Warning: non-integer x = 2.527255
#> Warning: non-integer x = 3.400306
#> Warning: non-integer x = 0.101489
#> Warning: non-integer x = 0.949103
#> Warning: non-integer x = 1.370549
#> Warning: non-integer x = 1.547692
#> Warning: non-integer x = 0.982809
#> Warning: non-integer x = 2.041183
#> Warning: non-integer x = 1.176935
#> Warning: non-integer x = 2.505857
#> Warning: non-integer x = 1.032161
#> Warning: non-integer x = 0.255538
#> Warning: non-integer x = 0.554103
#> Warning: non-integer x = 2.318883
#> Warning: non-integer x = 1.581815
#> Warning: non-integer x = 1.258035
#> Warning: non-integer x = 0.626950
#> Warning: non-integer x = 1.052591
#> Warning: non-integer x = 1.823221
#> Warning: non-integer x = 2.136558
#> Warning: non-integer x = 2.268230
#> Warning: non-integer x = 1.154560
#> Warning: non-integer x = 1.463050
#> Warning: non-integer x = 1.977804
#> Warning: non-integer x = 0.688215
#> Warning: non-integer x = 0.803976
#> Warning: non-integer x = 1.133609
#> Warning: non-integer x = 2.048016
#> Warning: non-integer x = 1.591686
#> Warning: non-integer x = 0.630718
#> Warning: non-integer x = 0.442541
#> Warning: non-integer x = 1.963653
#> Warning: non-integer x = 2.865891
#> Warning: non-integer x = 0.172310
#> Warning: non-integer x = 1.539405
#> Warning: non-integer x = 0.103812
#> Warning: non-integer x = 0.804238
#> Warning: non-integer x = 0.579874
#> Warning: non-integer x = 1.577218
#> Warning: non-integer x = 1.805362
#> Warning: non-integer x = 0.618176
#> Warning: non-integer x = 3.147417
#> Warning: non-integer x = 1.436101
#> Warning: non-integer x = 0.617873
#> Warning: non-integer x = 1.402494
#> Warning: non-integer x = 0.261107
#> Warning: non-integer x = 0.896189
#> Warning: non-integer x = 1.613488
#> Warning: non-integer x = 1.992288
#> Warning: non-integer x = 0.582935
#> Warning: non-integer x = 1.348051
#> Warning: non-integer x = 1.376386
#> Warning: non-integer x = 1.451526
#> Warning: non-integer x = 1.872200
#> Warning: non-integer x = 0.299953
#> Warning: non-integer x = 1.758038
#> Warning: non-integer x = 0.395442
#> Warning: non-integer x = 1.881569
#> Warning: non-integer x = 0.534007
#> Warning: non-integer x = 2.206453
#> Warning: non-integer x = 2.290139
#> Warning: non-integer x = 0.767969
#> Warning: non-integer x = 0.957595
#> Warning: non-integer x = 2.703339
#> Warning: non-integer x = 0.022748
#> Warning: non-integer x = 0.600785
#> Warning: non-integer x = 1.007690
#> Warning: non-integer x = 1.572937
#> Warning: non-integer x = 2.240754
#> Warning: non-integer x = 0.745491
#> Warning: non-integer x = 1.052310
#> Warning: non-integer x = 0.150261
#> Warning: non-integer x = 2.021730
#> Warning: non-integer x = 0.939316
#> Warning: non-integer x = 0.242851
#> Warning: non-integer x = 0.820111
#> Warning: non-integer x = 1.359250
#> Warning: non-integer x = 1.207177
#> Warning: non-integer x = 0.537418
#> Warning: non-integer x = 0.615447
#> Warning: non-integer x = 1.236314
#> Warning: non-integer x = 0.864699
#> Warning: non-integer x = 2.270620
#> Warning: non-integer x = 0.611869
#> Warning: non-integer x = 3.351781
#> Warning: non-integer x = 1.095238
#> Warning: non-integer x = 1.573811
#> Warning: non-integer x = 0.582918
#> Warning: non-integer x = 0.108724
#> Warning: non-integer x = 1.500399
#> Warning: non-integer x = 2.446774
#> Warning: non-integer x = 0.857122
#> Warning: non-integer x = 1.571801
#> Warning: non-integer x = 0.900941
#> Warning: non-integer x = 1.511035
#> Warning: non-integer x = 1.710121
#> Warning: non-integer x = 1.724784
#> Warning: non-integer x = 1.759073
#> Warning: non-integer x = 0.260178
#> Warning: non-integer x = 3.691025
#> Warning: non-integer x = 2.465245
#> Warning: non-integer x = 1.723462
#> Warning: non-integer x = 1.041336
#> Warning: non-integer x = 1.034860
#> Warning: non-integer x = 3.046367
#> Warning: non-integer x = 0.441660
#> Warning: non-integer x = 0.427638
#> Warning: non-integer x = 0.321433
#> Warning: non-integer x = 1.064792
#> Warning: non-integer x = 1.075398
#> Warning: non-integer x = 1.284791
#> Warning: non-integer x = 0.013125
#> Warning: non-integer x = 3.871649
#> Warning: non-integer x = 1.591251
#> Warning: non-integer x = 2.585643
#> Warning: non-integer x = 2.451748
#> Warning: non-integer x = 0.554863
#> Warning: non-integer x = 2.702629
#> Warning: non-integer x = 0.107189
#> Warning: non-integer x = 0.903112
#> Warning: non-integer x = 1.419434
#> Warning: non-integer x = 1.449750
#> Warning: non-integer x = 1.047670
#> Warning: non-integer x = 0.754163
#> Warning: non-integer x = 1.336825
#> Warning: non-integer x = 1.842128
#> Warning: non-integer x = 0.204903
#> Warning: non-integer x = 0.206200
#> Warning: non-integer x = 0.322530
#> Warning: non-integer x = 0.493468
#> Warning: non-integer x = 0.230874
#> Warning: non-integer x = 1.923258
#> Warning: non-integer x = 0.317364
#> Warning: non-integer x = 1.299677
#> Warning: non-integer x = 1.853916
#> Warning: non-integer x = 1.049010
#> Warning: non-integer x = 0.418033
#> Warning: non-integer x = 0.076965
#> Warning: non-integer x = 0.074537
#> Warning: non-integer x = 1.924100
#> Warning: non-integer x = 1.656912
#> Warning: non-integer x = 0.324883
#> Warning: non-integer x = 3.152866
#> Warning: non-integer x = 0.763424
#> Warning: non-integer x = 1.205884
#> Warning: non-integer x = 0.010879
#> Warning: non-integer x = 2.738527
#> Warning: non-integer x = 2.001791
#> Warning: non-integer x = 0.088246
#> Warning: non-integer x = 0.497766
#> Warning: non-integer x = 0.280407
#> Warning: non-integer x = 1.824386
#> Warning: non-integer x = 0.230732
#> Warning: non-integer x = 0.679641
#> Warning: non-integer x = 0.444531
#> Warning: non-integer x = 1.754621
#> Warning: non-integer x = 0.080701
#> Warning: non-integer x = 1.436088
#> Warning: non-integer x = 1.061638
#> Warning: non-integer x = 1.371316
#> Warning: non-integer x = 3.247686
#> Warning: non-integer x = 1.843836
#> Warning: non-integer x = 1.053081
#> Warning: non-integer x = 1.193401
#> Warning: non-integer x = 0.982402
#> Warning: non-integer x = 0.509473
#> Warning: non-integer x = 1.176840
#> Warning: non-integer x = 3.485852
#> Warning: non-integer x = 0.329578
#> Warning: non-integer x = 1.340226
#> Warning: non-integer x = 1.773971
#> Warning: non-integer x = 0.622917
#> Warning: non-integer x = 0.565593
#> Warning: non-integer x = 3.782684
#> Warning: non-integer x = 0.785680
#> Warning: non-integer x = 2.939343
#> Warning: non-integer x = 0.968658
#> Warning: non-integer x = 0.098522
#> Warning: non-integer x = 1.499640
#> Warning: non-integer x = 1.166491
#> Warning: non-integer x = 2.100875
#> Warning: non-integer x = 3.415012
#> Warning: non-integer x = 0.082367
#> Warning: non-integer x = 0.562594
#> Warning: non-integer x = 2.071395
#> Warning: non-integer x = 2.791976
#> Warning: non-integer x = 0.349181
#> Warning: non-integer x = 1.521057
#> Warning: non-integer x = 1.051849
#> Warning: non-integer x = 0.137587
#> Warning: non-integer x = 0.884075
#> Warning: non-integer x = 0.333104
#> Warning: non-integer x = 0.381285
#> Warning: non-integer x = 1.956441
#> Warning: non-integer x = 1.094644
#> Warning: non-integer x = 0.224991
#> Warning: non-integer x = 0.045563
#> Warning: non-integer x = 0.213900
#> Warning: non-integer x = 0.140311
#> Warning: non-integer x = 0.746692
#> Warning: non-integer x = 0.044708
#> Warning: non-integer x = 3.417612
#> Warning: non-integer x = 1.221895
#> Warning: non-integer x = 0.451257
#> Warning: non-integer x = 2.466978
#> Warning: non-integer x = 0.486431
#> Warning: non-integer x = 1.081419
#> Warning: non-integer x = 2.775543
#> Warning: non-integer x = 0.263508
#> Warning: non-integer x = 0.116675
#> Warning: non-integer x = 0.679879
#> Warning: non-integer x = 1.820269
#> Warning: non-integer x = 0.075109
#> Warning: non-integer x = 0.228919
#> Warning: non-integer x = 0.763983
#> Warning: non-integer x = 3.888822
#> Warning: non-integer x = 0.142457
#> Warning: non-integer x = 1.923523
#> Warning: non-integer x = 1.179941
#> Warning: non-integer x = 1.875280
#> Warning: non-integer x = 1.937406
#> Warning: non-integer x = 3.103519
#> Warning: non-integer x = 0.109372
#> Warning: non-integer x = 1.328523
#> Warning: non-integer x = 0.647469
#> Warning: non-integer x = 1.363052
#> Warning: non-integer x = 1.653795
#> Warning: non-integer x = 0.050592
#> Warning: non-integer x = 3.906338
#> Warning: non-integer x = 2.952124
#> Warning: non-integer x = 1.431925
#> Warning: non-integer x = 1.362165
#> Warning: non-integer x = 1.566830
#> Warning: non-integer x = 0.625926
#> Warning: non-integer x = 0.442604
#> Warning: non-integer x = 0.528990
#> Warning: non-integer x = 2.560141
#> Warning: non-integer x = 0.886397
#> Warning: non-integer x = 0.097122
#> Warning: non-integer x = 1.180777
#> Warning: non-integer x = 0.376879
#> Warning: non-integer x = 0.708302
#> Warning: non-integer x = 1.544002
#> Warning: non-integer x = 0.675773
#> Warning: non-integer x = 0.014015
#> Warning: non-integer x = 1.909731
#> Warning: non-integer x = 3.577139
#> Warning: non-integer x = 1.181422
#> Warning: non-integer x = 1.141539
#> Warning: non-integer x = 0.136585
#> Warning: non-integer x = 1.216826
#> Warning: non-integer x = 1.972940
#> Warning: non-integer x = 2.721045
#> Warning: non-integer x = 0.465475
#> Warning: non-integer x = 0.852263
#> Warning: non-integer x = 0.930621
#> Warning: non-integer x = 0.645601
#> Warning: non-integer x = 0.486252
#> Warning: non-integer x = 1.960096
#> Warning: non-integer x = 0.223505
#> Warning: non-integer x = 0.660963
#> Warning: non-integer x = 1.530735
#> Warning: non-integer x = 0.135101
#> Warning: non-integer x = 0.887153
#> Warning: non-integer x = 2.147292
#> Warning: non-integer x = 1.402982
#> Warning: non-integer x = 0.245947
#> Warning: non-integer x = 1.629245
#> Warning: non-integer x = 1.460164
#> Warning: non-integer x = 0.677930
#> Warning: non-integer x = 0.351463
#> Warning: non-integer x = 0.784556
#> Warning: non-integer x = 0.765673
#> Warning: non-integer x = 0.114217
#> Warning: non-integer x = 2.096411
#> Warning: non-integer x = 1.649703
#> Warning: non-integer x = 0.703420
#> Warning: non-integer x = 2.675788
#> Warning: non-integer x = 2.019042
#> Warning: non-integer x = 3.388165
#> Warning: non-integer x = 3.156717
#> Warning: non-integer x = 3.015257
#> Warning: non-integer x = 0.129410
#> Warning: non-integer x = 0.238041
#> Warning: non-integer x = 0.924140
#> Warning: non-integer x = 1.959355
#> Warning: non-integer x = 1.432433
#> Warning: non-integer x = 1.241995
#> Warning: non-integer x = 1.219607
#> Warning: non-integer x = 1.294912
#> Warning: non-integer x = 1.938988
#> Warning: non-integer x = 0.521650
#> Warning: non-integer x = 0.095651
#> Warning: non-integer x = 1.272661
#> Warning: non-integer x = 2.122060
#> Warning: non-integer x = 1.358496
#> Warning: non-integer x = 0.309256
#> Warning: non-integer x = 0.303190
#> Warning: non-integer x = 0.377733
#> Warning: non-integer x = 1.231469
#> Warning: non-integer x = 5.588045
#> Warning: non-integer x = 1.345007
#> Warning: non-integer x = 2.097465
#> Warning: non-integer x = 1.792323
#> Warning: non-integer x = 1.381083
#> Warning: non-integer x = 1.020355
#> Warning: non-integer x = 1.752176
#> Warning: non-integer x = 0.532354
#> Warning: non-integer x = 2.813969
#> Warning: non-integer x = 0.989421
#> Warning: non-integer x = 1.115803
#> Warning: non-integer x = 1.846251
#> Warning: non-integer x = 2.807942
#> Warning: non-integer x = 0.821149
#> Warning: non-integer x = 3.405174
#> Warning: non-integer x = 2.993638
#> Warning: non-integer x = 0.308531
#> Warning: non-integer x = 0.975037
#> Warning: non-integer x = 1.299555
#> Warning: non-integer x = 2.453670
#> Warning: non-integer x = 1.495205
#> Warning: non-integer x = 1.128619
#> Warning: non-integer x = 1.936797
#> Warning: non-integer x = 0.939193
#> Warning: non-integer x = 0.279640
#> Warning: non-integer x = 1.558999
#> Warning: non-integer x = 0.576308
#> Warning: non-integer x = 1.332025
#> Warning: non-integer x = 1.763559
#> Warning: non-integer x = 0.872125
#> Warning: non-integer x = 2.059863
#> Warning: non-integer x = 0.479680
#> Warning: non-integer x = 0.224110
#> Warning: non-integer x = 0.070838
#> Warning: non-integer x = 2.476789
#> Warning: non-integer x = 0.480777
#> Warning: non-integer x = 1.393439
#> Warning: non-integer x = 0.020210
#> Warning: non-integer x = 2.002864
#> Warning: non-integer x = 1.467683
#> Warning: non-integer x = 3.656217
#> Warning: non-integer x = 0.413573
#> Warning: non-integer x = 0.647674
#> Warning: non-integer x = 1.042901
#> Warning: non-integer x = 0.019979
#> Warning: non-integer x = 0.332589
#> Warning: non-integer x = 1.048956
#> Warning: non-integer x = 1.395840
#> Warning: non-integer x = 1.870283
#> Warning: non-integer x = 1.657997
#> Warning: non-integer x = 1.493567
#> Warning: non-integer x = 2.057752
#> Warning: non-integer x = 1.037974
#> Warning: non-integer x = 1.135708
#> Warning: non-integer x = 0.599363
#> Warning: non-integer x = 1.206189
#> Warning: non-integer x = 1.661028
#> Warning: non-integer x = 1.561536
#> Warning: non-integer x = 1.554752
#> Warning: non-integer x = 1.149658
#> Warning: non-integer x = 0.227185
#> Warning: non-integer x = 1.238444
#> Warning: non-integer x = 0.238482
#> Warning: non-integer x = 0.985704
#> Warning: non-integer x = 1.760832
#> Warning: non-integer x = 1.106213
#> Warning: non-integer x = 2.274584
#> Warning: non-integer x = 0.835814
#> Warning: non-integer x = 1.924842
#> Warning: non-integer x = 1.584041
#> Warning: non-integer x = 0.959321
#> Warning: non-integer x = 2.489982
#> Warning: non-integer x = 0.101692
#> Warning: non-integer x = 1.227002
#> Warning: non-integer x = 2.768952
#> Warning: non-integer x = 0.371837
#> Warning: non-integer x = 2.926846
#> Warning: non-integer x = 1.330996
#> Warning: non-integer x = 0.158957
#> Warning: non-integer x = 0.351620
#> Warning: non-integer x = 0.456512
#> Warning: non-integer x = 1.317927
#> Warning: non-integer x = 2.251201
#> Warning: non-integer x = 1.586826
#> Warning: non-integer x = 3.379397
#> Warning: non-integer x = 0.630204
#> Warning: non-integer x = 1.047012
#> Warning: non-integer x = 1.098924
#> Warning: non-integer x = 2.590514
#> Warning: non-integer x = 0.368805
#> Warning: non-integer x = 0.910730
#> Warning: non-integer x = 2.421846
#> Warning: non-integer x = 2.845380
#> Warning: non-integer x = 0.154445
#> Warning: non-integer x = 2.777103
#> Warning: non-integer x = 0.458127
#> Warning: non-integer x = 0.788107